### 2020-01-17 #2023 @whedon
[PRE REVIEW]: BGGM: Bayesian Gaussian Graphical Models in R
**Submitting author:** @donaldRwilliams (<a href="http://orcid.org/0000-0001-6735-8785">Donald Williams </a>)
**Repository:** <a href="https://github.com/donaldRwilliams/BGGM" target ="_blank">https://github.com/donaldRwilliams/BGGM</a>
**Version:** 1.0.0
