### 2020-01-17 #2023 @donaldRwilliams
Hi:
Here are two potential reviewers.
(1) Paul-Christian Bürkner
(2) Tristan Mahr

I also noted there was an error compiling the pdf. I used the `rticles` R package to make the JOSS submission, which provided a nice template for this journal. However, there are a whole bunch of additional files in the folder "joss_paper" other than paper.md. Perhaps this caused the error ?

thanks !
