### 2020-01-26 1b652be
make CRAN changes
