+++ R/ggm_compare_ppc.default.R
@@ -44,7 +44,7 @@
 #' see methods(class = "ggm_compare_ppc")
 #'
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' # Assume null is true
 #' Y1 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
 #' Y2 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
