### 2020-02-05 3a12cf2
change default priors
