+++ man/ggm_compare_bf.Rd
@@ -4,7 +4,7 @@
 \alias{ggm_compare_bf}
 \title{Compare GGMs with Bayesian Hypothesis Testing}
 \usage{
-ggm_compare_bf(..., prior_sd = 0.35, iter = 25000, cores = 2)
+ggm_compare_bf(..., prior_sd = 0.2, iter = 25000, cores = 2)
 }
 \arguments{
 \item{...}{data matrices (\emph{n} by  \emph{p}). Requires at least two.}
@@ -53,6 +53,16 @@ Y1 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
 Y2 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
 Y3 <- MASS::mvrnorm(500, rep(0, 16), Sigma = diag(16))
 
-bf_ggm <- ggm_compare_bf(Y1, Y2, Y3, prior_sd = .5, iter = 500, cores = 2)
+# fit model
+bf_ggm <- ggm_compare_bf(Y1, Y2, Y3,
+                         prior_sd = .2,
+                         iter = 500, cores = 2)
+
+# select graph
+sel <- select(bf_ggm, BF_cut = 3)
+
+# plot
+plot(sel)
+
 }
 }
