### 2020-02-05 9ae0a52
add robustness check and prior plot
