+++ man/plot_prior.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/plot_prior.R
+\name{plot_prior}
+\alias{plot_prior}
+\title{Prior Distribution Plot}
+\usage{
+plot_prior(prior_sd = 0.2, samples = 10000)
+}
+\arguments{
+\item{prior_sd}{hypothesized standard deviation for the edges or partial correlations.}
+
+\item{samples}{number of samples drawn from the prior distribution}
+}
+\value{
+\code{ggplot object}
+}
+\description{
+Visualize the prior distribution used for the
+Bayesian hypothesis testing methods. This is governed by \code{prior_sd}.
+}
+\examples{
+plot_prior(prior_sd = c(0.1, 0.2, 0.3), samples = 1000)
+}
