+++ NAMESPACE
@@ -5,6 +5,7 @@ S3method(fitted,estimate)
 S3method(ggm_compare_estimate,default)
 S3method(plot,ggm_compare_ppc)
 S3method(plot,metric)
+S3method(plot,robust_check)
 S3method(plot,select.estimate)
 S3method(plot,select.explore)
 S3method(plot,select.ggm_compare_bf)
@@ -24,6 +25,7 @@ S3method(print,map)
 S3method(print,metric)
 S3method(print,post.pred)
 S3method(print,predict.estimate)
+S3method(print,robust_check)
 S3method(print,select.estimate)
 S3method(print,select.explore)
 S3method(print,select.ggm_compare_bf)
@@ -68,9 +70,11 @@ export(map)
 export(mape)
 export(mse)
 export(plot_adjacency)
+export(plot_prior)
 export(posterior_predict)
 export(pp_check)
 export(rmse)
+export(robust_check)
 export(select)
 export(shiny_bggm)
 import(ggplot2)
