+++ README.Rmd
@@ -15,19 +15,21 @@ library(ggplot2)
 library(BGGM)
 ```
 
+# BGGM: Bayesian Gaussian Graphical Models
+
 [![CRAN Version](http://www.r-pkg.org/badges/version/BGGM)](https://cran.r-project.org/package=BGGM)
 [![Downloads](https://cranlogs.r-pkg.org/badges/BGGM)](https://cran.r-project.org/package=BGGM)
 
 
 
 
-# Summary
+## Summary
 Gaussian graphical models (GGM) allow for learning conditional (in)dependence structures that are encoded by partial correlations. Whereas there are several `R` packages for classical methods [see @Kuismin2017, Table 1], there are only two that implement a Bayesian approach [@Leday2018; @Mohammadi2015a]. These are exclusively focused on identifying the graphical structure. The package **BGGM** not only contains novel Bayesian methods for this purpose, but it also includes Bayesian methodology for extending inference beyond identifying non-zero relations. 
 
 **BGGM** is built around two approaches for *Bayesian* inference--estimation and hypothesis testing. This distinction is arbitrary [see @Rouder2018], but is used to organize this work.  The former focuses on the posterior distribution and includes extensions to assess predictability [@Haslbeck2018], as well as methodology to compare partial correlations. The latter includes methods for Bayesian hypothesis testing, in both exploratory and confirmatory contexts, with the novel matrix-$F$ prior distribution [@Mulder2018]. This allows for testing the null hypothesis of conditional independence, as well as inequality and equality constrained hypotheses. Further, there are several approaches for comparing GGMs across any number of groups.  The package also includes a suite of options for model checking. Together, **BGGM** is a comprehensive toolbox for Gaussian graphical modeling in `R`. 
 
 
-# Installation
+## Installation
 You can install BGGM from github as follows.
 
 ```{r gh-installation, eval = FALSE}
@@ -237,7 +239,7 @@ Hc  0   1
 --- 
 note: equal hypothesis prior probabilities
 ```
-Note that `A1--A2` denotes the partial correlation between variables A1 and A2. Any number of hypothesis can be tested. They just need to be separated by a a semi-colon, e.g., `hypothesis = c(A1--A2 > 0; A1--A2 = 0)`, which also demonstrates that it is possible to simultaneously test both inequality (`>` or `<`) and equality (`=`) restrictions.
+Note that `A1--A2` denotes the partial correlation between variables A1 and A2. Any number of hypothesis can be tested. They just need to be separated by a semi-colon, e.g., `hypothesis = c(A1--A2 > 0; A1--A2 = 0)`, which also demonstrates that it is possible to simultaneously test both inequality (`>` or `<`) and equality (`=`) restrictions.
 
 ### Comparing GGMs
 The Bayes factor approach can also be used to compare GGMs [see @williams_ggmcompare]. 
