+++ README.Rmd
@@ -15,7 +15,7 @@ library(ggplot2)
 library(BGGM)
 ```
 
-<img src="joss_paper/hex.png" width = 120/>
+<img src="joss_paper/hex.png" width = 250 />
 
 # BGGM: Bayesian Gaussian Graphical Models
 
