### 2020-02-21 #2111 @whedon
[REVIEW]: BGGM: Bayesian Gaussian Graphical Models in R
**Submitting author:** @donaldRwilliams (<a href="http://orcid.org/0000-0001-6735-8785">Donald Williams </a>)
**Repository:** <a href="https://github.com/donaldRwilliams/BGGM" target ="_blank">https://github.com/donaldRwilliams/BGGM</a>
**Version:** 1.0.0
