+++ NAMESPACE
@@ -55,6 +55,7 @@ S3method(summary,select.ggm_compare_bf)
 S3method(summary,select.ggm_compare_estimate)
 export(assess_predictability)
 export(bayes_R2)
+export(bf_groups)
 export(confirm)
 export(convergence)
 export(diagnostics)
