### 2020-03-07 #2111 @donaldRwilliams
Hi:
Apologies on the delay. I added the paper.md file to the root of the repository.
