### 2020-03-09 donaldRwilliams/BGGM#2 @paulgovan
LICENSE file
This issue is for [github.com/openjournals/joss-reviews/issues/2111](https://github.com/openjournals/joss-reviews/issues/2111).

There is no LICENSE file in the root directory.
