### 2020-03-09 donaldRwilliams/BGGM#3 @paulgovan
Statement of Need
This issue is for [github.com/openjournals/joss-reviews/issues/2111](https://github.com/openjournals/joss-reviews/issues/2111).

There is not a clear "statement of need" in either the package documentation or the journal paper. Both provide a description of what the package is useful for, but not necessarily why this package is necessary as opposed to other BGGM software.
