### 2020-03-09 donaldRwilliams/BGGM#4 @paulgovan
Real world example problems
This issue is for [github.com/openjournals/joss-reviews/issues/2111](https://github.com/openjournals/joss-reviews/issues/2111).

The package docs provide numerous examples, but ideally there would be some "real world" examples or case studies of how this package solved research problems.
