### 2020-03-09 donaldRwilliams/BGGM#5 @paulgovan
Automated tests
This issue is for [github.com/openjournals/joss-reviews/issues/2111](https://github.com/openjournals/joss-reviews/issues/2111).

There are no automated tests for this package that I'm aware of (e.g. travis). I would also recommend submitting this package to CRAN. This would provide a wider use-base for the package.
