### 2020-03-09 donaldRwilliams/BGGM#6 @paulgovan
Community guidelines
This issue is for [github.com/openjournals/joss-reviews/issues/2111](https://github.com/openjournals/joss-reviews/issues/2111).

There are no community guidelines for users or contributors who want to submit an issue or contribute to the package. This could be a simple statement referencing the built in issue and pull request features on github.
