### 2020-03-09 donaldRwilliams/BGGM#5 @donaldRwilliams
Hi:
Thank you for all the great suggestions ! One note is that the package is already on CRAN.

https://cran.r-project.org/web/packages/BGGM/index.html
