### 2020-03-09 donaldRwilliams/BGGM#7 @paulgovan
State of the field
This issue is for [github.com/openjournals/joss-reviews/issues/2111](https://github.com/openjournals/joss-reviews/issues/2111).

"Do the authors describe how this software compares to other commonly-used packages?"

It's pretty clear that the software is useful, but it's not really clear how this software compares to and is better in some way than other software that's currently available. In addition, the background on existing software seems incomplete, as there are numerous packages for building Bayesian graphical models.
