### 2020-03-09 donaldRwilliams/BGGM#7 @donaldRwilliams
We are happy to include this. To our knowledge, for GGMs in particular (and not Bayesian networks), there are only two R packages (BDgraph and beam). We are not aware of others, but can take another look.
