### 2020-03-09 donaldRwilliams/BGGM#8 @paulgovan
Paper length
This issue is for [github.com/openjournals/joss-reviews/issues/2111](https://github.com/openjournals/joss-reviews/issues/2111).

The paper is well over the max length of 1000 words. It looks like the paper is basically the same as the README. I think the paper is meant to be more high-level. Examples and detailed descriptions of the package are not really necessary and can be reserved for the README, wiki, etc.
