### 2020-03-09 donaldRwilliams/BGGM#9 @paulgovan
Paper references
This issue is for [github.com/openjournals/joss-reviews/issues/2111](https://github.com/openjournals/joss-reviews/issues/2111).

The references are not rendered in the paper. I see that they appear in the package docs, so I'm thinking that there is something wrong with the way the paper is being built.
