### 2020-03-09 donaldRwilliams/BGGM#8 @donaldRwilliams
what is your suggestion here ?

We wrote the paper specifically for JOSS and then figured it would also make a nice README. As opposed to just submitting to JOSS our previously written README, etc.
