### 2020-03-09 donaldRwilliams/BGGM#9 @donaldRwilliams
Hmm.
Not sure. Here is a link to the paper. There may be an issue now that the paper had to be moved to the root. will have to take a look later today

https://psyarxiv.com/t2cn7/
