### 2020-03-09 donaldRwilliams/BGGM#10 @paulgovan
Code errors
This issue is for [https://github.com/openjournals/joss-reviews/issues/2111](https://github.com/openjournals/joss-reviews/issues/2111).

After taking a cursory review of the code, there seem to be problems with some of the functions. For example `fit_analytic` does not seem to be a recognized function. The code/documentation needs to be updated before a more in-depth review is possible.

Also, I would recommend a more description naming convention for the package. Functions like 'fit' and 'estimate' are very vague and can be confusing when working with multiple packages at the same time.
