### 2020-03-09 donaldRwilliams/BGGM#10 @donaldRwilliams
`fit_analytic` is not a function in the package, so not really sure how to address this. If you are referring to the example in estimate.R, then `fit_analytic` is the name of an object--i.e.,

```#' # analytic approach (sample by setting analytic = FALSE)
#' fit_analytic <- estimate(Y, analytic = TRUE)
```

Note also we do not have a function named `fit`.  The function `estimate` estimates GGMs (and note changing names of functions would be quite the change considering the package is already on CRAN, researchers use it, etc.)
