### 2020-03-09 donaldRwilliams/BGGM#2 @donaldRwilliams
I added the LICENSE file.

https://github.com/donaldRwilliams/BGGM/blob/master/LICENSE
