### 2020-03-13 donaldRwilliams/BGGM#10 @paulgovan
The examples in the README should be updated so that they are stand alone. For example, the fit_analytic object is not defined in the README, so these examples are not reproducible. I would also recommend checking the other documentation for similar issues.
