### 2020-03-13 donaldRwilliams/BGGM#8 @paulgovan
Take a look at the journal's instructions for authors.
