### 2020-03-13 donaldRwilliams/BGGM#5 @paulgovan
Ok then my recommendation would be to update the documentation to reflect that.
