### 2020-03-13 donaldRwilliams/BGGM#10 @donaldRwilliams
I see that now. I will fix it and check all the examples !
