### 2020-03-13 donaldRwilliams/BGGM#8 @donaldRwilliams
in our experience there is typically flexibility surrounding the instructions for author. In this case, the packing includes the software implementation of novel methods that were introduced in 3 peer-reviewed papers. Hence why it is on the longer side. But happy to shorten if the editor also thinks it should be shorter.
