### 2020-03-14 donaldRwilliams/BGGM#10 @jayrobwilliams
Another option would be to include `fit_analytic` as a data object in the package, as `bfi` is currently. I'll also note that this issue is included in the paper as well as the README.
