### 2020-03-14 donaldRwilliams/BGGM#11 @jayrobwilliams
Installation instructions
This issue is for github.com/openjournals/joss-reviews/issues/2111.

While the package is already on CRAN, it is conventional to also include instructions to install the development version from GitHub in the README. This would be done with
```
library(remotes)
install_github("donaldRwilliams/BGGM")
```
