### 2020-03-14 donaldRwilliams/BGGM#4 @jayrobwilliams
Taking a look at the `bfi` and `rsa` data objects reveals they are real and not simulated data. I would be satisfied on this front if the paper discussed the substantive questions addressed in these examples.
