### 2020-03-14 #2111 @jayrobwilliams
I've completed my initial review. I largely agree with @paulgovan, but have noted a couple of other issues on the package repo. I'm also going to attempt to compile the paper as it's currently missing.
