### 2020-03-14 donaldRwilliams/BGGM#4 @donaldRwilliams
We will add this. Also, to address Paul's initial comment, we are planning to write an additional vignette to address his comment. This would include, for example, actual substantive applications as opposed to the examples just showing the implementation in the documentation.
