### 2020-03-14 529f109
updated readme and paper per JOSS recommendation
