+++ README.Rmd
@@ -32,12 +32,20 @@ Gaussian graphical models (GGM) allow for learning conditional (in)dependence st
 
 
 ## Installation
-You can install BGGM from github as follows.
+To install the latest release version from CRAN use
 
 ```{r gh-installation, eval = FALSE}
 install.packages("BGGM")
 ```
 
+The current developmental version can be installed with
+
+```{r, eval = FALSE}
+if (!requireNamespace("remotes")) {
+  install.packages("remotes")
+}
+remotes::install_github("donaldRwilliams/BGGM")
+```
 
 
 ## Estimation
@@ -56,7 +64,7 @@ fit <- estimate(Y = Y,
                 iter = 1000)
 
 # select graph
-E <- select(fit_analytic, ci_width = 0.95)
+E <- select(fit, ci_width = 0.95)
 
 # summary
 summary(E)
