### 2020-03-14 donaldRwilliams/BGGM#11 @donaldRwilliams
I just updated the README to add this and to correct the `fit_analytic` issue. If satisfied, I can close this issue ?
