### 2020-03-16 1240b16
update ggm compare estimate
