### 2020-04-14 #2111 @donaldRwilliams
@arfon Hi. thank you for this check in. We are currently working on updating the documentation per the reviewers suggestions to include actual examples of hypotheses one can test.
