+++ R/helpers.R
@@ -4,7 +4,1071 @@
 #' @importFrom utils combn
 #' @importFrom foreach %dopar% foreach
 #' @importFrom graphics plot
+#' @importFrom Rdpack reprompt
 #' @import ggplot2
+
+
+
+print_confirm <- function(x, ...){
+  cat("BGGM: Bayesian Gaussian Graphical Models \n")
+  cat("Type: Confirmatory Hypothesis Testing \n")
+  cat("--- \n")
+  cat("Call:\n")
+  print(x$call)
+  cat("--- \n")
+  cat("Hypotheses: \n")
+
+  if(length(x$hypotheses) == length(x$post_prob)){
+
+    hyps <- data.frame( t(t(names(x$post_prob))), c(t(x$hypotheses)))
+    colnames(hyps) <- NULL
+    print(hyps, row.names = F)
+  }
+
+
+
+  if(length(x$hypotheses) != length(x$post_prob)){
+
+    if(length(x$hypotheses) > 1){
+      hyps <- data.frame( t(t(names(x$post_prob))), c(t(x$hypotheses),
+                                                      paste("'not ", "H1-",
+                                                            length(x$hypotheses), "'", sep = "")))
+      colnames(hyps) <- NULL
+      print(hyps, row.names = FALSE, right = F)
+
+    } else{
+      hyps <- data.frame( t(t(names(x$post_prob))), c(t(x$hypotheses),
+                                                      paste("'not ", "H1", "'", sep = "")))
+      colnames(hyps) <- NULL
+      print(hyps, row.names = FALSE )
+
+    }
+
+  }
+
+  cat("--- \n")
+  cat("Posterior prob: \n")
+
+  temp <- data.frame( (paste("p(", names(x$post_prob), "|Y) = ", round(x$post_prob, 4),  sep = "")))
+  colnames(temp) <- ""
+
+  print(temp, row.names = F, right = F)
+  cat("--- \n")
+  cat('Bayes factor matrix: \n')
+  print(t(x$BF_matrix))
+  cat("--- \n")
+  cat("note: equal hypothesis prior probabilities")
+}
+
+
+print_ggm_compare_bf <- function(object, ...) {
+  x <- object
+  name_temp <- matrix(0, object$p, object$p)
+
+  name_temp[] <-
+    unlist(lapply(1:object$p , function(x)
+      paste(1:object$p, x, sep = "--")))
+
+  edge_name <- name_temp[upper.tri(name_temp)]
+
+  groups <- length(object$info$dat)
+
+  BF_10 <- 1 /  object$BF_01[upper.tri(object$BF_01)]
+  prob_H1 <- BF_10 / (1 + BF_10)
+  prob_H0 <- 1 - prob_H1
+
+  if (groups == 2) {
+    pcor_diff <- apply(object$post_samps[[1]], 2,  BGGM::fisher_z2r) -
+      apply(object$post_samps[[2]], 2,  BGGM::fisher_z2r)
+
+    sd_diff <- apply(pcor_diff, 2, sd)
+    results <- data.frame(
+      Edge = edge_name,
+      Estimate = round(unlist(object$mu_diff), 3),
+      Est.Error = round(sd_diff, 3),
+
+      Pr.H0 = round(prob_H0, 3),
+      Pr.H1 = round(prob_H1, 3)
+    )
+
+
+  } else {
+    results <- data.frame(
+      Edge = edge_name,
+
+      Pr.H0 = round(prob_H0, 3),
+      Pr.H1 = round(prob_H1, 3)
+    )
+
+  }
+  x <- list(results = results,
+                          object = object)
+  # class(returned_object) <- "summary.ggm_compare_bf"
+  # return(returned_object)
[…]
