+++ R/RcppExports.R
@@ -1,0 +1,41 @@
+# Generated by using Rcpp::compileAttributes() -> do not edit by hand
+# Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
+
+#' @title Encapsulates a double
+#' @name mvn_continuous
+#' @export
+NULL
+
+#' @title testing binary
+#' @name mvn_binary
+#' @export
+NULL
+
+rcpparma_hello_world <- function() {
+    .Call('_BGGM_rcpparma_hello_world', PACKAGE = 'BGGM')
+}
+
+rcpparma_outerproduct <- function(x) {
+    .Call('_BGGM_rcpparma_outerproduct', PACKAGE = 'BGGM', x)
+}
+
+rcpparma_innerproduct <- function(x) {
+    .Call('_BGGM_rcpparma_innerproduct', PACKAGE = 'BGGM', x)
+}
+
+rcpparma_bothproducts <- function(x) {
+    .Call('_BGGM_rcpparma_bothproducts', PACKAGE = 'BGGM', x)
+}
+
+mvn_continuous <- function(Y, X, delta, epsilon, iter) {
+    .Call('_BGGM_mvn_continuous', PACKAGE = 'BGGM', Y, X, delta, epsilon, iter)
+}
+
+trunc_mvn <- function(mu, rinv, z, y, cutpoints) {
+    .Call('_BGGM_trunc_mvn', PACKAGE = 'BGGM', mu, rinv, z, y, cutpoints)
+}
+
+mvn_binary <- function(Y, X, delta, epsilon, iter, beta_prior, cutpoints) {
+    .Call('_BGGM_mvn_binary', PACKAGE = 'BGGM', Y, X, delta, epsilon, iter, beta_prior, cutpoints)
+}
+

[…]
