+++ R/confirm.R
@@ -1,13 +1,27 @@
-#' Confirmatory Hypothesis Testing
+#' Confirmatory Hypothesis Testing in GGMs
 #'
 #' @description Traditionally, Gaussian graphical models (GGM) are inherently exploratory. That is, automated model selection is performed. A key aspect of \strong{BGGM} is the ability to extend inference beyond exploratory and to
 #' confirmatory hypothesis testing. This is accomplished by testing equality and/or inequality constraints for sets of edges (partial correlations).
 #'
-#' @param Y data matrix  (\emph{n} by  \emph{p}).
-#' @param hypothesis hypothesis (or hypotheses) to be tested
+#' @param Y  matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
+#'
+#' @param hypothesis character string. hypothesis (or hypotheses) to be tested. See notes for futher details.
+#'
 #' @param prior_sd hypothesized standard deviation of the prior distribution
-#' @param iter posterior and prior samples. 25,000 is the default, as it results in a more stable Bayes factor than using, say, 5,000.
-#' @param cores number of cores for parallel computing. The default is 2, but this can be adjusted
+#'
+#' @param formula an object of class \code{\link[stats]{formula}}. This allows for including
+#' control variables in the model (e.g.,, \code{~ gender * education}).
+#'
+#' @param data an optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
+#' to a data frame containing the variables in \code{formula}. This is required when controlling for variables.
+#'
+#' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
+#' \code{binary}, or \code{ordinal}. See the note for further details.
+#'
+#' @param iter posterior and prior samples. 25,000 is the default, as it results in a more stable Bayes factor than
+#' using, say, 5,000.
+#'
+#' @param cores integer. How many cores for parallel computing ? (default is 2).
 #'
 #' @return list of class \code{confirm}:
 #'
@@ -25,6 +39,7 @@
 #' }
 #'
 #' @importFrom MASS ginv
+#' @importFrom BFpack BF
 #' @importFrom stats cov rbeta
 #' @export
 #'
@@ -61,414 +76,119 @@
 #'                       cores = 2)
 #'# summary
 #' test_directions
+#'
+#'#############################
+#'######## Binary #############
+#'#############################
+#'
+#'
+#'
+#'
 #'}
-confirm <- function(Y, hypothesis, prior_sd = 0.25,
-                    iter = 25000,  cores = 2){
-
-  # code taken and adapted from (with permission):
-  # https://github.com/Jaeoc/lmhyp
+confirm <- function(Y, hypothesis,
+                    prior_sd = 0.25,
+                    formula = NULL,
+                    data = NULL,
+                    type = "continuous",
+                    iter = 25000,
+                    cores = 2){
 
   # set prior prob to 1--i.e., equal
   priorprob = 1
 
-  returned_mats <-list()
+  # hyperparameter
[…]
