### 2020-04-20 90740ac
add new data, posterior samples, and better documentation
