+++ R/confirm.R
@@ -1,7 +1,9 @@
 #' Confirmatory Hypothesis Testing in GGMs
 #'
-#' @description Traditionally, Gaussian graphical models (GGM) are inherently exploratory. That is, automated model selection is performed. A key aspect of \strong{BGGM} is the ability to extend inference beyond exploratory and to
-#' confirmatory hypothesis testing. This is accomplished by testing equality and/or inequality constraints for sets of edges (partial correlations).
+#' @description Confirmatory hypothesis testing in GGMs, for, say, comparing formalized models encoding
+#' scienctific expectations. Hypotheses are expressed as equality and/or ineqaulity contraints on the
+#' partial correlations of interest. Here the focus is \emph{not} on determining the graph (see \code{\link{explore}}) but testing specific hypotheses related to
+#' the conditional (in)dependence structure. These methods were introduced in \insertCite{Williams2019_bf;textual}{BGGM}.
 #'
 #' @param Y  matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
 #'
@@ -16,12 +18,15 @@
 #' to a data frame containing the variables in \code{formula}. This is required when controlling for variables.
 #'
 #' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
-#' \code{binary}, or \code{ordinal}. See the note for further details.
+#' \code{binary}, \code{ordinal}, or \code{mixed}. See the note for further details.
 #'
 #' @param iter posterior and prior samples. 25,000 is the default, as it results in a more stable Bayes factor than
 #' using, say, 5,000.
 #'
 #' @param cores integer. How many cores for parallel computing ? (default is 2).
+#'
+#' @references
+#' \insertAllCited{}
 #'
 #' @return list of class \code{confirm}:
 #'
@@ -41,10 +46,48 @@
 #' @importFrom MASS ginv
 #' @importFrom BFpack BF
 #' @importFrom stats cov rbeta
-#' @export
+
 #'
-#' @note Currently inequality and equality restrictions can be tested. The former is an ordering the respective edge sizes,
-#' whereas the latter allows for testing whether certain edges are exactly the same.
+#' @note
+#' \strong{Interpretation of conditional (in)dependence models for latent data:}
+#'
+#' A  tetrachoric correlation (binary data) is a special case of a polychoric correlation (ordinal data). Both relations are
+#' between "theorized normally distributed continuous latent variables"
+#' (\href{https://en.wikipedia.org/wiki/Polychoric_correlation}{Wikipedia})
+#' In both instances, the correpsonding partial correlation between observed variables is conditioned
+#' on the remaining variables in the \emph{latent} space. This implies that interpration is much the same as
+#' for continuous data, but with respect to latent variables. We refer interested reader to
+#' \insertCite{@page 2364, section 2.2, in  @webb2008bayesian;textual}{BGGM}.
+#'
+#'
+#' \strong{Hypothesis Examples:}
+#'
+#' The hypotheses can be written either with the respective column names or numbers. For example, \code{1--2} denotes the
+#' relation between variables in column 1 and 2. Note that these must correspond to the upper triangular elements of the correlation
+#' matrix. This is accomplished by ensuring that the first number is smaller than the second number. This also applies when using
+#' column names (e.g., in reference to the column number).
+#'
+#'One hypothesis:
+#' \itemize{
+#'
+#' \item  To test whether some relations are larger than others, while others are expected to be equal, this can be writting as
+#'
+#'  \code{hyp <-  c(1--2 > 1--3  = 1--4 > 0)},
+#'
[…]
