+++ data-raw/ptsd_data.R
@@ -1,4 +1,1 @@
 ptsd <- read.csv('ptsd.csv')
-
-
-
