+++ NAMESPACE
@@ -42,9 +42,10 @@ export(mae)
 export(map)
 export(mape)
 export(mse)
-export(mvn_binary)
-export(mvn_continuous)
-export(mvn_ordinal)
+export(mv_binary)
+export(mv_continuous)
+export(mv_ordinal_albert)
+export(mv_ordinal_cowles)
 export(plot_adjacency)
 export(plot_prior)
 export(posterior_predict)
