+++ R/helpers.R
@@ -1,15 +1,1404 @@
+#' @importFrom stats coef cov2cor var dnorm lm
+#' na.omit pnorm quantile rWishart
+#' sd qnorm residuals fitted density
+#' @importFrom utils combn
+#' @importFrom foreach %dopar% foreach
+#' @importFrom graphics plot
+#' @importFrom Rdpack reprompt
+#' @import ggplot2
+
+
+convert_hyps <- function(hypothesis, Y){
+
+  p <- ncol(Y)
+
+  col_names <- BGGM:::numbers2words(1:p)
+
+  mat_name_num <- sapply(1:p, function(x) paste0(1:p, "--", x, sep = ""))
+
+  mat_name_n2w <- sapply(col_names, function(x) paste(col_names,x, sep = ""))
+
+  mat_name_actual <- sapply(colnames(Y), function(x) paste0(colnames(Y), "--", x, sep = ""))
+
+  n_off_diag <- p*(p-1)*0.5
+
+  where <- lapply(1:n_off_diag, function(x) grep(mat_name_actual[upper.tri(diag(p))][x], hypothesis ))
+  where <- which(where == 1)
+
+  if(any(where)){
+
+    for (i in seq_along(where)) {
+      hypothesis <-  gsub(mat_name_actual[upper.tri(diag(p))][where[i]],
+                     replacement =  mat_name_n2w[upper.tri(diag(p))][where[i]], x = hypothesis )
+    }
+
+  } else if (!any(where)){
+
+    where <- lapply(1:n_off_diag, function(x) grep(mat_name_num[upper.tri(diag(p))][x], hypothesis ))
+    where <- which(where == 1)
+
+
+    for (i in seq_along(where)) {
+      hypothesis <-  gsub(mat_name_num[upper.tri(diag(p))][where[i]],
+                          replacement =  mat_name_n2w[upper.tri(diag(p))][where[i]], x = hypothesis )
+    }
+
+  } else {
+
+    stop("error in edge specification\n
+             Hints:\n
+             1) check that the first is smaller than the second (denoting the upper-triangular elements)-e.g., 1--2 and not 2--1\n
+             2) alternatively, the variable might not exist--e.g., p = 10 but 1--11 (11 > 10)")
+  }
+
+  hypothesis
+
+}
+
+
+
+prior_helper_2 <- function(p, delta , epsilon){
+
+  k <- p
+  nu <- 1/epsilon
+  parcorMat <- corMat <- ThetaMat <- SigmaMat <- array(0,dim=c(1e4,k,k))
+
+
+  for(s in 1:dim(parcorMat)[1]){
+
+    Psi <- rWishart(1,df = nu, Sigma = 1/nu*diag(k))[,,1]
+
+    Sigma <- rWishart(1,df=k-1+delta, Sigma=solve(Psi))[,,1]
+
+    Theta <- solve(Sigma)
+
+    ThetaMat[s,,] <- Theta
+    SigmaMat[s,,] <- Sigma
+
+    parcorMat[s,,] <- -diag(1/sqrt(diag(Theta)))%*%Theta%*%diag(1/sqrt(diag(Theta)))
+
+  }
+  parcorMat
+}
+
+
+sampling_helper_poly <- function(Y, delta, iter, type){
+
+  if(type == "binary"){
+
+  fit_mvn <- mvn_binary(Y, matrix(1, nrow = nrow(Y)),
+                        delta = delta,
+                        epsilon = 0.001,
+                        iter = iter,
+                        beta_prior = 0.0001,
+                        cutpoints = c(-Inf, 0, Inf) )
+
+  } else if(type == "ordinal"){
+
[…]
