+++ R/explore.default.R
@@ -1,274 +1,0 @@
-#' @title GGMs with Exploratory Bayesian Hypothesis Testing
-#' @name explore
-#' @description Learn the conditional (in)dependence structure with the Bayes factor computed from the matrix-F prior distribution. It is
-#' possible to test for only positive or negative edges, as well as two sided hypothesis testing (which is the customary approach). Further
-#' there is also an exhaustive option that provides the posterior probability of the null, greater than zero, and less than zero.
-#' These methods were introduced in \insertCite{Williams2019_bf;textual}{BGGM}.
-#'
-#' @param Y  matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
-#'
-#' @param formula an object of class \code{\link[stats]{formula}}. This allows for including
-#' control variables in the model (i.e., \code{~ gender}).
-#'
-#' @param data an optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
-#' to a data frame containing the variables in \code{formula}. This is required when controlling for variables.
-#'
-#' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
-#' \code{binary}, \code{ordinal}, or \code{mixed}. See the note for further details.
-#'
-#' @param iter number of iterations (posterior samples; defaults to 5000).
-#'
-#' @param cores number of cores for parallel computing. The default is 2, but this can be changed.
-#'
-#' @param prior_sd hypothesized standard deviation for the edges or partial correlations
-#'
-#' @param ... currently ignored.
-#'
-#' @references
-#' \insertAllCited{}
-#'
-#' @return list of class \code{explore}:
-#' \itemize{
-#' \item \code{parcors_mat} partial correlation matrix
-#' \item \code{parcors_sd} partial correlation standard deviations
-#' \item \code{samples} list of prior and posterior samples
-#' \itemize{
-#'  \item \code{fisher_z_post} Fisher z transformed posterior distributions (partial correlations)
-#'  \item  \code{pcor_post} partial correlation posterior distributions (not transformed)
-#'  \item \code{inv_cov_post} inverse covariance matrix posterior distribution
-#'  \item \code{pcor_prior} partial correlation prior distribution
-#'  \item \code{fisher_z_prior}  Fisher z transformed prior distributions (partial correlations)
-#'  }
-#' \item \code{delta} hyperparameter of matrix-F prior distribution (corresponds to \code{prior_sd})
-#' \item \code{iter} number of posterior and prior samples
-#' \item \code{dat} data matrix
-#' \item \code{call} match.call()
-#' \item \code{p} number of variables
-#' \item \code{cores} number of cores
-#' \item \code{edge} number of estimated edges
-#' }
-#'
-#'
-#'
-#'
-#' @note After sampling from the posterior distribution, use \code{select} to determine the edge set and \code{plot} for visualizing the
-#' edge set. see \code{methods(class = "explore")}
-#'
-#' \strong{Interpretation of conditional (in)dependence models for latent data:}
-#'
-#' A  tetrachoric correlation (binary data) is a special case of a polychoric correlation (ordinal data). Both relations are
-#' between "theorized normally distributed continuous latent variables"
-#' (\href{https://en.wikipedia.org/wiki/Polychoric_correlation}{Wikipedia})
-#' In both instances, the correpsonding partial correlation between observed variables is conditioned
-#' on the remaining variables in the \emph{latent} space. This implies that interpration is much the same as
[…]
