+++ joss_paper/MyJSSArticle.Rmd
@@ -1,0 +1,339 @@
+---
+# Example from https://joss.readthedocs.io/en/latest/submitting.html
+title: 'BGGM: Bayesian Gaussian Graphical Models in R'
+tags:
+- Gaussian graphical models
+- Bayesian
+- Bayes factor
+- partial correlation
+- R
+authors:
+  - name: Donald R. Williams
+    affiliation: 1 # (Multiple affiliations must be quoted)
+  - name: Joris Mulder
+    affiliation: 2
+affiliations:
+ - name: Department of Psychology, University of California, Davis
+   index: 1
+ - name: Department of Methodology and Statistics, Tilburg University
+   index: 2
+citation_author: Williams and Mulder
+date: 07 January 2020
+year: 2020
+bibliography: paper.bib
+output: rticles::joss_article
+csl: apa.csl
+journal: JOSS
+submitted: 07 January 2020
+published: (pre-print)
+---
+
+# Summary
+Gaussian graphical models (GGM) allow for learning conditional (in)dependence structures that are encoded by partial correlations. Whereas there are several `R` packages for classical methods [see @Kuismin2017, Table 1], there are only two that implement a Bayesian approach [@Leday2018; @Mohammadi2015a]. These are exclusively focused on identifying the graphical structure. The package **BGGM** not only contains novel Bayesian methods for this purpose, but it also includes Bayesian methodology for extending inference beyond identifying non-zero relations. 
+
+**BGGM** is built around two approaches for *Bayesian* inference--estimation and hypothesis testing. This distinction is arbitrary [see @Rouder2018], but is used to organize this work.  The former focuses on the posterior distribution and includes extensions to assess predictability [@Haslbeck2018], as well as methodology to compare partial correlations. The latter includes methods for Bayesian hypothesis testing, in both exploratory and confirmatory contexts, with the novel matrix-$F$ prior distribution [@Mulder2018]. This allows for testing the null hypothesis of conditional independence, as well as inequality and equality constrained hypotheses. Further, there are several approaches for comparing GGMs across any number of groups.  The package also includes a suite of options for model checking. Together, **BGGM** is a comprehensive toolbox for Gaussian graphical modeling in `R`. 
+
+## Estimation
+There are two possibilities for *estimating* GGMs. The first is an analytic solution and the second samples from the posterior distribution [described in @Williams2018bayes]. Sampling is recommended because the samples are required for various functions in **BGGM** (e.g., posterior predictive checks and prediction). 
+
+### Structure Learning
+Structure learning refers to determining which partial correlations are non-zero. This is implemented with:
+
+```r
+# data (p = 5 for demonstrative purposes)
+Y <- BGGM::bfi[,1:5]
+
+# fit model
+fit <- estimate(Y = Y, 
+                analytic = FALSE, 
+                iter = 1000)
+
+# select graph
+E <- select(fit_analytic, ci_width = 0.95)
+
+# summary
+summary(E)
+
+# output
+
+BGGM: Bayesian Gaussian Graphical Models 
+--- 
+Type: Selected Graph (Analytic Solution) 
+Credible Interval: 95 % 
+Connectivity: 80 % 
+--- 
+Call:
+select.estimate(x = fit_analytic, ci_width = 0.95)
+--- 
+Selected:
+ 
+Partial correlations 
+ 
+      1     2     3    4    5
+1  0.00 -0.24 -0.11 0.00 0.00
+2 -0.24  0.00  0.29 0.16 0.16
+3 -0.11  0.29  0.00 0.18 0.36
+4  0.00  0.16  0.18 0.00 0.12
+5  0.00  0.16  0.36 0.12 0.00
+--- 
+ 
+Adjacency 
+ 
+  1 2 3 4 5
+1 0 1 1 0 0
+2 1 0 1 1 1
[…]
