+++ man/posterior_samples.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/posterior_samples.R
+\name{posterior_samples}
+\alias{posterior_samples}
+\title{Extract posterior samples}
+\usage{
+posterior_samples(x, ...)
+}
+\arguments{
+\item{x}{A R object of class \code{estimate} or \code{explore}}
+
+\item{...}{currently ignored.}
+}
+\value{
+A matrix of posterior samples (corresponding to the upper-triangular elements)
+}
+\description{
+Extract posterior samples
+}
+\examples{
+Y <- BGGM::ptsd
+fit <- estimate(Y)
+posterior_samples(fit)
+}
