+++ man/ptsd.Rd
@@ -1,0 +1,41 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/datasets.R
+\docType{data}
+\name{ptsd}
+\alias{ptsd}
+\title{post-traumatic stress disorder dataset}
+\usage{
+data("ptsd")
+}
+\description{
+A dataset containing items that measure Post-traumatic stress disorder symptoms. There are 20 variables (p) and
+221 observations (n)
+}
+\details{
+\itemize{
+  \item Intrusive Thoughts
+  \item Nightmares
+  \item Flashbacks
+  \item Emotional cue reactivity
+  \item Psychological cue reactivity
+  \item Avoidance of thoughts
+  \item Avoidance of reminders
+  \item Trauma-related amnesia
+  \item Negative beliefs
+  \item Negative trauma-related emotions
+  \item Loss of interest
+  \item Detachment
+  \item Restricted affect
+  \item Irritability/anger
+  \item Self-destructive/reckless behavior
+  \item Hypervigilance
+  \item Exaggerated startle response
+  \item Difficulty concentrating
+  \item Sleep disturbance
+}
+}
+\references{
+Epskamp, S., & Fried, E. I. (2018). A tutorial on regularized partial correlation networks. Psychological methods.
+ @format A data frame with 221 rows and 20 variables
+}
+\keyword{datasets}
