+++ man/mv_ordinal_albert.Rd
@@ -1,0 +1,8 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/RcppExports.R
+\name{mv_ordinal_albert}
+\alias{mv_ordinal_albert}
+\title{testing ordinal albert and chibbs for thresholds}
+\description{
+testing ordinal albert and chibbs for thresholds
+}

[…]
