+++ NAMESPACE
@@ -29,6 +29,7 @@ export(bayes_R2)
 export(bf_groups)
 export(confirm)
 export(convergence)
+export(copula)
 export(diagnostics)
 export(estimate)
 export(explore)

[…]
