### 2020-04-23 0ad4d64
added summary for R2
