### 2020-04-23 #2111 @donaldRwilliams
@arfon @jayrobwilliams @paulgovan

I wanted to check in once more. We had received reviews for one of the methods that is implemented in the package (generally favorable) and we just finished that up. We also decided to follow the suggestions of @paulgovan to (1) make the paper more of a high level summary; and (2) also include real world examples (this was also suggested in the review).

Anyhow, we expect to have this completed by the end of next week.

Thanks !
