+++ inst/REFERENCES.bib
@@ -81,3 +81,41 @@ year = {2019}
   publisher={Elsevier}
 }
 
+
+@book{baum2006introduction,
+  title={An introduction to modern econometrics using Stata},
+  author={Baum, Christopher F and Christopher, F},
+  year={2006},
+  publisher={Stata press}
+}
+
+@article{bringmann2019centrality,
+  title={What do centrality measures measure in psychological networks?},
+  author={Bringmann, Laura F and Elmer, Timon and Epskamp, Sacha and Krause, Robert W and Schoch, David and Wichers, Marieke and Wigman, Johanna TW and Snippe, Evelien},
+  journal={Journal of Abnormal Psychology},
+  volume={128},
+  number={8},
+  pages={892},
+  year={2019},
+  publisher={American Psychological Association}
+}
+
+@article{hallquist2019problems,
+  title={Problems with centrality measures in psychopathology symptom networks: Why network psychometrics cannot escape psychometric theory},
+  author={Hallquist, Michael N and Wright, Aidan GC and Molenaar, Peter CM},
+  journal={Multivariate behavioral research},
+  pages={1--25},
+  year={2019},
+  publisher={Taylor \& Francis}
+}
+
+@article{leeb2006can,
+  title={Can one estimate the conditional distribution of post-model-selection estimators?},
+  author={Leeb, Hannes and P{\"o}tscher, Benedikt M and others},
+  journal={The Annals of Statistics},
+  volume={34},
+  number={5},
+  pages={2554--2591},
+  year={2006},
+  publisher={Institute of Mathematical Statistics}
+}

[…]
