+++ R/centrality.R
@@ -5,7 +5,7 @@
 #' correlations. Testing strength (e.g., differences or confirmatory testing of order contraints) is
 #' implemented in \code{test.centrality}.
 #'
-#' @name centrality.BGGM
+#' @name centrality
 #'
 #' @param object an object of class \code{estimate} or \code{explore}. All data types (e.g., ordinal, binary, etc.)
 #'               are supported.
@@ -16,9 +16,7 @@
 #' @param cluster charcter string. Which clusters or communities do the nodes belong to? This is
 #'                required when \code{type = bridge_stength}.
 #'
-#' @param select logical. Should strength be condtional on the selected model (default set to \code{NULL})? Note
-#'                 that is might seem desirable to compute strength from only the selected edges. However, this
-#'                 presents challenges for making statistical inference. See note for further details.
+#' @param select logical. Should strength be condtional on the selected model (default set to \code{NULL})? See note for further details.
 #'
 #' @param cred numeric. Credible interval width used for selecting the graph with an \code{estimate}
 #'             object (default is \code{0.95}).
@@ -46,30 +44,31 @@
 #'
 #' The default setting is \code{select = FALSE}. This means that \emph{all} of the partial correlation for a given
 #' node are summed and not only those that were determined to be non-zero. Although it may seem that including only
-#' the selected edges leads to desirable "inference," this is not the case. This would create a problem known
-#' as model selection bias that arises when \emph{conditioning} on the selected model \insertCite{@leeb2006can}{BGGM} .
+#' the selected edges leads to desirable "inference," there is an important caveat to consider.
+#' This induces a problem known as model selection bias that arises when \emph{conditioning} on the selected model..
 #'
 #' To make sense of this, note that:
 #'
 #' \enumerate{
 #' \item To be selected, the partial correlations must far enough way from zero.
-#' \item As a result of 1, this means that edges will often be upward biased when they are only considered
-#'          when they are selected \insertCite{@more specifically a truncated sampling distribution, p. 5 in  @leeb2006can}{BGGM}
-#' \item This then can translate into strength also having issues when only computing it from the selected edges.
+#' \item As a result of 1, edges will often be upward biased becauase they are only considered
+#'       when they are selected \insertCite{@more specifically a truncated sampling distribution, p. 5 in  @leeb2006can}{BGGM}
+#' \item This can translate into strength also having issues when computing it from only the selected edges.
 #' }
 #'
-#' Overcoming this issue is an active area of research that is often called 'selective' or 'post-selection' inference
+#' Overcoming this issue is often called 'selective' or 'post-selection' inference
 #' (see the references in \href{http://bactra.org/notebooks/post-model-selection-inference.html}{weblink}).
-#' However, a viable and simple approach is to \strong{not condition} strength on the selected model (i.e., \code{select = FALSE}).
-#' This then readily allows for comparing the sum of partial correlations, say, differences in bridge strength. In future
-#' versions (> 2.0), the option \code{select = FALSE} may be removed altogether due to the innate proclivity to compare estimates
-#' visually (\code{plot.centrality}).
+#' A viable and simple approach is to \strong{not condition} strength on the selected model
+#' (i.e., \code{select = FALSE}). This readily allows for comparing the sum of partial correlations
[…]
