+++ R/RcppExports.R
@@ -1,63 +1,0 @@
-# Generated by using Rcpp::compileAttributes() -> do not edit by hand
-# Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
-
-#' @title multivariate regression
-#' @name mv_continuous
-#' @export
-NULL
-
-#' @title testing binary
-#' @name mv_binary
-#' @export
-NULL
-
-#' @title testing ordinal cowles for thresholds
-#' @name mv_ordinal_cowles
-#' @export
-NULL
-
-#' @title testing ordinal albert and chibbs for thresholds
-#' @name mv_ordinal_albert
-#' @export
-NULL
-
-mv_continuous <- function(Y, X, delta, epsilon, iter) {
-    .Call('_BGGM_mv_continuous', PACKAGE = 'BGGM', Y, X, delta, epsilon, iter)
-}
-
-trunc_mvn <- function(mu, rinv, z, y, cutpoints) {
-    .Call('_BGGM_trunc_mvn', PACKAGE = 'BGGM', mu, rinv, z, y, cutpoints)
-}
-
-mv_binary <- function(Y, X, delta, epsilon, iter, beta_prior, cutpoints) {
-    .Call('_BGGM_mv_binary', PACKAGE = 'BGGM', Y, X, delta, epsilon, iter, beta_prior, cutpoints)
-}
-
-Sigma_i_not_i <- function(x, index) {
-    .Call('_BGGM_Sigma_i_not_i', PACKAGE = 'BGGM', x, index)
-}
-
-select_col <- function(x, index) {
-    .Call('_BGGM_select_col', PACKAGE = 'BGGM', x, index)
-}
-
-select_row <- function(x, index) {
-    .Call('_BGGM_select_row', PACKAGE = 'BGGM', x, index)
-}
-
-remove_row <- function(x, which) {
-    .Call('_BGGM_remove_row', PACKAGE = 'BGGM', x, which)
-}
-
-remove_col <- function(x, index) {
-    .Call('_BGGM_remove_col', PACKAGE = 'BGGM', x, index)
-}
-
-mv_ordinal_cowles <- function(Y, X, delta, epsilon, iter, MH) {
-    .Call('_BGGM_mv_ordinal_cowles', PACKAGE = 'BGGM', Y, X, delta, epsilon, iter, MH)
-}
-
-mv_ordinal_albert <- function(Y, X, iter, delta, epsilon, K) {
-    .Call('_BGGM_mv_ordinal_albert', PACKAGE = 'BGGM', Y, X, iter, delta, epsilon, K)
-}
-
