+++ R/print_BGGM.R
@@ -1,132 +1,0 @@
-#' @name print.BGGM
-#' @title  Print method for \code{BGGM} objects
-#'
-#' @param x An object of class \code{BGGM}
-#' @param ... currently ignored
-#' @examples
-#' # data
-#' Y <- BGGM::bfi[, 1:5]
-#' # analytic approach (sample by setting analytic = FALSE)
-#' fit <- estimate(Y, analytic = TRUE)
-#' fit
-#' @export
-
-print.BGGM <- function(x,...){
-
-  # print estimate methods
-
-  if(is(x, "estimate")){
-
-    if(is(x, "default")){
-
-      print_estimate(x, ...)
-
-    } else if (is(x, "fitted")) {
-
-      print_fitted(x, ...)
-
-    } else if (is(x, "predict")){
-
-      print_predict(x, ...)
-
-    } else if(is(x, "map")){
-
-      print_map(x,...)
-
-    } else if(is(x, "coef")){
-
-      print_coef(x,...)
-
-    } else if(is(x, "ggm_compare_estimate")){
-
-
-      if(is(x, "summary")){
-
-        print_summary_ggm_estimate_compare(x,...)
-
-      } else {
-
-        print_ggm_compare(x,...)
-
-      }
-
-
-    } else if(is(x, "ggm_compare_ppc")){
-
-      print_ggm_compare_ppc(x, ...)
-
-    } else if(is(x, "metric")) {
-
-      if(is(x, "summary")){
-
-        print_summary_metric(x)
-
-      } else {
-
-        print_summary_metric(summary(x, ...))
-
-      }
-
-      # end metric
-    }  else if(is(x, "summary.estimate")){
-
-      print_summary_estimate(x)
-
-    } else if(is(x, "post.pred")){
-
-      print_post_pred(x, ...)
-
-    }  else if(is(x, "select.estimate")){
-
-      print_select_estimate(x,...)
-
-    } else if(is(x, "select.ggm_compare_estimate")){
-
-
-      print_select_ggm_compare_estimate(x, ...)
-
-    }
-
-  }
-  # explore methods
-  if(is(x, "explore")){
-
-    if(is(x, "default")){
-
-      if(is(x, "summary.explore")){
-
-      print_summary_explore(x,...)
-
-    } else {
-
-      print_explore(x,...)
-
-    }
-  }
-
-    if(is(x, "select.explore")){
-
-      print_select_explore(x, ...)
-    }
-
-    if(is(x, "select.ggm_compare_bf")){
-
-      print_select_ggm_compare_bf(x, ...)
-    }
-
-
-    if(is(x, "ggm_compare_bf")){
-
-      print_ggm_compare_bf(x, ...)
-
-    }
-
-  } # end of explore
-
-  if(is(x, "confirm")){
-
-    print_confirm(x, ...)
-
-  }  # end confirm
-
-  }
