+++ data-raw/bfi_data.R
@@ -1,2 +1,0 @@
-bfi <- read.csv('bfi.csv')
-#devtools::use_data(bfi, overwrite = T)

[…]
