+++ joss_paper/paper.bib
@@ -1,0 +1,328 @@
+@article{Epskamp2018,
+abstract = {We discuss the Gaussian graphical model (GGM; an undirected network of partial correlation coefficients) and detail its utility as an exploratory data analysis tool. The GGM shows which variables predict one-another, allows for sparse modeling of covariance structures, and may highlight potential causal relationships between observed variables. We describe the utility in 3 kinds of psychological datasets: datasets in which consecutive cases are assumed independent (e.g., cross-sectional data), temporally ordered datasets (e.g., n = 1 time series), and a mixture of the 2 (e.g., n {\textgreater} 1 time series). In time-series analysis, the GGM can be used to model the residual structure of a vector-autoregression analysis (VAR), also termed graphical VAR. Two network models can then be obtained: a temporal network and a contemporaneous network. When analyzing data from multiple subjects, a GGM can also be formed on the covariance structure of stationary means---the between-subjects network. We discuss the interpretation of these models and propose estimation methods to obtain these networks, which we implement in the R packages graphicalVAR and mlVAR. The methods are showcased in two empirical examples, and simulation studies on these methods are included in the supplementary materials.},
+archivePrefix = {arXiv},
+arxivId = {1609.04156},
+author = {Epskamp, Sacha and Waldorp, Lourens J. and Mottus, Rene and Borsboom, Denny},
+doi = {10.1080/00273171.2018.1454823},
+eprint = {1609.04156},
+journal = {Multivariate Behavioral Research},
+keywords = {Time-series analysis,exploratory-data analysis,multilevel modeling,multivariate analysis,network modeling},
+month = {jul},
+number = {4},
+pages = {453--480},
+publisher = {Routledge},
+title = {{The Gaussian Graphical Model in Cross-Sectional and Time-Series Data}},
+volume = {53},
+year = {2018}
+}
+
+@article{Schafer2005,
+abstract = {Inferring large-scale covariance matrices from sparse genomic data is an ubiquitous problem in bioinformatics. Clearly, the widely used standard covariance and correlation estimators are ill-suited for this purpose. As statistically efficient and computationally fast alternative we propose a novel shrinkage covariance estimator that exploits the Ledoit-Wolf (2003) lemma for analytic calculation of the optimal shrinkage intensity.Subsequently, we apply this improved covariance estimator (which has guaranteed minimum mean squared error, is well-conditioned, and is always positive definite even for small sample sizes) to the problem of inferring large-scale gene association networks. We show that it performs very favorably compared to competing approaches both in simulations as well as in application to real expression data.},
+author = {Sch{\"{a}}fer, Juliane and Strimmer, Korbinian},
+doi = {10.2202/1544-6115.1175},
+file = {:C$\backslash$:/Users/willidon/Downloads/[Statistical Applications in Genetics and Molecular Biology] A Shrinkage Approach to Large-Scale Covariance Matrix Estimation and Implications for Functional Genomics.pdf:pdf},
+isbn = {1544-6115},
+issn = {1544-6115},
+journal = {Statistical Applications in Genetics and Molecular Biology},
+mendeley-groups = {cov{\_}inv{\_}matrix},
+number = {1},
+pmid = {16646851},
+title = {{A Shrinkage Approach to Large-Scale Covariance Matrix Estimation and Implications for Functional Genomics}},
+volume = {4},
+year = {2005}
+}
+
+@article{Das2017,
[…]
