+++ data-raw/ptsd_cor1.R
@@ -1,0 +1,2 @@
+ptsd_cor1 <-  read.csv('ptsd_cor1.csv')
+
