+++ data-raw/bfi_data.R
@@ -1,0 +1,2 @@
+bfi <- read.csv('bfi.csv')
+#devtools::use_data(bfi, overwrite = T)
