+++ doc/ppc1.Rmd
@@ -1,0 +1,144 @@
+---
+title: "Comparing GGMs with the Posterior Predicive Distribution"
+author: "Donald R. Williams"
+output: 
+  rmarkdown::html_vignette:
+    toc: yes
+vignette: >
+  %\VignetteIndexEntry{Comparing GGMs with the Posterior Predicive Distributions}
+  %\VignetteEngine{knitr::rmarkdown}
+  \usepackage[utf8]{inputenc}
+  \usepackage{amsmath}
+params:
+  EVAL: !r identical(Sys.getenv("NOT_CRAN"), "true")
+---
+
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(collapse = TRUE, comment = "#>", fig.width = 7, fig.height = 7, fig.align = "center")
+library(ggplot2)
+library(BGGM)
+```
+
+
+# Introduction
+
+The **BGGM** package provides several options for comparing Gaussian graphical model. The approach presented 
+here is based on the posterior predictive distribution. The idea is that generated data from the fitted model should look like the observed data ($\textbf{Y}$). In the case of a well fitting model, the replicated data, herein referred to as $\textbf{Y}^{rep}$, can be viewed as data that could have been observed (but were not) or as predictive data of future observations (Rubin, 1984). We adopt the latter perspective. This is summarized in Gelman (2006):
+
+
+> “as the data that would appear if the experiment that produced $\textbf{Y}$ today were replicated
+tomorrow with the same model, $\mathcal{M}$, [and] the same (unknown) value of $\theta$ that produced $\textbf{Y}$ (pp. 737).”
+
+
+Our approach extends "experiments" to the more general “data generating process." In
+the context of comparing GGM’s, say, between two groups, the approach is to first estimate the
+GGM (i.e., the precision matrix denoted $\boldsymbol{\Theta}$ ) conditional on all of the groups being equal. Then the posterior predictive distribution can be sampled from $\boldsymbol{\Theta}$. $\textbf{Y}^{rep}$ then represents the data that we expect to observe in the future, assuming that the fitted model of group equality was the underlying data generating
+process.
+
+Assuming that each group $g \in {1, ...,G}$ is a realization from the same multivariate normal
+distribution, the null model is defined as
+
+$$
+\mathcal{M}_0 : \boldsymbol{\Theta}_1 = ... =  \boldsymbol{\Theta}_G
+$$
+
+The posterior for the common precision matrix $\boldsymbol{\Theta}(= \boldsymbol{\Theta}_1 = . . . = \boldsymbol{\Theta}_G)$, given the observed data,
+can be written as $p(\boldsymbol{\Theta}|\textbf{Y}^{obs}_1 , . . . , \textbf{Y}^{obs}_G, \mathcal{M}_0)$. Under $\mathcal{M}_0$, a posterior draw ($s$) for $\boldsymbol{\Theta}^{(s)}$ is in fact a posterior draw for the precision matrix in all groups, i.e.,   $\boldsymbol{\Theta}^{(s)} = \boldsymbol{\Theta}^{(s)}_1,..., \boldsymbol{\Theta}^{(s)}_G$.
+
+
+In review, it was pointed out by Sacha Epskamp that focusing on the precision matrix is not ideal, because it includes the diagonal elements which are not all that important for network *infernece*. Hence, to address this concern, we followed the approach in X and normalized $\boldsymbol{\Theta}$ as
+
+$$
[…]
