+++ data-raw/ptsd_cor2.R
@@ -1,1 +1,0 @@
-ptsd_cor2 <-  read.csv('ptsd_cor2.csv')
