+++ data-raw/ptsd_data.R
@@ -1,4 +1,0 @@
-ptsd <- read.csv('ptsd.csv')
-
-
-
