+++ R/zzz.R
@@ -1,10 +1,0 @@
-#' BGGM
-#'
-#' Description of your package
-#'
-#' @docType package
-#' @author you
-#' @import Rcpp
-#' @importFrom Rcpp sourceCpp
-#' @useDynLib BGGM
-NULL
