+++ data-raw/autism_ocd.R
@@ -1,1 +1,0 @@
-austism_ocd <- read.csv("autism_ocd.csv")
