+++ data-raw/csws_data.R
@@ -1,1 +1,0 @@
-csws <- read.csv("csws.csv")

[…]
