+++ R/estimate.R
@@ -1,0 +1,563 @@
+#' @title Bayesian Estimation of Gaussian Graphical Models
+#'
+#' @description Estimate the conditional (in)dependence with either an analytic solution or efficiently
+#' sampling from the posterior distribution. These methods were introduced in \insertCite{Williams2019;textual}{BGGM}.
+#' The graph is then selected with \code{\link{select.estimate}}, with either directional posterior probabilities
+#' \insertCite{Marsman2017a}{BGGM}, credible intervals, or a region of practical equivalence \insertCite{Kruschke2017}{BGGM}.
+#' Bayesian hypothesis testing is implemented in \code{\link{explore}} and \code{\link{confirm}} \insertCite{Williams2019_bf}{BGGM}.
+#'
+#' @name estimate
+#' @param Y  matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
+#'
+#' @param formula an object of class \code{\link[stats]{formula}}. This allows for including
+#' control variables in the model (i.e., \code{~ gender}).
+#'
+#' @param data an optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
+#' to a data frame containing the variables in \code{formula}. This is required when controlling for variables.
+#'
+#' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
+#' \code{binary}, \code{ordinal}, or \code{mixed}. See the note for further details.
+#'
+#' @param mixed_type numeric vector. An indicator of length p for which varibles should be treated as ranks.
+#' (1 for rank and 0 to assume normality). The default is currently to treat all integer variables as ranks
+#' when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.
+#'
+#' Treating continous data as ranks is computationally expensive and can be avoided by assuming normality.
+#'
+#' @param iter number of iterations (posterior samples; defaults to 5000).
+#'
+#' @param analytic logical. Should the analytic solution be computed (default is \code{FALSE})?
+#'
+#'
+#' @param ... currently ignored.
+#'
+#' @references
+#' \insertAllCited{}
+#'
+#'
+#' @return  list of class \code{estimate}:
+#'
+#' \code{analytic = TRUE}:
+#' \itemize{
+#' \item \code{fit} list of analytic solution estimates
+#' \itemize{
+#' \item \code{inv_mu} inverse covariance matrix (mean)
+#' \item \code{inv_var} inverse covariance matrix (variance)
+#' \item \code{partial} partial correlation matrix
+#' }
+#' \item \code{analytic} TRUE
+#' \item \code{call} match.call()
+#' \item \code{dat} data matrix
+#' \item \code{p} number of variables
+#' }
+#'
+#' \code{analytic = FALSE}:
+#' \itemize{
+#' \item \code{parcors_mat} partial correlation matrix
+#' \item \code{inv_mat} inverse covariance matrix
+#' \item \code{posterior samples} posterior samples for partial correlations and inverse covariance matrix
+#' \item \code{p} number of variables
+#' \item \code{dat} data matrix
+#' \item \code{iter} number of posterior samples
+#' \item \code{call} match.call()
+#' \item \code{analytic} FALSE
+#' }
+#'
+#'
+#'
+#' @note The default is to draws samples from the posterior distribution (\code{analytic = FALSE}). The samples are
+#' required for computing edge differences (see \code{\link{ggm_compare_estimate}}), Bayesian R2 introduced in
[…]
