+++ R/bf_groups.R
@@ -1,0 +1,517 @@
+#' Compare GGMs with Bayesian Hypothesis Testing
+#'
+#' @description Compare GGMs with the Bayes factor. This method allows for
+#' assessing (relative) evidence for edge equality or edges differences across any number of groups. Further, confirmatory hypothesis testing
+#' can be used to test predictions or expectations regarding difference or similarities in different groups (e.g., male vs. female).
+#'
+#' @param ... data matrices (\emph{n} by \emph{p}). Requires at least two.
+#' @param prior_sd hypothesized standard deviation for the edges or partial correlations.
+#' @param iter number of posterior samples.
+#' @param cores number of cores for parallel computing. The default is 2, but this can be adjusted.
+#' @param hypothesis NULL
+#'
+#' @return list of class \code{ggm_compare_bf}
+#' \itemize{
+#' \item \code{BF_01} Bayes factors in favor of the null hypothesis
+#' \item \code{p} number of variables
+#' \item  \code{info} list of information about the data matrices
+#'
+#' \itemize{
+#' \item \code{dat} list containing the data matrices
+#' \item \code{dat_info} sample size for each data matrix
+#' \item \code{pairwise} matrix of pairwise combinations
+#' }
+#' \item \code{iter} number of posterior and prior samples
+#' \item \code{call} match.call()
+#' \item \code{delta} hyperparameter of matrix-F prior distribution (corresponds to \code{prior_sd})
+#' }
+#'
+#' @note After fitting, use \code{select} to determine which partial correlations were different or the same (i.e., evidence for the null hypothesis of
+#' equality). This assumes \code{hypothesis = NULL}. If a hypothesis is tested, then use \code{summary} which provides
+#' information including the Bayes factors and posterior probabilities for each hypothesis.
+#'
+bf_groups <- function(..., prior_sd = 0.35,
+                      hypothesis = NULL,
+                      iter = 25000, cores = 2){
+
+  priorprob <- 1
+
+  info <- BGGM:::Y_combine(...)
+
+  groups <- length(info$dat)
+
+  delta <- BGGM:::delta_solve(prior_sd)
+
+  fits <- lapply(1:groups, function(x)   BGGM:::sampling(info$dat[[x]], nu = 10000,
+                                                         delta = delta,
+                                                         n_samples = iter,
+                                                         cores = cores))
+
+  # p
+  p <- info$dat_info$p[1]
+
+  edges <- 0.5 * (p * (p -1))
+
+  prior_samps <- list()
+  post_samps <- list()
+
+  for(i in 1:groups){
+
+    prior_samps[[i]] <- do.call(rbind.data.frame, lapply(1:cores, function(x)   fits[[i]][[x]]$fisher_z_prior[,1:edges]))
+
+    post_samps[[i]] <- do.call(rbind.data.frame, lapply(1:cores, function(x)   fits[[i]][[x]]$fisher_z_post[,1:edges]))
+
+  }
+
+
+  if(is.null(hypothesis)){
+
+    post_string <- list()
+    prior_string <- list()
+
+
+
+    mu_diff <- list()
+
+
+    for(i in 1:groups){
+
[…]
