+++ README.md
@@ -1,0 +1,620 @@
+
+<img src="joss_paper/hex.png" width = 250 />
+
+# BGGM: Bayesian Gaussian Graphical Models
+
+[![CRAN
+Version](http://www.r-pkg.org/badges/version/BGGM)](https://cran.r-project.org/package=BGGM)
+[![Downloads](https://cranlogs.r-pkg.org/badges/BGGM)](https://cran.r-project.org/package=BGGM)
+
+## Summary
+
+Gaussian graphical models (GGM) allow for learning conditional
+(in)dependence structures that are encoded by partial correlations.
+Whereas there are several `R` packages for classical methods (see
+Kuismin and Sillanpää 2017, Table 1), there are only two that implement
+a Bayesian approach (Leday and Richardson 2018; Mohammadi and Wit 2015).
+These are exclusively focused on identifying the graphical structure.
+The package **BGGM** not only contains novel Bayesian methods for this
+purpose, but it also includes Bayesian methodology for extending
+inference beyond identifying non-zero relations.
+
+**BGGM** is built around two approaches for *Bayesian*
+inference–estimation and hypothesis testing. This distinction is
+arbitrary (see Rouder, Haaf, and Vandekerckhove 2018), but is used to
+organize this work. The former focuses on the posterior distribution and
+includes extensions to assess predictability (Haslbeck and Waldorp
+2018), as well as methodology to compare partial correlations. The
+latter includes methods for Bayesian hypothesis testing, in both
+exploratory and confirmatory contexts, with the novel matrix-\(F\) prior
+distribution (Mulder and Pericchi 2018). This allows for testing the
+null hypothesis of conditional independence, as well as inequality and
+equality constrained hypotheses. Further, there are several approaches
+for comparing GGMs across any number of groups. The package also
+includes a suite of options for model checking. Together, **BGGM** is a
+comprehensive toolbox for Gaussian graphical modeling in `R`.
+
+## Installation
+
+To install the latest release version from CRAN use
+
+``` r
+install.packages("BGGM")
+```
+
+The current developmental version can be installed with
+
+``` r
+if (!requireNamespace("remotes")) {
+  install.packages("remotes")
+}
+remotes::install_github("donaldRwilliams/BGGM")
+```
+
+## Estimation
+
+There are two possibilities for *estimating* GGMs. The first is an
+analytic solution and the second samples from the posterior distribution
+(described in Williams 2018). Sampling is recommended because the
+samples are required for various functions in **BGGM** (e.g., posterior
+predictive checks and prediction).
+
+### Structure Learning
+
+Structure learning refers to determining which partial correlations are
+non-zero. This is implemented with:
+
+``` r
+# data (p = 5 for demonstrative purposes)
+Y <- BGGM::bfi[,1:5]
+
+# fit model
+fit <- estimate(Y = Y, 
+                analytic = FALSE, 
+                iter = 1000)
+
+# select graph
+E <- select(fit, ci_width = 0.95)
+
+# summary
+summary(E)
+
+# output
+
+BGGM: Bayesian Gaussian Graphical Models 
+--- 
+Type: Selected Graph (Analytic Solution) 
+Credible Interval: 95 % 
+Connectivity: 80 % 
+--- 
+Call:
+select.estimate(x = fit_analytic, ci_width = 0.95)
+--- 
+Selected:
+ 
+Partial correlations 
+ 
[…]
