+++ data-raw/ptsd_data.R
@@ -1,0 +1,4 @@
+ptsd <- read.csv('ptsd.csv')
+
+
+
