+++ inst/REFERENCES.bib
@@ -1,0 +1,121 @@
+@article{Kruschke2017,
+author = {Kruschke, John K. and Liddell, Torrin M.},
+doi = {10.3758/s13423-016-1221-4},
+file = {:C$\backslash$:/Users/wdona/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Kruschke, Liddell - 2017 - The Bayesian New Statistics Hypothesis testing, estimation, meta-analysis, and power analysis from a Bayesian.pdf:pdf},
+issn = {1069-9384},
+journal = {Psychonomic Bulletin {\&} Review},
+keywords = {Bayesian infe,Null hypothesis significance testing,bayes factor,bayesian,confidence interval,credible,effect size,equivalence,highest density interval,inference,interval,meta-analysis,null hypothesis significance testing,power analysis,region of practical},
+mendeley-groups = {robust{\_}t},
+pmid = {28176294},
+publisher = {Psychonomic Bulletin {\&} Review},
+title = {{The Bayesian New Statistics: Hypothesis testing, estimation, meta-analysis, and power analysis from a Bayesian perspective}},
+year = {2017}
+}
+
+
+@article{Marsman2017a,
+author = {Marsman, Maarten and Wagenmakers, Eric Jan},
+doi = {10.1177/0013164416669201},
+file = {:C$\backslash$:/Users/wdona/Box Sync/website/10.1177{\_}0013164416669201.pdf:pdf},
+isbn = {1552-3888 (Electronic)
+0013-1644 (Linking)},
+issn = {15523888},
+journal = {Educational and Psychological Measurement},
+keywords = {Bayesian inference,estimation,hypothesis testing,null hypothesis},
+mendeley-groups = {bayes{\_}general},
+number = {3},
+pages = {529--539},
+pmid = {29795927},
+title = {{Three Insights from a Bayesian Interpretation of the One-Sided P Value}},
+volume = {77},
+year = {2017}
+}
+
+
+@article{Williams2019,
+author = {Williams, Donald R.},
+doi = {10.31234/OSF.IO/X8DPR},
+journal = {arXiv},
+keywords = {Explanation,Gaussian Graphical Model,Partial Correlation,Prediction,Psychometrics,Quantitative Methods,Quantitative Psychology,Social and Behavioral Sciences,Statistical Methods},
+publisher = {PsyArXiv},
+title = {{Bayesian Estimation for Gaussian Graphical Models: Structure Learning, Predictability, and Network Comparisons}},
+year = {2019}
+}
+
+@article{Williams2019_bf,
+author = {Williams, Donald R. and Mulder, Joris},
+doi = {10.31234/osf.io/ypxd8},
+file = {:C$\backslash$:/Users/wdona/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Williams, Mulder - 2019 - Bayesian Hypothesis Testing for Gaussian Graphical Models Conditional Independence and Order Constraints.pdf:pdf},
+journal = {PsyArXiv},
+keywords = {bayes factor,gaussian graphical model,matrix- f prior,order constraints,partial correlation},
+mendeley-groups = {bayes{\_}general},
+title = {{Bayesian Hypothesis Testing for Gaussian Graphical Models: Conditional Independence and Order Constraints}},
+year = {2019}
+}
+
+
+@article{gelman_r2_2019,
+author = {Gelman, Andrew and Goodrich, Ben and Gabry, Jonah and Vehtari, Aki},
+doi = {10.1080/00031305.2018.1549100},
+issn = {15372731},
+journal = {American Statistician},
+keywords = {Bayesian methods,R-squared,Regression},
+month = {jul},
+number = {3},
+pages = {307--309},
+publisher = {American Statistical Association},
[…]
