+++ src/samplers.cpp
@@ -1,0 +1,1198 @@
+// -*- mode: C++; c-indent-level: 4; c-basic-offset: 4; indent-tabs-mode: nil; -*-
+// we only include RcppArmadillo.h which pulls Rcpp.h in for us
+#include "RcppArmadillo.h"
+#include <truncnorm.h>
+
+// [[Rcpp::depends(RcppArmadillo, RcppDist)]]
+
+
+//' @title multivariate regression
+//' @name mv_continuous
+//' @export
+
+// [[Rcpp::export]]
+Rcpp::List mv_continuous(arma::mat Y,
+                          arma::mat X,
+                          float delta,
+                          float epsilon,
+                          int iter){
+
+  // number of rows
+  int n = Y.n_rows;
+
+  // number of dependent variables
+  int k = Y.n_cols;
+
+  // number of predictors
+  int p = X.n_cols;
+
+  int nu = 1/ epsilon;
+  // #nu in Mulder & Pericchi (2018) formula (30) line 1.
+  int nuMP = delta + k - 1 ;
+
+  // #delta in Mulder & Pericchi (2018) formula (30) line 1.
+  int deltaMP = nu - k + 1 ;
+
+  // k * k identity mat
+  arma::mat  I_k(k, k, arma::fill::eye);
+
+  // p * p identity mat
+  arma::mat  I_p(p, p, arma::fill::eye);
+
+  // scatter matrix X' * X
+  arma::mat S_X(X.t() * X + I_p * 0.000001);
+
+  // inv S_X
+  arma::mat Sinv_X(inv(S_X));
+
+  // Psi update
+  arma::cube Psi(k, k, 1, arma::fill::zeros);
+
+  // scatter matrix dependent variables
+  arma::mat S_Y(k, k, arma::fill::zeros);
+  arma::mat B(epsilon*I_k);
+  arma::mat BMP(inv(B));
+  arma::mat BMPinv(inv(BMP));
+
+  // precison matrix
+  arma::cube Theta(k, k, 1, arma::fill::zeros);
+  arma::cube Theta_mcmc(k, k, iter, arma::fill::zeros);
+
+  // partial correlations
+  arma::mat pcors(k,k);
+  arma::cube pcors_mcmc(k, k, iter, arma::fill::zeros);
+
+  // correlations
+  arma::mat  cors(k,k);
+  arma::cube cors_mcmc(k, k, iter, arma::fill::zeros);
+
+  // covariance matrix
+  arma::cube Sigma(k, k, 1, arma::fill::zeros);
+  arma::cube Sigma_mcmc(k, k, iter, arma::fill::zeros);
+
+  // coefficients
+  arma::mat beta(p, k, arma::fill::zeros);
+  arma::cube beta_mcmc(p, k, iter,  arma::fill::zeros);
+
+  // starting value
+  Sigma.slice(0).fill(arma::fill::eye);
+  Psi.slice(0).fill(arma::fill::eye);
+  Theta.slice(0).fill(arma::fill::eye);
+
+  for(int s = 0; s < iter; ++s){
+
+    // draw coefficients
+    beta = reshape(mvnrnd(reshape(Sinv_X * X.t() * Y, k*p , 1),
+                          kron(Sigma.slice(0), Sinv_X)),
+                          p, k);
+
+    // scatter matrix
+    S_Y = Y.t() * Y + I_k - beta.t() * S_X * beta;
+
+    // sample Psi
+    Psi.slice(0) = wishrnd(inv(BMPinv + Theta.slice(0)), nuMP + deltaMP + k - 1);
+
[…]

[…]
