+++ man/copula.Rd
@@ -1,0 +1,8 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/RcppExports.R
+\name{copula}
+\alias{copula}
+\title{testing copula GGM}
+\description{
+testing copula GGM
+}

[…]
