+++ man/summary.explore.Rd
@@ -1,0 +1,19 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/explore.default.R
+\name{summary.explore}
+\alias{summary.explore}
+\title{Summary method for \code{explore} objects}
+\usage{
+\method{summary}{explore}(object, ...)
+}
+\arguments{
+\item{object}{An object of class \code{explore}}
+
+\item{...}{currently ignored}
+}
+\description{
+Summary method for \code{explore} objects
+}
+\seealso{
+\code{\link{explore}}
+}
