+++ man/mv_binary.Rd
@@ -1,0 +1,8 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/RcppExports.R
+\name{mv_binary}
+\alias{mv_binary}
+\title{testing binary}
+\description{
+testing binary
+}
