+++ man/mape.Rd
@@ -1,0 +1,31 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/metrics.R
+\name{mape}
+\alias{mape}
+\title{Mean Absolute Percentage Error}
+\usage{
+mape(object, ...)
+}
+\arguments{
+\item{object}{object of class \code{post.pred} or  \code{predict.estimate}}
+
+\item{...}{currently ignored}
+}
+\value{
+object of class \code{metric}
+}
+\description{
+Mean Absolute Percentage Error
+}
+\examples{
+# data
+Y <- subset(tas, gender == "M")[,-ncol(tas)]
+
+# fit model
+fit <- estimate(Y)
+
+# predict (note summary = FALSE)
+pred <- predict(fit, iter = 50, summary = FALSE)
+
+mape(pred)
+}
