### 2020-04-24 d803178
remove BF group. note was removed long ago but added in for someone to try
