### 2020-04-24 1f8e3ca
fix priors thanks to result of Jonas
