### 2020-04-24 de885d1
fix ggm compare
