+++ R/helpers.R
@@ -27,7 +27,7 @@ rank_helper <- function(Y){
 }
 
 
-
+# convert hypothesis to words
 convert_hyps <- function(hypothesis, Y){
 
   p <- ncol(Y)
@@ -40,9 +40,10 @@ convert_hyps <- function(hypothesis, Y){
 
   mat_name_actual <- sapply(colnames(Y), function(x) paste0(colnames(Y), "--", x, sep = ""))
 
-  n_off_diag <- p*(p-1)*0.5
+  n_off_diag <- p *( p - 1)*0.5
 
   where <- lapply(1:n_off_diag, function(x) grep(mat_name_actual[upper.tri(diag(p))][x], hypothesis ))
+
   where <- which(where == 1)
 
   if(any(where)){
@@ -80,8 +81,10 @@ convert_hyps <- function(hypothesis, Y){
 prior_helper_2 <- function(p, delta , epsilon){
 
   k <- p
+
   nu <- 1/epsilon
-  parcorMat <- corMat <- ThetaMat <- SigmaMat <- array(0,dim=c(1e4,k,k))
+
+  parcorMat <- corMat <- ThetaMat <- SigmaMat <- array(0, dim=c(1e4 , k , k))
 
 
   for(s in 1:dim(parcorMat)[1]){
@@ -106,7 +109,7 @@ sampling_helper_poly <- function(Y, delta, iter, type, mixed_type= NULL){
 
   if(type == "binary"){
 
-  fit_mvn <- mvn_binary(Y, matrix(1, nrow = nrow(Y)),
+  fit_mvn <- mv_binary(Y, matrix(1, nrow = nrow(Y)),
                         delta = delta,
                         epsilon = 0.001,
                         iter = iter,
@@ -115,10 +118,12 @@ sampling_helper_poly <- function(Y, delta, iter, type, mixed_type= NULL){
 
   } else if(type == "ordinal"){
 
-    fit_mvn <- mvn_ordinal(Y, matrix(1, nrow = nrow(Y)),
+    K <- max(apply(Y, 2, FUN = function(x) {max(unique(Y))} ))
+
+    fit_mvn <- mv_ordinal_albert(Y, matrix(1, nrow = nrow(Y)),
                           delta = delta,
                           epsilon = 0.1,
-                          iter = iter, MH = 0.001)
+                          iter = iter, K = K)
 
   } else if(type == "mixed"){
 
@@ -374,324 +379,81 @@ print_select_ggm_compare_bf <- function(x,...){
   cat("note: matrices (e.g., selected partial correlations) are in the select object")
 }
 
-print_select_explore <- function(x, hyp = "H1",
-                                   log = TRUE, summarize = FALSE, ...){
+print_select_explore <- function(x,
+                                 ...){
 
+  p <- ncol(x$pcor_mat_zero)
   cat("BGGM: Bayesian Gaussian Graphical Models \n")
   cat("--- \n")
   cat("Type:", x$type, "\n")
   cat("Alternative:", x$alternative, "\n")
-
-  # exhaustive
-  if(x$alternative == "exhaustive"){
-    cat("Posterior probability:", x$prob, "\n")
-    cat("--- \n")
-    cat("Call:\n")
-    print(x$call)
-    cat("--- \n")
-    if(!isFALSE(summarize)){
-
-      summ <-  cbind.data.frame(Edge = x$post_prob$edge,
-                                Estimate = x$pcor_mat[upper.tri(x$pcor_mat)],
-                                Est.Error = x$pcor_sd[upper.tri(x$pcor_sd)],
-                                "Pr.H0" = x$post_prob[,2],
[…]
