+++ R/RcppExports.R
@@ -1,6 +1,10 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
+#' @title Testing mat F
+#' @name Theta_continuous
+NULL
+
 #' @title multivariate regression
 #' @name mv_continuous
 #' @export
@@ -26,47 +30,51 @@ NULL
 #' @export
 NULL
 
+Theta_continuous <- function(Y, iter, delta, epsilon, prior_only, explore) {
+    .Call(`_BGGM_Theta_continuous`, Y, iter, delta, epsilon, prior_only, explore)
+}
+
 mv_continuous <- function(Y, X, delta, epsilon, iter) {
-    .Call('_BGGM_mv_continuous', PACKAGE = 'BGGM', Y, X, delta, epsilon, iter)
+    .Call(`_BGGM_mv_continuous`, Y, X, delta, epsilon, iter)
 }
 
 trunc_mvn <- function(mu, rinv, z, y, cutpoints) {
-    .Call('_BGGM_trunc_mvn', PACKAGE = 'BGGM', mu, rinv, z, y, cutpoints)
+    .Call(`_BGGM_trunc_mvn`, mu, rinv, z, y, cutpoints)
 }
 
 mv_binary <- function(Y, X, delta, epsilon, iter, beta_prior, cutpoints) {
-    .Call('_BGGM_mv_binary', PACKAGE = 'BGGM', Y, X, delta, epsilon, iter, beta_prior, cutpoints)
+    .Call(`_BGGM_mv_binary`, Y, X, delta, epsilon, iter, beta_prior, cutpoints)
 }
 
 Sigma_i_not_i <- function(x, index) {
-    .Call('_BGGM_Sigma_i_not_i', PACKAGE = 'BGGM', x, index)
+    .Call(`_BGGM_Sigma_i_not_i`, x, index)
 }
 
 select_col <- function(x, index) {
-    .Call('_BGGM_select_col', PACKAGE = 'BGGM', x, index)
+    .Call(`_BGGM_select_col`, x, index)
 }
 
 select_row <- function(x, index) {
-    .Call('_BGGM_select_row', PACKAGE = 'BGGM', x, index)
+    .Call(`_BGGM_select_row`, x, index)
 }
 
 remove_row <- function(x, which) {
-    .Call('_BGGM_remove_row', PACKAGE = 'BGGM', x, which)
+    .Call(`_BGGM_remove_row`, x, which)
 }
 
 remove_col <- function(x, index) {
-    .Call('_BGGM_remove_col', PACKAGE = 'BGGM', x, index)
+    .Call(`_BGGM_remove_col`, x, index)
 }
 
 mv_ordinal_cowles <- function(Y, X, delta, epsilon, iter, MH) {
-    .Call('_BGGM_mv_ordinal_cowles', PACKAGE = 'BGGM', Y, X, delta, epsilon, iter, MH)
+    .Call(`_BGGM_mv_ordinal_cowles`, Y, X, delta, epsilon, iter, MH)
 }
 
 mv_ordinal_albert <- function(Y, X, iter, delta, epsilon, K) {
-    .Call('_BGGM_mv_ordinal_albert', PACKAGE = 'BGGM', Y, X, iter, delta, epsilon, K)
+    .Call(`_BGGM_mv_ordinal_albert`, Y, X, iter, delta, epsilon, K)
 }
 
 copula <- function(z0_start, levels, K, Sigma_start, iter, delta, epsilon, idx) {
[…]
