+++ R/zzz.R
@@ -1,10 +1,9 @@
 #' BGGM
 #'
 #' Description of your package
-#'
 #' @docType package
 #' @author you
 #' @import Rcpp
 #' @importFrom Rcpp sourceCpp
-#' @useDynLib BGGM
+#' @useDynLib BGGM, .registration = TRUE
 NULL
