### 2020-04-25 f7dd546
finish print_explore
