+++ man/summary.select.explore.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/select.explore.R
+\name{summary.select.explore}
+\alias{summary.select.explore}
+\title{Summary Method for \code{select.explore} Objects}
+\usage{
+\method{summary}{select.explore}(object)
+}
+\arguments{
+\item{object}{object of class \code{select.explore}.}
+}
+\value{
+a data frame including the posterior mean, standard deviation,
+and posterior hypothesis probabilities for each relation.
+}
+\description{
+Summary Method for \code{select.explore} Objects
+}
