### 2020-04-25 9246edd
add note about the updates
