### 2020-04-25 0529cd9
analytic solve notes
