### 2020-04-25 e05a335
finish ggm compare estimate
