### 2020-04-25 d3c8065
add type and analtyic to print
