### 2020-04-25 e6eb112
add analytic to print for ggm compare estimate
