### 2020-04-26 8129621
change ggm compare BF to new matrix-F for compatability with other data types
