+++ R/ggm_compare_bf.default.R
@@ -55,129 +55,150 @@
 #' plot(sel)
 #'
 #' }
-ggm_compare_bf <- function(..., prior_sd = 0.20,
-                           iter = 25000, cores = 2){
+ggm_compare_bf <- function(...,
+                           prior_sd = 0.20,
+                           iter = 5000,
+                           cores = 2){
 
-  priorprob <- 1
+  # group info (e.g., n, p, etc)
+  info <- BGGM:::Y_combine(Y1, Y2, Y3)
 
-  info <- BGGM:::Y_combine(Y1, Y2)
-
-  groups <- length(info$dat)
-
-  if(groups < 2){
-
-    stop("must have (at least) two groups")
-
-    }
-
-  delta <- delta_solve(prior_sd)
-
-  fits <- lapply(1:groups, function(x)   sampling(info$dat[[x]], nu = 10000,
-                                                         delta = delta,
-                                                         n_samples = iter,
-                                                         cores = cores))
-
+  # number of variables
   p <- info$dat_info$p[1]
 
-  edges <- 0.5 * (p * (p -1))
+  # groups
+  groups <- length(info$dat)
 
-  prior_samps <- list()
-  post_samps <- list()
-
-  for(i in 1:groups){
-
-    prior_samps[[i]] <- do.call(rbind.data.frame, lapply(1:cores, function(x)   fits[[i]][[x]]$fisher_z_prior[,1:edges]))
-
-    post_samps[[i]] <- do.call(rbind.data.frame,  lapply(1:cores, function(x)   fits[[i]][[x]]$fisher_z_post[,1:edges]))
-
-  }
-  post_string <- list()
-  prior_string <- list()
-
-  mu_diff <- list()
-  sd_diff <- list()
-
-    for(i in 1:groups){
-
-      post_string[[i]] <-   paste0("post_samps[[", i, "]][,",  1:edges, "]", sep = "")
-
-      prior_string[[i]] <-   paste0("prior_samps[[", i, "]][,",  1:edges, "]", sep = "")
-
+  # check at least two groups
+  if(groups < 2){
+    stop("must have (at least) two groups")
     }
 
-    groups_as_words <- numbers2words(1:groups)
+  # matrix-F hyperparameter
+  delta <- BGGM:::delta_solve(prior_sd)
 
-    hyp <- paste(groups_as_words, sep = " ", collapse = "=")
+  # sample posterior
+  post_samp <- lapply(1:groups, function(x) {
 
-    framed <- framer(hyp)
+    Y <- info$dat[[x]]
 
-    mats <- create_matrices(framed = framed, varnames = groups_as_words)
+    .Call(
+      '_BGGM_Theta_continuous',
+      PACKAGE = 'BGGM',
+      Y = Y,
+      iter = iter + 50,
+      delta = delta,
+      epsilon = 0.1,
+      prior_only = 0,
+      explore = 1
+    )
+  })
 
 
-    post_string <- do.call(rbind, post_string)
-    prior_string <- do.call(rbind, prior_string)
+  # matrix dimensions for prior
+  Y_dummy <- matrix(rnorm( groups * (groups+1) ),
+                    nrow = groups+1, ncol = groups)
 
-    BF <- NA
+  # sample prior
+  prior_samp <- lapply(1:groups, function(x) {
 
-    for(i in 1:edges){
+    # Y <- info$dat[[x]]
 
-      # temporary string
-      temp_string <- paste("cov(cbind(", paste(post_string[,i], sep = " ", collapse = ","), "))")
+    .Call(
+      '_BGGM_Theta_continuous',
+      PACKAGE = 'BGGM',
+      Y = Y_dummy,
+      iter = 10000,
[…]
