+++ R/centrality.R
@@ -1,201 +1,0 @@
-#' Bayesian Centrality
-#'
-#' @description Compute Bayesian centrality. Currently the only option is for strength (see note for details).
-#' For each variable, this provides a posterior distribution for the sum of the respective partial
-#' correlations. Testing strength (e.g., differences or confirmatory testing of order contraints) is
-#' implemented in \code{test.centrality}.
-#'
-#' @name centrality
-#'
-#' @param object an object of class \code{estimate} or \code{explore}. All data types (e.g., ordinal, binary, etc.)
-#'               are supported.
-#'
-#' @param type character string. Which type of centrality? Currently the only options are
-#'             \code{strength} and \code{bridge_strength}. See note for further details.
-#'
-#' @param cluster charcter string. Which clusters or communities do the nodes belong to? This is
-#'                required when \code{type = bridge_stength}.
-#'
-#' @param select logical. Should strength be condtional on the selected model (default set to \code{FALSE})? See note for further details.
-#'
-#' @param cred numeric. Credible interval width used for selecting the graph with an \code{estimate}
-#'             object (default is \code{0.95}).
-#'
-#' @param BF_cut numeric. Evidentiary threshold used for selected the graph with an \code{explore}
-#'               object (default is \code{3})
-#'
-#'
-#'@references
-#'\insertAllCited{}
-#'
-#' @note Centrality is a new feature to \strong{BGGM} (> 1.0). At this time, only centrality indices based
-#' on the partial correlations (e.g., strength) are implemented and this will not likely change
-#' \insertCite{@see critiques on centrality in @bringmann2019centrality; @hallquist2019problems}{BGGM}. This is because
-#' strength is \emph{merely} the sum of partial correlations, which have a known (approximate) distribution
-#' (\href{https://en.wikipedia.org/wiki/Partial_correlation}{Wikipedia}). This
-#' translates into the sum also having those same properties. This is ideal for making statistical inference. In fact,
-#' it is not uncommon to test the sum of coefficients with classical methods
-#' \insertCite{@page 96, section 4.5.2, in  @baum2006introduction;textual}{BGGM}.
-#'
-#' We are open to all suggestions for implementing strength related measures: perhaps there is a particular sum
-#' of theorectial and/or substantative interest.
-#'
-#' \strong{Beware of Model Selection Bias:}
-#'
-#' The default setting is \code{select = FALSE}. This means that \emph{all} of the partial correlation for a given
-#' node are summed and not only those that were determined to be non-zero. Although it may seem that including only
-#' the selected edges leads to desirable "inference," there is an important caveat to consider.
-#' This induces a problem known as model selection bias that arises when \emph{conditioning} on the selected model..
-#'
-#' To make sense of this, note that:
-#'
-#' \enumerate{
-#' \item To be selected, the partial correlations must far enough way from zero.
-#' \item As a result of 1, edges will often be upward biased becauase they are only considered
-#'       when they are selected \insertCite{@more specifically a truncated sampling distribution, p. 5 in  @leeb2006can}{BGGM}
-#' \item This can translate into strength also having issues when computing it from only the selected edges.
-#' }
-#'
-#' Overcoming this issue is often called 'selective' or 'post-selection' inference
-#' (see the references in \href{http://bactra.org/notebooks/post-model-selection-inference.html}{weblink}).
[…]
