+++ NAMESPACE
@@ -27,7 +27,6 @@ S3method(summary,metric)
 S3method(summary,select.explore)
 export(assess_predictability)
 export(bayes_R2)
-export(centrality)
 export(confirm)
 export(convergence)
 export(copula)
