### 2020-04-26 12d3b60
document ggm compare bf
