+++ inst/REFERENCES.bib
@@ -119,3 +119,38 @@ year = {2019}
   year={2006},
   publisher={Institute of Mathematical Statistics}
 }
+
+@article{williams2020comparing,
+  title={Comparing Gaussian graphical models with the posterior predictive distribution and Bayesian model selection.},
+  author={Williams, Donald R and Rast, Philippe and Pericchi, Luis R and Mulder, Joris},
+  journal={Psychological Methods},
+  year={2020},
+  publisher={American Psychological Association}
+}
+
+@article{Kass1995,
+author = {Kass, Robert E and Raftery, Adrian E},
+file = {::},
+journal = {Journal of the American Statistical Association},
+mendeley-groups = {general},
+number = {430},
+pages = {773--795},
+title = {{Bayes Factors}},
+volume = {90},
+year = {1995}
+}
+
+@article{Mulder2018,
+author = {Mulder, Joris and Pericchi, Luis},
+doi = {10.1214/17-BA1092},
+file = {::},
+issn = {19316690},
+journal = {Bayesian Analysis},
+keywords = {hierarchical models,horsehoe prior,intrinsic prior,matrix-variate F distribution,testing inequality constraints},
+mendeley-groups = {cov{\_}inv{\_}matrix},
+number = {4},
+pages = {1--22},
+title = {{The Matrix-F Prior for Estimating and Testing Covariance Matrices}},
+year = {2018}
+}
+

[…]
