### 2020-04-26 eece51f
add helper functions
