+++ R/ggm_compare_confirm.R
@@ -1,116 +1,197 @@
+#' GGM Compare: Confirmatory Hypothesis Testing
+#'
+#' @name ggm_compare_confirm
+#'
+#' @param ... at least two matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
+#'
+#' @param hypothesis character string. hypothesis (or hypotheses) to be tested. See notes for futher details.
+#'
+#' @param formula an object of class \code{\link[stats]{formula}}. This allows for including
+#' control variables in the model (i.e., \code{~ gender}).
+#'
+#' @param data an optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
+#' to a data frame containing the variables in \code{formula}. This is required when controlling for variables.
+#'
+#' @param prior_sd The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
+#' The `default` is 0.25. See note for further details.
+#'
+#' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
+#' \code{binary}, or \code{ordinal}. See the note for further details.
+#'
+#' @param mixed_type numeric vector. An indicator of length p for which varibles should be treated as ranks.
+#' (1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables
+#' as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.
+#'
+#' @param iter number of iterations (posterior samples; defaults to 5000).
+#'
+#' @param seed The seed for random number generation (default set to \code{1}).
+#'
+#' @return
+#' @export
+#'
+#' @examples
+ggm_compare_confirm <- function(...,
+                                hypothesis,
+                                formula = NULL,
+                                data = NULL,
+                                type = "continuous",
+                                mixed_type = NULL,
+                                prior_sd = 0.20,
+                                iter = 5000,
+                                seed = 1){
 
+  # set seed
+  set.seed(seed)
 
-# sample posterior
-post_samp <- lapply(1:groups, function(x) {
+  # prior prob
+  priorprob <- 1
 
-  Y <- as.matrix(scale(info$dat[[x]], scale = FALSE))
+  # delta parameter
+  delta <- BGGM:::delta_solve(prior_sd)
 
-  .Call(
-    '_BGGM_Theta_continuous',
-    PACKAGE = 'BGGM',
-    Y = Y,
-    iter = iter + 50,
-    delta = delta,
-    epsilon = 0.1,
-    prior_only = 0,
-    explore = 1
-  )
-})
+  # group info (e.g., n, p, etc)
+  info <- BGGM:::Y_combine(Y1, Y2)
 
-# sample prior
-prior_samp <- lapply(1:groups, function(x) {
+  # number of variables
+  p <- info$dat_info$p[1]
 
-   Y <- info$dat[[x]]
+  # groups
+  groups <- length(info$dat)
 
-  .Call(
-    '_BGGM_Theta_continuous',
-    PACKAGE = 'BGGM',
-    Y = Y,
-    iter = 10000,
-    delta = delta,
-    epsilon = 0.1,
-    prior_only = 1,
-    explore = 0
-  )$fisher_z
-})
+  # number of edges
[…]
