### 2020-04-27 2239e7b
add all data types and formula to ggm_compare_estimate
