+++ R/ggm_compare_estimate.default.R
@@ -7,8 +7,21 @@
 #' @param ... matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
 #' Requires at least two.
 #'
+#' @param formula an object of class \code{\link[stats]{formula}}. This allows for including
+#' control variables in the model (i.e., \code{~ gender}).
+#'
+#' @param data an optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
+#' to a data frame containing the variables in \code{formula}. This is required when controlling for variables.
+#'
+#' @param prior_sd The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
+#' The `default` is 0.25. See note for further details.
+#'
 #' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
 #' \code{binary}, or \code{ordinal}. See the note for further details.
+#'
+#' @param mixed_type numeric vector. An indicator of length p for which varibles should be treated as ranks.
+#' (1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables
+#' as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.
 #'
 #' @param iter number of iterations (posterior samples; defaults to 5000).
 #'
@@ -53,34 +66,52 @@
 #' sel$mat_adj
 #' @export
 ggm_compare_estimate <- function(...,
+                                 formula = NULL,
+                                 data = NULL,
                                  type = "continuous",
+                                 mixed_type = NULL,
                                  analytic = FALSE,
                                  prior_sd = 0.25,
-                                 iter = 5000){
+                                 iter = 5000,
+                                 seed = 1){
 
-  if(type == "continuous"){
 
-    info <- Y_combine(...)
+  # set seed
+  set.seed(seed)
 
-    p <- info$dat_info$p[1]
+  # combine data
+  info <- Y_combine(...)
 
-    groups <- length(info$dat)
+  # number of variables
+  p <- info$dat_info$p[1]
 
-    comparisons <- nrow(info$pairwise)
+  # number of observation
+  n = info$dat_info$n[1]
+
+  # number of groups
+  groups <- length(info$dat)
+
+  # number of comparisons
+  comparisons <- nrow(info$pairwise)
+
+  # delta rho ~ beta(delta/2, delta/2)
+  delta <- delta_solve(prior_sd)
 
   if(groups < 2){
-
     stop("must have (at least) two groups")
-
     }
 
-    if(!analytic){
+  # sample
+  if(!analytic){
+    # continuous
+    if(type == "continuous"){
+      # no control variables
+      if(is.null(formula)){
 
-      delta <- delta_solve(prior_sd)
+        # posterior sample
+        post_samp <- lapply(1:groups, function(x) {
 
-      post_samp <- lapply(1:groups, function(x) {
-
-        Y <- info$dat[[x]]
+        Y <- as.matrix(scale(info$dat[[x]], scale = F))
 
         .Call(
           '_BGGM_Theta_continuous',
@@ -92,12 +123,147 @@ ggm_compare_estimate <- function(...,
           prior_only = 0,
           explore = 1
         )$pcors
-    })
+        })
+
+        # control for variables
+        } else {
+
+          # model matrix
+          X <- model.matrix(formula, data)
+
+          post_samp <- lapply(1:groups, function(x) {
+
+            Y <- as.matrix(scale(info$dat[[x]], scale = F))
+
+            .Call(
+              "_BGGM_mv_continuous",
+              Y = Y,
+              X = X,
[…]
