+++ R/RcppExports.R
@@ -1,35 +1,6 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-#' @title Testing mat F
-#' @name Theta_continuous
-NULL
-
-#' @title multivariate regression
-#' @name mv_continuous
-#' @export
-NULL
-
-#' @title testing binary
-#' @name mv_binary
-#' @export
-NULL
-
-#' @title testing ordinal cowles for thresholds
-#' @name mv_ordinal_cowles
-#' @export
-NULL
-
-#' @title testing ordinal albert and chibbs for thresholds
-#' @name mv_ordinal_albert
-#' @export
-NULL
-
-#' @title testing copula GGM
-#' @name copula
-#' @export
-NULL
-
 Theta_continuous <- function(Y, iter, delta, epsilon, prior_only, explore) {
     .Call(`_BGGM_Theta_continuous`, Y, iter, delta, epsilon, prior_only, explore)
 }
