+++ man/mv_ordinal_cowles.Rd
@@ -1,8 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RcppExports.R
-\name{mv_ordinal_cowles}
-\alias{mv_ordinal_cowles}
-\title{testing ordinal cowles for thresholds}
-\description{
-testing ordinal cowles for thresholds
-}
