+++ man/mv_continuous.Rd
@@ -1,8 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RcppExports.R
-\name{mv_continuous}
-\alias{mv_continuous}
-\title{multivariate regression}
-\description{
-multivariate regression
-}
