+++ man/Theta_continuous.Rd
@@ -1,8 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/RcppExports.R
-\name{Theta_continuous}
-\alias{Theta_continuous}
-\title{Testing mat F}
-\description{
-Testing mat F
-}

[…]
