+++ R/ggm_compare_estimate.default.R
@@ -27,6 +27,8 @@
 #'
 #' @param analytic logical. Should the analytic solution be computed (default is \code{FALSE}) ?
 #'
+#' @param seed The seed for random number generation (default set to \code{1}).
+#'
 #' @return
 #' A list of class \code{ggm_compare_estimate} containing:
 #'  \itemize{
@@ -75,7 +77,6 @@ ggm_compare_estimate <- function(...,
                                  iter = 5000,
                                  seed = 1){
 
-
   # set seed
   set.seed(seed)
 
@@ -103,26 +104,30 @@ ggm_compare_estimate <- function(...,
 
   # sample
   if(!analytic){
+
     # continuous
+
     if(type == "continuous"){
+
       # no control variables
       if(is.null(formula)){
 
         # posterior sample
         post_samp <- lapply(1:groups, function(x) {
 
-        Y <- as.matrix(scale(info$dat[[x]], scale = F))
+          # data Y_gj
+          Y <- as.matrix(scale(info$dat[[x]], scale = F))
 
-        .Call(
-          '_BGGM_Theta_continuous',
-          PACKAGE = 'BGGM',
-          Y = Y,
-          iter = iter + 50,
-          delta = delta,
-          epsilon = 0.1,
-          prior_only = 0,
-          explore = 1
-        )$pcors
+          .Call(
+            '_BGGM_Theta_continuous',
+            PACKAGE = 'BGGM',
+            Y = Y,
+            iter = iter + 50,
+            delta = delta,
+            epsilon = 0.1,
+            prior_only = 0,
+            explore = 1
+          )$pcors
         })
 
         # control for variables
@@ -131,8 +136,10 @@ ggm_compare_estimate <- function(...,
           # model matrix
           X <- model.matrix(formula, data)
 
+          # posterior sample
           post_samp <- lapply(1:groups, function(x) {
 
+            # data Y_gj
             Y <- as.matrix(scale(info$dat[[x]], scale = F))
 
             .Call(
@@ -143,27 +150,28 @@ ggm_compare_estimate <- function(...,
               epsilon = 0.1,
               iter = iter + 50
             )$pcors
-
           })
+        }
 
-          }
-
+      # binary data
       } else if(type == "binary") {
 
-
+        # intercept only
         if(is.null(formula)){
 
           X <- matrix(1, n, 1)
 
+          # model matrix
           } else {
 
-            # model matrix
             X <- model.matrix(formula, data)
 
             }
 
+        # posterior sample
         post_samp <- lapply(1:groups, function(x) {
 
+          # Y_gj
           Y <- as.matrix(info$dat[[x]])
 
           .Call(
@@ -176,8 +184,7 @@ ggm_compare_estimate <- function(...,
             beta_prior = 0.0001,
             cutpoints = c(-Inf, 0, Inf)
           )$pcors
-
-        })
+          })
 
       } else if(type == "ordinal"){
 
@@ -193,12 +200,16 @@ ggm_compare_estimate <- function(...,
 
         }
 
+        # posterior sample
         post_samp <- lapply(1:groups, function(x) {
 
+          # Y_gj
           Y <- as.matrix(info$dat[[x]])
 
+          # categories
           K <- max(apply(Y, 2, function(x) { length(unique(x))   } ))
 
+          # call c ++
           .Call("_BGGM_mv_ordinal_albert",
               Y = Y + 1,
               X = X,
@@ -206,43 +217,41 @@ ggm_compare_estimate <- function(...,
[…]
