+++ R/ggm_compare_bf.default.R
@@ -25,6 +25,8 @@
 #' (1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables
 #' as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.
 #'
+#' @param analytic logical. Should the analytic solution be computed (default is \code{FALSE}) ? See note for details.
+#'
 #' @param iter number of iterations (posterior samples; defaults to 5000).
 #'
 #' @param seed The seed for random number generation (default set to \code{1}).
@@ -113,15 +115,23 @@ ggm_compare_explore <- function(...,
                            data = NULL,
                            type = "continuous",
                            mixed_type = NULL,
+                           analytic = FALSE,
                            prior_sd = 0.20,
                            iter = 5000,
                            seed = 1){
+
+
+  # set seed
+  set.seed(seed)
 
   # group info (e.g., n, p, etc)
   info <- BGGM:::Y_combine(...)
 
   # number of variables
   p <- info$dat_info$p[1]
+
+  # number of observation
+  n = info$dat_info$n[1]
 
   # groups
   groups <- length(info$dat)
@@ -134,7 +144,13 @@ ggm_compare_explore <- function(...,
   # matrix-F hyperparameter
   delta <- BGGM:::delta_solve(prior_sd)
 
-  # sample posterior
+  if(!analytic){
+
+  if(type == "continuous"){
+
+      if(is.null(formula)){
+
+      # sample posterior
   post_samp <- lapply(1:groups, function(x) {
 
     Y <- as.matrix(scale(info$dat[[x]], scale = FALSE))
@@ -150,6 +166,146 @@ ggm_compare_explore <- function(...,
       explore = 1
     )
   })
+
+  # control for variables
+  } else {
+
+    # model matrix
+    X <- model.matrix(formula, data)
+
+    # posterior sample
+    post_samp <- lapply(1:groups, function(x) {
+
+      # data Y_gj
+      Y <- as.matrix(scale(info$dat[[x]], scale = F))
+
+      .Call(
+        "_BGGM_mv_continuous",
+        Y = Y,
+        X = X,
+        delta = delta,
+        epsilon = 0.1,
+        iter = iter + 50
+      )
+    })
+  }
+
+  } else if (type == "binary") {
+
+
+    # intercept only
+    if(is.null(formula)){
+
+      X <- matrix(1, n, 1)
+
+      # model matrix
+    } else {
+
+      X <- model.matrix(formula, data)
+
+    }
+
+
+    # posterior sample
+    post_samp <- lapply(1:groups, function(x) {
+
+      # Y_gj
+      Y <- as.matrix(info$dat[[x]])
+
+      .Call(
+        "_BGGM_mv_binary",
+        Y = Y,
+        X = X,
+        delta = delta,
+        epsilon = 0.1,
+        iter = iter + 50,
+        beta_prior = 0.0001,
+        cutpoints = c(-Inf, 0, Inf)
+      )
+    })
+
+  } else if(type == "ordinal"){
+
+    # intercept only
+    if(is.null(formula)){
+
+      X <- matrix(1, n, 1)
+
+    } else {
+
+      # model matrix
+      X <- model.matrix(formula, data)
+
+    }
+
+    # posterior sample
+    post_samp <- lapply(1:groups, function(x) {
+
+      # Y_gj
+      Y <- as.matrix(info$dat[[x]])
+
+      # categories
+      K <- max(apply(Y, 2, function(x) { length(unique(x))   } ))
+
+      # call c ++
+      .Call("_BGGM_mv_ordinal_albert",
[…]
