+++ R/estimate.R
@@ -28,6 +28,8 @@
 #'
 #' @param analytic logical. Should the analytic solution be computed (default is \code{FALSE})?
 #'
+#' @param seed The seed for random number generation (default set to \code{1}).
+#'
 #'
 #' @param ... currently ignored.
 #'
@@ -81,13 +83,13 @@
 #'
 #' \strong{Interpretation of conditional (in)dependence models for latent data:}
 #'
-#' A  tetrachoric correlation (binary data) is a special case of a polychoric correlation (ordinal data). Both relations are
-#' between "theorized normally distributed continuous latent variables"
-#' (\href{https://en.wikipedia.org/wiki/Polychoric_correlation}{Wikipedia})
-#' In both instances, the correpsonding partial correlation between observed variables is conditioned
-#' on the remaining variables in the \emph{latent} space. This implies that interpration is much the same as
-#' for continuous data, but with respect to latent variables. We refer interested reader to
-#' \insertCite{@page 2364, section 2.2, in  @webb2008bayesian;textual}{BGGM}.
+#' A  tetrachoric correlation (binary data) is a special case of a polychoric correlation (ordinal data).
+#' Both relations are between "theorized normally distributed continuous latent variables"
+#' (\href{https://en.wikipedia.org/wiki/Polychoric_correlation}{Wikipedia}). In both instances,
+#' the correpsonding partial correlation between observed variables is conditioned
+#' on the remaining variables in the \emph{latent} space. This implies that interpration
+#' is much the same as for continuous data, but with respect to latent variables.
+#' We refer interested reader to \insertCite{@page 2364, section 2.2, in  @webb2008bayesian;textual}{BGGM}.
 #'
 #'
 #' see \code{methods("estimate")}
@@ -107,430 +109,356 @@ estimate  <- function(Y,
                       data = NULL,
                       type = "continuous",
                       mixed_type = NULL,
-                      iter = 5000,
                       analytic = FALSE,
-                      ep = 0.001,...){
+                      prior_sd = 0.25,
+                      iter = 5000,
+                      seed = 1,
+                      ...){
+
+  # set seed
+  set.seed(seed)
+
+  # obervations
+  n <- nrow(Y)
+
+  # nodes
+  p <- ncol(Y)
+
+  # na omit
+  Y <- as.matrix(na.omit(Y))
+
+  # delta rho ~ beta(delta/2, delta/2)
+  delta <- delta_solve(prior_sd)
+
+  # sample posterior
+  if(!analytic){
+
+    # continuous
+    if(type == "continuous"){
+
+      # scale Y
+      Y <- scale(Y, scale = T)
+
+      # no control
+      if(is.null(formula)){
+
+        # posterior sample
+        post_samp <- .Call(
+          '_BGGM_Theta_continuous',
+          PACKAGE = 'BGGM',
+          Y = Y,
+          iter = iter + 50,
+          delta = delta,
+          epsilon = 0.1,
+          prior_only = 0,
+          explore = 1
+          )
+
+        # control for variables
+        } else {
+
+        # model matrix
+        X <- model.matrix(formula, data)
+
+        # posterior sample
+        post_samp <- .Call(
+          "_BGGM_mv_continuous",
+          Y = Y,
+          X = X,
+          delta = delta,
+          epsilon = 0.1,
+          iter = iter + 50
+        )
+
+    } # end control
+
+      # binary
+    } else if (type == "binary"){
 
 
+      # intercept only
+      if(is.null(formula)){
+        X <- matrix(1, n, 1)
+        # model matrix
+      } else {
[…]
