+++ NAMESPACE
@@ -9,6 +9,7 @@ S3method(plot,select.estimate)
 S3method(plot,select.explore)
 S3method(plot,select.ggm_compare_bf)
 S3method(plot,summary.explore)
+S3method(plot,summary.ggm_compare_estimate)
 S3method(plot,summary.ggm_compare_explore)
 S3method(plot,summary_estimate)
 S3method(pp_check,estimate)

[…]
