### 2020-04-28 8e343fd
change estimate methods and generics to c ++
