+++ R/print_BGGM.R
@@ -11,158 +11,116 @@
 #' fit
 #' @export
 
-print.BGGM <- function(x,...){
-
+print.BGGM <- function(x, ...) {
   # print estimate methods
 
-  if(is(x, "estimate")){
-
-    if(is(x, "default")){
-
+  if (is(x, "estimate")) {
+    if (is(x, "default")) {
       print_estimate(x, ...)
 
     } else if (is(x, "fitted")) {
-
       print_fitted(x, ...)
 
-    } else if (is(x, "predict")){
-
+    } else if (is(x, "predict")) {
       print_predict(x, ...)
 
-    } else if(is(x, "map")){
+    } else if (is(x, "map")) {
+      print_map(x, ...)
 
-      print_map(x,...)
+    } else if (is(x, "coef")) {
+      print_coef(x, ...)
 
-    } else if(is(x, "coef")){
-
-      print_coef(x,...)
-
-    } else if(is(x, "ggm_compare_estimate")){
-
-
-      if(is(x, "summary")){
-
-        print_summary_ggm_estimate_compare(x,...)
+    } else if (is(x, "ggm_compare_estimate")) {
+      if (is(x, "summary")) {
+        print_summary_ggm_estimate_compare(x, ...)
 
       } else {
-
-        print_ggm_compare(x,...)
+        print_ggm_compare(x, ...)
 
       }
 
 
-    } else if(is(x, "ggm_compare_ppc")){
-
+    } else if (is(x, "ggm_compare_ppc")) {
       print_ggm_compare_ppc(x, ...)
 
-    } else if(is(x, "metric")) {
-
-
-
-      if(is(x, "summary")){
-
+    } else if (is(x, "metric")) {
+      if (is(x, "summary")) {
         print_summary_metric(x)
 
       } else {
-
-
-
-        if(is(x, "R2")){
-
+        if (is(x, "R2")) {
           print_summary_metric(summary(x, ...))
 
         } else {
+          print_summary_metric(summary(x, ...))
 
-
-        print_summary_metric(summary(x, ...))
-
+        }
       }
-}
       # end metric
-    }  else if(is(x, "summary.estimate")){
-
+    }  else if (is(x, "summary.estimate")) {
       print_summary_estimate(x)
 
-    } else if(is(x, "post.pred")){
-
+    } else if (is(x, "post.pred")) {
       print_post_pred(x, ...)
 
-    }  else if(is(x, "select.estimate")){
+    }  else if (is(x, "select.estimate")) {
+      print_select_estimate(x, ...)
 
-      print_select_estimate(x,...)
-
-    } else if(is(x, "select.ggm_compare_estimate")){
-
-
+    } else if (is(x, "select.ggm_compare_estimate")) {
       print_select_ggm_compare_estimate(x, ...)
 
     }
 
   }
   # explore methods
-  if(is(x, "explore")){
+  if (is(x, "explore")) {
+    if (is(x, "default")) {
+      if (is(x, "summary.explore")) {
+        print_summary_explore(x, ...)
 
-    if(is(x, "default")){
+      } else {
+        print_explore(x, ...)
 
-      if(is(x, "summary.explore")){
-
-      print_summary_explore(x,...)
-
-    } else {
-
-      print_explore(x,...)
-
+      }
     }
-  }
 
-    if(is(x, "select.explore")){
-
-      if(is(x, "summary")){
-
-
+    if (is(x, "select.explore")) {
+      if (is(x, "summary")) {
         print_summary_select_explore(x, ...)
 
 
-        } else {
-
-      print_select_explore(x, ...)
+      } else {
[…]
