+++ R/select.ggm_compare_bf.R
@@ -37,11 +37,21 @@
 #' summary(ggm_bf_sel)
 #' }
 
-select.ggm_compare_explore <- function(object, post_prob = 0.50,...){
+select.ggm_compare_explore <- function(object,
+                                       post_prob = 0.50,
+                                       BF = NULL,...){
 
   x <- object
 
-  BF_cut = (post_prob) / (1 - post_prob)
+  if(is.null(BF)){
+
+    BF_cut = (post_prob) / (1 - post_prob)
+
+    } else {
+
+      BF_cut <- post_prob
+
+  }
 
   BF_10 <- 1/ x$BF_01
 
@@ -74,109 +84,14 @@ select.ggm_compare_explore <- function(object, post_prob = 0.50,...){
                           iter = x$iter,
                           info = x$info,
                           post_prob = post_prob,
+                          BF = BF,
                           pcor_mat_10 = pcor_mat,
                           object = object)
 
   class(returned_object) <- c("BGGM",
                               "explore",
+                              "select",
                               "select.ggm_compare_bf")
   returned_object
 
 }
-
-
-#' @title Plot \code{select.ggm_compare_bf} Objects
-#'
-#' @description This function plots the selected graph when comparing GGMs with Bayesian hypothesis
-#' testing. There are two heatmap plots in total. The first includes egdes for which there was
-#' evidence for a diffirence. The second includes edges for which there is evidence for the null
-#' hypothesis of equality. The tiles in the heatmap correspond to the Bayes factor
-#'
-#'
-#' @param x object of class \code{select.ggm_compare_bf}
-#' @param H0_low tile color for the smallest Bayes factors in the null hypothesis heatmap
-#' @param H0_high tile color for the largest Bayes factors in the null hypothesis heatmap
-#' @param H1_low tile color for the smallest Bayes factors in the alternative hypothesis heatmap
-#' @param H1_high  tile color for the largest Bayes factors in the alternative hypothesis heatmap
-#' @param ... currently ignored
-#'
-#' @return list containing two \code{ggplot} objects
-#' @export
-#'
-#' @examples
-#' \donttest{
-#' # group 1
-#' Y1 <- MASS::mvrnorm(500, mu = rep(0,16),
-#'                     Sigma =  BGGM::ptsd_cor2,
-#'                     empirical = FALSE)
-#'
-#' # group 2
-#' Y2 <- MASS::mvrnorm(500, mu = rep(0,16),
-#'                     Sigma =  BGGM::ptsd_cor2,
-#'                     empirical = FALSE)
-#' # fit model
-#' fit <- ggm_compare_bf(Y1, Y2,
-#'                      prior_sd = 0.20,
-#'                      iter = 50,
-#'                      cores = 2)
-#' # select E
-#' E <- select(fit, BF_cut = 3)
-#'
-#' # plot E
-#' plot(E)
-#' }
-plot.select.ggm_compare_bf <- function(x, H0_low = "lightblue",
-                                       H0_high = "purple",
-                                       H1_low = "yellow",
-                                       H1_high = "red", ...){
-
-  # make alternative matrix
-  alt_df <- reshape::melt(x$BF_10_adj)
-  alt_df$BF <- ifelse(alt_df$value == 0, NA, alt_df$value)
-
-  min_scale <- round(log(x$BF),2)
-
-  max_scale <- round(max(log(na.omit(alt_df$BF))),2)
-
-  alt_mat <- ggplot(data = alt_df, aes(x=as.factor(X1),
-                                       y = as.factor(X2),
-                                       fill=log(BF))) +
-    geom_tile() +
-    ylim(rev(levels(as.factor(alt_df$X1)))) +
-    scale_fill_continuous(low= H1_low,
-                          high= H1_high,
[…]
