### 2020-04-30 501605b
remove seed from ggm_compare_explore
