+++ man/ggm_compare_explore.Rd
@@ -12,8 +12,7 @@ ggm_compare_explore(
   mixed_type = NULL,
   analytic = FALSE,
   prior_sd = 0.2,
-  iter = 5000,
-  seed = 1
+  iter = 5000
 )
 }
 \arguments{
