### 2020-04-30 0898742
change default prior for estiamte
