### 2020-04-30 004be0e
remove seed from ggm_compare_estimate
