+++ R/ggm_compare_estimate.default.R
@@ -2,7 +2,8 @@
 #'
 #' @name ggm_compare_estimate
 #'
-#' @description Compare edges (partial correlations) that are estimated from groups to, say, detect a differences or equivalence.
+#' @description Compare partial correlations that are estimated from any number of groups. This method works for
+#' continuous, binary, ordinal, and mixed data.
 #'
 #' @param ... matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
 #' Requires at least two.
@@ -14,7 +15,7 @@
 #' to a data frame containing the variables in \code{formula}. This is required when controlling for variables.
 #'
 #' @param prior_sd The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
-#' The `default` is 0.25. See note for further details.
+#' The `default` is 0.50. See note for further details.
 #'
 #' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
 #' \code{binary}, or \code{ordinal}. See the note for further details.
@@ -27,7 +28,6 @@
 #'
 #' @param analytic logical. Should the analytic solution be computed (default is \code{FALSE}) ?
 #'
-#' @param seed The seed for random number generation (default set to \code{1}).
 #'
 #' @return
 #' A list of class \code{ggm_compare_estimate} containing:
@@ -414,14 +414,13 @@ summary.ggm_compare_estimate <- function(object,
 
 
   if(col_names | is.null(cn)){
+
     mat_names <- sapply(1:p , function(x) paste(1:p, x, sep = "--"))[upper.tri(I_p)]
-
-
-
 
     } else {
 
       mat_names <-  sapply(cn , function(x) paste(cn, x, sep = "--"))[upper.tri(I_p)]
+
       }
 
   dat_results <- list()
