### 2020-05-01 4d02616
remove dependencies: Rcpp, etc.
