+++ R/ggm_compare_estimate.default.R
@@ -3,7 +3,8 @@
 #' @name ggm_compare_estimate
 #'
 #' @description Compare partial correlations that are estimated from any number of groups. This method works for
-#' continuous, binary, ordinal, and mixed data.  The approach (i.e., a difference between posterior distributions) was
+#' continuous, binary, ordinal, and mixed data (a combination of categorical and continuous variables).
+#' The approach (i.e., a difference between posterior distributions) was
 #' described in  \insertCite{Williams2019;textual}{BGGM}.
 #'
 #' @param ... matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
@@ -20,15 +21,17 @@
 #' The `default` is 0.50. See note for further details.
 #'
 #' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
-#' \code{binary}, or \code{ordinal}. See the note for further details.
+#' \code{binary}, \code{ordinal}, or \code{continuous}. See the note for further details.
 #'
-#' @param mixed_type numeric vector. An indicator of length p for which varibles should be treated as ranks.
-#' (1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables
+#' @param mixed_type numeric vector. An indicator of length \emph{p} for which varibles should be treated as ranks.
+#' (1 for rank and 0 to use the 'empirical' or observed distribution). The default is currently to treat all integer variables
 #' as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.
 #'
 #' @param iter number of iterations (posterior samples; defaults to 5000).
 #'
-#' @param analytic logical. Should the analytic solution be computed (default is \code{FALSE}) ?
+#' @param analytic logical. Should the analytic solution be computed (default is \code{FALSE})? This is only available
+#'                 for continous data. Note that if \code{type = "mixed"} and \code{analytic = TRUE}, the data will
+#'                 automatically be treated as continuous.
 #'
 #' @references
 #' \insertAllCited{}
@@ -47,8 +50,54 @@
 #' This is accomplished with pairwise comparisons for each relation. In the case of three groups,
 #' for example, group 1 and group 2 are compared, then group 1 and group 3 are compared, and then
 #' group 2 and group 3 are compared. There is a full distibution for each difference that can be
-#' summarized  \code{\link{summary.ggm_compare_estimate}} and then visualized.
-#' \code{\link{plot.summary.ggm_compare_estimate}}
+#' summarized (i.e., \code{\link{summary.ggm_compare_estimate}}) and then visualized
+#' (i.e., \code{\link{plot.summary.ggm_compare_estimate}}).
+#'
+#' \strong{Selecting the Graph of Differences:}
+#'
+#' Selecting the differences is implemented in \code{\link{select.ggm_compare_estimate}}. Currently the only available
+#' option is to detect differences with  credible interval exclusion of zero. This
+#' corresponds to a directional posterior probability \insertCite{Marsman2017a}{BGGM}. For example,
+#' the probability (conditional on the model) of a difference  is at least 97.5\eqn{%} when the 95\eqn{%}
+#' credible interval excludes zero.
+#'
+#'
+#' \strong{Interpretation of conditional (in)dependence models for latent data:}
+#'
+#' A  tetrachoric correlation (binary data) is a special case of a polychoric correlation (ordinal data).
+#' Both relations are between "theorized normally distributed continuous latent variables"
+#' (\href{https://en.wikipedia.org/wiki/Polychoric_correlation}{Wikipedia}). In both instances,
+#' the correpsonding partial correlation between observed variables is conditioned
[…]
