+++ inst/REFERENCES.bib
@@ -154,3 +154,31 @@ title = {{The Matrix-F Prior for Estimating and Testing Covariance Matrices}},
 year = {2018}
 }
 
+
+@article{hoff2007extending,
+  title={Extending the rank likelihood for semiparametric copula estimation},
+  author={Hoff, Peter D and others},
+  journal={The Annals of Applied Statistics},
+  volume={1},
+  number={1},
+  pages={265--283},
+  year={2007},
+  publisher={Institute of Mathematical Statistics}
+}
+
+
+@article{Gelman1996a,
+author = {Gelman, Andrew and Meng, Xiao-Li and Stern, Hal},
+doi = {10.1.1.142.9951},
+file = {:C$\backslash$:/Users/Donny/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Gelman, Meng, Stern - 1996 - Posterior predictive assessment of model fitness via realized discrepancies. Vol.6, No.4.pdf:pdf},
+isbn = {1017-0405},
+issn = {10170405},
+journal = {Statistica Sinica},
+keywords = {and phrases,bayesian p -value,discrepancy,graphical assess-,ment,mixture model,model criticism,p -value,posterior predictive p -value,prior predictive,realized discrepancy,$\chi$ 2 test},
+mendeley-groups = {robust{\_}t},
+number = {4},
+pages = {733--807},
+title = {{Posterior predictive assessment of model fitness via realized discrepancies. Vol.6, No.4.}},
+volume = {6},
+year = {1996}
+}

[…]
