+++ R/helpers.R
@@ -8,6 +8,54 @@
 #' @import ggplot2
 
 
+remove_predictors_helper <- function(Y_groups, formula){
+
+  # number of groups
+  groups <- length(Y_groups)
+
+
+  Y_groups <- lapply(seq_len(groups), function(x)  na.omit(Y_groups[[x]]) )
+
+
+  model_matrices <- lapply(seq_len(groups) , function(x) {
+    model.matrix(formula, Y_groups[[x]])
+  })
+
+  # model matrix terms
+  mm_terms <- attr(terms(formula), "term.labels")
+
+  if(length(mm_terms) == 0){
+
+    Y_groups <- Y_groups
+
+  } else {
+
+    Y_groups <- lapply(seq_len(groups), function(x){
+
+      # check for factors
+      factor_pred <- which(paste0("as.factor(", colnames(Y_groups[[x]]), ")") %in% mm_terms)
+
+      # check for scaled
+      scale_pred <- which(paste0("scale(", colnames(Y_groups[[x]]), ")") %in% mm_terms)
+
+      # check for non factors
+      cont_pred <- which(colnames(Y_groups[[x]]) %in% mm_terms)
+
+      # remove predictors
+      Y_groups[[x]][,-c(factor_pred, cont_pred, scale_pred)]
+
+    })
+  }
+
+
+
+
+  list(Y_groups = Y_groups, model_matrices = model_matrices)
+
+}
+
+
+
 rank_helper <- function(Y){
 
   # adapted from hoff (2008). See documentation.
@@ -243,7 +291,7 @@ print_ggm_confirm <- function(x, ...){
   cat("Posterior Samples:", x$iter, "\n")
 
   for(i in 1:groups){
-    cat("  Group", paste( i, ":", sep = "") ,info$dat_info$n[[i]], "\n")
+    cat("  Group", paste( i, ":", sep = "") , info$dat_info$n[[i]], "\n")
   }
   # number of variables
   cat("Variables (p):", x$p, "\n")
@@ -625,7 +673,7 @@ print_summary_metric <- function(x, digits = 2,...){
   cat("Estimates:\n\n")
   dat <- x$summary
   colnames(dat) <- c(colnames(dat)[1:3], "Cred.lb", "Cred.ub")
-  print(as.data.frame( sapply(dat , round, digits)),
+  print(as.data.frame(dat),
         row.names = FALSE)
 }
 
@@ -714,64 +762,6 @@ summary_ggm_compare_ppc <- function(object, ...){
 print_ggm_compare_ppc <- function(x,...){
 
   print_summary_ggm_compare_ppc((summary_ggm_compare_ppc(x)))
-}
-
-
-
-
-
-
-
-print_coef <- function(x,...){
-
-  res_sigma <- x$inv_2_beta$sigma
-
-  lb <- (1 - x$cred) / 2
-
-  ub <- 1 - lb
-
-  cred_int <- stats::quantile(res_sigma, prob = c(lb, ub))
-
-  res_sigma_summ <- data.frame(Estimate = mean(res_sigma),
-                               Est.Error = sd(res_sigma),
-                               t(cred_int))
-
-  # R2
-  ypred <- t(apply(as.matrix(x$inv_2_beta$betas)[1:x$iter,], 1,
-                   function(z)  z %*% t(as.matrix(x$data[,- x$node]))))
-
-  r2 <- R2_helper(ypred, x$data[,x$node], ci_width = x$cred)
-  cred_in <- stats::quantile(r2$R2, prob = c(lb, ub))
-
[…]
