### 2020-05-06 d39d658
add p to ggm_estimate_compare
