+++ DESCRIPTION
@@ -43,7 +43,8 @@ Imports:
 Suggests: 
   knitr, 
   rmarkdown, 
-  dplyr
+  mice,
+  abind
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
