+++ NAMESPACE
@@ -25,6 +25,7 @@ S3method(summary,ggm_compare_estimate)
 S3method(summary,ggm_compare_explore)
 S3method(summary,metric)
 S3method(summary,select.explore)
+export(bggm_missing)
 export(confirm)
 export(convergence)
 export(estimate)

[…]
