### 2020-05-09 f3e0814
change to bggm_fast
