+++ R/predict.estimate.R
@@ -1,11 +1,17 @@
-#' Prediction from an \code{estimate} Object
+#' Model Predictions for \coode{estimate} Objects
+#'
 #' @name predict.estimate
+#'
 #' @param object object of class \code{estimate}
-#' @param iter iters used to compute the residuals
+#'
+#' @param iter number of posterior samples (defaults to all in the object).
+#'
 #' @param cred credible interval used for summarizing
+#'
 #' @param newdata an optional data frame for obtaining predictions (e.g., on test data)
-#' @param summary summarize the posterior samples (Default is \code{TRUE}).
-#' Setting it to \code{FALSE} can be used to then compute performance metrics.
+#'
+#' @param summary summarize the posterior samples (defaults to \code{TRUE}).
+#'
 #' @param ... currently ignored
 #'
 #' @return \code{summary = TRUE}: 3D array of dimensions n (observations),
@@ -24,67 +30,293 @@
 #' predict(fit, iter = 25)
 #' }
 #' @export
-predict.estimate <- function(object, iter = 500, cred = 0.95,
-                             newdata = NULL, summary = TRUE,...){
+predict.estimate <- function(object,
+                           newdata = NULL,
+                           summary = TRUE,
+                           cred = 0.95,
+                           iter = NULL,
+                           ...){
 
-  if(object$iter < iter){
-    stop("iter cannot be larger than the number of samples used for fitting")
-  }
-  p <- object$p
-  if(is.null(newdata)){
-  dat <- object$dat
-  n <- nrow(dat)
-  } else {
-
-    dat <- newdata
-    n <- nrow(dat)
-    if(ncol(dat) != p){
-      stop("new data not allowed (number of variables differs)")
-    }
-  }
-
-
-
+  # lower bound
   lb <- (1 - cred) / 2
+
+  # uppder bound
   ub <- 1 - lb
 
-  predict_array <- array(0, dim = c(n, 4, p ),
-                        dimnames = list(1:n,
-                                        c("Post.mean", "Post.sd",
-                                          "Cred.lb", "Cred.ub"),
-                                        paste0("node_", 1:p)))
 
-  betas <- inverse_2_beta(object, samples = iter)
+  if(is.null(iter)){
 
-  if(isTRUE(summary)){
-  for(i in 1:p){
-    beta <- betas$betas[[i]]
-    pred <- t(sapply(1:iter, function(s) {dat[,-i] %*% t(beta[s,])}))
-    predict_array[,1,i] <- apply(pred, 2, mean)
-    predict_array[,2,i] <- apply(pred, 2, sd)
-    predict_array[,3:4,i] <- t(apply(pred, 2, quantile, prob = c(lb, ub)))
-  }
-
-  returned_object <-  predict_array
-
-  class(returned_object) <- c("BGGM", "estimate", "predict")
-
-  } else {
-
-    pred <- list()
-
-    for(i in 1:p){
-
-      beta <- betas$betas[[i]]
-      pred[[i]] <- t(sapply(1:iter, function(s) {dat[,-i] %*% t(beta[s,])}))
-
-      }
-
-  returned_object <- list(pred = pred, dat = dat)
-  class(returned_object) <- c("estimate", "predict")
+    iter <- object$iter
 
   }
-returned_object
+
+  # correlations
+  cors <- pcor_to_cor(object, iter = iter)$R
+
+  if(object$type == "continuous") {
+
+    if(is.null(newdata)){
+
+      # data matrix
+      Y <- object$Y
+
+      # nodes
+      p <- ncol(Y)
+
+      # observations
[…]
