+++ R/fisher_z2r.R
@@ -7,7 +7,7 @@
 #'
 #' @examples
 #' fisher_r2z(0.5)
-fisher_z2r <- function(z){
+fisher_z_to_r <- function(z){
   r <- z2r(z)
   return(r)
 }

[…]
