### 2020-05-10 89f5485
document estimate and explore
