+++ R/estimate.R
@@ -2,91 +2,111 @@
 #'
 #' @description Estimate the conditional (in)dependence with either an analytic solution or efficiently
 #' sampling from the posterior distribution. These methods were introduced in \insertCite{Williams2019;textual}{BGGM}.
-#' The graph is then selected with \code{\link{select.estimate}}, with either directional posterior probabilities
-#' \insertCite{Marsman2017a}{BGGM}, credible intervals, or a region of practical equivalence \insertCite{Kruschke2017}{BGGM}.
-#' Bayesian hypothesis testing is implemented in \code{\link{explore}} and \code{\link{confirm}} \insertCite{Williams2019_bf}{BGGM}.
+#' The graph is selected with \code{\link{select.estimate}} and then plotted with \code{\link{plot.select}}.
 #'
 #' @name estimate
 #'
-#' @param Y  matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
+#' @param Y  Matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
 #'
-#' @param formula an object of class \code{\link[stats]{formula}}. This allows for including
+#' @param formula An object of class \code{\link[stats]{formula}}. This allows for including
 #' control variables in the model (i.e., \code{~ gender}). See the note for further details.
 #'
-#' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
-#' \code{binary}, \code{ordinal}, or \code{mixed}. See the note for further details.
+#' @param type Character string. Which type of data for \strong{Y} ? The options include \code{continuous},
+#' \code{binary}, \code{ordinal}, or \code{mixed}. Note that mixed can be used for data with only
+#' ordinal variables. See the note for further details.
 #'
-#' @param mixed_type numeric vector. An indicator of length p for which varibles should be treated as ranks.
+#' @param mixed_type Numeric vector. An indicator of length \emph{p} for which varibles should be treated as ranks.
 #' (1 for rank and 0 to assume normality). The default is currently to treat all integer variables as ranks
 #' when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.
 #'
-#' Treating continous data as ranks is computationally expensive and can be avoided by assuming normality.
+#' @param iter Number of iterations (posterior samples; defaults to 5000).
 #'
-#' @param iter number of iterations (posterior samples; defaults to 5000).
+#' @param prior_sd Scale of the prior distribution, approximately the standard deviation of a beta distribution
+#' (defaults to 0.50).
 #'
-#' @param analytic logical. Should the analytic solution be computed (default is \code{FALSE})?
+#' @param analytic Logical. Should the analytic solution be computed (default is \code{FALSE})?
 #'
-#' @param ... currently ignored.
+#' @param ... Currently ignored.
 #'
 #' @references
 #' \insertAllCited{}
 #'
 #'
-#' @return  list of class \code{estimate}:
+#' @return The returned object of class \code{estimate} contains a lot of information that
+#'         is used for printing and plotting the results. For users of \strong{BGGM}, the following
+#'         are the useful objects:
 #'
-#' \code{analytic = TRUE}:
 #' \itemize{
-#' \item \code{fit} list of analytic solution estimates
-#' \itemize{
-#' \item \code{inv_mu} inverse covariance matrix (mean)
-#' \item \code{inv_var} inverse covariance matrix (variance)
-#' \item \code{partial} partial correlation matrix
-#' }
-#' \item \code{analytic} TRUE
[…]
