+++ R/explore.default.R
@@ -3,90 +3,295 @@
 #' @name explore
 #'
 #' @description Learn the conditional (in)dependence structure with the Bayes factor using the matrix-F prior distribution
-#' \insertCite{Mulder2018}{BGGM}. It is
-#' possible to test for only positive or negative edges, as well as two sided hypothesis testing (which is the customary approach). Further
-#' there is also an exhaustive option that provides the posterior probability of the null, greater than zero, and less than zero.
+#' \insertCite{Mulder2018}{BGGM}. It is possible to test for only positive or negative edges, as well as two sided hypothesis
+#' testing (which is the customary approach). Further there is also an exhaustive option that provides the posterior
+#' probability of the null, greater than zero, and less than zero.
 #' These methods were introduced in \insertCite{Williams2019_bf;textual}{BGGM}.
+#' The graph is selected with \code{\link{select.estimate}} and
+#' then plotted with \code{\link{plot.select}}.
 #'
-#' @param Y  matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
+#' @param Y  Matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
 #'
-#' @param formula an object of class \code{\link[stats]{formula}}. This allows for including
+#' @param formula An object of class \code{\link[stats]{formula}}. This allows for including
 #' control variables in the model (i.e., \code{~ gender}).
 #'
-#' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
+#' @param type Character string. Which type of data for \strong{Y} ? The options include \code{continuous},
 #' \code{binary}, \code{ordinal}, or \code{mixed} (semi-parametric copula). See the note for further details.
 #'
-#' @param mixed_type numeric vector. An indicator of length p for which varibles should be treated as ranks.
-#' (1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables
-#' as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.
+#' @param mixed_type Numeric vector. An indicator of length p for which varibles should be treated as ranks.
+#' (1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer
+#' variables as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.
 #'
-#' @param iter number of iterations (posterior samples; defaults to 5000).
+#' @param analytic Logical. Should the analytic solution be computed (default is \code{FALSE})?
+#'                 (currently not implemented)
 #'
-#' @param prior_sd hypothesized standard deviation for the edges or partial correlations
+#' @param prior_sd Scale of the prior distribution, approximately the standard deviation of a beta distribution.
+#' (defaults to 0.25).
+#'
+#' @param iter Number of iterations (posterior samples; defaults to 5000).
 #'
 #' @param ... currently ignored.
 #'
 #' @references
 #' \insertAllCited{}
 #'
-#' @return list of class \code{explore}:
+#' @return The returned object of class \code{explore} contains a lot of information that
+#'         is used for printing and plotting the results. For users of \strong{BGGM}, the following
+#'         are the useful objects:
+#'
 #' \itemize{
-#' \item \code{parcors_mat} partial correlation matrix
-#' \item \code{parcors_sd} partial correlation standard deviations
-#' \item \code{samples} list of prior and posterior samples
-#' \itemize{
[…]
