+++ R/helpers.R
@@ -702,87 +702,7 @@ print_post_pred <- function(x,...){
 
 
 
-print_summary_ggm_compare_ppc <- function(x, ...){
-  cat("BGGM: Bayesian Gaussian Graphical Models \n")
-  cat("--- \n")
-  if(x$object$type == "nodewise"){
-    cat("Type: GGM Comparison (Nodewise Predictive Check) \n")
-  } else{
-    cat("Type: GGM Comparison (Global Predictive Check) \n")
-  }
-  p <- x$object$info$dat_info$p[1]
-  cat("Posterior Samples:", x$object$iter, "\n")
 
-  groups <- length(x$object$info$dat)
-  for (i in 1:groups) {
-    cat("  Group",
-        paste(i, ":", sep = "") ,
-        x$object$info$dat_info$n[[i]],
-        "\n")
-  }
-  cat("Variables (p):", p, "\n")
-  cat("Edges:", .5 * (p * (p-1)), "\n")
-  cat("--- \n")
-  cat("Call: \n")
-  print(x$object$call)
-  cat("--- \n")
-  cat("Estimates: \n \n")
-  if(x$object$type == "global"){
-    print(x$results, right = T, row.names = F,...)
-    cat("--- \n")
-    cat("note: \np_value = p(T(Y_rep) > T(y)|Y)\nKLD = (symmetric) Kullback-Leibler divergence")
-  }
-  if(x$object$type == "nodewise"){
-    for(i in 1:length(x$object$obs_jsd)){
-      cat(do.call(rbind, x$object$names)[[i]], "\n")
-      print(   x$results[[i]],  row.names = F, ...)
-      cat("\n")
-    }
-
-    cat("--- \n")
-    cat("note: \np_value = p(T(Y_rep) > T(y)|Y)\nKLD = (symmetric) Kullback-Leibler divergence")
-  }
-}
-
-
-
-
-
-
-
-summary_ggm_compare_ppc <- function(object, ...){
-
-  p <- object$info$dat_info$p[1]
-
-  if (object$type == "global") {
-    results <- data.frame(
-      contrast = do.call(rbind, object$names),
-      KLD =  do.call(rbind, object$obs_jsd),
-      p_value = object$pvalue
-    )
-  }
-  if (object$type == "nodewise") {
-    results <- list()
-    for (i in 1:length(object$obs_jsd)) {
-      results[[i]] <-
-        data.frame(
-          node = 1:p ,
-          KLD =  round(do.call(rbind, object$obs_jsd[[i]]), 3),
-          p_value = unlist(object$pvalue[[i]])
-        )
-      names(results)[[i]] <- object$names[[i]]
-    }
-  }
-
-  returned_object <- list(results = results,
-                          object = object)
-}
-
-
-print_ggm_compare_ppc <- function(x,...){
-
-  print_summary_ggm_compare_ppc((summary_ggm_compare_ppc(x)))
-}
 
 
 print_map <- function(x,...){

[…]
