### 2020-05-10 8bfe892
add seed and group prior
