+++ R/RcppExports.R
@@ -1,24 +1,24 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-Theta_continuous <- function(Y, iter, delta, epsilon, prior_only, explore, start) {
-    .Call(`_BGGM_Theta_continuous`, Y, iter, delta, epsilon, prior_only, explore, start)
+Theta_continuous <- function(Y, iter, delta, epsilon, prior_only, explore, start, progress) {
+    .Call(`_BGGM_Theta_continuous`, Y, iter, delta, epsilon, prior_only, explore, start, progress)
 }
 
 sample_prior <- function(Y, iter, delta, epsilon, prior_only, explore) {
     .Call(`_BGGM_sample_prior`, Y, iter, delta, epsilon, prior_only, explore)
 }
 
-mv_continuous <- function(Y, X, delta, epsilon, iter) {
-    .Call(`_BGGM_mv_continuous`, Y, X, delta, epsilon, iter)
+mv_continuous <- function(Y, X, delta, epsilon, iter, start, progress) {
+    .Call(`_BGGM_mv_continuous`, Y, X, delta, epsilon, iter, start, progress)
 }
 
 trunc_mvn <- function(mu, rinv, z, y, cutpoints) {
     .Call(`_BGGM_trunc_mvn`, mu, rinv, z, y, cutpoints)
 }
 
-mv_binary <- function(Y, X, delta, epsilon, iter, beta_prior, cutpoints) {
-    .Call(`_BGGM_mv_binary`, Y, X, delta, epsilon, iter, beta_prior, cutpoints)
+mv_binary <- function(Y, X, delta, epsilon, iter, beta_prior, cutpoints, start, progress) {
+    .Call(`_BGGM_mv_binary`, Y, X, delta, epsilon, iter, beta_prior, cutpoints, start, progress)
 }
 
 Sigma_i_not_i <- function(x, index) {
@@ -45,12 +45,12 @@ mv_ordinal_cowles <- function(Y, X, delta, epsilon, iter, MH) {
     .Call(`_BGGM_mv_ordinal_cowles`, Y, X, delta, epsilon, iter, MH)
 }
 
-mv_ordinal_albert <- function(Y, X, iter, delta, epsilon, K) {
-    .Call(`_BGGM_mv_ordinal_albert`, Y, X, iter, delta, epsilon, K)
+mv_ordinal_albert <- function(Y, X, iter, delta, epsilon, K, start, progress) {
+    .Call(`_BGGM_mv_ordinal_albert`, Y, X, iter, delta, epsilon, K, start, progress)
 }
 
-copula <- function(z0_start, levels, K, Sigma_start, iter, delta, epsilon, idx) {
-    .Call(`_BGGM_copula`, z0_start, levels, K, Sigma_start, iter, delta, epsilon, idx)
+copula <- function(z0_start, levels, K, Sigma_start, iter, delta, epsilon, idx, progress) {
+    .Call(`_BGGM_copula`, z0_start, levels, K, Sigma_start, iter, delta, epsilon, idx, progress)
 }
 
 pcor_to_cor_internal <- function(x, p) {
