+++ R/ggm_compare_confirm.R
@@ -1,18 +1,24 @@
 #' GGM Compare: Confirmatory Hypothesis Testing
 #'
+#' #' @description Confirmatory hypothesis testing for comparing GGMs. Hypotheses are expressed as equality
+#' and/or ineqaulity contraints on the partial correlations of interest. Here the focus is \emph{not}
+#' on determining the graph (see \code{\link{explore}}) but testing specific hypotheses related to
+#' the conditional (in)dependence structure. These methods were introduced in
+#' \insertCite{Williams2019_bf;textual}{BGGM} and in \insertCite{williams2020comparing;textual}{BGGM}
+#'
 #' @name ggm_compare_confirm
 #'
-#' @param ... at least two matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (nodes).
+#' @param ... At least two matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (nodes).
 #'
-#' @param hypothesis character string. hypothesis (or hypotheses) to be tested. See notes for futher details.
+#' @param hypothesis Character string. The hypothesis (or hypotheses) to be tested. See notes for futher details.
 #'
 #' @param formula an object of class \code{\link[stats]{formula}}. This allows for including
 #' control variables in the model (i.e., \code{~ gender}).
 #'
-#' @param prior_sd The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
-#' The `default` is 0.25. See note for further details.
+#' @param prior_sd Numeric. The scale of the prior distribution (centered at zero),
+#'                 in reference to a beta distribtuion (defaults to 0.25).
 #'
-#' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
+#' @param type Character string. Which type of data for \strong{Y} ? The options include \code{continuous},
 #' \code{binary}, or \code{ordinal}. See the note for further details.
 #'
 #' @param mixed_type numeric vector. An indicator of length p for which varibles should be treated as ranks.
@@ -23,10 +29,113 @@
 #'
 #' @param seed The seed for random number generation (default set to \code{1}).
 #'
-#' @return
-#' @export
+#' @return The returned object of class \code{confirm} contains a lot of information that
+#'         is used for printing and plotting the results. For users of \strong{BGGM}, the following
+#'         are the useful objects:
+#'
+#' \itemize{
+#'
+#' \item \code{out_hyp_prob} Posterior hypothesis probabilities.
+#'
+#' \item \code{info} An object of class \code{BF} from the R package \strong{BFpack}.
+#'
+#' }
+#'
+#' @details
+#' The hypotheses can be written either with the respective column names or numbers.
+#' For example, \code{g1_1--2} denotes the relation between the variables in column 1 and 2 for group 1.
+#' The \code{g1_1} is required and the only difference from \code{\link{confirm}} (one group).
+#' Note that these must correspond to the upper triangular elements of the correlation
+#' matrix. This is accomplished by ensuring that the first number is smaller than the second number.
+#' This also applies when using column names (i.e,, in reference to the column number).
+#'
+#' \strong{One Hypothesis}:
+#'
+#' \itemize{
+#'
+#' \item  To test whether a relation in larger in one group, while both are expected
+#'        to be positive,  this can be written as
+#'
+#'  \code{hyp <-  c(g1_1--2 > g2_1--2 > 0)},
+#'
+#'  This is then compared to the complement.
+#'}
+#'
+#' \strong{More Than One Hypothesis}:
+#'
+#' \itemize{
+#'
+#' \item The above hypothesis can also be compared to, say, a null model by using ";"
+#'       to seperate the hypotheses, for example,
+#'
+#' \code{hyp <-  c(g1_1--2 > g2_1--2 > 0; g1_1--2 = g2_1--2 = 0)}.
+#'
[…]
