### 2020-05-11 4166d37
add progress bar to roll your own
