### 2020-05-12 e8bb6f3
add documentation to ggm_compare_explore
