+++ R/explore.default.R
@@ -74,10 +74,6 @@
 #'  vector of length \emph{p}. A one indicates to use the ranks, whereas a zero indicates to "ignore"
 #'  that variable. By default all integer variables are handled as ranks.
 #'
-#'
-#'
-#'
-#'
 #' \strong{Dealing with Errors}:
 #'
 #' An error is most likely to arise when \code{type = "ordinal"}. The are two common errors (although still rare):
