+++ R/ggm_compare_bf.default.R
@@ -2,28 +2,28 @@
 #'
 #' @name ggm_compare_explore
 #'
-#' @description Compare Gaussian graphical models with (exploratory) Bayesian hypothesis testing using the matrix-F prior
-#' distribution \insertCite{Mulder2018}{BGGM}. A test for each partial correlation in the model for any number of groups.
-#' This provides evidence for the null hypothesis of no difference and the alternative hypothesis
-#' of difference. With more than two groups, the test is for \emph{all} groups simultaneously (i.e., the relation is the same or different in all groups).
-#' This method was introduced in \insertCite{williams2020comparing;textual}{BGGM}. For confirmatory hypothesis testing see
-#' \code{confirm_groups}.
+#' @description Compare Gaussian graphical models with (exploratory) hypothesis testing using the matrix-F prior
+#' distribution \insertCite{Mulder2018}{BGGM}. A test for each partial correlation in the model for any number
+#' of groups. This provides evidence for the null hypothesis of no difference and the alternative hypothesis
+#' of difference. With more than two groups, the test is for \emph{all} groups simultaneously (i.e., the relation
+#' is the same or different in all groups). This method was introduced in \insertCite{williams2020comparing;textual}{BGGM}.
+#' For confirmatory hypothesis testing see \code{confirm_groups}.
 #'
-#' @param ... at least two matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
+#' @param ... At least two matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
 #'
-#' @param formula an object of class \code{\link[stats]{formula}}. This allows for including
+#' @param formula An object of class \code{\link[stats]{formula}}. This allows for including
 #' control variables in the model (i.e., \code{~ gender}).
 #'
-#' @param data an optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
+#' @param data An optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
 #' to a data frame containing the variables in \code{formula}. This is required when controlling for variables.
 #'
-#' @param prior_sd The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
+#' @param prior_sd Numeric. The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
 #' The `default` is 0.25. See note for further details.
 #'
-#' @param type character string. Which type of data for \strong{Y} ? The options include \code{continuous},
+#' @param type Character string. Which type of data for \code{Y} ? The options include \code{continuous},
 #' \code{binary}, or \code{ordinal}. See the note for further details.
 #'
-#' @param mixed_type numeric vector. An indicator of length p for which varibles should be treated as ranks.
+#' @param mixed_type Numeric vector. An indicator of length p for which varibles should be treated as ranks.
 #' (1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables
 #' as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.
 #'
@@ -39,80 +39,159 @@
 #' \insertAllCited{}
 #'
 #'
-#' @return list of class \code{ggm_compare_bf}
-#' \itemize{
-#' \item \code{BF_01}  Bayes factors in favor of the null hypothesis.
-#' \item \code{pcor_diff} partial correlation difference (for two groups only)
-#' \item \code{p} number of variables
-#' \item  \code{info} list of information about the data matrices
[…]
