+++ man/ggm_compare_explore.Rd
@@ -17,21 +17,21 @@ ggm_compare_explore(
 )
 }
 \arguments{
-\item{...}{at least two matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}
+\item{...}{At least two matrices (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}
 
-\item{formula}{an object of class \code{\link[stats]{formula}}. This allows for including
+\item{formula}{An object of class \code{\link[stats]{formula}}. This allows for including
 control variables in the model (i.e., \code{~ gender}).}
 
-\item{type}{character string. Which type of data for \strong{Y} ? The options include \code{continuous},
+\item{type}{Character string. Which type of data for \code{Y} ? The options include \code{continuous},
 \code{binary}, or \code{ordinal}. See the note for further details.}
 
-\item{mixed_type}{numeric vector. An indicator of length p for which varibles should be treated as ranks.
+\item{mixed_type}{Numeric vector. An indicator of length p for which varibles should be treated as ranks.
 (1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer variables
 as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.}
 
 \item{analytic}{logical. Should the analytic solution be computed (default is \code{FALSE}) ? See note for details.}
 
-\item{prior_sd}{The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
+\item{prior_sd}{Numeric. The scale of the prior distribution (centered at zero), in reference to a beta distribtuion.
 The `default` is 0.25. See note for further details.}
 
 \item{iter}{number of iterations (posterior samples; defaults to 5000).}
@@ -40,87 +40,167 @@ The `default` is 0.25. See note for further details.}
 
 \item{seed}{An integer for the random seed.}
 
-\item{data}{an optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
+\item{data}{An optional data frame, list or environment (or an object coercible by \code{\link[base]{as.data.frame}})
 to a data frame containing the variables in \code{formula}. This is required when controlling for variables.}
 }
 \value{
-list of class \code{ggm_compare_bf}
-\itemize{
-\item \code{BF_01}  Bayes factors in favor of the null hypothesis.
-\item \code{pcor_diff} partial correlation difference (for two groups only)
-\item \code{p} number of variables
-\item  \code{info} list of information about the data matrices
+The returned object of class \code{ggm_compare_explore} contains a lot of information that
+        is used for printing and plotting the results. For users of \strong{BGGM}, the following
+        are the useful objects:
 
 \itemize{
-\item \code{dat} list containing the data matrices
-\item \code{dat_info} sample size for each data matrix
-\item \code{pairwise} matrix of pairwise combinations
-}
-\item \code{iter} number of posterior and prior samples
-\item \code{call} match.call()
-\item \code{delta} hyperparameter of matrix-F prior distribution (corresponds to \code{prior_sd})
 
-\item \code{groups} number of groups
-\item \code{post_samp} matrix of posterior samples
+\item \code{BF_01} A \emph{p} by \emph{p} matrix including
+                    the Bayes factor for the null hypothesis.
+
+\item \code{pcor_diff} A \emph{p} by \emph{p} matrix including
+                       the difference in partial correlations (only for two groups).
+
+\item \code{samp} A list containing the fitted models (of class \code{explore}) for each group.
+
 }
 }
 \description{
-Compare Gaussian graphical models with (exploratory) Bayesian hypothesis testing using the matrix-F prior
[…]
