+++ R/roll_your_own.R
@@ -24,7 +24,6 @@
 #' The user has complete control of this function. Hence, care must be taken as to what \code{FUN} returns
 #' and in what format.
 #'
-#' @export
 #'
 #' @examples
 #' \donttest{
@@ -88,6 +87,7 @@
 #' # full distribution (node 1)
 #' hist(net_stat[1,])
 #' }
+#' @export
 roll_your_own <- function(object,
                           FUN,
                           iter = NULL,
@@ -131,6 +131,162 @@ roll_your_own <- function(object,
     est
   })
 
-  results
+  returned_object <- list(results = results, iter = iter)
+
+  class(returned_object) <- c("BGGM",
+                              "roll_your_own")
+
+  return(returned_object)
+  }
+
+
+print_roll_your_own <- function(x, cred = 0.95, ...) {
+  cat("BGGM: Bayesian Gaussian Graphical Models \n")
+  cat("--- \n")
+  cat("Network Stats: Roll Your Own\n")
+  cat("Posterior Samples:", x$iter, "\n")
+  cat("--- \n")
+  cat("Estimates: \n\n")
+  lb <- (1-cred) / 2
+  ub <- 1 - lb
+  dims <- dim(x$results)
+  if(is.null(dims)){
+    mu <- mean(x$results)
+    scale <- sd(x$results)
+    res <- data.frame(Post.mean = round(mean(x$results), 3),
+                      Post.sd =    round(sd(x$results), 3),
+                      Cred.lb = round(quantile(x$results, probs = lb), 3),
+                      Cred.ub = round(quantile(x$results, probs = lb), 3) )
+  } else {
+
+    mu <-  apply( x$results, 1, mean)
+    p <- length(mu)
+    scale <- apply( x$results, 1, sd)
+    ci_lb <- apply( x$results, 1, quantile, lb)
+    ci_ub <- apply( x$results, 1, quantile, ub)
+
+    res<- data.frame(Node = 1:p, Post.mean = round(mu, 3),
+                     Post.sd = round(scale, 3),
+                     Cred.lb = round(ci_lb, 3),
+                     Cred.ub = round(ci_ub, 3))
+    }
+
+  print(res, row.names = FALSE)
+  cat("--- \n")
 }
 
+
+#' Plot \code{roll_your_own} Objects
+#'
+#' @param x An object of class \code{roll_your_own}
+#'
+#' @param fill Character string specifying the color for the ridges.
+#'
+#' @param alpha Numeric. Transparancey of the ridges
+#'
+#' @param ... Currently ignored
+#'
+#' @return An object of class \code{ggplot}
+#'
+#' @importFrom ggridges stat_density_ridges
+#'
+#' @examples
+#' \donttest{
+#' ####################################
+#' ###### example 1: assortment #######
+#' ####################################
+#' library(assortnet)
+#'
+#' Y <- BGGM::bfi[,1:10]
+#' membership <- c(rep("a", 5), rep("c", 5))
+#'
+#' # fit model
+#' fit <- estimate(Y = Y,
+#'                 analytic = FALSE,
+#'                 iter = 1000)
+#'
+#' # list of columns belowinging in each group
+#' e.g., first 5 are "a", last 5 are "c"
+#'
+#' membership <- c(rep("a", 5), rep("c", 5))
+#'
+#'f <- function(x,...){
+#' assortment.discrete(x, ...)$r
+#'}
+#'
+#'
+#' net_stat <- roll_your_own(object = fit,
+#'                           FUN = f,
+#'                           types = membership,
+#'                           weighted = TRUE,
+#'                           SE = FALSE, M = 1)
+#'
+#' # print
+#' net_stat
+#'
+#' # plot
+#' plot(net_stat)
+#'
+#'
[…]
