+++ NAMESPACE
@@ -4,6 +4,7 @@ S3method(coef,estimate)
 S3method(coef,explore)
 S3method(plot,ggm_compare_ppc)
 S3method(plot,predictability)
+S3method(plot,roll_your_own)
 S3method(plot,select)
 S3method(plot,summary.estimate)
 S3method(plot,summary.explore)
