### 2020-05-12 d0a2ac3
finish roll_your_own, including a check for zeros
