+++ R/roll_your_own.R
@@ -21,9 +21,15 @@
 #'
 #' @details
 #'
-#' The user has complete control of this function. Hence, care must be taken as to what \code{FUN} returns
-#' and in what format.
+#' The user has complete control of this function. Hence, care must be taken as to what \code{FUN}
+#' returns and in what format. The function should return a single number (one for the entire GGM)
+#' or a vector (one for each node). This ensures that the print and \code{\link{plot.roll_your_own}}
+#' will work.
 #'
+#'
+#' When \code{select = TRUE}, the graph is selected and then the network statistics are computed based on
+#' the weigthed adjacency matrix. This is accomplished internally by multiplying each of the sampled
+#' partial correlation matrices by the adjacency matrix.
 #'
 #' @examples
 #' \donttest{
@@ -31,6 +37,7 @@
 #' ####################################
 #' ###### example 1: assortment #######
 #' ####################################
+#' # assortment from this package
 #' library(assortnet)
 #'
 #' Y <- BGGM::bfi[,1:10]
@@ -47,7 +54,7 @@
 #' membership <- c(rep("a", 5), rep("c", 5))
 #'
 #'f <- function(x,...){
-#' assortment.discrete(x, ...)$r
+#' assortnet::assortment.discrete(x, ...)$r
 #'}
 #'
 #'
@@ -57,12 +64,16 @@
 #'                           weighted = TRUE,
 #'                           SE = FALSE, M = 1)
 #'
-#' hist(net_stat)
+#' # print
+#' net_stat
 #'
+#' # plot
+#' plot(net_stat)
 #'
 #' ############################################
 #' ###### example 2: expected influence #######
 #' ############################################
+#' # expected influence from this package
 #' library(networkTools)
 #'
 #' # data
@@ -73,19 +84,51 @@
 #'
 #' # define function
 #' f <- function(x,...){
-#'   expectedInf(x,...)$step1
+#'   networktools::expectedInf(x,...)$step1
 #' }
 #'
 #' # compute
 #' net_stat <- roll_your_own(object = fit,
 #'                           FUN = f,
 #'                           iter = 1000)
+#' # print
+#' net_stat
 #'
-#' colmeans
-#' colMeans(t(net_stat))
+#' # plot
+#' plot(net_stat)
 #'
-#' # full distribution (node 1)
-#' hist(net_stat[1,])
+#'
+#' #######################################
+#' ### example 3: mixed data & bridge ####
+#' #######################################
+#' # bridge from this package
+#' library(networkTools)
+#'
+#' # data
+#' Y <- ptsd
+#'
+#' fit <- estimate(Y,
+#'                 type = "mixed",
+#'                 iter = 1000)
+#'
+#' # clusters
+#' communities <- substring(colnames(Y), 1, 1)
+#'
+#' # function is slow
+#' f <- function(x, ...){
+#'  networktools::bridge(x, ...)$`Bridge Strength`
+#' }
+#'
+#' net_stat <- roll_your_own(fit,
+#'                           FUN = f,
+#'                           select = TRUE,
+#'                           communities = communities)
+#'
+#' # print
+#' net_stat
+#'
+#' #plot
+#' plot(net_stat)
 #' }
 #' @export
 roll_your_own <- function(object,
@@ -147,16 +190,22 @@ print_roll_your_own <- function(x, cred = 0.95, ...) {
   cat("Posterior Samples:", x$iter, "\n")
   cat("--- \n")
   cat("Estimates: \n\n")
+
   lb <- (1-cred) / 2
   ub <- 1 - lb
+
   dims <- dim(x$results)
+
[…]
