+++ DESCRIPTION
@@ -36,7 +36,9 @@ Suggests:
   MASS,
   mice,
   mvnfast,
-  rmarkdown
+  rmarkdown,
+  assortnet,
+  networkTools
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
