+++ R/ggm_compare_confirm.R
@@ -1,6 +1,6 @@
 #' GGM Compare: Confirmatory Hypothesis Testing
 #'
-#' #' @description Confirmatory hypothesis testing for comparing GGMs. Hypotheses are expressed as equality
+#' @description Confirmatory hypothesis testing for comparing GGMs. Hypotheses are expressed as equality
 #' and/or ineqaulity contraints on the partial correlations of interest. Here the focus is \emph{not}
 #' on determining the graph (see \code{\link{explore}}) but testing specific hypotheses related to
 #' the conditional (in)dependence structure. These methods were introduced in
@@ -175,12 +175,18 @@ ggm_compare_confirm <- function(...,
 
       post_samp <- lapply(1:groups, function(x) {
 
-        message("BGGM: Posterior Sampling ", "(Group ",x ,")")
+        if(isTRUE(progress)){
+
+          message("BGGM: Posterior Sampling ", "(Group ",x ,")")
+
+        }
 
         # data
         Y <- as.matrix(scale(dat_list[[x]], scale = F))
 
         Y <- na.omit(Y)
+
+        start <- solve(cov(Y))
 
         .Call(
           '_BGGM_Theta_continuous',
@@ -188,9 +194,11 @@ ggm_compare_confirm <- function(...,
           Y = Y,
           iter = iter + 50,
           delta = delta,
-          epsilon = 0.1,
+          epsilon = 0.01,
           prior_only = 0,
-          explore = 1
+          explore = 1,
+          start = start,
+          progress = progress
         )
       })
 
@@ -199,7 +207,11 @@ ggm_compare_confirm <- function(...,
 
       post_samp <- lapply(1:groups, function(x) {
 
-        message("BGGM: Posterior Sampling ", "(Group ",x ,")")
+        if(isTRUE(progress)){
+
+          message("BGGM: Posterior Sampling ", "(Group ",x ,")")
+
+        }
 
         control_info <- remove_predictors_helper(list(as.data.frame(dat_list[[x]])),
                                                  formula = formula)
@@ -216,14 +228,18 @@ ggm_compare_confirm <- function(...,
         # model matrix
         X <- as.matrix(control_info$model_matrices[[1]])
 
+        start <- solve(cov(Y))
+
         # posterior sample
         .Call(
           "_BGGM_mv_continuous",
           Y = Y,
           X = X,
           delta = delta,
-          epsilon = 0.1,
-          iter = iter + 50
+          epsilon = 0.01,
+          iter = iter + 50,
+          start = start,
+          progress = progress
         )
       })
     }
@@ -235,7 +251,12 @@ ggm_compare_confirm <- function(...,
 
 
       post_samp <- lapply(1:groups, function(x) {
-        message("BGGM: Posterior Sampling ", "(Group ",x ,")")
+
+        if(isTRUE(progress)){
+
+          message("BGGM: Posterior Sampling ", "(Group ",x ,")")
+
+        }
 
         # data
         Y <- as.matrix(na.omit(dat_list[[x]]))
@@ -248,16 +269,20 @@ ggm_compare_confirm <- function(...,
 
         X <- matrix(1, n, 1)
 
+        start <- solve(cov(Y))
+
         # posterior sample
       .Call(
         "_BGGM_mv_binary",
         Y = Y,
         X = X,
         delta = delta,
-        epsilon = 0.1,
+        epsilon = 0.01,
         iter = iter + 50,
         beta_prior = 0.0001,
-        cutpoints = c(-Inf, 0, Inf)
+        cutpoints = c(-Inf, 0, Inf),
+        start = start,
+        progress = progress
       )
       })
 
[…]
