+++ DESCRIPTION
@@ -38,7 +38,7 @@ Suggests:
   mvnfast,
   rmarkdown,
   assortnet,
-  networkTools
+  networktools
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
