+++ R/helpers.R
@@ -2,9 +2,9 @@
 #' na.omit pnorm quantile rWishart
 #' sd qnorm residuals fitted density
 #' @importFrom utils combn
-#' @importFrom foreach %dopar% foreach
 #' @importFrom graphics plot
 #' @importFrom Rdpack reprompt
+#' @importFrom MASS ginv
 #' @import ggplot2
 
 
@@ -67,8 +67,12 @@ binary_latent_helper <- function(x){
   thresholds <- c(-Inf, 0, Inf)
 
   # latent data
-  latent_data <- sapply(1:p, function(z) qnorm(runif(n, min = pnorm(thresholds[x[,z]], mean = 0, sd = 1),
-                                                     max = pnorm(thresholds[x[,z]+1], mean = 0, sd = 1)),
+  latent_data <- sapply(1:p, function(z) qnorm(runif(n, min = pnorm(thresholds[x[,z]],
+                                                                    mean = 0,
+                                                                    sd = 1),
+                                                     max = pnorm(thresholds[x[,z]+1],
+                                                                 mean = 0,
+                                                                 sd = 1)),
                                                mean = 0, sd = 1))
   # latent data (sd = 1)
   latent_data <- scale(latent_data)
@@ -86,8 +90,12 @@ ordinal_latent_helper <- function(x, thresholds){
 
   thresholds <- t(sapply(1:p, function(x) colMeans(thresholds[,,x])))
 
-  latent_data <- sapply(1:p, function(z)  qnorm(runif(n, min=pnorm(thresholds[z, x[,z]], mean = 0, sd = 1),
-                                                      max=pnorm(thresholds[z, x[,z]+1], mean = 0, sd = 1)),
+  latent_data <- sapply(1:p, function(z)  qnorm(runif(n, min = pnorm(thresholds[z, x[,z]],
+                                                                     mean = 0,
+                                                                     sd = 1),
+                                                         max = pnorm(thresholds[z, x[,z]+1],
+                                                                     mean = 0,
+                                                                     sd = 1)),
                                                 mean = 0, sd = 1))
 
   latent_data <- scale(latent_data)
@@ -119,7 +127,7 @@ rank_helper <- function(Y){
 
 group_hyp_helper <- function(hypothesis, x){
 
-  hyp <- gsub(BGGM:::hyp_converter(BGGM:::convert_hyps(hypothesis = hypothesis, cbind(x)))$hyp_converted,
+  hyp <- gsub(hyp_converter(convert_hyps(hypothesis = hypothesis, cbind(x)))$hyp_converted,
               pattern = "_", replacement = "")
   hyp
 }
@@ -130,7 +138,7 @@ convert_hyps <- function(hypothesis, Y){
 
   p <- ncol(Y)
 
-  col_names <- BGGM:::numbers2words(1:p)
+  col_names <- numbers2words(1:p)
 
   mat_name_num <- sapply(1:p, function(x) paste0(1:p, "--", x, sep = ""))
 
@@ -316,140 +324,6 @@ sampling_helper_poly <- function(Y, delta, iter, type, mixed_type= NULL){
 
 
 }
-
-
-
-
-print_ggm_confirm <- function(x, ...){
[…]
