+++ .github/workflows/try_conversion.yml
@@ -1,46 +1,0 @@
-on:
-  push:
-    branches:
-      - master
-  pull_request:
-    branches:
-      - master
-
-name: R-CMD-check
-
-jobs:
-  R-CMD-check:
-    runs-on: ${{ matrix.config.os }}
-
-    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
-
-    strategy:
-      fail-fast: false
-      matrix:
-        config:
-        - { os: windows-latest, r: 'devel'}
-        - { os: macOS-latest, r: 'devel'}
-        - { os: ubuntu-16.04, r: '3.6', cran: "https://demo.rstudiopm.com/all/__linux__/xenial/latest"}
-
-    env:
-      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
-      CRAN: ${{ matrix.config.cran }}
-
-    steps:
-      - name: Configure git
-        run: git config --global core.autocrlf false
-
-      - uses: actions/checkout@v1
-
-      - uses: r-lib/actions/setup-r@master
-        with:
-          r-version: ${{ matrix.config.r }}
-
-      - uses: r-lib/actions/setup-pandoc@master
-
-      - name: Install dependencies
-        run: |
-          Rscript -e "install.packages(c('BiocManager'))"
-          Rscript -e "BiocManager::install(c('rcmdcheck', 'covr', 'rmarkdown', 'knitr', 'Rhdf5lib', 'testthat', 'BiocStyle'))"
-      - name: Check
-        run: Rscript -e "rcmdcheck::rcmdcheck(args = c('--no-manual'), error_on = 'warning', check_dir = 'check')"
