### 2020-05-13 d070154
try to fix make vars
