+++ man/plot.roll_your_own.Rd
@@ -40,7 +40,7 @@ membership <- c(rep("a", 5), rep("c", 5))
 
 # define function
 f <- function(x,...){
-assortment::assortment.discrete(x, ...)$r
+ assortment.discrete(x, ...)$r
 }
 
 
@@ -71,7 +71,7 @@ fit <- estimate(Y = Y, iter = 5000)
 
 # define function
 f <- function(x,...){
-  networktools::expectedInf(x,...)$step1
+  expectedInf(x,...)$step1
 }
 
 # compute
@@ -103,7 +103,7 @@ fit <- estimate(Y,
 communities <- substring(colnames(Y), 1, 1)
 
 f <- function(x, ...){
- networktools::bridge(x, ...)$`Bridge Strength`
+      bridge(x, ...)$`Bridge Strength`
 }
 
 net_stat <- roll_your_own(fit,
