+++ src/Makevars.in
@@ -9,8 +9,5 @@
 ## And with R 3.4.0, and RcppArmadillo 0.7.960.*, we turn C++11 on as OpenMP
 ## support within Armadillo prefers / requires it
 CXX_STD = CXX11
-
-PKG_CXXFLAGS = -I../inst/include @OPENMP_FLAG@
-
+PKG_CXXFLAGS = @OPENMP_FLAG@ -I../inst/include
 PKG_LIBS= @OPENMP_FLAG@ $(LAPACK_LIBS) $(BLAS_LIBS) $(FLIBS)
-
