+++ man/plot.summary.ggm_compare_explore.Rd
@@ -16,7 +16,7 @@
 \item{...}{Currently ignored}
 }
 \value{
-
+A \code{ggplot} object
 }
 \description{
 Visualize posterior hypothesis probabilities.
