+++ NAMESPACE
@@ -41,7 +41,6 @@ export(ggm_compare_ppc)
 export(map)
 export(pcor_to_cor)
 export(plot_prior)
-export(posterior_predict)
 export(posterior_samples)
 export(predictability)
 export(roll_your_own)
@@ -71,6 +70,7 @@ importFrom(stats,density)
 importFrom(stats,dnorm)
 importFrom(stats,fitted)
 importFrom(stats,lm)
+importFrom(stats,model.matrix)
 importFrom(stats,na.omit)
 importFrom(stats,pnorm)
 importFrom(stats,qnorm)
@@ -78,7 +78,6 @@ importFrom(stats,quantile)
 importFrom(stats,rWishart)
 importFrom(stats,rbeta)
 importFrom(stats,residuals)
-importFrom(stats,rnorm)
 importFrom(stats,sd)
 importFrom(stats,var)
 importFrom(utils,combn)
