### 2020-05-13 6e7e198
remove all examples
