+++ R/gen_ordinal.R
@@ -33,7 +33,6 @@
 #' ################################
 #' ######### example 1 ############
 #' ################################
-#' library(BGGM)
 #'
 #' main <-  ptsd_cor1[1:5,1:5]
 #' p <- ncol(main)
