+++ man/plot.ggm_compare_ppc.Rd
@@ -34,42 +34,3 @@ An object (or list of objects) of class \code{ggplot}
 Plot the predictive check with
 \href{https://CRAN.R-project.org/package=ggridges/vignettes/introduction.html}{ggridges}.
 }
-\examples{
-\donttest{
-library(BGGM)
-
-# data
-Y <- bfi
-
-#############################
-######### global ############
-#############################
-
-# males
-Ym <- subset(Y, gender == 1,
-             select = - c(gender, education))
-
-# females
-
-Yf <- subset(Y, gender == 2,
-             select = - c(gender, education))
-
-
-global_test <- ggm_compare_ppc(Ym, Yf)
-
-global_test
-
-plot(global_test)
-
-#############################
-######### nodewise ##########
-#############################
-
-nodewise_test <- ggm_compare_ppc(Ym, Yf,
-                                 test = "nodewise")
-
-nodewise_test
-
-plot(nodewise_test)
-}
-}
