+++ R/explore.default.R
@@ -1,652 +1,0 @@
-#' @title GGM: Exploratory Bayesian Hypothesis Testing
-#'
-#' @name explore
-#'
-#' @description Learn the conditional (in)dependence structure with the Bayes factor using the matrix-F prior distribution
-#' \insertCite{Mulder2018}{BGGM}. It is possible to test for only positive or negative edges, as well as two sided hypothesis
-#' testing (which is the customary approach). Further there is also an exhaustive option that provides the posterior
-#' probability of the null, greater than zero, and less than zero.
-#' These methods were introduced in \insertCite{Williams2019_bf;textual}{BGGM}.
-#' The graph is selected with \code{\link{select.estimate}} and
-#' then plotted with \code{\link{plot.select}}.
-#'
-#' @param Y  Matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
-#'
-#' @param formula An object of class \code{\link[stats]{formula}}. This allows for including
-#' control variables in the model (i.e., \code{~ gender}).
-#'
-#' @param type Character string. Which type of data for \strong{Y} ? The options include \code{continuous},
-#' \code{binary}, \code{ordinal}, or \code{mixed} (semi-parametric copula). See the note for further details.
-#'
-#' @param mixed_type Numeric vector. An indicator of length p for which varibles should be treated as ranks.
-#' (1 for rank and 0 to assume normality). The default is currently (dev version) to treat all integer
-#' variables as ranks when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.
-#'
-#' @param analytic Logical. Should the analytic solution be computed (default is \code{FALSE})?
-#'                 (currently not implemented)
-#'
-#' @param prior_sd Scale of the prior distribution, approximately the standard deviation of a beta distribution.
-#' (defaults to 0.25).
-#'
-#' @param iter Number of iterations (posterior samples; defaults to 5000).
-#'
-#' @param progress Logical. Should a progress bar be included (defaults to \code{TRUE}) ?
-#'
-#' @param seed An integer for the random seed.
-#'
-#' @param ... Currently ignored (leave empty).
-#'
-#' @references
-#' \insertAllCited{}
-#'
-#' @return The returned object of class \code{explore} contains a lot of information that
-#'         is used for printing and plotting the results. For users of \strong{BGGM}, the following
-#'         are the useful objects:
-#'
-#' \itemize{
-#'
-#' \item \code{pcor_mat} partial correltion matrix (posterior mean).
-#'
-#' \item \code{post_samp} an object containing the posterior samples.
-#'
-#' }
-#'
-#' @details
-#'
-#' \strong{Controlling for Variables}:
-#'
-#' When controlling for variables, it is assumed that \code{Y} includes \emph{only}
-#' the nodes in the GGM and the control variables. Internally, \code{only} the predictors
-#' that are included in \code{formula} are removed from \code{Y}. This is not behavior of, say,
-#' \code{\link{lm}}, but was adopted to ensure  users do not have to write out each variable that
-#' should be included in the GGM. An example is provided below.
-#'
-#' \strong{Mixed Type}:
-#'
-#'  The term "mixed" is somewhat of a misnomer, because the method can be used for data including \emph{only}
-#'  continuous or \emph{only} discrete variables. This is based on the ranked likelihood which requires sampling
[…]
