+++ R/ggm_compare_ppc.default.R
@@ -1,759 +1,0 @@
-#' @title GGM Compare: Posterior Predictive Check
-#'
-#' @description
-#' Compare GGMs with a posterior predicitve check \insertCite{gelman1996posterior}{BGGM}.
-#' This method was introduced in \insertCite{williams2020comparing;textual}{BGGM}. Currently,
-#' there is a \code{global} (the entire GGM) and a \code{nodewise} test. The default
-#' is to compare GGMs with respect to the posterior predictive distribution of Kullback
-#' Leibler divergence and the sum of squared errors. It is also possible to compare the
-#' GGMs with a user defined test-statistic.
-#'
-#' @name ggm_compare_ppc
-#'
-#'
-#' @param ... At least two matrices (or data frames) of dimensions \emph{n} (observations) by  \emph{p} (variables).
-#'
-#' @param test Which test should be performed (defaults to \code{"global"}) ? The options include
-#'             \code{global} and \code{nodewise}.
-#'
-#'
-#' @param iter Number of replicated datasets used to construct the predictivie distribution
-#'            (defaults to 5000).
-#'
-#' @param FUN An optional function for comparing GGMs that returns a number. See \strong{Details}.
-#'
-#' @param custom_obs Number corresponding to the observed score for comparing the GGMs. This is
-#'                   required if a function is provided in \code{FUN}. See \strong{Details}.
-#'
-#' @param loss Logical. If a function is provided, is the measure a "loss function"
-#'             (i.e., a large score is bad thing). This determines how the \emph{p}-value
-#'             is computed. See \strong{Details}.
-#'
-#' @references
-#' \insertAllCited{}
-#'
-#' @details
-#'
-#' The \code{FUN} argument allows for a user defined test-statisic (the measure used to compare the GGMs).
-#' The function must include only two agruments, each of which corresponds to a dataset. For example,
-#' \code{f <- function(Yg1, Yg2)}, where each Y is dataset of dimensions \emph{n} by \emph{p}. The
-#' groups are then compare within the function, returning a single number. An example is provided below.
-#'
-#' Further, when using a custom function care must be taken when specifying the argument \code{loss}.
-#' We recommended to visualize the results with \code{plot} to ensure the \emph{p}-value was computed
-#' in the right direction.
-#'
-#' @note
-#'
-#' \strong{Interpretation}:
-#'
-#' The primary test-statistic is symmetric KL-divergence that is termed Jensen-Shannon divergence (JSD).
-#' This is in essence a likelihood ratio that provides the "distance" between two multivariate normal
-#' distributions. The basic idea is to (1) compute the posterior predictive distribution, assuming group equality
-#' (the null model). This provides the error that we would expect to see under the null model; (2) compute
-#' JSD for the observed groups; and (3) compare the observed JSD to the posterior predictive distribution,
-#' from which a posterior predictive \emph{p}-value is computed.
-#'
-#' For the \code{global} check, the sum of squared error is also provided.
-#' This is computed from the partial correlation matrices and it is analagous
-#' to the strength test in \insertCite{van2017comparing;textual}{BGGM}. The \code{nodewise}
[…]

[…]
