+++ R/datasets.R
@@ -1,0 +1,827 @@
+#' post-traumatic stress disorder dataset
+#'
+#' A dataset containing items that measure Post-traumatic stress disorder symptoms. There are 20 variables (p) and
+#' 221 observations (n)
+#'
+#' \itemize{
+#'   \item Intrusive Thoughts
+#'   \item Nightmares
+#'   \item Flashbacks
+#'   \item Emotional cue reactivity
+#'   \item Psychological cue reactivity
+#'   \item Avoidance of thoughts
+#'   \item Avoidance of reminders
+#'   \item Trauma-related amnesia
+#'   \item Negative beliefs
+#'   \item Negative trauma-related emotions
+#'   \item Loss of interest
+#'   \item Detachment
+#'   \item Restricted affect
+#'   \item Irritability/anger
+#'   \item Self-destructive/reckless behavior
+#'   \item Hypervigilance
+#'   \item Exaggerated startle response
+#'   \item Difficulty concentrating
+#'   \item Sleep disturbance
+#' }
+#'
+#' @docType data
+#' @keywords datasets
+#' @name ptsd
+#' @usage data("ptsd")
+#' @references
+#' Epskamp, S., & Fried, E. I. (2018). A tutorial on regularized partial correlation networks. Psychological methods.
+#'  @format A data frame with 221 rows and 20 variables
+NULL
+
+
+
+#' post-traumatic stress disorder correlation matrix (# 1)
+#'
+#' A dataset containing items that measure Post-traumatic stress disorder symptoms. There are 16 variables in total.
+#' The correlation matrix was estimated from 526 individuals.
+#'
+#' \itemize{
+#'   \item Intrusive Thoughts
+#'   \item Nightmares
+#'   \item Flashbacks
+#'   \item Physiological/psychological reactivity
+#'   \item Avoidance of thoughts
+#'   \item Avoidance of situations
+#'   \item Amnesia
+#'   \item Disinterest in activities
+#'   \item Feeling detached
+#'   \item Emotional numbing
+#'   \item Foreshortened future
+#'   \item Sleep problems
+#'   \item Irritability
+#'   \item Concentration problems
+#'   \item Hypervigilance
+#'   \item Startle response
+#' }
+#'
+#' @docType data
+#' @keywords datasets
+#' @name ptsd_cor1
+#' @examples
+#' data(ptsd_cor1)
+#' Y <- MASS::mvrnorm(n = 965, mu = rep(0, 16),
+#'                    Sigma = ptsd_cor1,  empirical = TRUE)
+#' @references
+#' Fried, E. I., Eidhof, M. B., Palic, S., Costantini, G., Huisman-van Dijk, H. M., Bockting, C. L., ... & Karstoft, K. I. (2018).
+#' Replicability and generalizability of posttraumatic stress disorder (PTSD) networks: a cross-cultural multisite study of PTSD
+#' symptoms in four trauma patient samples. Clinical Psychological Science, 6(3), 335-351.
+#' @format A (correlation) matrix with 16 variables
+NULL
+
+#' post-traumatic stress disorder correlation matrix (# 2)
+#'
+#' A dataset containing items that measure Post-traumatic stress disorder symptoms. There are 16 variables in total.
+#' The correlation matrix was estimated from 365 individuals.
+#'
+#' \itemize{
+#'   \item Intrusive Thoughts
+#'   \item Nightmares
+#'   \item Flashbacks
+#'   \item Physiological/psychological reactivity
+#'   \item Avoidance of thoughts
+#'   \item Avoidance of situations
+#'   \item Amnesia
+#'   \item Disinterest in activities
+#'   \item Feeling detached
+#'   \item Emotional numbing
[…]
