+++ R/datasets.R
@@ -565,11 +565,9 @@ NULL
 #'   \item \code{VR}  Verbal rituals
 #' }
 #'
-#'   \itemize{
-#'
 #'   \strong{OCD}
-#'
-#'
+
+#' \itemize{
 #'   \item \code{CD} Concern with things touched due to dirt/bacteria
 #'   \item \code{TB} Thoughts of doing something bad around others
 #'   \item \code{CT} Continual thoughts that do not go away
@@ -636,9 +634,9 @@ NULL
 #'                      in some way?
 #' }
 #'
+#'   \strong{Anxiety}
 #'   \itemize{
 #'
-#'   \strong{Anxiety}
 #'
 #'
 #'   \item \code{GAD1} Feeling nervous, anxious, or on edge
@@ -709,9 +707,9 @@ NULL
 #'                      in some way?
 #' }
 #'
+#'   \strong{Anxiety}
 #'   \itemize{
 #'
-#'   \strong{Anxiety}
 #'
 #'
 #'   \item \code{GAD1} Feeling nervous, anxious, or on edge
