+++ R/datasets.R
@@ -1,7 +1,7 @@
-#' post-traumatic stress disorder dataset
+#' Data: Post-Traumatic Stress Disorder
 #'
-#' A dataset containing items that measure Post-traumatic stress disorder symptoms. There are 20 variables (p) and
-#' 221 observations (n)
+#' A dataset containing items that measure Post-traumatic stress disorder symptoms \insertCite{armour2017network}{BGGM}.
+#' There are 20 variables (\emph{p}) and  221 observations (\emph{n}).
 #'
 #' \itemize{
 #'   \item Intrusive Thoughts
@@ -26,20 +26,26 @@
 #' }
 #'
 #' @docType data
+#'
 #' @keywords datasets
+#'
 #' @name ptsd
+#'
 #' @usage data("ptsd")
+#'
 #' @references
-#' Epskamp, S., & Fried, E. I. (2018). A tutorial on regularized partial correlation networks. Psychological methods.
-#'  @format A data frame with 221 rows and 20 variables
+#'
+#' \insertAllCited{}
+#'
+#' @format A dataframe with 221 rows and 20 variables
 NULL
 
 
 
-#' post-traumatic stress disorder correlation matrix (# 1)
+#' Data: Post-Traumatic Stress Disorder (Sample # 1)
 #'
-#' A dataset containing items that measure Post-traumatic stress disorder symptoms. There are 16 variables in total.
-#' The correlation matrix was estimated from 526 individuals.
+#' A correlation matrix that includes 16 variables. The correlation matrix was estimated from 526
+#' individuals \insertCite{fried2018replicability}{BGGM}.
 #'
 #' \itemize{
 #'   \item Intrusive Thoughts
@@ -61,23 +67,31 @@ NULL
 #' }
 #'
 #' @docType data
+#'
 #' @keywords datasets
+#'
 #' @name ptsd_cor1
+#'
 #' @examples
+#'
 #' data(ptsd_cor1)
-#' Y <- MASS::mvrnorm(n = 965, mu = rep(0, 16),
-#'                    Sigma = ptsd_cor1,  empirical = TRUE)
+#'
+#' Y <- MASS::mvrnorm(n = 526,
+#'                    mu = rep(0, 16),
+#'                    Sigma = ptsd_cor1,
+#'                    empirical = TRUE)
+#'
 #' @references
-#' Fried, E. I., Eidhof, M. B., Palic, S., Costantini, G., Huisman-van Dijk, H. M., Bockting, C. L., ... & Karstoft, K. I. (2018).
-#' Replicability and generalizability of posttraumatic stress disorder (PTSD) networks: a cross-cultural multisite study of PTSD
-#' symptoms in four trauma patient samples. Clinical Psychological Science, 6(3), 335-351.
-#' @format A (correlation) matrix with 16 variables
+#'
+#' \insertAllCited{}
+#'
+#' @format A correlation matrix with 16 variables
 NULL
 
-#' post-traumatic stress disorder correlation matrix (# 2)
+#' Data: Post-Traumatic Stress Disorder (Sample # 2)
 #'
-#' A dataset containing items that measure Post-traumatic stress disorder symptoms. There are 16 variables in total.
-#' The correlation matrix was estimated from 365 individuals.
+#' A correlation matrix that includes 16 variables. The correlation matrix
+#' was estimated from 365 individuals \insertCite{fried2018replicability}{BGGM}.
 #'
 #' \itemize{
 #'   \item Intrusive Thoughts
@@ -99,25 +113,31 @@ NULL
 #' }
 #'
 #' @docType data
+#'
 #' @keywords datasets
+#'
 #' @name ptsd_cor2
+#'
 #' @examples
 #' data(ptsd_cor2)
[…]
