+++ inst/REFERENCES.bib
@@ -368,3 +368,35 @@ year = {1998}
   journal={arXiv preprint arXiv:1911.07728},
   year={2019}
 }
+
+@article{armour2017network,
+  title={A network analysis of DSM-5 posttraumatic stress disorder symptoms and correlates in US military veterans},
+  author={Armour, Cherie and Fried, Eiko I and Deserno, Marie K and Tsai, Jack and Pietrzak, Robert H},
+  journal={Journal of anxiety disorders},
+  volume={45},
+  pages={49--59},
+  year={2017},
+  publisher={Elsevier}
+}
+
+
+@article{fried2018replicability,
+  title={Replicability and generalizability of posttraumatic stress disorder (PTSD) networks: a cross-cultural multisite study of PTSD symptoms in four trauma patient samples},
+  author={Fried, Eiko I and Eidhof, Marloes B and Palic, Sabina and Costantini, Giulio and Huisman-van Dijk, Hilde M and Bockting, Claudi LH and Engelhard, Iris and Armour, Cherie and Nielsen, Anni BS and Karstoft, Karen-Inge},
+  journal={Clinical Psychological Science},
+  volume={6},
+  number={3},
+  pages={335--351},
+  year={2018},
+  publisher={Sage Publications Sage CA: Los Angeles, CA}
+}
+
+@Manual{psych,
+    title = {psych: Procedures for Psychological, Psychometric, and Personality Research},
+    author = {William Revelle},
+    organization = { Northwestern University},
+    address = { Evanston, Illinois},
+    year = {2019},
+    note = {R package version 1.9.12},
+    url = {https://CRAN.R-project.org/package=psych},
+  }
