+++ NAMESPACE
@@ -8,6 +8,7 @@ importFrom(MASS,ginv)
 importFrom(Rcpp,sourceCpp)
 importFrom(Rdpack,reprompt)
 importFrom(graphics,plot)
+importFrom(methods,is)
 importFrom(stats,coef)
 importFrom(stats,cov2cor)
 importFrom(stats,density)
@@ -21,7 +22,10 @@ importFrom(stats,qnorm)
 importFrom(stats,quantile)
 importFrom(stats,rWishart)
 importFrom(stats,residuals)
+importFrom(stats,rnorm)
+importFrom(stats,runif)
 importFrom(stats,sd)
 importFrom(stats,var)
+importFrom(stats,weighted.mean)
 importFrom(utils,combn)
 useDynLib(BGGM, .registration = TRUE)
