+++ R/print_BGGM.R
@@ -1,7 +1,8 @@
 #' @name print.BGGM
-#' @title  Print method for \code{BGGM} objects
+#' @title  Print Method for \code{BGGM} Objects
 #'
-#' @param x An object of class \code{BGGM}
+#' @param x An object of class \code{BGGM}.
+#'
 #' @param ... currently ignored
 #'
 #' @importFrom methods is
