+++ DESCRIPTION
@@ -17,9 +17,6 @@ Depends: R (>= 3.5.0)
 License: GPL-2
 Imports: 
     BFpack,
-    GGally (>= 1.4.0),
-    ggplot2 (>= 3.2.1),
-    ggridges (>= 0.5.1),
     MASS,
     Matrix,
     mvtnorm (>= 1.0.11),
@@ -31,7 +28,8 @@ Imports:
     sna (>= 2.5),
     stats,
     utils,
-    stringr
+    stringr, 
+    method
 Suggests: 
  knitr
 Encoding: UTF-8
