+++ DESCRIPTION
@@ -26,12 +26,21 @@ Imports:
     RcppProgress (>= 0.1),
     Rdpack,
     sna (>= 2.5),
+    GGally (>= 1.4.0),
+    ggplot2 (>= 3.2.1),
+    ggridges (>= 0.5.1)
     stats,
     utils,
     stringr, 
     methods
 Suggests: 
- knitr
+ knitr,
+ mice,
+ mvnfast,
+ networktools,
+ rmarkdown,
+ abind,
+ assortnet
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
