+++ R/helpers.R
@@ -1,6 +1,6 @@
 #' @importFrom stats coef cov2cor var dnorm lm
 #' na.omit pnorm quantile rWishart runif dnorm rnorm
-#' sd qnorm residuals fitted density weighted.mean cov terms
+#' sd qnorm residuals fitted density weighted.mean
 #' @importFrom utils combn
 #' @importFrom graphics plot
 #' @importFrom Rdpack reprompt
@@ -1157,8 +1157,7 @@ create_matrices <- function(framed, varnames) {
         if (all.equal(framed$right[r], framed$left[r + 1])) {
           if (substring(framed$right[r], 1, 1) == "(") {
             framed$right[r] <- sub("),.+", ")", framed$right[r])
-            framed$left[r + 1] <-
-              sub(".+),", "", framed$left[r + 1])
+            framed$left[r + 1] <- sub(".+),", "", framed$left[r + 1])
 
           } else{
             framed$right[r] <- sub(",.+", "", framed$right[r])
@@ -1189,7 +1188,7 @@ create_matrices <- function(framed, varnames) {
       multiples <- vector("list", length = length(commas))
 
       for (r in seq_along(commas)) {
-        several <- framed[commas,][r, ]
+        several <- framed[commas, ][r,]
 
         if (several$comp == "=") {
           several <- c(several$left, several$right)
@@ -1221,31 +1220,31 @@ create_matrices <- function(framed, varnames) {
         }
       }
 
-      framed <- framed[-commas,]
+      framed <- framed[-commas, ]
       multiples <- do.call(rbind, multiples)
       framed <- rbind(multiples, framed)
     }
   }
 
-  equality <- framed[framed$comp == "=",]
-  inequality <- framed[!framed$comp == "=",]
+  equality <- framed[framed$comp == "=", ]
+  inequality <- framed[!framed$comp == "=", ]
 
   #****Equality part string-to-matrix
   if (nrow(equality) == 0) {
     R_e <- r_e <- NULL
   } else{
-    outcomes <- suppressWarnings(apply(equality[, -2], 2, as.numeric))
+    outcomes <- suppressWarnings(apply(equality[,-2], 2, as.numeric))
     outcomes <- matrix(outcomes, ncol = 2, byrow = FALSE)
     if (any(rowSums(is.na(outcomes)) == 0))
       stop("Value compared with value rather than variable, e.g., '2 = 2', check hypotheses")
     rows <- which(rowSums(is.na(outcomes)) < 2)
-    specified <- t(outcomes[rows,])
+    specified <- t(outcomes[rows, ])
     specified <- specified[!is.na(specified)]
     r_e <- ifelse(rowSums(is.na(outcomes)) == 2, 0, specified)
     r_e <- matrix(r_e)
 
     var_locations <-
-      apply(equality[, -2], 2, function(x)
+      apply(equality[,-2], 2, function(x)
         ifelse(x %in% varnames, match(x, varnames), 0))
     var_locations <- matrix(var_locations, ncol = 2)
 
@@ -1253,10 +1252,10 @@ create_matrices <- function(framed, varnames) {
       matrix(rep(0, nrow(equality) * length(varnames)), ncol = length(varnames))
 
     for (i in seq_along(r_e)) {
-      if (!all(var_locations[i, ] > 0)) {
-        R_e[i, var_locations[i,]] <- 1
+      if (!all(var_locations[i,] > 0)) {
+        R_e[i, var_locations[i, ]] <- 1
       } else{
-        R_e[i, var_locations[i,]] <- c(1, -1)
+        R_e[i, var_locations[i, ]] <- c(1,-1)
       }
     }
   }
[…]
