+++ .travis.yml
@@ -1,5 +1,3 @@
 language: r
 
 cache: packages
-
-warnings_are_errors: false
