+++ NAMESPACE
@@ -10,7 +10,6 @@ importFrom(Rdpack,reprompt)
 importFrom(graphics,plot)
 importFrom(methods,is)
 importFrom(stats,coef)
-importFrom(stats,cov)
 importFrom(stats,cov2cor)
 importFrom(stats,density)
 importFrom(stats,dnorm)
@@ -26,7 +25,6 @@ importFrom(stats,residuals)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
 importFrom(stats,sd)
-importFrom(stats,terms)
 importFrom(stats,var)
 importFrom(stats,weighted.mean)
 importFrom(utils,combn)
