+++ DESCRIPTION
@@ -17,6 +17,9 @@ Depends: R (>= 3.5.0)
 License: GPL-2
 Imports: 
     BFpack,
+    GGally (>= 1.4.0),
+    ggplot2 (>= 3.2.1),
+    ggridges (>= 0.5.1),
     MASS,
     Matrix,
     mvtnorm (>= 1.0.11),
@@ -26,21 +29,11 @@ Imports:
     RcppProgress (>= 0.1),
     Rdpack,
     sna (>= 2.5),
-    GGally (>= 1.4.0),
-    ggplot2 (>= 3.2.1),
-    ggridges (>= 0.5.1),
     stats,
     utils,
-    stringr, 
-    methods
+    stringr
 Suggests: 
- knitr,
- mice,
- mvnfast,
- networktools,
- rmarkdown,
- abind,
- assortnet
+ knitr
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
