+++ R/estimate.R
@@ -1,0 +1,671 @@
+#' @title GGM: Estimation
+#'
+#' @description Estimate the conditional (in)dependence with either an analytic solution or efficiently
+#' sampling from the posterior distribution. These methods were introduced in \insertCite{Williams2019;textual}{BGGM}.
+#' The graph is selected with \code{\link{select.estimate}} and then plotted with \code{\link{plot.select}}.
+#'
+#' @name estimate
+#'
+#' @param Y  Matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
+#'
+#' @param formula An object of class \code{\link[stats]{formula}}. This allows for including
+#' control variables in the model (i.e., \code{~ gender}). See the note for further details.
+#'
+#' @param type Character string. Which type of data for \strong{Y} ? The options include \code{continuous},
+#' \code{binary}, \code{ordinal}, or \code{mixed}. Note that mixed can be used for data with only
+#' ordinal variables. See the note for further details.
+#'
+#' @param mixed_type Numeric vector. An indicator of length \emph{p} for which varibles should be treated as ranks.
+#' (1 for rank and 0 to assume normality). The default is currently to treat all integer variables as ranks
+#' when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.
+#'
+#' @param iter Number of iterations (posterior samples; defaults to 5000).
+#'
+#' @param prior_sd Scale of the prior distribution, approximately the standard deviation of a beta distribution
+#' (defaults to 0.50).
+#'
+#' @param analytic Logical. Should the analytic solution be computed (default is \code{FALSE})?
+#'
+#' @param progress Logical. Should a progress bar be included (defaults to \code{TRUE}) ?
+#'
+#' @param seed An integer for the random seed.
+#'
+#' @param ... Currently ignored.
+#'
+#' @references
+#' \insertAllCited{}
+#'
+#'
+#' @return The returned object of class \code{estimate} contains a lot of information that
+#'         is used for printing and plotting the results. For users of \strong{BGGM}, the following
+#'         are the useful objects:
+#'
+#' \itemize{
+#'
+#' \item \code{pcor_mat} Partial correltion matrix (posterior mean).
+#'
+#' \item \code{post_samp} An object containing the posterior samples.
+#'
+#' }
+#'
+#'
+#' @details
+#'
+#' The default is to draw samples from the posterior distribution (\code{analytic = FALSE}). The samples are
+#' required for computing edge differences (see \code{\link{ggm_compare_estimate}}), Bayesian R2 introduced in
+#' \insertCite{gelman_r2_2019;textual}{BGGM} (see \code{\link{predictability}}), etc. If the goal is to *only* determine
+#' the non-zero effects, this can be accomplished by setting \code{analytic = TRUE}. This is particularly useful
+#' when a fast solution is needed (see the examples in \code{\link{ggm_compare_ppc}})
+#'
+#' \strong{Controlling for Variables}:
+#'
+#' When controlling for variables, it is assumed that \code{Y} includes \emph{only}
+#' the nodes in the GGM and the control variables. Internally, \code{only} the predictors
+#' that are included in \code{formula} are removed from \code{Y}. This is not behavior of, say,
+#' \code{\link{lm}}, but was adopted to ensure  users do not have to write out each variable that
+#' should be included in the GGM. An example is provided below.
+#'
+#' \strong{Mixed Type}:
+#'
[…]
