+++ man/summary.predictability.Rd
@@ -1,0 +1,18 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/predictability.R
+\name{summary.predictability}
+\alias{summary.predictability}
+\title{Summary Method for \code{predictability} Objects}
+\usage{
+\method{summary}{predictability}(object, cred = 0.95, ...)
+}
+\arguments{
+\item{object}{an object of class \code{predictability}}
+
+\item{cred}{credible interval}
+
+\item{...}{currently ignored}
+}
+\description{
+Summary Method for \code{predictability} Objects
+}
