+++ man/fisher_r_to_z.Rd
@@ -1,0 +1,20 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/fisher_r2z.R
+\name{fisher_r_to_z}
+\alias{fisher_r_to_z}
+\title{Fisher Z Transformation}
+\usage{
+fisher_r_to_z(r)
+}
+\arguments{
+\item{r}{correlation (can be a vector)}
+}
+\value{
+Fisher Z transformed correlation(s)
+}
+\description{
+Tranform correlations to Fisher's Z
+}
+\examples{
+fisher_r_to_z(0.5)
+}

[…]
