+++ man/coef.estimate.Rd
@@ -30,6 +30,122 @@ There is a direct correspondence between the inverse covariance matrix and
  multiple regression \insertCite{kwan2014regression,Stephens1998}{BGGM}. This readily allows
  for converting the GGM paramters to regression coefficients. All data types are supported.
 }
+\examples{
+\donttest{
+# note: iter = 250 for demonstrative purposes
+
+#########################################
+### example 1: continuous and ordinal ###
+#########################################
+# data
+Y <- ptsd
+
+# continuous
+
+# fit model
+fit <- estimate(Y, type = "continuous")
+
+# regression
+reg <- coef(fit)
+
+# summary(reg)
+
+
+# ordinal
+
+# fit model (note + 1, due to zeros)
+fit <- estimate(Y + 1, type = "ordinal")
+
+# summarize the partial correlations
+reg <- coef(fit)
+
+# summary
+summary(reg)
+
+#########################
+### example 2: binary ###
+#########################
+# data
+Y <- women_math
+
+# fit model
+fit <- estimate(Y, type = "binary")
+
+# summarize the partial correlations
+reg <- coef(fit)
+
+# summary
+summary(reg)
+
+########################################
+### example 3: control  with formula ###
+########################################
+# (the following works with all data types)
+
+# controlling for gender
+Y <- bfi
+
+# Y contains two control variables
+# (gender and education)
+
+# the following is incorrect, as education is
+# automatically included in Y !
+
+incorrect <- estimate(Y, formula = ~ gender)
+
+# to control for only gender
+# (remove education)
+
+Y <- subset(Y, select = - education)
+
+ # fit model
+fit <- estimate(Y, formula = ~ gender,
+                iter = 250)
+
+# regression
+reg <- coef(fit)
+
+# summary
+summary(reg)
+
+
+# control for an intercation (for some reason ?)
+# (gender by education)
+
+# data
+Y <- bfi
+
+# fit model
+fit <- estimate(Y, formula = ~ gender * education,
+                iter = 250)
+
+# regression
+reg <- coef(fit)
+
+# summary
+summary(reg)
+
+########################################
+### example 4: control  with "mixed" ###
+########################################
+# control with mixed data approach
+# (all variables included in Y)
+
+# data
+Y <- bfi
+
+# fit model
+fit <- estimate(Y, type = "mixed",
+                iter = 250)
+
+# regression
+reg <- coef(fit)
+
+# summary
+summary(reg)
+
+}
+}
 \references{
 \insertAllCited{}
 }
