+++ NAMESPACE
@@ -55,6 +55,7 @@ importFrom(Rcpp,sourceCpp)
 importFrom(Rdpack,reprompt)
 importFrom(ggridges,stat_density_ridges)
 importFrom(graphics,plot)
+importFrom(methods,is)
 importFrom(network,"%e%")
 importFrom(network,"%v%<-")
 importFrom(network,"network.vertex.names<-")
