+++ NAMESPACE
@@ -65,6 +65,7 @@ importFrom(reshape,melt)
 importFrom(sna,gplot.layout.circle)
 importFrom(stats,acf)
 importFrom(stats,coef)
+importFrom(stats,cor)
 importFrom(stats,cov)
 importFrom(stats,cov2cor)
 importFrom(stats,density)
@@ -82,7 +83,9 @@ importFrom(stats,residuals)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
 importFrom(stats,sd)
+importFrom(stats,terms)
 importFrom(stats,var)
 importFrom(stats,weighted.mean)
+importFrom(utils,capture.output)
 importFrom(utils,combn)
 useDynLib(BGGM, .registration = TRUE)
