### 2020-05-14 bf7df67
change data compression
