+++ R/helpers.R
@@ -1649,7 +1649,7 @@ globalVariables(c('Y1','Y2',
                   'print_summary_ggm_estimate_compare',
                   'print_summary_metric',
                   'print_summary_select_explore',
-                  '..'))
+                  '..', 'ppc', 'rope', 'y'))
 
 
 gen_pcors <- function (p = 20, edge_prob = 0.3, lb = 0.05, ub = 0.3) {
