+++ NAMESPACE
@@ -53,6 +53,7 @@ importFrom(GGally,ggnet2)
 importFrom(MASS,ginv)
 importFrom(Rcpp,sourceCpp)
 importFrom(Rdpack,reprompt)
+importFrom(ggridges,geom_density_ridges)
 importFrom(ggridges,stat_density_ridges)
 importFrom(graphics,plot)
 importFrom(methods,is)
