### 2020-05-14 f57fb7f
coef documentation
