+++ R/coef.estimate.R
@@ -11,7 +11,7 @@
 #'
 #' @param iter Number of iterations (posterior samples; defaults to the number in the object).
 #'
-#' @param ... Not currently used.
+#' @param ... Currently ignored.
 #'
 #' @references
 #' \insertAllCited{}
@@ -27,6 +27,55 @@
 #' \item \code{object} An object of class \code{estimate} (the fitted model).
 #' }
 #'
+#' @examples
+#' \donttest{
+#' # note: iter = 250 for demonstrative purposes
+#'
+#' data
+#' Y <- ptsd
+#' #########################################
+#' ### example 1: continuous and ordinal ###
+#' #########################################
+#'
+#' # fit model
+#' fit <- estimate(Y, type = "continuous",
+#'                iter = 250)
+#'
+#' # regression
+#' reg <- coef(fit)
+#'
+#' # summary
+#' summ <- summary(reg)
+#'
+#' # ordinal
+#'
+#' # fit model (note + 1, due to zeros)
+#' fit <- estimate(Y + 1, type = "ordinal",
+#'                 iter = 250)
+#'
+#' # summarize the partial correlations
+#' reg <- coef(fit)
+#'
+#' # summary
+#' summ <- summary(reg)
+#'
+#' #########################
+#' ### example 2: binary ###
+#' #########################
+#' # data
+#' Y <- women_math
+#'
+#' # fit model
+#' fit <- estimate(Y, type = "binary",
+#'                 iter = 250)
+#'
+#' # summarize the partial correlations
+#' reg <- coef(fit)
+#'
+#' # summary
+#' summ <- summary(reg)
+#'
+#'}
 #' @export
 coef.estimate <- function(object, iter = NULL,...) {
 
@@ -97,11 +146,11 @@ coef.estimate <- function(object, iter = NULL,...) {
 #'
 #' @name coef.explore
 #'
-#' @param object An Object of class \code{estimate}
+#' @param object An Object of class \code{explore}.
 #'
 #' @param iter Number of iterations (posterior samples; defaults to the number in the object).
 #'
-#' @param ... Not currently used.
+#' @param ... Currently ignored.
 #'
 #' @references
 #' \insertAllCited{}
@@ -116,9 +165,60 @@ coef.estimate <- function(object, iter = NULL,...) {
 #' \item \code{object} An object of class \code{explore} (the fitted model).
 #' }
 #'
-
+#' @examples
+#' \donttest{
+#' # note: iter = 250 for demonstrative purposes
+#'
+#' data
+#' Y <- ptsd
+#' #########################################
+#' ### example 1: continuous and ordinal ###
+#' #########################################
+#'
+#' # fit model
+#' fit <- explore(Y, type = "continuous",
+#'                iter = 250)
+#'
+#' # regression
+#' reg <- coef(fit)
+#'
+#' # summary
+#' summ <- summary(reg)
+#'
+#' summ
+#'
+#' # ordinal
+#'
+#' # fit model (note + 1, due to zeros)
+#' fit <- explore(Y + 1, type = "ordinal",
+#'                iter = 250)
+#'
+#' # summarize the partial correlations
+#' reg <- coef(fit)
+#'
+#' # summary
+#' summ <- summary(reg)
+#'
+#' #########################
+#' ### example 2: binary ###
+#' #########################
+#' # data
+#' Y <- women_math
+#'
+#' # fit model
+#' fit <- explore(Y, type = "binary",
+#'                iter = 250)
+#'
+#' # summarize the partial correlations
+#' reg <- coef(fit)
+#'
+#' # summary
+#' summ <- summary(reg)
+#'
+#'}
 #' @export
-coef.explore <- function(object, iter = NULL,...) {
+coef.explore <- function(object,
[…]
