### 2020-05-14 2b841f5
add estimate examples
