+++ R/convergence.R
@@ -18,7 +18,66 @@
 #'
 #' @importFrom stats acf
 #'
-
+#' @examples
+#'
+#' \donttest{
+#' # note: iter = 250 for demonstrative purposes
+#'
+#' # data
+#' Y <- ptsd
+#'
+#' #########################
+#' ###### continuous #######
+#' #########################
+#' fit <- estimate(Y, iter = 250)
+#'
+#' # print names first
+#' convergence(fit, print_names = TRUE)
+#'
+#' # trace plots
+#' convergence(fit, type = "trace",
+#'             param = c("B1--B2", "B1--B3"))
+#'
+#' # acf plots
+#' convergence(fit, type = "acf",
+#'             param = c("B1--B2", "B1--B3"))
+#'
+#' #########################
+#' ######## mixed ##########
+#' #########################
+#' # copula
+#'
+#' fit <- estimate(Y, type = "mixed",
+#'                 iter = 250)
+#'
+#' # print names first
+#' convergence(fit, print_names = TRUE)
+#'
+#' # trace plots
+#' convergence(fit, type = "trace",
+#'             param = c("B1--B2", "B1--B3"))
+#'
+#' # acf plots
+#' convergence(fit, type = "acf",
+#'             param = c("B1--B2", "B1--B3"))
+#'
+#' #########################
+#' ######## ordinal ########
+#' #########################
+#' fit <- estimate(Y + 1, type = "ordinal",
+#'                 iter = 250)
+#'
+#' # print names first
+#' convergence(fit, print_names = TRUE)
+#'
+#' # trace plots
+#' convergence(fit, type = "trace",
+#'             param = c("B1--B2", "B1--B3"))
+#'
+#' # acf plots
+#' convergence(fit, type = "acf",
+#'             param = c("B1--B2", "B1--B3"))
+#' }
 #' @export
 convergence <- function(object,
                         param = NULL,
