+++ man/rsa.Rd
@@ -3,7 +3,7 @@
 \docType{data}
 \name{rsa}
 \alias{rsa}
-\title{Resilience Scale of Adults (RSA)}
+\title{Data: Resilience Scale of Adults (RSA)}
 \format{
 A data frame with 28 variables and 1973 observations (5 point Likert scale)
 }
