+++ man/summary.estimate.Rd
@@ -7,26 +7,23 @@
 \method{summary}{estimate}(object, col_names = TRUE, cred = 0.95, ...)
 }
 \arguments{
-\item{object}{an object of class \code{estimate}}
+\item{object}{An object of class \code{estimate}}
 
-\item{col_names}{logical. Should the summary include the column names (default is \code{TRUE})?
+\item{col_names}{Logical. Should the summary include the column names (default is \code{TRUE})?
 Setting to \code{FALSE} includes the column numbers (e.g., \code{1--2}).}
 
-\item{cred}{credible interval width}
+\item{cred}{Numeric. The credible interval width for summarizing the posterior
+distributions (defaults to 0.95; must be between 0 and 1).}
 
-\item{...}{currently ignored}
+\item{...}{Currently ignored.}
 }
 \value{
-a list containing the summarized posterior distributions
-# data
-Y <- BGGM::bfi[, 1:5]
-# analytic approach (sample by setting analytic = FALSE)
-fit <- estimate(Y, analytic = TRUE)
-summary(fit)
+A dataframe containing the summarized posterior distributions.
 }
 \description{
-Summary method for \code{estimate.default} objects
+Summarize the posterior distribution of each partial correlation
+with the posterior mean and standard deviation.
 }
 \seealso{
-\code{\link{select.estimate}}
+\code{\link{estimate}}
 }

[…]
