### 2020-05-14 2cf3af6
add ggm_compare_estimate documentation
