### 2020-05-14 e0a16b6
update documentation: ggm_compare_confirm
