+++ R/ggm_compare_ppc.default.R
@@ -119,7 +119,9 @@
 #'
 #' }
 #'
-#'  @examples
+#'
+#' @examples
+#'
 #' \donttest{
 #' # note: iter = 250 for demonstrative purposes
 #'
