### 2020-05-14 ad63662
fix T to TRUE in ggm_comapre_ppc
