### 2020-05-14 f4e3e06
fix TRUE in ggm_compare_ppc
