+++ R/ggm_compare_ppc.default.R
@@ -638,7 +638,7 @@ print_ggm_compare_ppc <- function(x, ...){
 
 #' Plot \code{ggm_compare_ppc} Objects
 #'
-#' @description Plot the predictive check with\code{\link[ggrides]{ggridges}}
+#' @description Plot the predictive check with\code{\link[ggridges]{ggridges}}
 #'
 #' @param x An object of class \code{ggm_compare_ppc}
 #'
