+++ man/plot_prior.Rd
@@ -20,6 +20,8 @@ Visualize the implied prior distribution for the partial correlations. This is
              particularly useful for the Bayesian hypothesis testing methods.
 }
 \examples{
-plot_prior(prior_sd = 0.25, iter = 5000)
+# note: iter = 250 for demonstrative purposes
+
+plot_prior(prior_sd = 0.25, iter = 250)
 
 }
