+++ man/zero_order_cors.Rd
@@ -7,14 +7,14 @@
 zero_order_cors(Y, type = "continuous", iter = 5000, mixed_type = NULL)
 }
 \arguments{
-\item{Y}{matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}
+\item{Y}{Matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}
 
-\item{type}{character string. Which type of data for \strong{Y} ? The options include \code{continuous},
+\item{type}{Character string. Which type of data for \code{Y} ? The options include \code{continuous},
 \code{binary}, \code{ordinal}, or \code{mixed}. See the note for further details.}
 
-\item{iter}{number of iterations (posterior samples; defaults to 5000).}
+\item{iter}{Number of iterations (posterior samples; defaults to 5000).}
 
-\item{mixed_type}{numeric vector. An indicator of length p for which varibles should be treated as ranks.
+\item{mixed_type}{Numeric vector. An indicator of length p for which varibles should be treated as ranks.
 (1 for rank and 0 to assume normality). The default is currently to treat all integer variables as ranks
 when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.}
 }
@@ -32,4 +32,78 @@ Estimate zero-order correlations for any type of data. Note zero-order refers to
 no variables are controlled for (i.e., bivariate correlations). To our knowledge, this is the only Bayesian
 implementation in \code{R} that can estiamte Pearson's,  tetrachoric (binary), polychoric
 (ordinal with more than two cateogries), and rank based correlation coefficients.
+}
+\details{
+\strong{Mixed Type}:
+
+ The term "mixed" is somewhat of a misnomer, because the method can be used for data including \emph{only}
+ continuous or \emph{only} discrete variables. This is based on the ranked likelihood which requires sampling
+ the ranks for each variable (i.e., the data is not merely transformed to ranks). This is computationally
+ expensive when there are many levels. For example, with continuous data, there are as many ranks
+ as data points!
+
+ The option \code{mixed_type} allows the user to determine  which variable should be treated as ranks
+ and the "emprical" distribution is used otherwise \insertCite{hoff2007extending}{BGGM}. This is
+ accomplished by specifying an indicator vector of length \emph{p}. A one indicates to use the ranks,
+ whereas a zero indicates to "ignore" that variable. By default all integer variables are treated as ranks.
+
+\strong{Dealing with Errors}:
+
+An error is most likely to arise when \code{type = "ordinal"}. The are two common errors (although still rare):
+
+\itemize{
+
+\item The first is due to sampling the thresholds, especially when the data is heavily skewed.
+      This can result in an ill-defined matrix. If this occurs, we recommend to first try
+      decreasing \code{prior_sd} (i.e., a more informative prior). If that does not work, then
+      change the data type to \code{type = mixed} which then estimates a copula GGM
+      (this method can be used for data containing \strong{only} ordinal variable). This should
+      work without a problem.
+
+\item  The second is due to how the ordinal data are categorized. For example, if the error states
+       that the index is out of bounds, this indicates that the first category is a zero. This is not allowed, as
+       the first category must be one. This is addressed by adding one (e.g., \code{Y + 1}) to the data matrix.
+
+}
+}
+\examples{
+\donttest{
+# note: iter = 250 for demonstrative purposes
+
+Y <- ptsd[,1:3]
+
+#################################
+####### example 1: Pearson's ####
+#################################
+
+fit <- zero_order_cors(Y, type = "continuous",
+                       iter = 250)
+
+
+#################################
+###### example 2: polychoric ####
+#################################
+
[…]
