+++ NAMESPACE
@@ -60,6 +60,7 @@ importFrom(methods,is)
 importFrom(network,"%e%")
 importFrom(network,"%v%<-")
 importFrom(network,"network.vertex.names<-")
+importFrom(network,network)
 importFrom(network,set.edge.attribute)
 importFrom(network,set.edge.value)
 importFrom(reshape,melt)
