### 2020-05-14 1aff03c
changed MAP to use analytic_solve from the helpers file
