### 2020-05-14 95ee2b7
completel plot.select documentation
