+++ man/plot.select.Rd
@@ -17,30 +17,95 @@
 )
 }
 \arguments{
-\item{x}{object of class \code{select}.}
+\item{x}{An object of class \code{select}.}
 
-\item{layout}{a character string specifying the graph layout (default is \code{circle}). See \link[sna]{gplot.layout}.}
+\item{layout}{Character string. Which graph layout (defaults is \code{circle}) ?.
+See \link[sna]{gplot.layout}.}
 
-\item{pos_col}{a character string specifying the color for the positive edges.}
+\item{pos_col}{Character string. Color for the positive edges (defaults to \code{green}).}
 
-\item{neg_col}{a character string specifying the color for the negative edges.}
+\item{neg_col}{Character string.  Color for the negative edges (defaults to \code{green}).}
 
-\item{node_size}{numeric. Node size.}
+\item{node_size}{Numeric. The size of the nodes (defaults to \code{10}).}
 
-\item{edge_magnify}{numeric. A value that is multiplied by the edge weights. This can increase (> 1) or
+\item{edge_magnify}{Numeric. A value that is multiplied by the edge weights. This can increase (> 1) or
 derease (< 1) the line widths.}
 
-\item{groups}{a character string of length \emph{p} (the number of nodes in the model). This indicates
-groups of nodes that should be the same color (e.g., clusters).}
+\item{groups}{A character string of length \emph{p} (the number of nodes in the model).
+This indicates groups of nodes that should be the same color
+(e.g., "clusters" or "communities").}
 
-\item{palette}{a character string sepcifying the palette for the \code{groups} color (default is \code{Set3}). See
-\href{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}{palette options here}.}
+\item{palette}{A character string sepcifying the palette for the \code{groups}.
+(default is \code{Set3}). See \href{http://www.cookbook-r.com/Graphs/Colors_(ggplot2)/}{palette options here}.}
 
-\item{...}{additional options passed to \link[GGally]{ggnet2}}
+\item{...}{Additional options passed to \link[GGally]{ggnet2}}
 }
 \value{
-object of class \code{ggplot}
+An object (or list of objects) of class \code{ggplot}
 }
 \description{
-Network Plot for \code{select} Objects
+Visualize the conditional (in)dependence structure.
+}
+\examples{
+\donttest{
+#########################
+### example 1: one ggm ##
+#########################
+# data
+Y <- bfi[,1:25]
+
+# estimate
+fit <- estimate(Y, iter = 250)
+
+# "communities"
+comm <- substring(colnames(Y), 1, 1)
+
+# edge set
+E <- select(fit)
+
+# plot edge set
+plt_E <- plot(E, edge_magnify = 5,
+              palette = "Set1",
+              groups = comm)
+
+# ggplot object
+plt_E$plt +
+  ggtitle("Personality")
+
+
+
+#############################
+### example 2: ggm compare ##
+#############################
+# compare males vs. females
+
+# data
+Y <- bfi[,1:26]
+
+Ym <- subset(Y, gender == 1,
+             select = -gender)
+
+Yf <- subset(Y, gender == 2,
+              select = -gender)
+
+# estimate
+fit <- ggm_compare_estimate(Ym, Yf, iter = 250)
+
+# "communities"
+comm <- substring(colnames(Ym), 1, 1)
+
+# edge set
+E <- select(fit)
+
+# plot edge set
+plt_E <- plot(E, edge_magnify = 5,
+              palette = "Set1",
+              groups = comm)
+
+# ggplot object
+plt_E[[1]] +
+ggtitle("Personality")
+
+}
+
 }
