### 2020-05-14 3a9eef4
add estimation to BGGM-package.R
