+++ man/BGGM-package.Rd
@@ -17,7 +17,16 @@ The central contribution of \strong{BGGM} is to extend those approaches:
 
 \enumerate{
 
-\item Bayesian hypothesis testing with the novel matrix-F prior distribution \insertCite{Mulder2018}{BGGM}
+\item Bayesian estimation with the novel matrix-F prior distribution \insertCite{Mulder2018}{BGGM}.
+
+ \itemize{
+
+ \item Estimation \code{\link{estimate}}.
+
+ }
+
+
+\item Bayesian hypothesis testing with the novel matrix-F prior distribution \insertCite{Mulder2018}{BGGM}.
 
  \itemize{
 
