### 2020-05-14 c546511
plot select updated to remove ggplot
