### 2020-05-14 3488bca
fix posterior samples to includes @examples
