+++ R/roll_your_own.R
@@ -31,7 +31,105 @@
 #' the weigthed adjacency matrix. This is accomplished internally by multiplying each of the sampled
 #' partial correlation matrices by the adjacency matrix.
 #'
-
+#' @examples
+#' \donttest{
+#'
+#' ####################################
+#' ###### example 1: assortment #######
+#' ####################################
+#' # assortment from this package
+#' library(assortnet)
+#'
+#' Y <- BGGM::bfi[,1:10]
+#' membership <- c(rep("a", 5), rep("c", 5))
+#'
+#' # fit model
+#' fit <- estimate(Y = Y,
+#'                 analytic = FALSE,
+#'                 iter = 1000)
+#'
+#' # list of columns belonging in each group
+#'
+#' membership <- c(rep("a", 5), rep("c", 5))
+#'
+#'f <- function(x,...){
+#'     assortment.discrete(x, ...)$r
+#'}
+#'
+#'
+#' net_stat <- roll_your_own(object = fit,
+#'                           FUN = f,
+#'                           types = membership,
+#'                           weighted = TRUE,
+#'                           SE = FALSE, M = 1)
+#'
+#' # print
+#' net_stat
+#'
+#' # plot
+#' plot(net_stat)
+#'
+#' ############################################
+#' ###### example 2: expected influence #######
+#' ############################################
+#' # expected influence from this package
+#' library(networktools)
+#'
+#' # data
+#' Y <- depression
+#'
+#' # fit model
+#' fit <- estimate(Y = Y, iter = 5000)
+#'
+#' # define function
+#' f <- function(x,...){
+#'      expectedInf(x,...)$step1
+#' }
+#'
+#' # compute
+#' net_stat <- roll_your_own(object = fit,
+#'                           FUN = f,
+#'                           iter = 1000)
+#' # print
+#' net_stat
+#'
+#' # plot
+#' plot(net_stat)
+#'
+#'
+#' #######################################
+#' ### example 3: mixed data & bridge ####
+#' #######################################
+#' # bridge from this package
+#' library(networktools)
+#'
+#' # data
+#' Y <- ptsd
+#'
+#' fit <- estimate(Y,
+#'                 type = "mixed",
+#'                 iter = 1000)
+#'
+#' # clusters
+#' communities <- substring(colnames(Y), 1, 1)
+#'
+#' # function is slow
+#' f <- function(x, ...){
+#'  bridge(x, ...)$`Bridge Strength`
+#' }
+#'
+#' net_stat <- roll_your_own(fit,
+#'                           FUN = f,
+#'                           select = TRUE,
+#'                           communities = communities)
+#'
+#' # print
+#' net_stat
+#'
+#' #plot
+#' plot(net_stat)
+#' }
+#'
 #' @export
 roll_your_own <- function(object,
                           FUN,
