+++ DESCRIPTION
@@ -39,7 +39,9 @@ Imports:
     utils,
     stringr
 Suggests: 
- knitr
+ knitr,
+ assortment,
+ networktools
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
