+++ man/roll_your_own.Rd
@@ -42,43 +42,8 @@ partial correlation matrices by the adjacency matrix.
 \examples{
 \donttest{
 
-####################################
-###### example 1: assortment #######
-####################################
-# assortment from this package
-library(assortnet)
-
-Y <- BGGM::bfi[,1:10]
-membership <- c(rep("a", 5), rep("c", 5))
-
-# fit model
-fit <- estimate(Y = Y,
-                analytic = FALSE,
-                iter = 1000)
-
-# list of columns belonging in each group
-
-membership <- c(rep("a", 5), rep("c", 5))
-
-f <- function(x,...){
-    assortment.discrete(x, ...)$r
-}
-
-
-net_stat <- roll_your_own(object = fit,
-                          FUN = f,
-                          types = membership,
-                          weighted = TRUE,
-                          SE = FALSE, M = 1)
-
-# print
-net_stat
-
-# plot
-plot(net_stat)
-
 ############################################
-###### example 2: expected influence #######
+###### example 1: expected influence #######
 ############################################
 # expected influence from this package
 library(networktools)
@@ -106,7 +71,7 @@ plot(net_stat)
 
 
 #######################################
-### example 3: mixed data & bridge ####
+### example 2: mixed data & bridge ####
 #######################################
 # bridge from this package
 library(networktools)
