+++ man/roll_your_own.Rd
@@ -41,9 +41,41 @@ partial correlation matrices by the adjacency matrix.
 }
 \examples{
 \donttest{
+####################################
+###### example 1: assortment #######
+####################################
+# assortment
+library(assortnet)
+
+Y <- BGGM::bfi[,1:10]
+membership <- c(rep("a", 5), rep("c", 5))
+
+# fit model
+fit <- estimate(Y = Y, iter = 1000)
+
+# membership
+membership <- c(rep("a", 5), rep("c", 5))
+
+# define function
+f <- function(x,...){
+ assortment.discrete(x, ...)$r
+}
+
+
+net_stat <- roll_your_own(object = fit,
+                          FUN = f,
+                          types = membership,
+                          weighted = TRUE,
+                          SE = FALSE, M = 1)
+
+# print
+net_stat
+
+# plot
+plot(net_stat)
 
 ############################################
-###### example 1: expected influence #######
+###### example 2: expected influence #######
 ############################################
 # expected influence from this package
 library(networktools)
@@ -71,7 +103,7 @@ plot(net_stat)
 
 
 #######################################
-### example 2: mixed data & bridge ####
+### example 3: mixed data & bridge ####
 #######################################
 # bridge from this package
 library(networktools)
