### 2020-05-15 19d4c11
finish select.ggm_compare documentation
