### 2020-05-15 d9c6978
complete select documentation
