+++ R/select.ggm_compare_bf.R
@@ -33,6 +33,17 @@
 #' ##################
 #' ### example 1: ###
 #' ##################
+#' # data
+#' Y <- bfi
+#'
+#' # males and females
+#' Ymale <- subset(Y, gender == 1,
+#'                    select = -c(gender,
+#'                                education))[,1:10]
+#'
+#' Yfemale <- subset(Y, gender == 2,
+#'                      select = -c(gender,
+#'                                  education))[,1:10]
 #'
 #' # fit model
 #' fit <- ggm_compare_explore(Ymale, Yfemale,
