+++ man/select.estimate.Rd
@@ -1,10 +1,14 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/select.estimate.R
+% Please edit documentation in R/select.estimate.R,
+%   R/select.ggm_compare_estimate.R
 \name{select.estimate}
 \alias{select.estimate}
+\alias{select.ggm_compare_estimate}
 \title{Graph Selection for \code{estimate} Objects}
 \usage{
 \method{select}{estimate}(object, cred = 0.95, alternative = "two.sided", ...)
+
+\method{select}{ggm_compare_estimate}(object, cred = 0.95, ...)
 }
 \arguments{
 \item{object}{An object of class \code{estimate.default}.}
@@ -16,7 +20,7 @@
 must be one of "two.sided" (default), "greater"  or "less".
 See note for futher details.}
 
-\item{...}{Currently ignored.}
+\item{...}{not currently used}
 }
 \value{
 The returned object of class \code{select.estimate} contains a lot of information that
@@ -30,9 +34,28 @@ The returned object of class \code{select.estimate} contains a lot of informatio
 \item \code{object} An object of class \code{estimate} (the fitted model).
 
 }
+
+The returned object of class \code{select.ggm_compare_estimate} contains a lot of information that
+        is used for printing and plotting the results. For users of \strong{BGGM}, the following
+        are the useful objects:
+
+
+\itemize{
+
+\item \code{mean_diff} A list of matrices for each group comparsion (partial correlation differences).
+
+\item \code{pcor_adj} A list of weighted adjacency matrices for each group comparsion.
+
+\item \code{adj} A list of adjacency matrices for each group comparsion.
+
+}
 }
 \description{
 Provides the selected graph based on credible intervals for
+the partial correlations that did not contain zero
+\inserCite{Williams2019}{BGGM}.
+
+Provides the selected graph (of differences) based on credible intervals for
 the partial correlations that did not contain zero
 \inserCite{Williams2019}{BGGM}.
 }
@@ -58,6 +81,26 @@ E <- select(fit)
 
 }
 
+\donttest{
+
+##################
+### example 1: ###
+##################
+data
+Y <- bfi
+males and females
+
+Ymale <- subset(Y, gender == 1,
+               select = -c(gender,
+                           education))
+# fit model
+fit <- ggm_compare_estimate(Ymale, Yfemale,
+                           type = "continuous")
+
+
+E <- select(fit)
+
+}
 }
 \references{
 \insertAllCited{}
