### 2020-05-15 3de8adf
fix insertcite errors that were producing latex errors
