+++ R/select.explore.R
@@ -1,7 +1,7 @@
 #' @title Graph selection for \code{explore} Objects
 #'
 #' @description Provides the selected graph based on the Bayes factor
-#' \inserCite{Williams2019_bf}{BGGM}.
+#' \insertCite{Williams2019_bf}{BGGM}.
 #'
 #' @name select.explore
 #'
@@ -19,7 +19,7 @@
 #' \insertAllCited{}
 #'
 #' @details Exhaustive provides the posterior hypothesis probabilities for
-#' a positive, negative, or null relation \inserCite{@see Table 3 in @Williams2019_bf}{BGGM}.
+#' a positive, negative, or null relation \insertCite{@see Table 3 in @Williams2019_bf}{BGGM}.
 #'
 #' @note Care must be taken with the options \code{alternative = "less"} and
 #'       \code{alternative = "greater"}. This is because the full parameter space is not included,
