+++ README.Rmd
@@ -1,6 +1,6 @@
 ---
 output: github_document
-bibliography: joss_paper/paper.bib
+bibliography: inst/REFERENCES.bib
 ---
 ```{r, echo = FALSE, message=F}
 knitr::opts_chunk$set(
@@ -24,334 +24,15 @@ library(BGGM)
 [![Build Status](https://travis-ci.org/donaldRwilliams/BGGM.svg?branch=master)](https://travis-ci.org/donaldRwilliams/BGGM)
 
 
-
-
-## Summary
-Gaussian graphical models (GGM) allow for learning conditional (in)dependence structures that are encoded by partial correlations. Whereas there are several `R` packages for classical methods [see @Kuismin2017, Table 1], there are only two that implement a Bayesian approach [@Leday2018; @Mohammadi2015a]. These are exclusively focused on identifying the graphical structure. The package **BGGM** not only contains novel Bayesian methods for this purpose, but it also includes Bayesian methodology for extending inference beyond identifying non-zero relations. 
-
-**BGGM** is built around two approaches for *Bayesian* inference--estimation and hypothesis testing. This distinction is arbitrary [see @Rouder2018], but is used to organize this work.  The former focuses on the posterior distribution and includes extensions to assess predictability [@Haslbeck2018], as well as methodology to compare partial correlations. The latter includes methods for Bayesian hypothesis testing, in both exploratory and confirmatory contexts, with the novel matrix-$F$ prior distribution [@Mulder2018]. This allows for testing the null hypothesis of conditional independence, as well as inequality and equality constrained hypotheses. Further, there are several approaches for comparing GGMs across any number of groups.  The package also includes a suite of options for model checking. Together, **BGGM** is a comprehensive toolbox for Gaussian graphical modeling in `R`. 
-
-
-## Installation
-To install the latest release version from CRAN use
-
-```{r gh-installation, eval = FALSE}
-install.packages("BGGM")
-```
-
-The current developmental version can be installed with
-
-```{r, eval = FALSE}
-if (!requireNamespace("remotes")) {
-  install.packages("remotes")
-}
-remotes::install_github("donaldRwilliams/BGGM")
-```
-
-Note that currently all of the methods are being extended to ordinal, binary, and mixed data (as of April 2020). Further, all the functions are being changed to `c++`. These are much more effecient, in some case reducing the computing time from 20 seconds to 1 second. Hence, the developmental version is a work in progress. We hope to have version `2.0.0` of BGGM submitted to CRAN by the end of May, 2020.
-
-
-## Estimation
-There are two possibilities for *estimating* GGMs. The first is an analytic solution and the second samples from the posterior distribution [described in @Williams2018bayes]. Sampling is recommended because the samples are required for various functions in **BGGM** (e.g., posterior predictive checks and prediction). 
-
-### Structure Learning
-Structure learning refers to determining which partial correlations are non-zero. This is implemented with:
-
-```r
-# data (p = 5 for demonstrative purposes)
-Y <- BGGM::bfi[,1:5]
-
-# fit model
-fit <- estimate(Y = Y, 
-                analytic = FALSE, 
-                iter = 1000)
-
-# select graph
-E <- select(fit, ci_width = 0.95)
-
-# summary
-summary(E)
-
-# output
-
-BGGM: Bayesian Gaussian Graphical Models 
---- 
-Type: Selected Graph (Analytic Solution) 
-Credible Interval: 95 % 
-Connectivity: 80 % 
---- 
-Call:
-select.estimate(x = fit_analytic, ci_width = 0.95)
[…]
