+++ README.md
@@ -9,620 +9,36 @@ Version](http://www.r-pkg.org/badges/version/BGGM)](https://cran.r-project.org/p
 [![Build
 Status](https://travis-ci.org/donaldRwilliams/BGGM.svg?branch=master)](https://travis-ci.org/donaldRwilliams/BGGM)
 
-## Summary
+The `R` package **BGGM** provides tools for making Bayesian inference in
+Gaussian graphical models (GGM). The methods are organized around two
+general approaches for Bayesian inference: (1) estimation (Williams
+2019) and (2) (Williams and Mulder 2019).
 
-Gaussian graphical models (GGM) allow for learning conditional
-(in)dependence structures that are encoded by partial correlations.
-Whereas there are several `R` packages for classical methods (see
-Kuismin and Sillanpää 2017, Table 1), there are only two that implement
-a Bayesian approach (Leday and Richardson 2018; Mohammadi and Wit 2015).
-These are exclusively focused on identifying the graphical structure.
-The package **BGGM** not only contains novel Bayesian methods for this
-purpose, but it also includes Bayesian methodology for extending
-inference beyond identifying non-zero relations.
+<!-- and (2) hypothesis testing -->
 
-**BGGM** is built around two approaches for *Bayesian*
-inference–estimation and hypothesis testing. This distinction is
-arbitrary (see Rouder, Haaf, and Vandekerckhove 2018), but is used to
-organize this work. The former focuses on the posterior distribution and
-includes extensions to assess predictability (Haslbeck and Waldorp
-2018), as well as methodology to compare partial correlations. The
-latter includes methods for Bayesian hypothesis testing, in both
-exploratory and confirmatory contexts, with the novel matrix-\(F\) prior
-distribution (Mulder and Pericchi 2018). This allows for testing the
-null hypothesis of conditional independence, as well as inequality and
-equality constrained hypotheses. Further, there are several approaches
-for comparing GGMs across any number of groups. The package also
-includes a suite of options for model checking. Together, **BGGM** is a
-comprehensive toolbox for Gaussian graphical modeling in `R`.
+<!-- \insertCite{Williams2019_bf}{BGGM}. The key distinction is that the former focuses on either the -->
 
-## Installation
+<!-- posterior or posterior predictive distribution, whereas the latter focuses on model comparison -->
 
-To install the latest release version from CRAN use
-
-``` r
-install.packages("BGGM")
-```
-
-The current developmental version can be installed with
-
-``` r
-if (!requireNamespace("remotes")) {
-  install.packages("remotes")
-}
-remotes::install_github("donaldRwilliams/BGGM")
-```
-
-Note that currently all of the methods are being extended to ordinal,
-binary, and mixed data (as of April 2020). Further, all the functions
-are being changed to `c++`. These are much more effecient, in some case
-reducing the computing time from 20 seconds to 1 second. Hence, the
-developmental version is a work in progress. We hope to have version
-`2.0.0` of BGGM submitted to CRAN by the end of May, 2020.
-
-## Estimation
-
-There are two possibilities for *estimating* GGMs. The first is an
-analytic solution and the second samples from the posterior distribution
-(described in Williams 2018). Sampling is recommended because the
[…]
