### 2020-05-15 9dff665
change readme to basic. Then begin work on branch readme
