+++ README.md
@@ -10,36 +10,9 @@ Version](http://www.r-pkg.org/badges/version/BGGM)](https://cran.r-project.org/p
 Status](https://travis-ci.org/donaldRwilliams/BGGM.svg?branch=master)](https://travis-ci.org/donaldRwilliams/BGGM)
 
 The `R` package **BGGM** provides tools for making Bayesian inference in
-Gaussian graphical models (GGM). The methods are organized around two
-general approaches for Bayesian inference: (1) estimation (Williams
-2019) and (2) (Williams and Mulder 2019).
-
-<!-- and (2) hypothesis testing -->
-
+Gaussian graphical models (GGM). <!-- and (2) hypothesis testing -->
 <!-- \insertCite{Williams2019_bf}{BGGM}. The key distinction is that the former focuses on either the -->
-
 <!-- posterior or posterior predictive distribution, whereas the latter focuses on model comparison -->
-
 <!-- with the Bayes factor. The methods in \strong{BGGM} build upon existing algorithms that are well-known in the literature. The central contribution of \strong{BGGM} is to extend those approaches: -->
 
 ## References
-
-<div id="refs" class="references">
-
-<div id="ref-Williams2019">
-
-Williams, Donald R. 2019. “Bayesian Estimation for Gaussian Graphical
-Models: Structure Learning, Predictability, and Network Comparisons.”
-*arXiv*. <https://doi.org/10.31234/OSF.IO/X8DPR>.
-
-</div>
-
-<div id="ref-Williams2019_bf">
-
-Williams, Donald R., and Joris Mulder. 2019. “Bayesian Hypothesis
-Testing for Gaussian Graphical Models: Conditional Independence and
-Order Constraints.” *PsyArXiv*. <https://doi.org/10.31234/osf.io/ypxd8>.
-
-</div>
-
-</div>
