+++ README.Rmd
@@ -58,13 +58,16 @@ support for all data types (e.g., ordinal). As of 5/15/2020, the developmental v
 The methods in **BGGM** build upon existing algorithms that are well-known in the literature.
 The central contribution of **BGGM** is to extend those approaches:
 
+1.  Bayesian estimation with the novel matrix$-F$ prior distribution \insertCite{Mulder2018}.
+  
+  + Estimation [`estimate`](#Estimate)
 
 
 
-<!-- and (2) hypothesis testing -->
-<!-- \insertCite{Williams2019_bf}{BGGM}. The key distinction is that the former focuses on either the -->
-<!-- posterior or posterior predictive distribution, whereas the latter focuses on model comparison -->
-<!-- with the Bayes factor. The methods in \strong{BGGM} build upon existing algorithms that are well-known in the literature. The central contribution of \strong{BGGM} is to extend those approaches: -->
+# Illustrative Examples
+
+## Estimate
+
 
 ## References
 
