### 2020-05-15 bc37797
add two more examples to readme
