+++ inst/REFERENCES.bib
@@ -122,7 +122,7 @@ year = {2019}
 
 @article{williams2020comparing,
   title={Comparing Gaussian graphical models with the posterior predictive distribution and Bayesian model selection.},
-  author={Williams, Donald R and Rast, Philippe and Pericchi, Luis R and Mulder, Joris},
+  author={Williams, Donald R. and Rast, Philippe and Pericchi, Luis R and Mulder, Joris},
   journal={Psychological Methods},
   year={2020},
   publisher={American Psychological Association}
