+++ README.md
@@ -11,11 +11,11 @@ Status](https://travis-ci.org/donaldRwilliams/BGGM.svg?branch=master)](https://t
 
 The `R` package **BGGM** provides tools for making Bayesian inference in
 Gaussian graphical models (GGM). The methods are organized around two
-general approaches for Bayesian inference: (1) estimation (Williams
-2019) and (2) hypothesis testing (Williams and Mulder 2019). The key
-distinction is that the former focuses on either the posterior or
-posterior predictive distribution, whereas the latter focuses on model
-comparison with the Bayes factor.
+general approaches for Bayesian inference: (1) estimation (D. R.
+Williams 2019) and (2) hypothesis testing (D. R. Williams and Mulder
+2019). The key distinction is that the former focuses on either the
+posterior or posterior predictive distribution, whereas the latter
+focuses on model comparison with the Bayes factor.
 
 ## Installation
 
@@ -48,30 +48,49 @@ extend those approaches:
 1.  Bayesian estimation with the novel matrix-F prior distribution
     (Mulder and Pericchi 2018)
     
-      - [Estimation](#bayesian-estimation) (Williams 2019)
+      - [Estimation](#bayesian-estimation) (D. R. Williams 2019)
 
 2.  Bayesian hypothesis testing with the matrix-F prior distribution
     (Mulder and Pericchi 2018).
     
-      - [Exploratory hypothesis testing](#Exploratory) (Williams and
-        Mulder 2019).
+      - [Exploratory hypothesis testing](#Exploratory) (D. R. Williams
+        and Mulder 2019).
     
-      - [Confirmatory hypothesis testing](#Confirmatory) (Williams and
-        Mulder 2019).
+      - [Confirmatory hypothesis testing](#Confirmatory) (D. R. Williams
+        and Mulder 2019).
 
-3.  Comparing Gaussian Graphical Models
+3.  Comparing Gaussian graphical models
     
       - [Partial correlation
-        differences](#Partial%20Correlation%20Differences)(Williams
+        differences](#partial-correlation-differences) (D. R. Williams
         2019)
     
-      - [Posterior predictive
-        check](#Posterior%20Predictive%20Check)(Williams et al. 2020)
+      - [Posterior predictive check](#posterior-predictive%20check)
+        (Williams et al. 2020)
     
-      - [Exploratory hypothesis
-        testing](#Exploratory%20\(groups\))(Williams et al. 2020)
+      - [Exploratory hypothesis testing](#exploratory-\(groups\))
+        (Williams et al. 2020)
     
-      - 
+      - [Confirmatory hypothesis testing](#confirmatory-\(groups\))
+        (Williams et al. 2020)
+
+4.  Extending inference beyond the conditional (in)dependence structure
+    
+      - [Predictability](#Predictability) (D. R. Williams 2019)
+    
+      - [Posterior uncertainty intervals](#posterior-uncertainty) in the
+        partial correlations (D. R. Williams 2019)
+    
+      - [Custom Network Statistics](#custom-network-statistics)
+
+The computationally intensive tasks are written in `c++` via the `R`
+package **Rcpp** (Eddelbuettel et al. 2011) and the `c++` library
+**Armadillo** (Sanderson and Curtin 2016). The Bayes factors are
+computed with the `R` package **BFpack** (Mulder et al. 2019).
+Furthermore, there are plotting functions for each method, control
[…]
