+++ inst/REFERENCES.bib
@@ -32,7 +32,7 @@ year = {2017}
 }
 
 
-@article{Williams2019a,
+@article{Williams2019,
 author = {Williams, Donald R.},
 doi = {10.31234/OSF.IO/X8DPR},
 journal = {arXiv},
@@ -42,7 +42,7 @@ title = {{Bayesian Estimation for Gaussian Graphical Models: Structure Learning,
 year = {2019}
 }
 
-@article{Williams2019b,
+@article{Williams2019_bf,
 author = {Williams, Donald R. and Mulder, Joris},
 doi = {10.31234/osf.io/ypxd8},
 file = {:C$\backslash$:/Users/wdona/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Williams, Mulder - 2019 - Bayesian Hypothesis Testing for Gaussian Graphical Models Conditional Independence and Order Constraints.pdf:pdf},
