+++ README.Rmd
@@ -91,19 +91,19 @@ The central contribution of **BGGM** is to extend those approaches:
 The computationally intensive tasks are written in `c++` via the `R` package **Rcpp** [@eddelbuettel2011rcpp] and the `c++` library **Armadillo** [@sanderson2016armadillo]. The Bayes factors are computed with the `R` package **BFpack** [@mulder2019bfpack]. Furthermore, there are plotting functions
 for each method, control variables can be included in the model (e.g., `~ gender`), and there is support for missing values (see `bggm_missing`).
 
-## Supported Data Types:
+## Supported Data Types
 
-* **Continuous**--the continuous method was described [in  @Williams2019_bf].
+* **Continuous**: the continuous method was described [in  @Williams2019].
 
-* **Binary**--the binary method builds directly upon [-@talhouk2012efficient]
+* **Binary**: the binary method builds directly upon [-@talhouk2012efficient]
   that, in turn, built upon the approaches of [-lawrence2008bayesian] and
   [-webb2008bayesian;textual] (to name a few).
   
-* **Ordinal**--the ordinal methods require sampling thresholds. There are two approach 
+* **Ordinal**: the ordinal methods require sampling thresholds. There are two approach 
    included in **BGGM**. The customary approach described in [-@albert1993bayesian] 
    (the default) andthe 'Cowles' algorithm described in [- @cowles1996accelerating].
    
-* **Mixed**--the mixed data (a combination of discrete and continuous) method was introduced
+* **Mixed**: the mixed data (a combination of discrete and continuous) method was introduced
  in [-@hoff2007extending;textual]. This is a semi-parametric copula model
  (i.e., a copula GGM) based on the ranked likelihood. Note that this can be used for data
  consisting entirely of ordinal data (not restricted to "mixed" data).
