### 2020-05-15 f474871
fix in text citations
