+++ inst/REFERENCES.bib
@@ -39,7 +39,7 @@ journal = {arXiv},
 keywords = {Explanation,Gaussian Graphical Model,Partial Correlation,Prediction,Psychometrics,Quantitative Methods,Quantitative Psychology,Social and Behavioral Sciences,Statistical Methods},
 publisher = {PsyArXiv},
 title = {{Bayesian Estimation for Gaussian Graphical Models: Structure Learning, Predictability, and Network Comparisons}},
-year = {2019}
+year = {2018}
 }
 
 @article{Williams2019_bf,
