+++ .travis.yml
@@ -3,3 +3,8 @@ language: r
 cache: packages
 
 warnings_are_errors: false
+
+os:
+  - windows
+  - linux
+  - osx
