+++ .travis.yml
@@ -5,6 +5,5 @@ cache: packages
 warnings_are_errors: false
 
 os:
-  - windows
   - linux
   - osx
