+++ README.Rmd
@@ -66,7 +66,7 @@ The central contribution of **BGGM** is to extend those approaches:
 
     + [Exploratory hypothesis testing](#Exploratory)
   
-    + [Confirmatory hypothesis testing](#Confirmatory) [@Williams2019_bf]
+    + [Confirmatory hypothesis testing](#Confirmatory)
     
 3. Comparing Gaussian graphical models [@Williams2019; @williams2020comparing]
     
@@ -83,7 +83,7 @@ The central contribution of **BGGM** is to extend those approaches:
     +  [Predictability](#Predictability) 
     
     +  [Posterior uncertainty intervals](#posterior-uncertainty) in the 
-       partial correlations [@Williams2019]
+       partial correlations
        
     +  [Custom Network Statistics](#custom-network-statistics)
     
@@ -93,20 +93,22 @@ for each method, control variables can be included in the model (e.g., `~ gender
 
 ## Supported Data Types
 
-* **Continuous**: the continuous method was described in  @Williams2019.
+* **Continuous**: the continuous method was described in  @Williams2019. Note that 
+                  this is based on the customary Wishart distribution                 
+                  ([Wikipedia](https://en.wikipedia.org/wiki/Wishart_distribution))
 
-* **Binary**: the binary method builds directly upon [-@talhouk2012efficient]
-  that, in turn, built upon the approaches of [-lawrence2008bayesian] and
-  [-webb2008bayesian;textual] (to name a few).
+* **Binary**: the binary method builds directly upon @talhouk2012efficient
+  that, in turn, built upon the approaches of lawrence2008bayesian and
+  @webb2008bayesian (to name a few).
   
 * **Ordinal**: the ordinal methods require sampling thresholds. There are two approach 
    included in **BGGM**. The customary approach described in [-@albert1993bayesian] 
    (the default) andthe 'Cowles' algorithm described in [- @cowles1996accelerating].
    
 * **Mixed**: the mixed data (a combination of discrete and continuous) method was introduced
- in [-@hoff2007extending;textual]. This is a semi-parametric copula model
- (i.e., a copula GGM) based on the ranked likelihood. Note that this can be used for data
- consisting entirely of ordinal data (not restricted to "mixed" data).
+ in @hoff2007extending. This is a semi-parametric copula model
+ (i.e., a copula GGM) based on the ranked likelihood. Note that this can be used for *only* ordinal data 
+ (not restricted to "mixed" data).
 
 # Illustrative Examples
 
