+++ .travis.yml
@@ -16,3 +16,9 @@ os:
 addons:
   apt:
     packages: libgsl0-dev
+
+
+install:
+  - Rscript -e "if (!require('devtools')) install.packages('devtools')"
+  - Rscript -e "devtools::install_deps(dep = TRUE)"
+  - Rscript -e "devtools::install_github('cjvanlissa/bain')"
