+++ .travis.yml
@@ -20,5 +20,4 @@ addons:
 
 install:
   - Rscript -e "if (!require('devtools')) install.packages('devtools')"
-  - Rscript -e "devtools::install_deps(dep = TRUE)"
   - Rscript -e "devtools::install_github('cjvanlissa/bain')"
