+++ .travis.yml
@@ -6,18 +6,5 @@ warnings_are_errors: false
 
 os:
   - linux
-  - osx
-
- homebrew:
-    packages:
-    - lavaan
-    - bain
-
-addons:
-  apt:
-    packages: libgsl0-dev
 
 
-install:
-  - Rscript -e "if (!require('devtools')) install.packages('devtools')"
-  - Rscript -e "devtools::install_github('cjvanlissa/bain')"
