+++ .travis.yml
@@ -7,17 +7,3 @@ warnings_are_errors: false
 os:
   - linux
   - osx
-
- homebrew:
-    packages:
-    - lavaan
-    - bain
-
-addons:
-  apt:
-    packages: libgsl0-dev
-
-
-install:
-  - Rscript -e "if (!require('devtools')) install.packages('devtools')"
-  - Rscript -e "devtools::install_github('cjvanlissa/bain')"
