### 2020-05-15 5b0e6df
update travis.yml to include bain devtools install
