### 2020-05-15 d62dd67
add r_packages in a different order
