+++ inst/REFERENCES.bib
@@ -400,3 +400,24 @@ year = {1998}
     note = {R package version 1.9.12},
     url = {https://CRAN.R-project.org/package=psych},
   }
+
+
+@article{rubin1984bayesianly,
+  title={Bayesianly justifiable and relevant frequency calculations for the applied statistician},
+  author={Rubin, Donald B},
+  journal={The Annals of Statistics},
+  pages={1151--1172},
+  year={1984},
+  publisher={JSTOR}
+}
+
+@book{Jeffreys1961,
+address = {Oxford},
+author = {Jeffreys, Harold},
+booktitle = {Theory of Probability},
+isbn = {0191589675},
+mendeley-groups = {symptomsSchizo/paul},
+publisher = {Oxford University Press},
+title = {{The theory of probability}},
+year = {1961}
+}
