+++ README.Rmd
@@ -155,7 +155,7 @@ The residue is a suite of useful methods not explicitly for GGMs:
 * Bivariate correlations for binary (tetrachoric), ordinal (polychoric), mixed (rank based),
   and continous (Pearson's) data.
   
-  + Tetrachoric
+    + Tetrachoric
 
 ```{r, echo=FALSE}
 load(file = "readme_models/binary_cors.rda")
@@ -169,15 +169,16 @@ cors <- zero_order_cors(Y, type = "binary",
                         iter = 250)
 ```
 
-```{r, echo=FALSE}
+```{r, echo=FALSE, results='asis'}
 row.names(cors$R_mean) <- 1:6
-knitr::kable(round(cors$R_mean, 3), 
+knitr::kable(round(cors$R_mean, 3), caption = "Tetrachoric",
              col.names = c("1", "2", "3","4","5", "6"), 
-             row.names = TRUE,
-             caption = "Tetrachoric Correlations")
+             row.names = TRUE)
 ```
 
 Note that the object `cors` also includes the sampled correlation matrices (in this case 250) in an array.
+
+    + Polychoric
   
 * Multivariate regression for binary (probit), ordinal (probit),
   mixed (rank likelihood), and continous data.
