+++ README.md
@@ -170,10 +170,12 @@ cors <- zero_order_cors(Y, type = "binary",
 | 5 | \-0.140 | \-0.146 | \-0.343 | \-0.363 |   1.000 | \-0.194 |
 | 6 |   0.098 | \-0.146 |   0.351 |   0.169 | \-0.194 |   1.000 |
 
-Tetrachoric Correlations
+Tetrachoric
 
 Note that the object `cors` also includes the sampled correlation
 matrices (in this case 250) in an array.
+
+    + Polychoric
 
   - Multivariate regression for binary (probit), ordinal (probit), mixed
     (rank likelihood), and continous data.
