### 2020-05-15 7246625
add the binary correlation example
