+++ README.Rmd
@@ -152,10 +152,10 @@ and there is support for missing values (see `bggm_missing`).
 The primary focus of **BGGM** is Gaussian graphical modeling (the inverse covariance matrix).
 The residue is a suite of useful methods not explicitly for GGMs:
   
-* Bivariate correlations for binary (tetrachoric), ordinal (polychoric), mixed (rank based),
-  and continous (Pearson's) data.
+* Bivariate correlations for `binary` (tetrachoric), `ordinal` (polychoric), `mixed` (rank based),
+  and `continuous` (Pearson's) data.
   
-    + Tetrachoric
+Here is an example for computing tetrachoric correlations:
 
 ```{r, echo=FALSE}
 load(file = "readme_models/binary_cors.rda")
@@ -165,8 +165,7 @@ load(file = "readme_models/binary_cors.rda")
 # binary data
 Y <- women_math[1:500,]
 
-cors <- zero_order_cors(Y, type = "binary", 
-                        iter = 250)
+cors <- zero_order_cors(Y, type = "binary")
 ```
 
 ```{r, echo=FALSE, results='asis'}
@@ -176,9 +175,10 @@ knitr::kable(round(cors$R_mean, 3), caption = "Tetrachoric",
              row.names = TRUE)
 ```
 
-Note that the object `cors` also includes the sampled correlation matrices (in this case 250) in an array.
+The object `cors` also includes the sampled correlation matrices (in this case 250) in an array. 
 
-    + Polychoric
+
+
   
 * Multivariate regression for binary (probit), ordinal (probit),
   mixed (rank likelihood), and continous data.
