### 2020-05-16 f2fa87b
add regression summary for multivariate probit for the README
