+++ NAMESPACE
@@ -43,6 +43,7 @@ export(pcor_to_cor)
 export(plot_prior)
 export(posterior_samples)
 export(predictability)
+export(regression_summary)
 export(roll_your_own)
 export(select)
 export(zero_order_cors)
