+++ R/regression_summary.R
@@ -14,7 +14,7 @@
 #'
 #' @examples
 #' \donttest{
-#' note: iter = 250 for demonstrative purposes
+#' # note: iter = 250 for demonstrative purposes
 #'
 #' # data
 #' # Y <- bfi
