+++ README.Rmd
@@ -187,7 +187,7 @@ Here is an example for a multivarate probit model with an ordinal outcome, where
 `E5` ("take charge") and `N5` ("panic easily") are predicted by `gender` and `education`:
 
 ```{r,echo=FALSE}
-load("readme_models/fit_mvn_ordinal.rda")
+load("readme_models/mv_probit.rda")
 ```
 
 ```{r, eval = F}
@@ -202,13 +202,15 @@ Y <- subset(Y, select = c("E5", "N5",
 mv_probit <- estimate(Y, formula = ~ gender + as.factor(education), 
                       type = "ordinal")
 
-
-regression_summary(mv_probit)
 ```
 
 Note that **BGGM** does not use the customary `model.matrix` formulation. This is for good reason, as 
 each variable in the GGM does not need to be written out. Here we effectively "tricked" **BGGM** to 
-fit a multivariate probit model (each variable included in `formula` is removed from `Y`).
+fit a multivariate probit model (each variable included in `formula` is removed from `Y`). 
+
+```{r}
+regression_summary(mv_probit)
+```
 
 
 ## Note on Conditional (In)dependence Models for Latent Data
