+++ README.Rmd
@@ -151,9 +151,14 @@ and there is support for missing values (see `bggm_missing`).
 
 The primary focus of **BGGM** is Gaussian graphical modeling (the inverse covariance matrix).
 The residue is a suite of useful methods not explicitly for GGMs. For example, 
-  
-* Bivariate correlations for `binary` (tetrachoric), `ordinal` (polychoric), `mixed` (rank based),
-  and `continuous` (Pearson's) data.
+
+<br>
+<br>
+
+### Bivariate Correlations
+
+Bivariate correlations for `binary` (tetrachoric), `ordinal` (polychoric), `mixed` (rank based),
+and `continuous` (Pearson's) data.
   
 Here is an example for computing tetrachoric correlations:
 
@@ -179,9 +184,13 @@ knitr::kable(round(cors$R_mean, 3),
 
 The object `cors` also includes the sampled correlation matrices (in this case 250) in an array. 
 
+<br>
+<br>
 
-* Multivariate regression for binary (probit), ordinal (probit),
-  mixed (rank likelihood), and continous data.
+### Multivariate Regresssion
+
+Multivariate regression for binary (probit), ordinal (probit),
+mixed (rank likelihood), and continous data.
   
 Here is an example for a multivarate probit model with an ordinal outcome, where 
 `E5` ("take charge") and `N5` ("panic easily") are predicted by `gender` and `education`:
