### 2020-05-16 978b820
fix typos in README
