+++ README.md
@@ -36,7 +36,7 @@ remotes::install_github("donaldRwilliams/BGGM")
 ```
 
 Note that the developmental version is recommended, due to the
-algorithims being written in `c++` and support for all data types. The
+algorithms being written in `c++` and support for all data types. The
 developmental version is essentially **BGGM** version 2.0.0.
 
 ## Overview
@@ -136,7 +136,7 @@ support for missing values (see `bggm_missing`).
 
 #### Predictability
 
-#### Posterior Uncertatiny
+#### Posterior Uncertainty
 
 #### Custom Network Statistics
 
@@ -174,12 +174,12 @@ cors$R
 The object `cors` also includes the sampled correlation matrices (in
 this case 250) in an array.
 
-### Multivariate Regresssion
+### Multivariate Regression
 
 Multivariate regression for binary (probit), ordinal (probit), mixed
-(rank likelihood), and continous data.
+(rank likelihood), and continuous data.
 
-Here is an example for a multivarate probit model with an ordinal
+Here is an example for a multivariate probit model with an ordinal
 outcome, where `E5` (“take charge”) and `N5` (“panic easily”) are
 predicted by `gender` and `education`:
 
@@ -242,22 +242,22 @@ outcome.
 ## Note on Conditional (In)dependence Models for Latent Data
 
 All of the data types (besides continuous) model latent data. That is,
-unoboserved data that is assumed to be Gaussian distributed. For
-example, a tetrachoric correlation (binary data) is a special case of a
+unobserved data that is assumed to be Gaussian distributed. For example,
+a tetrachoric correlation (binary data) is a special case of a
 polychoric correlation (ordinal data). Both relations are between
 "theorized normally distributed continuous *latent* variables
 [Wikepedia](https://en.wikipedia.org/wiki/Polychoric_correlation). In
-both instances, the correpsonding partial correlation between observed
+both instances, the corresponding partial correlation between observed
 variables is conditioned on the remaining variables in the *latent*
-space. This implies that interpration is similar to continuous data, but
-with respect to latent variables. We refer interested users to (see page
-2364, section 2.2, in Webb and Forster 2008).
+space. This implies that interpretation is similar to continuous data,
+but with respect to latent variables. We refer interested users to (see
+page 2364, section 2.2, in Webb and Forster 2008).
 
 ## High Dimensional Data?
 
-**BGGM** was builit specificially for social-behvarioal scientists. Of
+**BGGM** was built specifically for social-behavioral scientists. Of
 course, the methods can be used by all researchers. However, there is
-currently *not* support for high-dimensonal data (i.e., more variables
+currently *not* support for high-dimensional data (i.e., more variables
 than observations) that are common place in, say, the genetics
 literature. These data are rare in the social-behavioral sciences. In
 the future, support for high-dimensional data may be added to **BGGM**.
