+++ man/pcor_mat.Rd
@@ -1,0 +1,34 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/pcor_mat.R
+\name{pcor_mat}
+\alias{pcor_mat}
+\title{Extract the Partial Correlation Matrix}
+\usage{
+pcor_mat(object, ...)
+}
+\arguments{
+\item{object}{A model estimated with \strong{BGGM}. All classes are supported, assuming
+there is matrix to extract.}
+
+\item{...}{Currently ignored.}
+}
+\value{
+The estimated partial correlation matrix.
+}
+\description{
+Extract the Partial Correlation Matrix
+}
+\examples{
+\donttest{
+# note: iter = 250 for demonstrative purposes
+
+# data
+Y <- ptsd[,1:5] + 1
+
+# ordinal
+fit <- estimate(Y, type = "ordinal",
+                iter = 250)
+
+pcor_mat(fit)
+}
+}
