+++ NAMESPACE
@@ -39,6 +39,7 @@ export(ggm_compare_estimate)
 export(ggm_compare_explore)
 export(ggm_compare_ppc)
 export(map)
+export(pcor_mat)
 export(pcor_to_cor)
 export(plot_prior)
 export(posterior_samples)
