### 2020-05-16 3c0f39f
add posterior sampling example to readme
