### 2020-05-16 9a36ae8
change figure size in readme for posterior sampling
