+++ README.Rmd
@@ -159,7 +159,7 @@ select(fit)
 
 and then plotted
 
-```{r,  out.width = '40%'}
+```{r,  out.width = '50%'}
 # "communities"
 comm <- substring(colnames(Y), 1, 1)
 
@@ -174,7 +174,20 @@ This basic "workflow" can be used with all methods and data types. A more involv
 is provided below.
 
 #### Analytic
+There is also an analytic solution that is based on the Wishart distribution. This 
+simple solution provides competivie performance with "state-of-the-art" methods, assuming that
+*n* (observations) > *p* (variables) The one 
+caveat is that it works only for `type = "continuous"` (the default).
 
+```{r, eval=FALSE}
+# analytic
+fit <- estimate(Y, analytic = TRUE)
+
+# network plot
+plot(select(fit))
+```
+
+This is quite handy when (1) only the conditional dependence structure is of interest and (2) an immediate solution is desirable. An example of (2) is provided  in [Posterior Predictive Check](#posterior-predictive-check).
 
 ### Bayesian Hypothesis Testing
 
