### 2020-05-16 d0dad87
add bayes explore to the readme
