### 2020-05-16 7e143dc
add summary output from the exhaustive test
