+++ inst/REFERENCES.bib
@@ -421,3 +421,13 @@ publisher = {Oxford University Press},
 title = {{The theory of probability}},
 year = {1961}
 }
+
+
+@book{Hoijtink2011,
+author = {Hoijtink, H},
+file = {:C$\backslash$:/Users/Donny/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Hoijtink - 2011 - Informative hypotheses Theory and practice for behavioral and social scientists.pdf:pdf},
+mendeley-groups = {bayes{\_}general},
+publisher = {Chapman and Hall/CRC},
+title = {{Informative hypotheses: Theory and practice for behavioral and social scientists}},
+year = {2011}
+}
