### 2020-05-16 7f0630f
removed wordiness from the readme
