### 2020-05-16 0f7e9ea
update read me for the confirmatory example
