### 2020-05-16 6c027c7
add support for all classes for extracting the partial correlation matrix
