+++ R/pcor_mat.R
@@ -1,7 +1,17 @@
 #' @title Extract the Partial Correlation Matrix
 #'
+#' @description Extract the partial correlation matrix (posterior mean)
+#' from \code{\link{estimate}}, \code{\link{explore}}, \code{\link{ggm_compare_estimate}},
+#' and \code{\link{ggm_compare_explore}} objects. It is also possible to extract the
+#' partial correlation differences for \code{\link{ggm_compare_estimate}} and
+#' \code{\link{ggm_compare_explore}} objects.
+#'
 #' @param object A model estimated with \strong{BGGM}. All classes are supported, assuming
-#' there is matrix to extract.
+#' there is matrix to be extracted.
+#'
+#' @param difference Logical. Should the difference be returned (defaults to \code{FALSE}) ? Note
+#' that this assumes there is a difference (e.g., an object of class \code{ggm_compare_estimate})
+#' and ignored otherwise.
 #'
 #' @param ... Currently ignored.
 #'
@@ -21,18 +31,181 @@
 #' pcor_mat(fit)
 #' }
 #' @export
-pcor_mat <- function(object, ...){
+pcor_mat <- function(object, difference = FALSE, ...){
 
-  cn <- colnames(object$Y)
+  # estimate default
+  if(all(c("estimate", "default") %in% class(object))){
 
-  colnames(object$pcor_mat) <- cn
+    cn <- colnames(object$Y)
 
-  rownames(object$pcor_mat) <- cn
+    if(is.null(cn)) {
+      cn <- 1:object$p
+    }
 
-  if(is.null(cn)) {
-    cn <- 1:object$p
+    colnames(object$pcor_mat) <- cn
+    rownames(object$pcor_mat) <- cn
+
+    pcor_mat <- round(object$pcor_mat, 3)
+
+    # return
+    pcor_mat
+
+    # explore default
+  } else if(all(c("explore", "default") %in% class(object))){
+
+    cn <- colnames(object$Y)
+
+    if(is.null(cn)) {
+      cn <- 1:object$p
+    }
+
+    colnames(object$pcor_mat) <- cn
+    rownames(object$pcor_mat) <- cn
+
+    pcor_mat <- round(object$pcor_mat, 3)
+
+    pcor_mat
+
+    # ggm compare estimate
+  } else if(is(object, "ggm_compare_estimate")){
+
+    # analytic is false
+    if(isFALSE(object$analytic)){
+
+      cn <- colnames(object$post_samp$Y)
+      comparisons <- length(object$pcor_mats)
+
+      if(is.null(cn)) {
+        cn <- 1:object$p
+      }
+
+      # difference ?
+      if(isTRUE(difference)){
+
+        # name matrices rows and columns
+        for(i in seq_len(comparisons)){
+
+          colnames(object$pcor_mats[[i]]) <- cn
+          rownames(object$pcor_mats[[i]]) <- cn
+
+        }
+
+        pcor_mat <- object$pcor_mats
+
+
+      } else {
+
+        pcor_mat <- list()
+
+        for(i in seq_len(comparisons)){
+
+          pcor_mat[[i]] <-  round(object$post_samp[[i]]$pcor_mat, 3)
+
+        }
+
+        # name for clarity
+        names(pcor_mat) <- paste0("Y_g", seq_len(comparisons))
+
+      }
+
+      # return
+      pcor_mat
+
+      # analytic
+    } else {
+
+      # difference
+      if(isTRUE(difference)){
+
+        # names
+        cn <- colnames(object$info$dat[[1]])
+
+        # comparisons
+        comparisons <- length(object$diff)
+
+        # name matrices rows and columns
+        for(i in seq_len(comparisons)){
[…]
