+++ README.Rmd
@@ -141,7 +141,7 @@ summary(fit)
 ```
 
 The returned object can also be plotted, which allows for visualizing the posterior uncertainty interval for each relation. An example is provided below in [Posterior uncertainty intervals](#posterior-uncertatiny). 
-The partial correlation matrix is accedes with 
+The partial correlation matrix is accessed with 
 
 ```{r, eval=FALSE}
 pcor_mat(fit)
@@ -159,7 +159,7 @@ select(fit)
 
 and then plotted
 
-```{r,  out.width = '50%'}
+```{r,  out.width = '65%'}
 # "communities"
 comm <- substring(colnames(Y), 1, 1)
 
@@ -176,7 +176,7 @@ is provided below.
 #### Analytic
 There is also an analytic solution that is based on the Wishart distribution. This 
 simple solution provides competitive performance with "state-of-the-art" methods, assuming that
-*n* (observations) > *p* (variables) The one 
+*n* (observations) > *p* (variables). The one 
 caveat is that it works only for `type = "continuous"` (the default).
 
 ```{r, eval=FALSE}
@@ -213,6 +213,7 @@ relation; and (3) a negative relation.
 ```{r}
 summary(E)
 ```
+The posterior hypothesis probabilites are provided in the last three columns.
 When using `plot(E)`, there is a network plot for each hypothesis.
 
 
@@ -242,7 +243,6 @@ fit <- fit_hyp1
 # (try to) confirm
 fit <- confirm(Y = Y, hypothesis = hyp, 
                type = "ordinal")
-
 ```
 
 The object `fit` is then printed
@@ -252,7 +252,20 @@ fit
 
 The posterior hypothesis probability is `0.895` which provides some evidence for the order 
 constraint. The Bayes factor matrix then divides the posterior probabilities. This provide a measure
-of relative support for which hypothesis the data were more likely under. 
+of *relative* support for which hypothesis the data were more likely under. 
+
+Finally, the results can be plotted
+
+```{r, out.width = '65%'}
+plot(fit) + 
+  scale_fill_brewer(palette = "Set2", 
+                    name = "Posterior Prob")
+```
+
+This demonstrates that all the `plot()` functions in **BGGM** return `ggplot` objects that can be futher customized.
+Note that **BGGM** is not focused on making publication ready plots. Typically the bare mimumium is provided 
+that can then be honed in.
+
 
 ### Comparing Gaussian Graphical Models
 
@@ -264,7 +277,7 @@ of relative support for which hypothesis the data were more likely under.
 
 #### Confirmatory (groups)
 
-### Beyond the Conditional (in)dependence Structure
+### Beyond the Conditional (In)dependence Structure
 
 #### Predictability
 
