### 2020-05-16 dd0054a
fix typos in names
