### 2020-05-16 e802a7a
fix spacing in BGGM readme
