+++ README.md
@@ -10,7 +10,6 @@ Version](http://www.r-pkg.org/badges/version/BGGM)](https://cran.r-project.org/p
 Status](https://travis-ci.org/donaldRwilliams/BGGM.svg?branch=master)](https://travis-ci.org/donaldRwilliams/BGGM)
 
 The `R` package **BGGM** provides tools for making Bayesian inference in
-
 Gaussian graphical models (GGM). The methods are organized around two
 general approaches for Bayesian inference: (1) estimation and (2)
 hypothesis testing. The key distinction is that the former focuses on
