+++ README.Rmd
@@ -25,8 +25,7 @@ library(BGGM)
 [![Build Status](https://travis-ci.org/donaldRwilliams/BGGM.svg?branch=master)](https://travis-ci.org/donaldRwilliams/BGGM)
 
 
-The `R` package **BGGM** provides tools for making Bayesian inference in
-
+The `R` package **BGGM** provides tools for making Bayesian inference in 
 Gaussian graphical models (GGM). The methods are organized around two general 
 approaches for Bayesian inference: (1) estimation and (2) hypothesis 
 testing. The key distinction is that the former focuses on either 
