### 2020-05-17 7664f91
add first ggm compare for KL divergence and sse
