+++ README.Rmd
@@ -273,6 +273,34 @@ that can then be honed in.
 #### Partial Correlation Differences
 
 #### Posterior Predictive Check
+The predictive check method uses Jensen-Shannon divergence (symmetric KL divergence) or the Sum of squared error (for the partial correlation matrices) to compare groups [@williams2020comparing].
+
+In this first example, personality networks are compared for males and females.
+
+```{r}
+# data
+Y <- bfi
+# males
+Ymales <- subset(Y, gender == 1, 
+                 select = -c(gender, education))
+# females
+Yfemales <- subset(Y, gender == 2, 
+                 select = -c(gender, education))
+```
+
+
+```{r, echo=FALSE, warning = FALSE}
+load(file = "readme_models/fit_ppc1.rda")
+fit <- fit_hyp1
+```
+
+The following compares the groups
+```{r, eval=2}
+fit <- ggm_compare_ppc(Ymales, Yfemales)
+
+fit
+```
+
 
 #### Exploratory (groups)
 
