### 2020-05-17 f73e566
add custom Hamming distance
