### 2020-05-17 be8362b
add links to KL divergence
