+++ README.Rmd
@@ -276,7 +276,7 @@ that can then be honed in.
 
 
 #### Posterior Predictive Check
-The predictive check method uses Jensen-Shannon divergence (symmetric KL divergence) or the Sum of squared error (for the partial correlation matrices) to compare groups [@williams2020comparing].
+The predictive check method uses Jensen-Shannon divergence (i.e., symmetric Kullback-Leibler divergence [Wikipedia](https://en.wikipedia.org/wiki/Kullback%E2%80%93Leibler_divergence)) and the Sum of squared error (for the partial correlation matrices) to compare groups [@williams2020comparing].
 
 In this first example, personality networks are compared for males and females.
 
