### 2020-05-17 236258e
add more involved network plot to README
