+++ inst/REFERENCES.bib
@@ -431,3 +431,11 @@ publisher = {Chapman and Hall/CRC},
 title = {{Informative hypotheses: Theory and practice for behavioral and social scientists}},
 year = {2011}
 }
+
+@Manual{sna,
+    title = {sna: Tools for Social Network Analysis},
+    author = {Carter T. Butts},
+    year = {2019},
+    note = {R package version 2.5},
+    url = {https://CRAN.R-project.org/package=sna},
+  }
