+++ README.Rmd
@@ -82,7 +82,7 @@ The central contribution of **BGGM** is to extend those approaches:
 
     +  [Predictability](#Predictability) 
     
-    +  [Posterior uncertainty intervals](#posterior-uncertatiny) in the 
+    +  [Posterior uncertainty intervals](#partial-correlation-differences) for the 
        partial correlations
        
     +  [Custom Network Statistics](#custom-network-statistics)
@@ -190,6 +190,7 @@ plot(select(fit))
 This is quite handy when (1) only the conditional dependence structure is of interest and (2) an immediate solution is desirable. An example of (2) is provided  in [Posterior Predictive Check](#posterior-predictive-check).
 
 <br>
+
 ### Bayesian Hypothesis Testing
 The Bayes factor based methods allow for determining the conditional **in**dependence structure 
 (evidence for the null hypothesis).
@@ -265,8 +266,7 @@ plot(fit) +
 ```
 
 This demonstrates that all the `plot()` functions in **BGGM** return `ggplot` objects that can be futher customized.
-Note that **BGGM** is not focused on making publication ready plots. Typically the bare mimumium is provided 
-that can then be honed in.
+Note that **BGGM** is not focused on making publication ready plots. Typically the bare mimumium is provided that can then be honed in.
 
 <br>
 
@@ -275,10 +275,11 @@ that can then be honed in.
 #### Partial Correlation Differences
 
 
-#### Posterior Predictive Check
-The predictive check method uses Jensen-Shannon divergence (i.e., symmetric Kullback-Leibler divergence [Wikipedia](https://en.wikipedia.org/wiki/Kullback%E2%80%93Leibler_divergence)) and the sum of squared error (for the partial correlation matrices) to compare groups [@williams2020comparing].
+This method compares groups by computing the difference for each relation in the model. In other 
+words, there are pairwise contrasts for each partial correlation, resulting in a posterior 
+distribution for each difference.
 
-In this first example, personality networks are compared for males and females.
+In all examples in this section, personality networks are compared for males and females.
 
 ```{r}
 # data
@@ -292,6 +293,40 @@ Ymales <- subset(Y, gender == 1,
 Yfemales <- subset(Y, gender == 2, 
                  select = -c(gender, education))
 ```
+
+```{r, echo=FALSE, warning = FALSE}
+load(file = "readme_models/fit_ggm_compare.rda")
+fit <- fit_ggm_compare
+```
+
+Fit the model
+```{r, eval=FALSE}
+fit <- ggm_compare_estimate(Ymales, Yfemales)
+```
+
+Then plot the results, in this case the posterior distribution for each difference
+
+```{r, eval=FALSE}
+# plot summary
+plot(summary(fit))[[1]] +
+  theme_bw() +
+  theme(panel.grid = element_blank(), 
+        axis.text.x = element_blank()) +
+  geom_hline(yintercept = 0, 
+             linetype = "dotted")
+
+```
+
+![](readme_models/ggm_compare_estimate.png)
+
+Note that it is also possible to use `select` for the object `fit` and then plot the results. This 
+produces a network plot including the selected differences.
+Furthermore, it is also possible to plot the partial correlations (not the differences). This
+is accomplised by using `plot` with the summary computed from an `estimate` object 
+([see above](#bayesian-estimation)).
+
+#### Posterior Predictive Check
+The predictive check method uses Jensen-Shannon divergence (i.e., symmetric Kullback-Leibler divergence [Wikipedia](https://en.wikipedia.org/wiki/Kullback%E2%80%93Leibler_divergence)) and the sum of squared error (for the partial correlation matrices) to compare groups [@williams2020comparing].
 
 
 ```{r, echo=FALSE, warning = FALSE}
@@ -319,9 +354,7 @@ plot(fit,
[…]
