### 2020-05-17 cc93642
try adding output with copy and paste. This can save space by not having to save the models
