+++ README.Rmd
@@ -10,7 +10,8 @@ knitr::opts_chunk$set(
   fig.path = "man/figures/README-",
   dev = "png",
   dpi = 500,
-  fig.align = "center"
+  fig.align = "center",
+  knitr::opts_chunk$set(comment = NA)
   )
 library(ggplot2)
 library(BGGM)
@@ -192,8 +193,8 @@ This is quite handy when (1) only the conditional dependence structure is of int
 <br>
 
 ### Bayesian Hypothesis Testing
-The Bayes factor based methods allow for determining the conditional **in**dependence structure 
-(evidence for the null hypothesis).
+The Bayes factor based methods allow for determining the conditional 
+**in**dependence structure (evidence for the null hypothesis).
 
 #### Exploratory
 ```{r, echo=FALSE, warning = FALSE}
@@ -274,7 +275,6 @@ Note that **BGGM** is not focused on making publication ready plots. Typically t
 
 #### Partial Correlation Differences
 
-
 This method compares groups by computing the difference for each relation in the model. In other 
 words, there are pairwise contrasts for each partial correlation, resulting in a posterior 
 distribution for each difference.
@@ -292,11 +292,6 @@ Ymales <- subset(Y, gender == 1,
 # females
 Yfemales <- subset(Y, gender == 2, 
                  select = -c(gender, education))
-```
-
-```{r, echo=FALSE, warning = FALSE}
-load(file = "readme_models/fit_ggm_compare.rda")
-fit <- fit_ggm_compare
 ```
 
 Fit the model
@@ -327,12 +322,6 @@ is accomplised by using `plot` with the summary computed from an `estimate` obje
 
 #### Posterior Predictive Check
 The predictive check method uses Jensen-Shannon divergence (i.e., symmetric Kullback-Leibler divergence [Wikipedia](https://en.wikipedia.org/wiki/Kullback%E2%80%93Leibler_divergence)) and the sum of squared error (for the partial correlation matrices) to compare groups [@williams2020comparing].
-
-
-```{r, echo=FALSE, warning = FALSE}
-load(file = "readme_models/fit_ppc1.rda")
-fit <- fit_ppc1
-```
 
 The following compares the groups
 ```{r, eval=FALSE}
@@ -421,8 +410,86 @@ Hamming distance asks a much different question related to the adjacency matrice
 information, such as edge weigths, is considered).
 
 #### Exploratory (groups)
+The Bayes factor based methods allow for determining the conditional 
+**in**dependence structure (evidence for the null hypothesis), in this case for group equality.
+
+Fit the model
+```{r, eval=FALSE}
+fit <- ggm_compare_explore(Ymales, Yfemales)
+```
+
+Then plot the results
+```{r, eval=FALSE}
+plot(summary(fit)) +
+  theme_bw() +
+  theme(panel.grid = element_blank(), 
+        axis.text.y = element_blank()) 
+```
+
+![](readme_models/plt_ggm_compare_explore.png)
+
+Here the posterior probability for a difference is visualized for each relation in the GGM. Note 
+that it is also possible to use `select` for the object `fit` and then plot the results. This 
+produces a network plot including the selected differences, in addition to a plot depicting the 
+relations for which there was evidence for the null hypothesis.
 
 #### Confirmatory (groups)
+A central contribution of **BGGM** is confirmatory hypothesis testing of (in)equality constraints [@Hoijtink2011],
[…]
