+++ inst/REFERENCES.bib
@@ -439,3 +439,11 @@ year = {2011}
     note = {R package version 2.5},
     url = {https://CRAN.R-project.org/package=sna},
   }
+
+
+@article{rodriguez2020formalizing,
+  title={On Formalizing Theoretical Expectations: Bayesian Testing of Central Structures in Psychological Networks},
+  author={Rodriguez, Josue E and Williams, Donald R and Rast, Philippe and Mulder, Joris},
+  year={2020},
+  publisher={PsyArXiv}
+}
