### 2020-05-17 1815fff
add weighted adjacency matrix
