+++ NAMESPACE
@@ -47,6 +47,7 @@ export(predictability)
 export(regression_summary)
 export(roll_your_own)
 export(select)
+export(weighted_adj_mat)
 export(zero_order_cors)
 import(Rcpp)
 import(ggplot2)
