### 2020-05-17 c09b47a
Add network plot as saved image for estiamtion method
