### 2020-05-17 bfba669
add plot to confirm hyp
