+++ README.md
@@ -278,7 +278,7 @@ summary(E)
 #>  B4--B5    0.348    0.072   0.000 1.000 0.000
 ```
 
-The posterior hypothesis probabilites are provided in the last three
+The posterior hypothesis probabilities are provided in the last three
 columns. When using `plot(E)`, there is a network plot for each
 hypothesis.
 
@@ -370,9 +370,9 @@ plot(fit) +
 ![](readme_models/confirm_hyp.png)
 
 This demonstrates that all the `plot()` functions in **BGGM** return
-`ggplot` objects that can be futher customized. Note that **BGGM** is
+`ggplot` objects that can be further customized. Note that **BGGM** is
 not focused on making publication ready plots. Typically the bare
-mimumium is provided that can then be honed in.
+minimum is provided that can then be honed in.
 
 <br>
 
@@ -412,12 +412,7 @@ difference
 
 ``` r
 # plot summary
-plot(summary(fit))[[1]] +
-  theme_bw() +
-  theme(panel.grid = element_blank(), 
-        axis.text.x = element_blank()) +
-  geom_hline(yintercept = 0, 
-             linetype = "dotted")
+plot(summary(fit))0
 ```
 
 ![](readme_models/ggm_compare_estimate.png)
@@ -425,8 +420,8 @@ plot(summary(fit))[[1]] +
 Note that it is also possible to use `select` for the object `fit` and
 then plot the results. This produces a network plot including the
 selected differences. Furthermore, it is also possible to plot the
-partial correlations (not the differences). This is accomplised by using
-`plot` with the summary computed from an `estimate` object ([see
+partial correlations (not the differences). This is accomplished by
+using `plot` with the summary computed from an `estimate` object ([see
 above](#bayesian-estimation)).
 
 #### Posterior Predictive Check
@@ -447,6 +442,32 @@ Then print the summary output with
 
 ``` r
 fit
+
+#> BGGM: Bayesian Gaussian Graphical Models 
+#> --- 
+#> Test: Global Predictive Check 
+#> Posterior Samples: 500 
+#>   Group 1: 805 
+#>   Group 2: 1631 
+#> Nodes:  25 
+#> Relations: 300 
+#> --- 
+#> Call: 
+#> ggm_compare_ppc(Ymales, Yfemales, iter = 500)
+#> --- 
+#> Symmetric KL divergence (JSD): 
+#>  
+#>    contrast JSD.obs p_value
+#>  Yg1 vs Yg2   0.442       0
+#> --- 
+#>  
+#> Sum of Squared Error: 
+#>  
+#>    contrast SSE.obs p.value
+#>  Yg1 vs Yg2   0.759       0
+#> --- 
+#> note:
+#> JSD is Jensen-Shannon divergence 
 ```
 
 In this case, there seems to be decisive evidence that the networks are
@@ -457,6 +478,8 @@ predictive distribution can also be plotted
 plot(fit, 
      critical = 0.05)$plot_jsd
 ```
+
+![](readme_models/ppc_1.png)
 
 where the red region is the “critical” area and the black point is the
 observed KL divergence for the networks. This again shows that the
@@ -517,12 +540,10 @@ then compare the groups
 fit <- ggm_compare_ppc(Ymales, Yfemales,
                              FUN = f,
                              custom_obs  = obs)
-```
 
-and finally print the results
-
-``` r
+# print
 fit
+
 #> BGGM: Bayesian Gaussian Graphical Models 
 #> --- 
 #> Test: Global Predictive Check 
@@ -539,14 +560,14 @@ fit
 #>  
 #>    contrast custom.obs p.value
 #>  Yg1 vs Yg2         75   0.576
-#> ---
+#> --- 
 ```
 
[…]
