+++ inst/REFERENCES.bib
@@ -445,5 +445,27 @@ year = {2011}
   title={On Formalizing Theoretical Expectations: Bayesian Testing of Central Structures in Psychological Networks},
   author={Rodriguez, Josue E and Williams, Donald R and Rast, Philippe and Mulder, Joris},
   year={2020},
-  publisher={PsyArXiv}
+  publisher={PsyArXiv},
+  journal={PsyArXiv}
+}
+
+
+@article{haslbeck2018well,
+  title={How well do network models predict observations? On the importance of predictability in network models},
+  author={Haslbeck, Jonas MB and Waldorp, Lourens J},
+  journal={Behavior research methods},
+  volume={50},
+  number={2},
+  pages={853--861},
+  year={2018},
+  publisher={Springer}
+}
+
+@article{jones2019bridge,
+  title={Bridge centrality: A network approach to understanding comorbidity},
+  author={Jones, Payton J and Ma, Ruofan and McNally, Richard J},
+  journal={Multivariate behavioral research},
+  pages={1--15},
+  year={2019},
+  publisher={Taylor \& Francis}
 }
