+++ README.Rmd
@@ -259,7 +259,7 @@ summary(E)
 #>  B4--B5    0.348    0.072   0.000 1.000 0.000
 ```
 
-The posterior hypothesis probabilites are provided in the last three columns.
+The posterior hypothesis probabilities are provided in the last three columns.
 When using `plot(E)`, there is a network plot for each hypothesis.
 
 
@@ -335,9 +335,10 @@ plot(fit) +
   ggtitle("Confirmatory: Comorbidity Network")
 ```
 
+
 ![](readme_models/confirm_hyp.png)
 
-This demonstrates that all the `plot()` functions in **BGGM** return `ggplot` objects that can be futher customized. Note that **BGGM** is not focused on making publication ready plots. Typically the bare mimumium is provided that can then be honed in.
+This demonstrates that all the `plot()` functions in **BGGM** return `ggplot` objects that can be further customized. Note that **BGGM** is not focused on making publication ready plots. Typically the bare minimum is provided that can then be honed in.
 
 <br>
 
@@ -373,12 +374,7 @@ Then plot the results, in this case the posterior distribution for each differen
 
 ```{r, eval=FALSE}
 # plot summary
-plot(summary(fit))[[1]] +
-  theme_bw() +
-  theme(panel.grid = element_blank(), 
-        axis.text.x = element_blank()) +
-  geom_hline(yintercept = 0, 
-             linetype = "dotted")
+plot(summary(fit))0
 
 ```
 
@@ -387,7 +383,7 @@ plot(summary(fit))[[1]] +
 Note that it is also possible to use `select` for the object `fit` and then plot the results. This 
 produces a network plot including the selected differences.
 Furthermore, it is also possible to plot the partial correlations (not the differences). This
-is accomplised by using `plot` with the summary computed from an `estimate` object 
+is accomplished by using `plot` with the summary computed from an `estimate` object 
 ([see above](#bayesian-estimation)).
 
 #### Posterior Predictive Check
@@ -402,6 +398,32 @@ fit <- ggm_compare_ppc(Ymales, Yfemales)
 Then print the summary output with
 ```{r, eval=FALSE}
 fit
+
+#> BGGM: Bayesian Gaussian Graphical Models 
+#> --- 
+#> Test: Global Predictive Check 
+#> Posterior Samples: 500 
+#>   Group 1: 805 
+#>   Group 2: 1631 
+#> Nodes:  25 
+#> Relations: 300 
+#> --- 
+#> Call: 
+#> ggm_compare_ppc(Ymales, Yfemales, iter = 500)
+#> --- 
+#> Symmetric KL divergence (JSD): 
+#>  
+#>    contrast JSD.obs p_value
+#>  Yg1 vs Yg2   0.442       0
+#> --- 
+#>  
+#> Sum of Squared Error: 
+#>  
+#>    contrast SSE.obs p.value
+#>  Yg1 vs Yg2   0.759       0
+#> --- 
+#> note:
+#> JSD is Jensen-Shannon divergence 
 ```
 
 In this case, there seems to be decisive evidence that the networks are different (as indicated by the
@@ -410,6 +432,8 @@ posterior predictive *p*-value). The predictive distribution can also be plotted
 plot(fit, 
      critical = 0.05)$plot_jsd
 ```
+
+![](readme_models/ppc_1.png)
 
 where the red region is the "critical" area and the black point is the observed KL divergence for the networks. This again shows that the "distance" between the networks is much more than expected, assuming that the groups were actually the same.
 
@@ -457,28 +481,38 @@ load(file = "readme_models/obs.rda")
 obs <- f(Ymales, Yfemales)
 ```
 
-```{r, echo=FALSE, warning = FALSE}
-load(file = "readme_models/global_hd.rda")
[…]
