+++ README.Rmd
@@ -374,7 +374,7 @@ Then plot the results, in this case the posterior distribution for each differen
 
 ```{r, eval=FALSE}
 # plot summary
-plot(summary(fit))0
+plot(summary(fit))
 
 ```
 
@@ -546,7 +546,7 @@ in this case for comparing groups. By this we are referring to testing expectati
 In this example, the focus is on agreeableness in a personality network. Here is a set of hypotheses
 ```{r}
 hyp <- c("g1_A2--A4 > g2_A2--A4 > 0 & g1_A4--A5 > g2_A4--A5 > 0;
-         g1_A4--A5 = g2_A4--A5 = 0  & g1_A2--A4 = g2_A2--A4 = 0")
+          g1_A4--A5 = g2_A4--A5 = 0  & g1_A2--A4 = g2_A2--A4 = 0")
 ```
 
 where the variables are `A2` ("inquire about others' well being"), `A4` ("love children"), 
@@ -554,13 +554,11 @@ and `A5` ("make people feel at ease"). The first hypothesis states that the cond
 dependent effects are larger for female than males (note the `&`), with the additional 
 constraint to positive values, whereas the second hypothesis is a "null" model.
 
-Fit the model
+The hypothesis is tested with the following
 ```{r, eval=FALSE}
-fit <- ggm_compare_confirm(Yfemales, Ymales, hypothesis = hyp)
-```
+fit <- ggm_compare_confirm(Yfemales, Ymales, 
+                           hypothesis = hyp)
 
-Then print the results
-```{r, eval=FALSE}
 # print
 fit
 
