### 2020-05-17 1a4f72f
finish rough of readme
