+++ README.Rmd
@@ -645,7 +645,7 @@ f <- function(x, ...){
 }
 ```
 
-Note that `x` takes the matrix and `f` can return either a single number of a number for each node. The next step is to fit the model and compute the network statistics
+Note that `x` takes the matrix and `f` can return either a single number of a number for each node. The next step is to fit the model and compute the network statistic
 
 ```{r, eval=FALSE}
 # data
@@ -664,8 +664,7 @@ net_stat <- roll_your_own(fit,
                           communities = communities)
 ```
 
-The function `f` is provided to `FUN` and `communities` is passed to `brigde` via `...`. The results can 
-be printed
+The function `f` is provided to `FUN` and `communities` is passed to `brigde` (inside of `f`) via `...`. The results can be printed
 
 ```{r, eval=FALSE}
 # print
