+++ DESCRIPTION
@@ -16,16 +16,13 @@ Description: Fit Bayesian Gaussian graphical models. The methods are separated i
 Depends: R (>= 3.5.0)
 License: GPL-2
 Imports:
-    abind,
     BFpack,
     GGally (>= 1.4.0),
     ggplot2 (>= 3.2.1),
     ggridges (>= 0.5.1),
     grDevices,
-    MASS,
     Matrix,
     methods,
-    mice,
     mvtnorm (>= 1.0.11),
     mvnfast,
     network (>= 1.15),
@@ -39,9 +36,11 @@ Imports:
     utils,
     stringr
 Suggests: 
+ abind,
  knitr,
  assortnet,
- networktools
+ networktools,
+ mice
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
