+++ DESCRIPTION
@@ -16,6 +16,7 @@ Description: Fit Bayesian Gaussian graphical models. The methods are separated i
 Depends: R (>= 3.5.0)
 License: GPL-2
 Imports:
+    abind,
     BFpack,
     GGally (>= 1.4.0),
     ggplot2 (>= 3.2.1),
