### 2020-05-18 dbc889a
dontest plot_prior
