+++ R/plot_prior.R
@@ -13,10 +13,11 @@
 #' @return A \code{ggplot} object.
 #'
 #' @examples
+#' \donttest{
 #' # note: iter = 250 for demonstrative purposes
 #'
 #' plot_prior(prior_sd = 0.25, iter = 250)
-#'
+#'}
 #' @export
 plot_prior <- function(prior_sd = 0.2, iter = 5000){
 
