+++ R/explore.default.R
@@ -294,15 +294,15 @@ explore <- function(Y,
 
   dot_dot_dot <- list(...)
 
-  # check for groups
-  if(length(dot_dot_dot) == 0){
-
+  # # check for groups
+  # if(length(dot_dot_dot) == 0){
+  #
     eps <- 0.01
-
-  } else {
-
-    eps <- 0.1
-  }
+  #
+  # } else {
+  #
+  #   eps <- 0.1
+  # }
 
   # delta parameter
   delta <- delta_solve(prior_sd)
