+++ DESCRIPTION
@@ -16,32 +16,29 @@ Description: Fit Bayesian Gaussian graphical models. The methods are separated i
 Depends: R (>= 3.5.0)
 License: GPL-2
 Imports:
-    abind,
-    BFpack,
+    BFpack (>= 0.2.1),
     GGally (>= 1.4.0),
     ggplot2 (>= 3.2.1),
     ggridges (>= 0.5.1),
     grDevices,
-    MASS,
-    Matrix,
+    MASS (>= 7.3-51.5),
     methods,
-    mice,
-    mvtnorm (>= 1.0.11),
-    mvnfast,
+    mvnfast (>= 0.2.5),
     network (>= 1.15),
-    pracma,
-    reshape,
+    reshape (>= 0.8.8),
     Rcpp (>= 1.0.4.6),
     RcppProgress (>= 0.1),
-    Rdpack,
+    Rdpack (>= 0.11-1),
     sna (>= 2.5),
+    stringr (>= 1.4.0),
     stats,
     utils,
-    stringr
 Suggests: 
- knitr,
- assortnet,
- networktools
+    abind,
+    assortnet,
+    knitr,
+    networktools,
+    mice
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
