### 2020-05-18 0ab8120
remove plot select due to change post prob decision rule
