+++ DESCRIPTION
@@ -30,15 +30,14 @@ Imports:
     RcppProgress (>= 0.1),
     Rdpack (>= 0.11-1),
     sna (>= 2.5),
-    stringr (>= 1.4.0),
     stats,
     utils,
 Suggests: 
-    abind,
-    assortnet,
-    knitr,
-    networktools,
-    mice
+    abind (>= 1.4-5),
+    assortnet (>= 0.12),
+    knitr (>= 1.28),
+    networktools (>= 1.2.3),
+    mice (>= 3.8.0)
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
