### 2020-05-18 514dba8
Bug: fix plotting function for networks selected with ggm_compare_explore
