+++ R/helpers.R
@@ -37,9 +37,8 @@ remove_predictors_helper <- function(Y_groups, formula){
   # number of groups
   groups <- length(Y_groups)
 
-
+  # Y groups
   Y_groups <- lapply(seq_len(groups), function(x)  na.omit(Y_groups[[x]]) )
-
 
   model_matrices <- lapply(seq_len(groups) , function(x) {
     stats::model.matrix(formula, Y_groups[[x]])
@@ -71,12 +70,9 @@ remove_predictors_helper <- function(Y_groups, formula){
     })
   }
 
-
-
-
   list(Y_groups = Y_groups, model_matrices = model_matrices)
 
-}
+  }
 
 
 
@@ -100,6 +96,8 @@ binary_latent_helper <- function(x){
                                                mean = 0, sd = 1))
   # latent data (sd = 1)
   latent_data <- scale(latent_data)
+
+  # latent data
   latent_data
 }
 
@@ -112,6 +110,7 @@ ordinal_latent_helper <- function(x, thresholds){
   # variables
   p <- ncol(x)
 
+  # mean of thresholds
   thresholds <- t(sapply(1:p, function(x) colMeans(thresholds[,,x])))
 
   latent_data <- sapply(1:p, function(z)  qnorm(runif(n, min = pnorm(thresholds[z, x[,z]],
@@ -121,25 +120,33 @@ ordinal_latent_helper <- function(x, thresholds){
                                                                      mean = 0,
                                                                      sd = 1)),
                                                 mean = 0, sd = 1))
-
+  # latent data
   latent_data <- scale(latent_data)
 
+  # latent data
   latent_data
 
-}
-
+  }
 
 rank_helper <- function(Y){
 
   # adapted from hoff (2008). See documentation.
   p <- ncol(Y)
+
   levels  <- apply(Y, 2, function(x) {match(x, sort(unique(x)))})
+
   K <-  apply(levels, 2, max, na.rm = TRUE)
+
   ranks <- apply(Y, 2, rank, ties.method = "max", na.last = "keep")
+
   n_complete <- apply(!is.na(ranks), 2, sum)
+
   U <- t(t(ranks)/(n_complete + 1))
+
   Z <- qnorm(U)
+
   S <- cov(Z)
+
   list(K = K,
        levels = levels,
        Sigma_start = S,
@@ -216,7 +223,6 @@ prior_helper_2 <- function(p, delta , epsilon){
 
   parcorMat <- corMat <- ThetaMat <- SigmaMat <- array(0, dim=c(1e4 , k , k))
 
-
   for(s in 1:dim(parcorMat)[1]){
 
     Psi <- rWishart(1,df = nu, Sigma = 1/nu*diag(k))[,,1]
@@ -234,352 +240,9 @@ prior_helper_2 <- function(p, delta , epsilon){
   parcorMat
 }
 
-
-sampling_helper_poly <- function(Y, delta, iter, type, mixed_type= NULL){
-
-  if(type == "binary"){
-
-  fit_mvn <- mv_binary(Y, matrix(1, nrow = nrow(Y)),
-                        delta = delta,
-                        epsilon = 0.001,
-                        iter = iter,
-                        beta_prior = 0.0001,
-                        cutpoints = c(-Inf, 0, Inf) )
-
-  } else if(type == "ordinal"){
-
-    K <- max(apply(Y, 2, FUN = function(x) {max(unique(Y))} ))
-
-    fit_mvn <- mv_ordinal_albert(Y, matrix(1, nrow = nrow(Y)),
[…]
