### 2020-05-18 e72f715
DOCS: add example to predictability
