### 2020-05-18 1676a83
DOC: example in summary.estimate
