### 2020-05-18 f94e192
DOC: finish examples for explore.default
