+++ R/explore.default.R
@@ -661,6 +661,17 @@ explore <- function(Y,
 #'
 #' @return A dataframe containing the summarized posterior distributions.
 #'
+#' @examples
+#' \donttest{
+#' # note: iter = 250 for demonstrative purposes
+#'
+#' Y <- ptsd
+#'
+#' fit <- explore(Y, iter = 250)
+#'
+#' summary(fit)
+#'
+#' }
 #' @export
 summary.explore <- function(object,
                              col_names = TRUE, ...) {
@@ -774,6 +785,17 @@ print_explore <- function(x,...){
 #'
 #' @seealso \code{\link{explore}}
 #'
+#' @examples
+#' \donttest{
+#' # note: iter = 250 for demonstrative purposes
+#'
+#' Y <- ptsd
+#'
+#' fit <- explore(Y, iter = 250)
+#'
+#' plot(summary(fit))
+#'
+#' }
 #' @export
 plot.summary.explore <- function(x,
                                  color = "black",
