### 2020-05-18 0dbaaad
DOC: improve documentation for plot.predictability to be consistent with the other functions.
