+++ R/estimate.R
@@ -826,6 +826,20 @@ print_estimate <- function(x, ...){
 #' @seealso \code{\link{estimate}}
 #'
 #' @return A \code{ggplot} object.
+#'
+#' @examples
+#' \donttest{
+#' # data
+#' Y <- ptsd
+#'
+#' fit <- estimate(Y, iter = 250)
+#'
+#' summary(fit)
+#'
+#' plot(summary(fit))
+#'
+#' }
+#'
 #' @export
 plot.summary.estimate <- function(x,
                                   color = "black",
