### 2020-05-18 ee7045b
DOC: fixed unclosed quote in plot.predictability
