### 2020-05-18 12c7cc7
change how latent data is sampled
