+++ R/RcppExports.R
@@ -1,26 +1,6 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-Theta_continuous <- function(Y, iter, delta, epsilon, prior_only, explore, start, progress) {
-    .Call(`_BGGM_Theta_continuous`, Y, iter, delta, epsilon, prior_only, explore, start, progress)
-}
-
-sample_prior <- function(Y, iter, delta, epsilon, prior_only, explore, progress) {
-    .Call(`_BGGM_sample_prior`, Y, iter, delta, epsilon, prior_only, explore, progress)
-}
-
-mv_continuous <- function(Y, X, delta, epsilon, iter, start, progress) {
-    .Call(`_BGGM_mv_continuous`, Y, X, delta, epsilon, iter, start, progress)
-}
-
-trunc_mvn <- function(mu, rinv, z, y, cutpoints) {
-    .Call(`_BGGM_trunc_mvn`, mu, rinv, z, y, cutpoints)
-}
-
-mv_binary <- function(Y, X, delta, epsilon, iter, beta_prior, cutpoints, start, progress) {
-    .Call(`_BGGM_mv_binary`, Y, X, delta, epsilon, iter, beta_prior, cutpoints, start, progress)
-}
-
 Sigma_i_not_i <- function(x, index) {
     .Call(`_BGGM_Sigma_i_not_i`, x, index)
 }
@@ -39,6 +19,26 @@ remove_row <- function(x, which) {
 
 remove_col <- function(x, index) {
     .Call(`_BGGM_remove_col`, x, index)
+}
+
+Theta_continuous <- function(Y, iter, delta, epsilon, prior_only, explore, start, progress) {
+    .Call(`_BGGM_Theta_continuous`, Y, iter, delta, epsilon, prior_only, explore, start, progress)
+}
+
+sample_prior <- function(Y, iter, delta, epsilon, prior_only, explore, progress) {
+    .Call(`_BGGM_sample_prior`, Y, iter, delta, epsilon, prior_only, explore, progress)
+}
+
+mv_continuous <- function(Y, X, delta, epsilon, iter, start, progress) {
+    .Call(`_BGGM_mv_continuous`, Y, X, delta, epsilon, iter, start, progress)
+}
+
+trunc_mvn <- function(mu, rinv, z, y, cutpoints) {
+    .Call(`_BGGM_trunc_mvn`, mu, rinv, z, y, cutpoints)
+}
+
+mv_binary <- function(Y, X, delta, epsilon, iter, beta_prior, cutpoints, start, progress) {
+    .Call(`_BGGM_mv_binary`, Y, X, delta, epsilon, iter, beta_prior, cutpoints, start, progress)
 }
 
 mv_ordinal_cowles <- function(Y, X, delta, epsilon, iter, MH) {
