+++ README.Rmd
@@ -279,7 +279,6 @@ hyp <- c("PHQ2--PHQ9 > PHQ1--PHQ9 > 0;
 
 There are two hypotheses separated by (`;`). The first expresses that the relation `PHQ2--PHQ9` ("feeling down, depressed, or hopeless" and "suicidal thoughts") is larger than `PHQ1--PHQ9` ("little interest or pleasure in doing things" and "suicidal thoughts"). In other words, that the partial correlation is larger for `PHQ2--PHQ9`. There is an additional constraint to positive values (`> 0`) for both relations. The second hypothesis is then a "null" model.
 
-=======
 
 ```{r, echo=FALSE, warning = FALSE}
 load(file = "readme_models/fit_hyp1.rda")
