+++ DESCRIPTION
@@ -1,4 +1,4 @@
-Package: BGGM 
+Package: BGGM
 Type: Package
 Title: Bayesian Gaussian Graphical Models
 Version: 2.0.0
