+++ _pkgdown.yml
@@ -1,0 +1,6 @@
+title: MODIStsp
+url: https://donaldrwilliams.github.io/BGGM/
+template:
+  params:
+    bootswatch: flatly
+    ganalytics: UA-12345678-0
