+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-title: MODIStsp
+title: BGGM
 url: https://donaldrwilliams.github.io/BGGM/
 template:
   params:
