### 2020-05-19 546bc0a
fix coef documentation
