+++ src-i386/bggm_fast.cpp
@@ -1,1973 +1,0 @@
-// -*- mode: C++; c-indent-level: 4; c-basic-offset: 4; indent-tabs-mode: nil; -*-
-// we only include RcppArmadillo.h which pulls Rcpp.h in for us
-
-#include "RcppArmadillo.h"
-#include <progress.hpp>
-#include <progress_bar.hpp>
-#include <truncnorm.h>
-
-// [[Rcpp::depends(RcppArmadillo, RcppDist, RcppProgress)]]
-
-
-// [[Rcpp::export]]
-arma::mat Sigma_i_not_i(arma::mat x, int index) {
-  arma::mat sub_x = x.row(index);
-  sub_x.shed_col(index);
-  return(sub_x);
-}
-
-
-// [[Rcpp::export]]
-arma::vec select_col(arma::mat x, int index){
-  arma::vec z = x.col(index);
-  return(z);
-}
-
-// [[Rcpp::export]]
-arma::mat select_row(arma::mat x, int index){
-  arma::mat z = x.row(index);
-  return(z);
-}
-
-// [[Rcpp::export]]
-arma::mat remove_row(arma::mat x, int which){
-  x.shed_row(which);
-  return(x);
-}
-
-// [[Rcpp::export]]
-arma::mat remove_col(arma::mat x, int index){
-  x.shed_col(index);
-  return(x);
-}
-
-
-// matrix F continous sampler
-// Williams, D. R., & Mulder, J. (2019). Bayesian hypothesis testing for Gaussian
-// graphical models:  Conditional independence and order constraints.
-
-// [[Rcpp::export]]
-Rcpp::List Theta_continuous(arma::mat Y,
-                            int iter,
-                            float delta,
-                            float epsilon,
-                            int prior_only,
-                            int explore,
-                            arma::mat start,
-                            bool progress) {
-
-
-
-  // note p changed to k to be consistent
-  //with the multivariate regression samplers
-  Progress  p(iter, progress);
-
-  // number of rows
-  float n = Y.n_rows;
-
-  int k = 1;
-
-  // sample prior
-  if(prior_only == 1){
-
-    if(explore == 1){
-
-      k = 3;
-
-    } else {
-
-      k = Y.n_cols;
-
-    }
-
-  } else {
-
-    // number of columns
-    k = Y.n_cols;
-
-  }
-
-  // k by k identity mat
-  arma::mat  I_k(k, k, arma::fill::eye);
-
-  int nu = 1/ epsilon;
-  // // #nu in Mulder & Pericchi (2018) formula (30) line 1.
-  int nuMP = delta + k - 1 ;
-  //
-  // // #delta in Mulder & Pericchi (2018) formula (30) line 1.
-  int deltaMP = nu - k + 1 ;
-
-  // Psi update
-  arma::cube Psi(k, k, 1, arma::fill::zeros);
-
-  arma::mat B(epsilon * I_k);
-  arma::mat BMP(inv(B));
-  arma::mat BMPinv(inv(BMP));
-
-  // precison matrix
-  arma::cube Theta(k, k, 1, arma::fill::zeros);
-  arma::cube Theta_mcmc(k, k, iter, arma::fill::zeros);
-
-  // partial correlations
-  arma::mat pcors(k,k);
-  arma::cube pcors_mcmc(k, k, iter, arma::fill::zeros);
-
[…]
