+++ src-i386/RcppExports.cpp
@@ -1,424 +1,0 @@
-// Generated by using Rcpp::compileAttributes() -> do not edit by hand
-// Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
-
-#include <RcppArmadillo.h>
-#include <Rcpp.h>
-
-using namespace Rcpp;
-
-// Sigma_i_not_i
-arma::mat Sigma_i_not_i(arma::mat x, int index);
-RcppExport SEXP _BGGM_Sigma_i_not_i(SEXP xSEXP, SEXP indexSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::mat >::type x(xSEXP);
-    Rcpp::traits::input_parameter< int >::type index(indexSEXP);
-    rcpp_result_gen = Rcpp::wrap(Sigma_i_not_i(x, index));
-    return rcpp_result_gen;
-END_RCPP
-}
-// select_col
-arma::vec select_col(arma::mat x, int index);
-RcppExport SEXP _BGGM_select_col(SEXP xSEXP, SEXP indexSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::mat >::type x(xSEXP);
-    Rcpp::traits::input_parameter< int >::type index(indexSEXP);
-    rcpp_result_gen = Rcpp::wrap(select_col(x, index));
-    return rcpp_result_gen;
-END_RCPP
-}
-// select_row
-arma::mat select_row(arma::mat x, int index);
-RcppExport SEXP _BGGM_select_row(SEXP xSEXP, SEXP indexSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::mat >::type x(xSEXP);
-    Rcpp::traits::input_parameter< int >::type index(indexSEXP);
-    rcpp_result_gen = Rcpp::wrap(select_row(x, index));
-    return rcpp_result_gen;
-END_RCPP
-}
-// remove_row
-arma::mat remove_row(arma::mat x, int which);
-RcppExport SEXP _BGGM_remove_row(SEXP xSEXP, SEXP whichSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::mat >::type x(xSEXP);
-    Rcpp::traits::input_parameter< int >::type which(whichSEXP);
-    rcpp_result_gen = Rcpp::wrap(remove_row(x, which));
-    return rcpp_result_gen;
-END_RCPP
-}
-// remove_col
-arma::mat remove_col(arma::mat x, int index);
-RcppExport SEXP _BGGM_remove_col(SEXP xSEXP, SEXP indexSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< arma::mat >::type x(xSEXP);
-    Rcpp::traits::input_parameter< int >::type index(indexSEXP);
-    rcpp_result_gen = Rcpp::wrap(remove_col(x, index));
-    return rcpp_result_gen;
-END_RCPP
-}
-// Theta_continuous
[…]
