+++ R/convergence.R
@@ -27,56 +27,21 @@
 #' Y <- ptsd
 #'
 #' #########################
-#' ###### continuous #######
-#' #########################
-#' fit <- estimate(Y, iter = 250)
-#'
-#' # print names first
-#' convergence(fit, print_names = TRUE)
-#'
-#' # trace plots
-#' convergence(fit, type = "trace",
-#'             param = c("B1--B2", "B1--B3"))
-#'
-#' # acf plots
-#' convergence(fit, type = "acf",
-#'             param = c("B1--B2", "B1--B3"))
-#'
-#' #########################
-#' ######## mixed ##########
-#' #########################
-#' # copula
-#'
-#' fit <- estimate(Y, type = "mixed",
-#'                 iter = 250)
-#'
-#' # print names first
-#' convergence(fit, print_names = TRUE)
-#'
-#' # trace plots
-#' convergence(fit, type = "trace",
-#'             param = c("B1--B2", "B1--B3"))
-#'
-#' # acf plots
-#' convergence(fit, type = "acf",
-#'             param = c("B1--B2", "B1--B3"))
-#'
-#' #########################
 #' ######## ordinal ########
 #' #########################
 #' fit <- estimate(Y + 1, type = "ordinal",
 #'                 iter = 250)
 #'
 #' # print names first
-#' convergence(fit, print_names = TRUE)
+#' param_names <- convergence(fit, print_names = TRUE)
 #'
 #' # trace plots
 #' convergence(fit, type = "trace",
-#'             param = c("B1--B2", "B1--B3"))
+#'             param = c("B1--B2", "B1--B3"))[[1]]
 #'
 #' # acf plots
 #' convergence(fit, type = "acf",
-#'             param = c("B1--B2", "B1--B3"))
+#'             param = c("B1--B2", "B1--B3"))[[1]]
 #' }
 #' @export
 convergence <- function(object,
