+++ R/coef.estimate.R
@@ -47,34 +47,6 @@
 #' # summary
 #' summ <- summary(reg)
 #'
-#' # ordinal
-#'
-#' # fit model (note + 1, due to zeros)
-#' fit <- estimate(Y + 1, type = "ordinal",
-#'                 iter = 250)
-#'
-#' # summarize the partial correlations
-#' reg <- coef(fit)
-#'
-#' # summary
-#' summ <- summary(reg)
-#'
-#' #########################
-#' ### example 2: binary ###
-#' #########################
-#' # data
-#' Y <- women_math
-#'
-#' # fit model
-#' fit <- estimate(Y, type = "binary",
-#'                 iter = 250)
-#'
-#' # summarize the partial correlations
-#' reg <- coef(fit)
-#'
-#' # summary
-#' summ <- summary(reg)
-#'
 #'}
 #' @export
 coef.estimate <- function(object, iter = NULL,...) {
@@ -168,40 +140,6 @@ coef.estimate <- function(object, iter = NULL,...) {
 #' @examples
 #' \donttest{
 #' # note: iter = 250 for demonstrative purposes
-#'
-#' # data
-#' Y <- ptsd
-#' #########################################
-#' ### example 1: continuous and ordinal ###
-#' #########################################
-#'
-#' # fit model
-#' fit <- explore(Y, type = "continuous",
-#'                iter = 250)
-#'
-#' # regression
-#' reg <- coef(fit)
-#'
-#' # summary
-#' summ <- summary(reg)
-#'
-#' summ
-#'
-#' # ordinal
-#'
-#' # fit model (note + 1, due to zeros)
-#' fit <- explore(Y + 1, type = "ordinal",
-#'                iter = 250)
-#'
-#' # summarize the partial correlations
-#' reg <- coef(fit)
-#'
-#' # summary
-#' summ <- summary(reg)
-#'
-#' #########################
-#' ### example 2: binary ###
-#' #########################
 #' # data
 #' Y <- women_math
 #'
