+++ index.Rmd
@@ -1,8 +1,5 @@
 ---
-title: "index"
-author: "Donny Williams"
-date: "5/18/2020"
-output: html_document
+output: github_document
 ---
 
 ```{r, echo = FALSE, message=F}
@@ -19,10 +16,9 @@ library(ggplot2)
 library(BGGM)
 ```
 
-![](joss_paper/hex.png)
+#  BGGM: Bayesian Gaussian Graphical Models <img src="joss_paper/hex.pn" align="right" alt="" width="120" />
 
 
-# BGGM: Bayesian Gaussian Graphical Models
 
 [![CRAN Version](http://www.r-pkg.org/badges/version/BGGM)](https://cran.r-project.org/package=BGGM)
 [![Downloads](https://cranlogs.r-pkg.org/badges/BGGM)](https://cran.r-project.org/package=BGGM)
