+++ index.md
@@ -1,0 +1,250 @@
+
+# Bayesian Gaussian Graphical Models <img src="joss_paper/hex.png" align="right" alt="" width="240" />
+
+[![CRAN
+Version](http://www.r-pkg.org/badges/version/BGGM)](https://cran.r-project.org/package=BGGM)
+[![Downloads](https://cranlogs.r-pkg.org/badges/BGGM)](https://cran.r-project.org/package=BGGM)
+[![Build
+Status](https://travis-ci.org/donaldRwilliams/BGGM.svg?branch=master)](https://travis-ci.org/donaldRwilliams/BGGM)
+
+The `R` package **BGGM** provides tools for making Bayesian inference in
+Gaussian graphical models (GGM). The methods are organized around two
+general approaches for Bayesian inference: (1) estimation and (2)
+hypothesis testing. The key distinction is that the former focuses on
+either the posterior or posterior predictive distribution (Gelman, Meng,
+and Stern 1996; see section 5 in Rubin 1984) , whereas the latter
+focuses on model comparison with the Bayes factor (Jeffreys 1961; Kass
+and Raftery 1995).
+
+## Installation
+
+To install the latest release version (1.0.0) from CRAN use
+
+``` r
+install.packages("BGGM")    
+```
+
+The current developmental version can be installed with
+
+``` r
+if (!requireNamespace("remotes")) { 
+  install.packages("remotes")   
+}   
+remotes::install_github("donaldRwilliams/BGGM")
+```
+
+Note that the developmental version is recommended, due to the
+algorithms being written in `c++` and support for all data types. The
+developmental version is essentially **BGGM** version 2.0.0.
+
+## Overview
+
+The methods in **BGGM** build upon existing algorithms that are
+well-known in the literature. The central contribution of **BGGM** is to
+extend those approaches:
+
+1.  Bayesian estimation with the novel matrix-F prior distribution
+    (Mulder and Pericchi 2018)
+    
+      - Estimation (Williams 2018)
+
+2.  Bayesian hypothesis testing with the matrix-F prior distribution
+    (Williams and Mulder 2019)
+    
+      - Exploratory hypothesis testing
+    
+      - Confirmatory hypothesis testing
+
+3.  Comparing Gaussian graphical models (Williams 2018; Williams et al.
+    2020)
+    
+      - Partial correlation differences
+    
+      - Posterior predictive check
+    
+      - Exploratory hypothesis testing
+    
+      - Confirmatory hypothesis testing
+
+4.  Extending inference beyond the conditional (in)dependence structure
+    (Williams 2018)
+    
+      - Predictability
+    
+      - Posterior uncertainty intervals for the partial correlations
+    
+      - Custom Network Statistics
+
+The computationally intensive tasks are written in `c++` via the `R`
+package **Rcpp** (Eddelbuettel et al. 2011) and the `c++` library
+**Armadillo** (Sanderson and Curtin 2016). The Bayes factors are
+computed with the `R` package **BFpack** (Mulder et al. 2019).
+Furthermore, there are plotting functions for each method, control
+variables can be included in the model (e.g., `~ gender`), and there is
+support for missing values (see `bggm_missing`).
+
+## Supported Data Types
+
+  - **Continuous**: The continuous method was described in Williams
[…]
