+++ _pkgdown.yml
@@ -1,6 +1,2 @@
 title: BGGM
 url: https://donaldrwilliams.github.io/BGGM/
-template:
-  params:
-    bootswatch: flatly
-    ganalytics: UA-12345678-0
