+++ index.Rmd
@@ -17,14 +17,14 @@ library(ggplot2)
 library(BGGM)
 ```
 
-# Bayesian Gaussian Graphical Models <img src="joss_paper/hex.png" align="right" alt="" width="240" />
+# Bayesian Gaussian Graphical Models <img src="man/figures/hex.png" align="right" alt="" width="120" />
 
 
-
+<!-- badges: start -->
 [![CRAN Version](http://www.r-pkg.org/badges/version/BGGM)](https://cran.r-project.org/package=BGGM)
 [![Downloads](https://cranlogs.r-pkg.org/badges/BGGM)](https://cran.r-project.org/package=BGGM)
 [![Build Status](https://travis-ci.org/donaldRwilliams/BGGM.svg?branch=master)](https://travis-ci.org/donaldRwilliams/BGGM)
-
+<!-- badges: end -->
 
 
 The `R` package **BGGM** provides tools for making Bayesian inference in 
