### 2020-05-19 f699e50
change BGGM size
