+++ NEWS.md
@@ -1,0 +1,3 @@
+# BGGM (developmental version)
+
+* Testing news
