### 2020-05-19 5adba92
add BGGM reference
