+++ inst/REFERENCES.bib
@@ -469,3 +469,10 @@ year = {2011}
   year={2019},
   publisher={Taylor \& Francis}
 }
+
+@article{williams2020bggm,
+  title={BGGM: Bayesian Gaussian Graphical Models in R},
+  author={Williams, Donald R and Mulder, Joris},
+  year={2020},
+  publisher={PsyArXiv}
+}
