### 2020-05-19 9cf20a4
remove arma::mat
