### 2020-05-19 b98e376
add arma:: to trimatu
