+++ vignettes/test_vig.Rmd
@@ -1,12 +1,0 @@
----
-title: "Put the title of your vignette here"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Put the title of your vignette here}
-  %\VignetteEngine{knitr::rmarkdown}
-  \usepackage[utf8]{inputenc}
----
-
-```{r}
-1 + 1
-```
