+++ vignettes/ppc_custom.Rmd
@@ -1,0 +1,441 @@
+---
+title: "Custom Network Comparisons"
+author: "Donny Williams"
+date: "5/19/2020"
+bibliography: ../inst/REFERENCES.bib
+output: 
+  rmarkdown::html_vignette:
+    toc: yes
+vignette: >
+  %\VignetteIndexEntry{Custom Network Comparisons}
+  %\VignetteEngine{knitr::rmarkdown}
+  \usepackage[utf8]{inputenc}
+---
+
+# Background
+It is quite common to have partial correlation networks (GGMs) for various subgroups, 
+say, males and females, a control and treatment group, or perhaps several educational 
+levels. In this case, it is important to not only determine whether the groups are 
+different, but actually compare the groups in a way that answers a specific question 
+of interest. 
+
+To date, most `R` packages provide a few ways to compare groups, including **BGGM** (version `1.0.0`). 
+In version `2.0.0`, however, **BGGM** includes a new feature for the function `ggm_compare_ppc` that enables 
+users to **compare networks in any way the want**. 
+
+# Basic Idea
+The techical details of the approach are described in [@williams2020comparing]. The basic idea is to
+
+1. Draw samples from the posterior distribution, assuming the groups are equal (i.e., the "null" model).
+
+2. Generate the posterior **predictive** distribution for the chosen test-statistic (how the groups 
+   are being compared)
+
+    + This can be understood as what we would expect to observe in the future 
+    (e.g., in replication), assuming the groups were in fact equal.
+
+3. Compute the test-statistic for the observed groups.
+
+4. Then compare the observed test-statisic to the predictive distribution 
+   (what is expected under the "null" model). 
+
+    + If the observed error is larger than the model assuming group equality, this 
+      suggests that the groups are different.
+
+
+In **BGGM**, the default is to compare the groups with respect to (symmetric) Kullback-Leibler 
+divergence (i.e., "distance" between multivariate normal distributions)  and the sum of 
+squared error (for the partial correlation matrix). This was shown to be quite powerful in @williams2020comparing, while also having a 
+low false positive rate. 
+
+In the following, the focus is on defining custom functions
+and using them with `ggm_compare_ppc`. In all examples, post-traumatic stress disorder
+networks are compared [@fried2018replicability].
+
+### R packages
+```{r, eval = FALSE}
+# need the developmental version
+if (!requireNamespace("remotes")) { 
+  install.packages("remotes")   
+}   
+
+# install from github
+remotes::install_github("donaldRwilliams/BGGM")
+
+```
+
+
+### Data
+Only the correlation matrices are available. Hence, multivarite normal data is generated with that *exact* 
+correlation structure via the `R` package **MASS**.
+```{r, warning =FALSE, message=FALSE}
+# need these packages
+library(BGGM)
+library(ggplot2)
+library(assortnet)
+library(networktools)
+library(MASS)
+
+# group 1
+Yg1 <- MASS::mvrnorm(n = 926, 
+                     mu = rep(0, 16), 
+                     Sigma = ptsd_cor3, 
+                     empirical = TRUE)
+
+# group 2
+Yg2 <- MASS::mvrnorm(n = 956, 
+                     mu = rep(0, 16), 
+                     Sigma = ptsd_cor4, 
+                     empirical = TRUE)
+```
+
+
+# Illustrative Examples
+
+## Correlation
+This first example looks at the correlation between partial correlations of the two networks. Note that
[…]
