+++ inst/REFERENCES.bib
@@ -476,3 +476,47 @@ year = {2011}
   year={2020},
   publisher={PsyArXiv}
 }
+
+@article{meng1994posterior,
+  title={Posterior predictive $ p $-values},
+  author={Meng, Xiao-Li and others},
+  journal={The annals of statistics},
+  volume={22},
+  number={3},
+  pages={1142--1160},
+  year={1994},
+  publisher={Institute of Mathematical Statistics}
+}
+
+@article{gelman2013two,
+  title={Two simple examples for understanding posterior p-values whose distributions are far from uniform},
+  author={Gelman, Andrew and others},
+  journal={Electronic Journal of Statistics},
+  volume={7},
+  pages={2595--2602},
+  year={2013},
+  publisher={The Institute of Mathematical Statistics and the Bernoulli Society}
+}
+
+
+@article{newman2003mixing,
+  title={Mixing patterns in networks},
+  author={Newman, Mark EJ},
+  journal={Physical Review E},
+  volume={67},
+  number={2},
+  pages={026126},
+  year={2003},
+  publisher={APS}
+}
+
+@article{robinaugh2016identifying,
+  title={Identifying highly influential nodes in the complicated grief network.},
+  author={Robinaugh, Donald J and Millner, Alexander J and McNally, Richard J},
+  journal={Journal of Abnormal Psychology},
+  volume={125},
+  number={6},
+  pages={747},
+  year={2016},
+  publisher={American Psychological Association}
+}
