+++ inst/REFERENCES.bib
@@ -474,7 +474,8 @@ year = {2011}
   title={BGGM: Bayesian Gaussian Graphical Models in R},
   author={Williams, Donald R and Mulder, Joris},
   year={2020},
-  publisher={PsyArXiv}
+  publisher={PsyArXiv},
+   journal={PsyArXiv}
 }
 
 @article{meng1994posterior,
@@ -519,4 +520,14 @@ year = {2011}
   pages={747},
   year={2016},
   publisher={American Psychological Association}
+}
+
+@inproceedings{herdin2005correlation,
+  title={Correlation matrix distance, a meaningful measure for evaluation of non-stationary MIMO channels},
+  author={Herdin, Markus and Czink, Nicolai and Ozcelik, H{\"u}seyin and Bonek, Ernst},
+  booktitle={2005 IEEE 61st Vehicular Technology Conference},
+  volume={1},
+  pages={136--140},
+  year={2005},
+  organization={IEEE}
 }
