### 2020-05-19 212dba3
website: fix typos
