### 2020-05-20 #2111 @arfon
:wave: @donaldRwilliams - just a friendly check in to see if you've managed to make these updates yet. JOSS is now reopen for submissions so we're actively trying to wrap up existing reviews.
