### 2020-05-20 #2111 @donaldRwilliams
Hi:
Sounds good. Just now finished making the rather large changes to the documentation, README, etc. Will have the paper updated by end of the week
