### 2020-05-20 024e7c9
webiste: fix typos in custom net stat
