### 2020-05-20 5906776
webiste: fix typos in custom network stats
