### 2020-05-20 93a5277
website: fix typos
