+++ vignettes/netstat_custom.Rmd
@@ -20,7 +20,7 @@ called `roll_your_own` and it was suggested by a user of **BGGM**  ([see feature
 
 # Basic Idea
 The basic idea is to compute the chosen network statistic for each of the sampled partial 
-correlations, resulting in a distribution. All that is required is to define a function
+correlation matrices, resulting in a distribution. All that is required is to define a function
 that takes either a partial correlation matrix or a weighted adjacency matrix 
 (the partial correlation matrix with values set to zero) as the first argument. 
 Several examples are provided below.
@@ -48,10 +48,9 @@ library(BGGM)
 library(ggplot2)
 library(assortnet)
 library(networktools)
-library(MASS)
 
+# data
 Y <- ptsd[,1:7]
-
 ```
 
 ### Fit Model
@@ -59,14 +58,11 @@ For these data, the GGM is estimated with a semi-parametric copula [@hoff2007ext
 In **BGGM**, this implemented with `type = mixed` which is kind of a misnomer because the data do not 
 have to be "mixed" (consisting of continuous and discrete variables). 
 Note that the model is fitted only once which highlights that only the posterior samples 
-are needed to compute any network statistic based on the sampled partial correlation matrices.
-
-```{r, echo=FALSE}
-load(file = "netstat_fit.rda")
-```
+are needed to compute any network statistic.
 
 ```{r, message=FALSE, warning=FALSE, eval=FALSE}
 library(BGGM)
+
 # copula ggm
 fit <- estimate(Y, type = "mixed", iter = 1000)
 ```
@@ -87,7 +83,7 @@ Note that `x` takes the matrix which is then passed to `expectedInf`. The `...` 
 passing additional arguments to the `expectedInf` function. An example is provided below. 
 With the function defined, the next step is to compute the network statistic.
 
-```{r, eval = TRUE, message=FALSE, results='hide'}
+```{r, eval = FALSE, message=FALSE, results='hide'}
 # iter = 250 for demonstrative purposes
 # (but note even 1000 iters takes less than 1 second)
 # compute
@@ -95,35 +91,69 @@ net_stat <- roll_your_own(object = fit,
                           FUN = f,
                           select = FALSE,
                           iter = 250)
-```
-```{r}
+# print
 net_stat
+
+#> BGGM: Bayesian Gaussian Graphical Models 
+#> --- 
+#> Network Stats: Roll Your Own
+#> Posterior Samples: 250 
+#> --- 
+#> Estimates: 
+#> 
+#>  Node Post.mean Post.sd Cred.lb Cred.ub
+#>     1     0.701   0.099   0.508   0.871
+#>     2     0.912   0.113   0.722   1.179
+#>     3     0.985   0.112   0.742   1.199
+#>     4     1.056   0.105   0.851   1.247
+#>     5     1.056   0.116   0.862   1.288
+#>     6     0.491   0.092   0.329   0.679
+#>     7     0.698   0.098   0.521   0.878
+#> --- 
 ```
+
 
 
 The option `select = FALSE` indicates to compute the statistics from the partial correlation matrices (nothing set to zero). This can be changed with `select = TRUE`. Internally, each of the sampled
 partial correlation matrices is multiplied by the adjacency matrix.
-```{r, eval = TRUE, results='hide'}
+```{r, eval = FALSE, results='hide'}
 net_stat <- roll_your_own(object = fit,
                           FUN = f,
                           select = TRUE,
                           iter = 250)
-```
-```{r}
+
+# print
 net_stat
+
+#> BGGM: Bayesian Gaussian Graphical Models 
[…]
