### 2020-05-20 03406ad
webiste: fix typos in predictability
