### 2020-05-21 ad4d226
webiste: mcmc diagnostics
