+++ vignettes/mcmc_diagnostics.Rmd
@@ -31,7 +31,7 @@ if (!requireNamespace("remotes")) {
 remotes::install_github("donaldRwilliams/BGGM")
 library(BGGM)
 ```
-```{r, echo=FALSE}
+```{r, echo=FALSE, message=FALSE}
 library(BGGM)
 ```
 
