### 2020-05-21 c078971
website: add reference to quote for acf plot
