+++ vignettes/mcmc_diagnostics.Rmd
@@ -38,7 +38,7 @@ library(BGGM)
 # ACF plot
 This first example includes an "acf" plot that looks at the auto correlation of the samples. In general,
 we do not want the samples to be strongly correlated or related to the previous samples (or lags $k$). 
-I am not sure there are general guidelines, but typically we do not want "auto correlation...for higher values of k, [because] this indicates a high degree of correlation between our samples and slow mixing."
+I am not sure there are general guidelines, but typically we do not want "auto correlation...for higher values of k, [because] this indicates a high degree of correlation between our samples and slow mixing " [source](http://sbfnk.github.io/mfiidd/mcmc_diagnostics.html)
 
 Here is an example for ordinal data. 
 
