### 2020-05-21 d4c72f1
website: fix typos in mcmc diagnostics
