+++ vignettes/mcmc_diagnostics.Rmd
@@ -74,7 +74,7 @@ convergence(fit, param = "B1--B2", type = "acf")
 
 ```
 
-![](../man/figures/acf.png)
+![](../man/figures/acf_problem.png)
 
 The argument `param` can take any number of parameters and a plot will be made for each (e.g.., `param = c("B1--B2", B1--B3)`). In this case, the auto correlations looks acceptable and actually really good (note the drop to zero). A problematic `acf` plot would have the black lines start at `1.0`
 and perhaps never go below `0.20`.  
