+++ vignettes/mcmc_diagnostics.Rmd
@@ -74,7 +74,9 @@ convergence(fit, param = "B1--B2", type = "acf")
 
 ```
 
-![](../man/figures/acf_problem.png)
+
+![](../man/figures/acf.png)
+[link if broken image](https://github.com/donaldRwilliams/BGGM/blob/master/man/figures/acf.png)
 
 The argument `param` can take any number of parameters and a plot will be made for each (e.g.., `param = c("B1--B2", B1--B3)`). In this case, the auto correlations looks acceptable and actually really good (note the drop to zero). A problematic `acf` plot would have the black lines start at `1.0`
 and perhaps never go below `0.20`.  
@@ -87,6 +89,7 @@ ts.sim <- arima.sim(list(order = c(1,1,0), ar = 0.7), n = 200)
 
 acf(ts.sim)
 ```
+
 ![](../man/figures/acf_problem.png)
 
 This would be considered problematic. If this occurs, one solution could be to thin the samples manually
