+++ _pkgdown.yml
@@ -5,7 +5,7 @@ template:
     bootswatch: flatly
   opengraph:
     image:
-      src: man/figures/logo.jpg
+      src: man/figures/hex.jpg
     twitter:
-      creator: "wdonald_1985"
+      creator: "@wdonald_1985"
       card: summary
