### 2020-05-21 64eb6b0
JOSS: remove old paper with updated
