+++ paper.md
@@ -18,322 +18,72 @@ affiliations:
  - name: Department of Methodology and Statistics, Tilburg University
    index: 2
 citation_author: Williams and Mulder
-date: 07 January 2020
+date: 05 May 2020
 year: 2020
-bibliography: joss_paper/paper.bib
-output: rticles::joss_article
-csl: joss_paper/apa.csl
-journal: JOSS
-submitted: 07 January 2020
-published: (pre-print)
+bibliography: inst/REFERENCES.bib
 ---
 
-# Summary
-Gaussian graphical models (GGM) allow for learning conditional (in)dependence structures that are encoded by partial correlations. Whereas there are several `R` packages for classical methods [see @Kuismin2017, Table 1], there are only two that implement a Bayesian approach [@Leday2018; @Mohammadi2015a]. These are exclusively focused on identifying the graphical structure. The package **BGGM** not only contains novel Bayesian methods for this purpose, but it also includes Bayesian methodology for extending inference beyond identifying non-zero relations. 
+The `R` package **BGGM** provides tools for making Bayesian inference in 
+Gaussian graphical models (GGM). The methods are organized around two general 
+approaches for Bayesian inference: (1) estimation and (2) hypothesis 
+testing. The key distinction is that the former focuses on either 
+the posterior or posterior predictive distribution [@Gelman1996a; see section 
+5 in @rubin1984bayesianly] , whereas the latter focuses on model comparison with the Bayes factor [@Jeffreys1961; @Kass1995].
 
-**BGGM** is built around two approaches for *Bayesian* inference--estimation and hypothesis testing. This distinction is arbitrary [see @Rouder2018], but is used to organize this work.  The former focuses on the posterior distribution and includes extensions to assess predictability [@Haslbeck2018], as well as methodology to compare partial correlations. The latter includes methods for Bayesian hypothesis testing, in both exploratory and confirmatory contexts, with the novel matrix-$F$ prior distribution [@Mulder2018]. This allows for testing the null hypothesis of conditional independence, as well as inequality and equality constrained hypotheses. Further, there are several approaches for comparing GGMs across any number of groups.  The package also includes a suite of options for model checking. Together, **BGGM** is a comprehensive toolbox for Gaussian graphical modeling in `R`. 
+# Overview
+The methods in **BGGM** build upon existing algorithms that are well-known in the literature.
+The central contribution of **BGGM** is to extend those approaches:
 
-## Estimation
-There are two possibilities for *estimating* GGMs. The first is an analytic solution and the second samples from the posterior distribution [described in @Williams2018bayes]. Sampling is recommended because the samples are required for various functions in **BGGM** (e.g., posterior predictive checks and prediction). 
+1.  Bayesian estimation with the novel matrix-F prior distribution [@Mulder2018]
+  
+    + [Estimation](https://github.com/donaldRwilliams/BGGM#bayesian-estimation) [@Williams2019]
 
-### Structure Learning
-Structure learning refers to determining which partial correlations are non-zero. This is implemented with:
+2. Bayesian hypothesis testing with the matrix-F prior distribution [@Williams2019_bf]
 
-```r
-# data (p = 5 for demonstrative purposes)
-Y <- BGGM::bfi[,1:5]
+    + [Exploratory hypothesis testing](https://github.com/donaldRwilliams/BGGM#Exploratory)
+  
[…]
