### 2020-05-21 200843f
JOSS: add subheading "BGGM"
