+++ NEWS.md
@@ -1,3 +1,1 @@
 # BGGM (developmental version)
-
-* Testing news
