### 2020-05-21 #2111 @donaldRwilliams
@arfon
We updated the paper. Note the package has undergone extensive changes, including rewriting all the algorithms in `c++`, adding support for binary, ordinal, and mixed data, linked to Travis, etc.

As for the reviewer comments, we extensively rewrote the documentation to include many examples, added to Travis, and most notably shortened the paper to be a "high-level" overview.
