### 2020-05-21 6caf7ac
README: update hex
