+++ README.Rmd
@@ -17,7 +17,7 @@ library(ggplot2)
 library(BGGM)
 ```
 
-<img src="joss_paper/hex.png" width = 250 />
+<img src="man/figures/hex.png" width = 250 />
 
 # BGGM: Bayesian Gaussian Graphical Models
 
