+++ paper.md
@@ -87,4 +87,8 @@ The computationally intensive tasks are written in `c++` via the `R` package **R
 for each method, control variables can be included in the model (e.g., `~ gender`), 
 and there is support for missing values (see `bggm_missing`).
 
+# Acknowledgements
+DRW was supported by a National Science Foundation Graduate Research Fellowship
+under Grant No. 1650042 and JM was supported by a ERC Starting Grant (758791).
+
 # References
