### 2020-05-21 cc7cbac
README: fix logo location
