### 2020-05-21 cd8bd6d
JOSS: Comparsion to other packages
