### 2020-05-21 donaldRwilliams/BGGM#4 @donaldRwilliams
I added these to the readme, as to keep the paper short as you suggested @paulgovan

see here for real world examples

suicidal thoughts:
https://github.com/donaldRwilliams/BGGM#Confirmatory

comparing agreeableness for males and females
https://github.com/donaldRwilliams/BGGM#confirmatory-groups
