### 2020-05-21 donaldRwilliams/BGGM#8 @donaldRwilliams
We took your advice and now provide an overview of the package..
