### 2020-05-21 donaldRwilliams/BGGM#5 @donaldRwilliams
We now have travis and the badge is on the readme
