### 2020-05-21 donaldRwilliams/BGGM#7 @donaldRwilliams
We added a note to the README. To our knowledge, there is no R package that includes all the algorithms and methods.
