+++ README.Rmd
@@ -862,4 +862,7 @@ can be used by all researchers. However, there is currently *not* support for hi
 These data are rare in the social-behavioral sciences. In the future, support for high-dimensional
 data may be added to **BGGM**.
 
+## Bug Reports and Feature Requests
+Bug reports and feature requests can be made by opening an issue on [Github](https://github.com/donaldRwilliams/BGGM/issues).
+
 ## References
