### 2020-05-21 donaldRwilliams/BGGM#6 @donaldRwilliams
We have addressed this in the readme:

https://github.com/donaldRwilliams/BGGM#bug-reports-and-feature-requests
