### 2020-05-21 f5c7cd3
README: add JOSS status badge
