+++ inst/REFERENCES.bib
@@ -531,3 +531,13 @@ year = {2011}
   year={2005},
   organization={IEEE}
 }
+
+@article{epskamp2012qgraph,
+  title={qgraph: Network visualizations of relationships in psychometric data},
+  author={Epskamp, Sacha and Cramer, Ang{\'e}lique OJ and Waldorp, Lourens J and Schmittmann, Verena D and Borsboom, Denny and others},
+  journal={Journal of Statistical Software},
+  volume={48},
+  number={4},
+  pages={1--18},
+  year={2012}
+}
