+++ vignettes/netplot.Rmd
@@ -71,9 +71,7 @@ plot(E)
 
 ### Customize Plot
 The above is `ggplot` that can be futher honed in. Here is an example.
-
-
-```{r, eval=FALSE}
+```r
 # extract communities
 comm <- substring(colnames(Y), 1, 1)
 
@@ -96,14 +94,11 @@ plot(E,
                      palette = "Set2")
 
 ```
+![](../man/figures/netplot_2.png)
 
 The `edge_magnify` is a value that is multiplied by the edges, `groups` allows for grouping the 
 variables (e.g., those thought to belong to the same "community" will be the same color), and the
 `scale_color_brewer` is from the package `ggplot2` (`pallete` controls the color of the `groups`).
-
-
-![](../man/figures/netplot_2.png)
-
 
 This is just scratching the surface of possibilities, as essentially any change 
 can be made to the plot. There is lots of support for making nice plots readily available
@@ -132,4 +127,12 @@ plot(E,
 ```
 ![](../man/figures/netplot_3.png)
 
+# Hypothesis Testing
+Forthcoming
 
+# Note
+**BGGM** provides a publication ready plot, but it is also limited compared to **qgraph** [@epskamp2012qgraph]. The one advantage of **BGGM**  is that all plots are `ggplots` which then
+allows for combining them rather easily. An example is included in another vignette that shows how to 
+combine several plots made with various methods in **BGGM**
+
+# References
