### 2020-05-21 donaldRwilliams/BGGM#7 @jayrobwilliams
The updated README and paper note that there is no R package that implements all of these methods. Are there other R packages that implement different subsets of these methods? If so, they would be important to mention here. Is the new c++ implementation faster than these packages? This would also be good to mention.
