### 2020-05-21 0bb66e1
website: add note about colors
