+++ inst/REFERENCES.bib
@@ -541,3 +541,10 @@ year = {2011}
   pages={1--18},
   year={2012}
 }
+
+@article{mohammadi2015bdgraph,
+  title={BDgraph: An R package for Bayesian structure learning in graphical models},
+  author={Mohammadi, Reza and Wit, Ernst C},
+  journal={arXiv preprint arXiv:1501.05108},
+  year={2015}
+}
