+++ README.Rmd
@@ -120,6 +120,9 @@ The following includes brief examples for *some* of the methods in **BGGM**.
 An ordinal GGM is estimated with
 
 ```{r, eval = FALSE}
+library(BGGM)
+library(ggplot2)
+
 # data
 Y <- ptsd[,1:5] + 1
 
@@ -220,8 +223,11 @@ The Bayes factor based methods allow for determining the conditional
 #### Exploratory
 
 ```{r, eval=FALSE}
+# now 10 nodes
+Y <- ptsd[,1:10]
+
 # exploratory hypothesis testing
-fit<- explore(Y, type = "ordinal")
+fit<- explore(Y)
 
 # select 
 E <- select(fit, alternative = "exhaustive")
