### 2020-05-21 ccb14a2
readme: update information about the other packages
