+++ README.Rmd
@@ -871,4 +871,8 @@ data may be added to **BGGM**.
 ## Bug Reports and Feature Requests
 Bug reports and feature requests can be made by opening an issue on [Github](https://github.com/donaldRwilliams/BGGM/issues).
 
+## Comparsion to Other Software
+**BGGM** is the only `R` package to implement all of these algorithms and methods. The `mixed` data approach 
+is also implemeted in the package **sbgcop** [base `R`, @hoff2007extending]. The `R` package **BDgraph** implements a Gaussian copula graphical model in `c++` [@mohammadi2015bdgraph], but not the binary or ordinal approaches. Futhermore, **BGGM** is the only package for confirmatory testing and comparing graphical models with the methods described in @williams2020comparing.
+
 ## References
