### 2020-05-21 donaldRwilliams/BGGM#3 @jayrobwilliams
It's not immediately clear from the updated README and paper what the main use and targeted audience is for the package. You can figure this out by reading through the examples, but there should be an explicit setence(s) that spell out what types of problems in which fields this package is intended for.
