### 2020-05-21 donaldRwilliams/BGGM#3 @donaldRwilliams
Ok. I will add this to the paper and Readme
later today.
