### 2020-05-21 donaldRwilliams/BGGM#6 @jayrobwilliams
The README tells users where they can submit a bug report or feature request, but the review checklist includes information for people who wish to "Contribute to the software." Many projects include a `CONTRIBUTING` file in the repository. Here is the `ggplot2` one as an example: https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md. While this is an especially involved document due to the size of the project, a `CONTRIBUTING` file should list etiquette for submitting pull requests and list any coding conventions the project uses.
