### 2020-05-21 donaldRwilliams/BGGM#3 @donaldRwilliams
To follow up on this @jayrobwilliams , I am having a difficult time coming up with a target audience that does not also limit the audience. Really any one that would want to estimate a GGM can use the methods, although it was developed by someone in the social behavioral sciences, if this makes sense.

Anyhow, open to suggestions.
