### 2020-05-21 36c16e1
Update and rename CONTRIBUTING to CONTRIBUTING.md
