### 2020-05-21 cda6abf
Delete CONTRIBUTING.md
