+++ README.Rmd
@@ -869,10 +869,13 @@ These data are rare in the social-behavioral sciences. In the future, support fo
 data may be added to **BGGM**.
 
 ## Bug Reports and Feature Requests
-Bug reports and feature requests can be made by opening an issue on [Github](https://github.com/donaldRwilliams/BGGM/issues).
+Bug reports and feature requests can be made by opening an issue on [Github](https://github.com/donaldRwilliams/BGGM/issues). To contribute towards
+the development of **BGGM**, you can start a branch with a pull request and we can 
+discuss the proposed changes there. All code *must* be written in base `R`, `c ++`, 
+or `fortran` (preferably `c ++`). 
 
-## Comparsion to Other Software
+## Comparison to Other Software
 **BGGM** is the only `R` package to implement all of these algorithms and methods. The `mixed` data approach 
-is also implemeted in the package **sbgcop** [base `R`, @hoff2007extending]. The `R` package **BDgraph** implements a Gaussian copula graphical model in `c++` [@mohammadi2015bdgraph], but not the binary or ordinal approaches. Futhermore, **BGGM** is the only package for confirmatory testing and comparing graphical models with the methods described in @williams2020comparing.
+is also implemented in the package **sbgcop** [base `R`, @hoff2007extending]. The `R` package **BDgraph** implements a Gaussian copula graphical model in `c++` [@mohammadi2015bdgraph], but not the binary or ordinal approaches. Furthermore, **BGGM** is the only package for confirmatory testing and comparing graphical models with the methods described in @williams2020comparing.
 
 ## References
