### 2020-05-21 donaldRwilliams/BGGM#6 @donaldRwilliams
I decided against a contributing file, although I now include loose guidelines for those that want to contribute where I note the code convention (not sure I really have one, however, other than to only use base R)
