### 2020-05-21 d9313ee
README: update the bug report, feature request, and contributing section
