+++ README.md
@@ -966,7 +966,7 @@ than observations) that are common place in, say, the genetics
 literature. These data are rare in the social-behavioral sciences. In
 the future, support for high-dimensional data may be added to **BGGM**.
 
-## Bug Reports and Feature Requests
+## Bug Reports, Feature Requests, and Contributing
 
 Bug reports and feature requests can be made by opening an issue on
 [Github](https://github.com/donaldRwilliams/BGGM/issues). To contribute
