+++ README.Rmd
@@ -49,7 +49,7 @@ if (!requireNamespace("remotes")) {
 remotes::install_github("donaldRwilliams/BGGM")
 ```
 
-Note that the developmental version is recommended, due to the algorithms being written in `c++` and 
+Note that the developmental version is recommended (and required to run the following examples), due to the algorithms being written in `c++` and 
 support for all data types. The developmental version is essentially
 **BGGM** version 2.0.0.
 
