### 2020-05-21 8130734
fix typos in README
