### 2020-05-22 bb18e14
website: add bayes testing to netplot
