+++ vignettes/netplot.Rmd
@@ -25,6 +25,7 @@ if (!requireNamespace("remotes")) {
 # install from github
 remotes::install_github("donaldRwilliams/BGGM")
 library(BGGM)
+library(cowplot)
 ```
 ```{r, echo=FALSE, message=FALSE}
 library(BGGM)
@@ -129,13 +130,56 @@ plot(E,
 ```
 ![](../man/figures/netplot_3.png)
 
-# Hypothesis Testing
-Forthcoming
+# Bayesian Hypothesis Testing
+The Bayesian hypothesis testing methods offer several advantages, for example, that 
+evidence for the null hypothesis of conditional independence is formally evaluated. 
+As a result, the `explore` method in **BGGM** provides plots for both the conditional 
+dependence and independence structure, in addition to a plot for which the evidence was 
+ambiguous.
+
+To highlight this advantage, `ptsd` data is used that has a relatively small sample size.
+
+```r
+# fit model
+fit <- explore(Y)
+
+E <- select(fit, BF_cut = 3)
+```
+
+Then plot the results. Note that there are three plots, so the package **cowplot** is used
+to combine them into one plot.
+
+```r
+plts <- plot(E, 
+             edge_magnify = 5, 
+             groups = comm)
+
+plot_grid(
+plts$H1_plt + 
+  ggtitle("Conditional Dependence") + 
+  theme(legend.position =  "none"),
+
+plts$H0_plt + 
+  ggtitle("Conditional Independence") + 
+  theme(legend.position =  "none"),
+plts$ambiguous_plt + 
+  ggtitle("Ambiguous"), 
+nrow = 1,
+rel_widths = c(1, 1, 1.1)
+)
+```
+
+![](../man/figures/hyp_plot.png)
+
+As can be seen, there is not evidence for conditional independence for any of the relations. And
+the ambiguous network makes clear there is large uncertainty as to what or what might not be the "true" network structure. This basic idea of having three adjacency matrices was proposed in @Williams2019_bf.
+
 
 # Note
-**BGGM** provides a publication ready plot, but it is also limited compared to **qgraph** [@epskamp2012qgraph]. The one advantage of **BGGM**  is that all plots are `ggplots` which then
-allows for combining them rather easily. An example is included in another vignette that shows how to 
-combine several plots made with various methods in **BGGM**
+**BGGM** provides a publication ready plot, but it is also limited compared to **qgraph** 
+[@epskamp2012qgraph]. The one advantage of **BGGM**  is that all plots are `ggplots` 
+which then allows for combining them rather easily. An example is included in another 
+vignette that shows how to combine several plots made with various methods in **BGGM**
 
 
 # References
