+++ index.md
@@ -13,13 +13,13 @@ Status](https://travis-ci.org/donaldRwilliams/BGGM.svg?branch=master)](https://t
 <!-- badges: end -->
 
 The `R` package **BGGM** provides tools for making Bayesian inference in
-Gaussian graphical models (GGM). The methods are organized around two
-general approaches for Bayesian inference: (1) estimation and (2)
-hypothesis testing. The key distinction is that the former focuses on
-either the posterior or posterior predictive distribution (Gelman, Meng,
-and Stern 1996; see section 5 in Rubin 1984), whereas the latter focuses
-on model comparison with the Bayes factor (Jeffreys 1961; Kass and
-Raftery 1995).
+Gaussian graphical models (GGM, Donald R Williams and Mulder 2020). The
+methods are organized around two general approaches for Bayesian
+inference: (1) estimation and (2) hypothesis testing. The key
+distinction is that the former focuses on either the posterior or
+posterior predictive distribution (Gelman, Meng, and Stern 1996; see
+section 5 in Rubin 1984), whereas the latter focuses on model comparison
+with the Bayes factor (Jeffreys 1961; Kass and Raftery 1995).
 
 ## Installation
 
@@ -56,29 +56,38 @@ extend those approaches:
 2.  Bayesian hypothesis testing with the matrix-F prior distribution
     (Williams and Mulder 2019)
     
-      - Exploratory hypothesis testing
+      - [Exploratory hypothesis
+        testing](https://github.com/donaldRwilliams/BGGM#Exploratory)
     
-      - Confirmatory hypothesis testing
+      - [Confirmatory hypothesis
+        testing](https://github.com/donaldRwilliams/BGGM#confirmatory)
 
-3.  Comparing Gaussian graphical models (Williams 2018; Williams et al.
-    2020)
+3.  Comparing Gaussian graphical models (Williams 2018; Donald R.
+    Williams et al. 2020)
     
-      - Partial correlation differences
+      - [Partial correlation
+        differences](https://github.com/donaldRwilliams/BGGM#partial-correlation-differences)
     
-      - Posterior predictive check
+      - [Posterior predictive
+        check](https://github.com/donaldRwilliams/BGGM#posterior-predictive-check)
     
-      - Exploratory hypothesis testing
+      - [Exploratory hypothesis
+        testing](https://github.com/donaldRwilliams/BGGM#exploratory-groups)
     
-      - Confirmatory hypothesis testing
+      - [Confirmatory hypothesis
+        testing](https://github.com/donaldRwilliams/BGGM#confirmatory-groups)
 
 4.  Extending inference beyond the conditional (in)dependence structure
     (Williams 2018)
     
-      - Predictability
+      - [Predictability](https://github.com/donaldRwilliams/BGGM#predictability)
     
-      - Posterior uncertainty intervals for the partial correlations
+      - [Posterior uncertainty
+        intervals](https://github.com/donaldRwilliams/BGGM#posterior-uncertainty)
+        for the partial correlations
     
-      - Custom Network Statistics
+      - [Custom Network
+        Statistics](https://github.com/donaldRwilliams/BGGM#custom-network-statistics)
 
 The computationally intensive tasks are written in `c++` via the `R`
[…]
