+++ index.Rmd
@@ -67,28 +67,28 @@ The central contribution of **BGGM** is to extend those approaches:
 
 2. Bayesian hypothesis testing with the matrix-F prior distribution [@Williams2019_bf]
 
-    + Exploratory hypothesis testing
+    + [Exploratory hypothesis testing](https://github.com/donaldRwilliams/BGGM#Exploratory)
   
-    + Confirmatory hypothesis testing
+    + [Confirmatory hypothesis testing](https://github.com/donaldRwilliams/BGGM#confirmatory)
     
 3. Comparing Gaussian graphical models [@Williams2019; @williams2020comparing]
     
-    + Partial correlation differences
+    + [Partial correlation differences](https://github.com/donaldRwilliams/BGGM#partial-correlation-differences)
     
-    + Posterior predictive check
+    + [Posterior predictive check](https://github.com/donaldRwilliams/BGGM#posterior-predictive-check)
     
-    + Exploratory hypothesis testing
+    + [Exploratory hypothesis testing](https://github.com/donaldRwilliams/BGGM#exploratory-groups)
     
-    + Confirmatory hypothesis testing
+    + [Confirmatory hypothesis testing](https://github.com/donaldRwilliams/BGGM#confirmatory-groups)
 
 4. Extending inference beyond the conditional (in)dependence structure [@Williams2019]
 
-    +  Predictability 
+    +  [Predictability](https://github.com/donaldRwilliams/BGGM#predictability)
     
-    +  Posterior uncertainty intervals for the 
+    +  [Posterior uncertainty intervals](https://github.com/donaldRwilliams/BGGM#posterior-uncertainty) for the 
        partial correlations
        
-    +  Custom Network Statistics
+    +  [Custom Network Statistics](https://github.com/donaldRwilliams/BGGM#custom-network-statistics)
     
     
 The computationally intensive tasks are written in `c++` via the `R` package **Rcpp** [@eddelbuettel2011rcpp] and the `c++` library **Armadillo** [@sanderson2016armadillo]. The Bayes factors are computed with the `R` package **BFpack** [@mulder2019bfpack]. Furthermore, there are plotting functions
@@ -116,5 +116,94 @@ and there is support for missing values (see `bggm_missing`).
 
 ## Illustrative Examples
 There are several vignettes in the [Articles](https://donaldrwilliams.github.io/BGGM/articles/) section.
+
+## Basic Usage
+It is common to have some combination of continuous and discrete (e.g., ordinal, binary, etc.) variables. **BGGM** (as of version `2.0.0`) can readily be used for these kinds of data. In this example, a model is fitted for the `gss` data in **BGGM**. 
+
+### Visualize
+The data are first visualized with the **psych** package
+```r
+library(psych)
+
+# data
+Y <- gss
+
+# histogram for each node
+psych::multi.hist(Y, density = FALSE)
+```
+![](man/figures/index_hist.png)
+which readily shows the data is "mixed". 
+
+### Fit Model
+A Gaussian copula graphical model is estimated as follows
+```r
+fit <- estimate(Y, type = "mixed")
+```
+`type` can be `continuous`, `binary`, `ordinal`, or `mixed`. Note that `type` is a misnomer, as the data can consist of *only* ordinal variables (for example).
+
+
[…]
