+++ NEWS.md
@@ -1,1 +1,5 @@
 # BGGM (developmental version)
+
+* Highlighting of empty expressions was restored (#1310)
+
+datra
