+++ index.Rmd
@@ -115,13 +115,14 @@ and there is support for missing values (see `bggm_missing`).
 
 
 ## Illustrative Examples
-There are several vignettes in the [Articles](https://donaldrwilliams.github.io/BGGM/articles/) section.
+There are several vignettes in the [Vignettes](https://donaldrwilliams.github.io/BGGM/articles/) section.
 
 ## Basic Usage
 It is common to have some combination of continuous and discrete (e.g., ordinal, binary, etc.) variables. **BGGM** (as of version `2.0.0`) can readily be used for these kinds of data. In this example, a model is fitted for the `gss` data in **BGGM**. 
 
 ### Visualize
 The data are first visualized with the **psych** package, which readily shows the data are "mixed". 
+
 ```r
 # dev version
 library(BGGM)
