+++ index.md
@@ -1,7 +1,7 @@
 
-# Bayesian Gaussian Graphical Models
+# <i class="fas fa-headset"></i> Bayesian Gaussian Graphical Models
 
-<!-- <img src="man/figures/hex.png" align="right" alt="" width="150" /> -->
+\< img src=“man/figures/logo.png” align=“right” alt="" width=“150” /\>
 
 <!-- badges: start -->
 
@@ -42,7 +42,7 @@ Note that the developmental version is recommended, due to the
 algorithms being written in `c++` and support for all data types. The
 developmental version is essentially **BGGM** version 2.0.0.
 
-## Overview
+## <i class="fas fa-terminal"></i> Overview
 
 The methods in **BGGM** build upon existing algorithms that are
 well-known in the literature. The central contribution of **BGGM** is to
@@ -120,7 +120,7 @@ support for missing values (see `bggm_missing`).
 
 ## Illustrative Examples
 
-There are several vignettes in the
+There are several examples in the
 [Vignettes](https://donaldrwilliams.github.io/BGGM/articles/) section.
 
 ## Basic Usage
