+++ index.Rmd
@@ -17,7 +17,7 @@ library(ggplot2)
 library(BGGM)
 ```
 
-# <i class="fas fa-headset"></i> Bayesian Gaussian Graphical Models <img src="man/figures/logo.png" align="right" alt="" width="150" />
+# <i class="fas fa-headphones"></i> Bayesian Gaussian Graphical Models <img src="man/figures/logo.png" align="right" alt="" width="150" />
 
 
 <!-- badges: start -->
@@ -56,7 +56,7 @@ support for all data types. The developmental version is essentially
 
 
 
-##  Overview
+## <i class="fas fa-clipboard-list"></i> Overview
 The methods in **BGGM** build upon existing algorithms that are well-known in the literature.
 The central contribution of **BGGM** is to extend those approaches:
 
