+++ index.Rmd
@@ -94,7 +94,7 @@ The computationally intensive tasks are written in `c++` via the `R` package **R
 for each method, control variables can be included in the model (e.g., `~ gender`), 
 and there is support for missing values (see `bggm_missing`).
 
-## Supported Data Types
+## <i class="fas fa-clipboard-list"></i> Supported Data Types
 
 * **Continuous**: The continuous method was described in  @Williams2019. Note that 
                   this is based on the customary [Wishart distribution](https://en.wikipedia.org/wiki/Wishart_distribution).
@@ -113,13 +113,13 @@ and there is support for missing values (see `bggm_missing`).
  *only* ordinal data (not restricted to "mixed" data).
 
 
-## Illustrative Examples
+##  <i class="fas fa-external-link-square-alt"></i> Illustrative Examples
 There are several examples in the [Vignettes](https://donaldrwilliams.github.io/BGGM/articles/) section.
 
 ## <i class="far fa-play-circle"></i>  Basic Usage
 It is common to have some combination of continuous and discrete (e.g., ordinal, binary, etc.) variables. **BGGM** (as of version `2.0.0`) can readily be used for these kinds of data. In this example, a model is fitted for the `gss` data in **BGGM**. 
 
-### Visualize
+### <i class="fas fa-terminal"></i> Visualize
 The data are first visualized with the **psych** package, which readily shows the data are "mixed". 
 
 ```r
@@ -138,7 +138,7 @@ psych::multi.hist(Y, density = FALSE)
 
 
 
-### Fit Model
+### <i class="fas fa-terminal"></i> Fit Model
 A Gaussian copula graphical model is estimated as follows
 ```r
 fit <- estimate(Y, type = "mixed")
@@ -146,7 +146,7 @@ fit <- estimate(Y, type = "mixed")
 `type` can be `continuous`, `binary`, `ordinal`, or `mixed`. Note that `type` is a misnomer, as the data can consist of *only* ordinal variables (for example).
 
 
-### Summarize Relations
+### <i class="fas fa-terminal"></i> Summarize Relations
 The estimated relations are summarized with
 ```r
 summary(fit)
@@ -196,7 +196,7 @@ plot(summary(fit))
 ```
 ![](man/figures/index_summ.png)
 
-### Graph Selection
+### <i class="fas fa-terminal"></i> Graph Selection
 The graph is selected and plotted with
 ```r
 E <- select(fit)
@@ -210,4 +210,4 @@ plot(E, node_size = 12,
 
 The Bayes factor testing approach is readily implemented by changing `estimate` to `explore`. 
 
-## References
+## <i class="fas fa-pen-square"></i> References
