### 2020-05-23 968891e
website: finish icons (removed as too many)
