+++ _pkgdown.yml
@@ -23,7 +23,7 @@ navbar:
       href: reference/index.html
     - text: "News"
       href: news/index.html
-
+    - text: "dplyr"
 
 articles:
   - title: "Getting Started"
@@ -35,3 +35,5 @@ articles:
       - netstat_custom
       - ppc_custom
       - predictability
+
+
