### 2020-05-23 65e1b6d
bggm_missing: remove continuous and warn.conflicts
