### 2020-05-23 204eaeb
documentation: update coef
