+++ R/coef.estimate.R
@@ -31,43 +31,16 @@
 #' \donttest{
 #' # note: iter = 250 for demonstrative purposes
 #'
-#' data
-#' Y <- ptsd
-#' #########################################
-#' ### example 1: continuous and ordinal ###
-#' #########################################
-#'
-#' # fit model
-#' fit <- estimate(Y, type = "continuous",
-#'                iter = 250)
-#'
-#' # regression
-#' reg <- coef(fit)
-#'
-#' # summary
-#' summ <- summary(reg)
-#'
-#' # ordinal
-#'
-#' # fit model (note + 1, due to zeros)
-#' fit <- estimate(Y + 1, type = "ordinal",
-#'                 iter = 250)
-#'
-#' # summarize the partial correlations
-#' reg <- coef(fit)
-#'
-#' # summary
-#' summ <- summary(reg)
-#'
 #' #########################
-#' ### example 2: binary ###
+#' ### example 1: binary ###
 #' #########################
 #' # data
-#' Y <- women_math
+#' Y <- women_math[1:500, ]
 #'
 #' # fit model
 #' fit <- estimate(Y, type = "binary",
-#'                 iter = 250)
+#'                 iter = 250,
+#'                 progress = FALSE)
 #'
 #' # summarize the partial correlations
 #' reg <- coef(fit)
@@ -75,6 +48,7 @@
 #' # summary
 #' summ <- summary(reg)
 #'
+#' summ
 #'}
 #' @export
 coef.estimate <- function(object, iter = NULL,...) {
@@ -169,52 +143,26 @@ coef.estimate <- function(object, iter = NULL,...) {
 #' \donttest{
 #' # note: iter = 250 for demonstrative purposes
 #'
-#' data
-#' Y <- ptsd
-#' #########################################
-#' ### example 1: continuous and ordinal ###
-#' #########################################
+#' # data
+#' Y <- ptsd[,1:4]
 #'
-#' # fit model
-#' fit <- explore(Y, type = "continuous",
-#'                iter = 250)
+#' ##########################
+#' ### example 1: ordinal ###
+#' ##########################
 #'
-#' # regression
+#' # fit model (note + 1, due to zeros)
+#' fit <- explore(Y + 1,
+#'                type = "ordinal",
+#'                iter = 250,
+#'                progress = FALSE)
+#'
+#' # summarize the partial correlations
 #' reg <- coef(fit)
 #'
 #' # summary
 #' summ <- summary(reg)
 #'
 #' summ
-#'
-#' # ordinal
-#'
-#' # fit model (note + 1, due to zeros)
-#' fit <- explore(Y + 1, type = "ordinal",
-#'                iter = 250)
-#'
-#' # summarize the partial correlations
-#' reg <- coef(fit)
-#'
-#' # summary
-#' summ <- summary(reg)
-#'
-#' #########################
-#' ### example 2: binary ###
-#' #########################
-#' # data
-#' Y <- women_math
-#'
-#' # fit model
-#' fit <- explore(Y, type = "binary",
-#'                iter = 250)
-#'
-#' # summarize the partial correlations
-#' reg <- coef(fit)
-#'
-#' # summary
-#' summ <- summary(reg)
-#'
 #'}
 #' @export
 coef.explore <- function(object,
