### 2020-05-23 7359402
feature: added progress bar to coef
