+++ man/coef.estimate.Rd
@@ -48,14 +48,12 @@ fit <- estimate(Y, type = "binary",
                 progress = FALSE)
 
 # summarize the partial correlations
-
-reg <- coef(fit)
+reg <- coef(fit, progress = FALSE)
 
 # summary
 summ <- summary(reg)
 
 summ
-
 }
 }
 \references{
