### 2020-05-23 d81cbc3
three ways to test hyp vignette
