+++ vignettes/hyp_3_ways.Rmd
@@ -7,7 +7,7 @@ output:
 rmarkdown::html_vignette:
   toc: yes
 vignette: >
-  %\VignetteIndexEntry{hree Ways to Test the Same Hypothesis}
+  %\VignetteIndexEntry{Three Ways to Test the Same Hypothesis}
   %\VignetteEngine{knitr::rmarkdown}
   \usepackage[utf8]{inputenc}
 ---
@@ -15,9 +15,9 @@ vignette: >
 # Introduction
 On a Facebook methods group, there was a question about testing hypotheses in network model. In 
 the comments, it was suggested that **BGGM** could be used to test the hypothesis. And it turns
-out that **BGGM** really shines for testing expections or predictions. In this vignette, I demonstrate 
+out that **BGGM** really shines for testing exceptions or predictions. In this vignette, I demonstrate 
 three ways to go about testing the same hypothesis, which was essentially testing for a difference
-in the **sum** of partrial correlations between groups.
+in the **sum** of partial correlations between groups.
 
 ### R package
 ```{r, eval = FALSE, message=FALSE}
@@ -46,4 +46,270 @@ Y_females <- subset(Y, gender == 2, select = -c(education, gender))[,1:5]
 ```
 
 # Approach 1: Posterior Difference
-The first approach
+The first approach is rather straightforward, with the caveat that the method needs to be implemented by the user
+Note that I could certainly implement this in **BGGM**, assuming there is enough interest. Please make a feature request here.
+
+## Hypothesis
+The hypothesis was that a sum of relations was larger in one group, for example,
+
+$$
+\begin{align}
+\mathcal{H}_0: (\rho^{male}_{A1--A2}\; + \; \rho^{female}_{A1--A3}) = (\rho^{female}_{A1--A2}\; + \; \rho^{female}_{A1--A3}) \\
+\mathcal{H}_1: (\rho^{male}_{A1--A2}\; + \; \rho^{female}_{A1--A3}) > (\rho^{female}_{A1--A2}\; + \; \rho^{female}_{A1--A3})
+\end{align}
+$$
+Note that the hypothesis is related to the sum of relations, which is readily tested in **BGGM**. 
+
+## Fit Models
+The first step is to estimate the model for each group
+
+```r
+# fit female
+fit_female <- estimate(Y_females, seed = 2)
+
+# fit males
+fit_male <- estimate(Y_males, seed = 1)
+```
+For an example, I used the default which is to assume the data is Gaussian. This can be chaged with `type = ` either `binary`, `ordinal`, or `mixed`.
+
+## Extract the Samples
+The next step is to extract the posterior samples for each relation
+
+```r
+post_male <- posterior_samples(fit_male)[,c("A1--A2", "A1--A3")]
+
+post_female <- posterior_samples(fit_female)[,c("A1--A2", "A1--A3")]
+```
+
+## Sum and Compute Difference
+The next step is to sum the relations and compute the difference
+
+```r
+# sum males
+sum_male <- rowSums(post_male) 
+
+# sum females
+sum_female <- rowSums(post_female)
+
+# difference
+diff <- sum_male - sum_female
+```
+which can then be plotted
+```r
+par(mfrow=c(1,3))
+hist(sum_male)
+hist(sum_female)
+hist(diff)
+```
+![](../man/figures/hyp_3ways_hist.png)
+
+
+
+
[…]
