### 2020-05-23 997e25b
website: fix typos
