### 2020-05-23 71694d9
vignette: fix typos in three hyps
