### 2020-05-23 4c8acbd
vignette : fix typos 3 hypotheses
