### 2020-05-23 900f738
vignette: fix typos
