### 2020-05-23 f305255
vigenette: correct typos
