### 2020-05-23 ab139cc
vignette: fix 3 ways typos and add the info about other data types
