+++ man/convergence.Rd
@@ -27,21 +27,19 @@ Monitor convergence of the MCMC algorithms.
 An overview of MCMC diagnostics can be found \href{http://sbfnk.github.io/mfiidd/mcmc_diagnostics.html}{here}.
 }
 \examples{
-
 \donttest{
 # note: iter = 250 for demonstrative purposes
 
 # data
-Y <- ptsd
+Y <- ptsd[,1:5]
 
 #########################
-######## ordinal ########
+###### continuous #######
 #########################
-fit <- estimate(Y + 1, type = "ordinal",
-                iter = 250)
+fit <- estimate(Y, iter = 250)
 
 # print names first
-param_names <- convergence(fit, print_names = TRUE)
+convergence(fit, print_names = TRUE)
 
 # trace plots
 convergence(fit, type = "trace",
