### 2020-05-24 7fa20d3
website: add troubleshoot page
