+++ index.Rmd
@@ -17,7 +17,7 @@ library(ggplot2)
 library(BGGM)
 ```
 
-# <i class="fas fa-headphones"></i> Bayesian Gaussian Graphical Models <img src="man/figures/logo.png" align="right" alt="" width="150" />
+# Bayesian Gaussian Graphical Models <img src="man/figures/logo.png" align="right" alt="" width="150" />
 
 
 <!-- badges: start -->
@@ -54,7 +54,11 @@ Note that the developmental version is recommended, due to the algorithms being 
 support for all data types. The developmental version is essentially
 **BGGM** version 2.0.0.
 
-
+### <i class="fas fa-skull-crossbones"></i> Dealing with Errors
+There are automatic checks for **BGGM**. This can be seen in the right panel "build". That badge is linked
+to the build history for **BGGM**. However, that only checks for Linux and **BGGM** is built on Windows. 
+The most common installation errors occur on osx (Mac). An evolving guide to address these
+issues is provided in the [Troubleshoot Section](https://donaldrwilliams.github.io/BGGM/articles/installation.html). 
 
 ## <i class="fas fa-clipboard-list"></i> Overview
 The methods in **BGGM** build upon existing algorithms that are well-known in the literature.
