+++ _pkgdown.yml
@@ -38,5 +38,4 @@ articles:
       - ppc_custom
       - predictability
       - hyp_3_ways
-
-
+      - in_tandem
