### 2020-05-24 2a6cff9
website: update troubleshoot
