+++ _pkgdown.yml
@@ -41,6 +41,110 @@ articles:
       - in_tandem
 
 reference:
-- title: Missing Data
-  decs: Handle Missing Values.
-- contents: bggm_missing
+  - title: "Missing Data"
+    desc: >
+     Handle Missing Values.
+    contents:
+    - bggm_missing
+  - title: "Estimation Based Methods"
+    contents:
+    - estimate
+    - coef.estimate
+    - predict.estimate
+    - plot.summary.estimate
+    - select.estimate
+    - summary.estimate
+  - title: "Exploratory Hypothesis Testing"
+    contents:
+    - explore
+    - coef.explore
+    - predict.explore
+    - plot.summary.explore
+    - select.explore
+    - summary.explore
+  - title: "Confirmatory Hypothesis Testing"
+    contents:
+    - confirm
+    - plot.confirm
+  - title: "Compare Gaussian Graphical Models"
+  - subtitle: "Posterior Predictive Check"
+    desc: >
+     Compare groups with a posterior predictive check, where the null model
+     is that the groups are equal. This works with any number of groups. There
+     is also an option to compare the groups with a user defined test-statistic.
+    contents:
+    - ggm_compare_ppc
+    - plot.ggm_compare_ppc
+  - subtitle: "Partial Correlation Differences"
+    desc: >
+     Pairwise comparisons for each partial correlation in the respective models.
+     This can be used for any number of groups. There is also an analytical solution.
+    contents:
+    - ggm_compare_estimate
+    - plot.summary.ggm_compare_estimate
+    - select.ggm_compare_estimate
+    - summary.ggm_compare_estimate
+  - subtitle: "Exploratory Hypothesis Testing"
+    desc: >
+      Pairwise comparisions with exploratory hypothesis testing. This method can be used to
+      compare several groups simultaneously.
+    contents:
+    - ggm_compare_explore
+    - plot.summary.ggm_compare_explore
+    - select.ggm_compare_explore
+    - summary.ggm_compare_explore
+  - subtitle: "Confirmatory Hypothesis Testing"
+    desc: >
+      Confirmatory hypothesis testing for groups.
+    contents:
+    - ggm_compare_confirm
+    - plot.confirm
+  - title: "Predictability"
+    contents:
+    - predictability
+    - plot.predictability
+    - summary.predictability
+  - title: "Network Statistics"
+    contents:
+    - roll_your_own
+    - plot.roll_your_own
+  - title: "Network Plot"
+    desc: >
+     Network plot for the selected graphs. This works with all method for which there is
+     a selected graph.
+    contents:
+    - plot.select
+  - title: Miscellaneous
+    contents:
+    - convergence
+    - fisher_z_to_r
+    - fisher_r_to_z
+    - gen_ordinal
+    - pcor_to_cor
+    - pcor_mat
+    - plot_prior
+    - posterior_samples
+    - regression_summary
+    - map
+    - weighted_adj_mat
+    - zero_order_cors
+  - title: "Data"
+    desc: >
+     Example dataests and correlation matrices.
+    contents:
+    - asd_ocd
+    - bfi
+    - depression_anxiety_t1
+    - depression_anxiety_t2
+    - gss
+    - iri
+    - ptsd
+    - ptsd_cor1
+    - ptsd_cor2
+    - ptsd_cor3
+    - ptsd_cor4
+    - rsa
+    - Sachs
+    - tas
+    - women_math
+
