### 2020-05-25 c336af6
new feature: pcor_sum
