+++ R/posterior_sum.R
@@ -1,0 +1,340 @@
+#' @title Partial Correlation Sum
+#'
+#' @name pcor_sum
+#'
+#' @description Compute and test partial correlation sums either within or between GGMs
+#' (e.g., different groups), resulting in a posterior distribution.
+#'
+#' @param ... An object of class \code{estimate}. This can be either one or two fitted objects.
+#'
+#' @param relations Character string. Which partial correlations should be summed?
+#'
+#' @param iter Number of iterations (posterior samples; defaults to the number in the object).
+#'
+#' @return An object of class \code{posterior_sum}, including the sum and possibly the difference for
+#' two sums.
+#'
+#' @details
+#' Some care must be taken when writing the string for \code{partial_sum}. Below are several examples
+#'
+#' \strong{Just a Sum}:
+#' Perhaps a sum is of interest, and not necessarily the difference of two sums. This can be written as
+#'
+#' \itemize{
+#' \item \code{partial_sum <-  c("A1--A2 + A1--A3 + A1--A4")}
+#' }
+#'
+#' which will sum those relations.
+#'
+#' \strong{Comparing Sums}:
+#' When comparing sums, each must be seperated by "\code{;}". For example,
+#'
+#' \itemize{
+#' \item \code{partial_sum <-  c("A1--A2 + A1--A3; A1--A2 + A1--A4")}
+#' }
+#'
+#' which will sum both and compute the difference. Note that there cannot be more than two sums, such
+#' that \code{c("A1--A2 + A1--A3; A1--A2 + A1--A4; A1--A2 + A1--A5")} will result in an error.
+#'
+#' \strong{Comparing Groups}:
+#'
+#' When more than one fitted object is suppled to \code{object} it is assumed that the groups
+#' should be compared for the same sum. Hence, in this case, only the sum needs to be written.
+#'
+#' \itemize{
+#' \item \code{partial_sum <-  c("A1--A2 + A1--A3 + A1--A4")}
+#' }
+#'
+#' The above results in that sum being computed for each group and then compared.
+#'
+#' @export
+#'
+#' @examples
+#' \donttest{
+#' # data
+#' Y <- bfi
+#'
+#' # males
+#' Y_males <- subset(Y, gender == 1, select = -c(education, gender))[,1:5]
+#'
+#' # females
+#' Y_females <- subset(Y, gender == 2, select = -c(education, gender))[,1:5]
+#'
+#' # males
+#' fit_males <- estimate(Y_males, seed = 1)
+#'
+#' # fit females
+#' fit_females <- estimate(Y_females, seed = 2)
+#'
+#'
+#' sums <- pcor_sum(fit_males,
+#'                  fit_females,
+#'                  relations = "A1--A2 + A1--A3")
+#' # print
+#' sums
+#'
+#' # plot difference
+#' plot(sums)[[3]]
+#' }
+pcor_sum <- function(..., iter = NULL, relations){
+
+  # collect ...
+  collect_objects <- list(...)
+
+  # number of groups
+  groups <- length(collect_objects)
+
+  # partial_sum_i
+  partial_sum_i <- list()
+
+  if(is.null(iter)){
+    iter <- collect_objects[[1]]$iter
+  }
+
+  # separate to count
+  count_sums <- strsplit(relations, "\\;")[[1]]
+
+  # how many sums ?
+  n_sums <- length(count_sums)
+
+  remove_space <- gsub("[[:space:]]", "", count_sums)
+
+  remove_plus <- gsub("\\+", replacement = " ",  remove_space)
+
+  each_sum <- strsplit(remove_plus, split = "[[:space:]]")
+
+  if (n_sums > 2) {
[…]
