+++ NAMESPACE
@@ -4,6 +4,7 @@ S3method(coef,estimate)
 S3method(coef,explore)
 S3method(plot,confirm)
 S3method(plot,ggm_compare_ppc)
+S3method(plot,pcor_sum)
 S3method(plot,predictability)
 S3method(plot,roll_your_own)
 S3method(plot,select)
@@ -40,6 +41,7 @@ export(ggm_compare_explore)
 export(ggm_compare_ppc)
 export(map)
 export(pcor_mat)
+export(pcor_sum)
 export(pcor_to_cor)
 export(plot_prior)
 export(posterior_samples)
