+++ man/plot.pcor_sum.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/posterior_sum.R
 \name{plot.pcor_sum}
 \alias{plot.pcor_sum}
-\title{Plot \code{posterior_sum} Object}
+\title{Plot \code{pcor_sum} Object}
 \usage{
 \method{plot}{pcor_sum}(x, fill = "#CC79A7", ...)
 }
@@ -18,7 +18,7 @@
 A list of \code{ggplot} objects
 }
 \description{
-Plot \code{posterior_sum} Object
+Plot \code{pcor_sum} Object
 }
 \note{
 \strong{Examples}:
