### 2020-05-26 548213b
webiste: fix typos in testing sums
