### 2020-05-26 e33a861
website: fix typos in testing sums
