+++ man/plot.predictability.Rd
@@ -49,11 +49,13 @@ Plot \code{predictability} Objects
 }
 \examples{
 \donttest{
-Y <- ptsd
+Y <- ptsd[,1:5]
 
-fit <- explore(Y, iter = 250)
+fit <- explore(Y, iter = 250,
+               progress = FALSE)
 
-r2 <- predictability(fit, iter = 250)
+r2 <- predictability(fit, iter = 250,
+                     progress = FALSE)
 
 plot(r2)
 }
