### 2020-05-26 35f589d
feature: add progress bar to predictability
