+++ R/pcor_mat.R
@@ -26,7 +26,8 @@
 #'
 #' # ordinal
 #' fit <- estimate(Y, type = "ordinal",
-#'                 iter = 250)
+#'                 iter = 250,
+#'                 progress = FALSE)
 #'
 #' pcor_mat(fit)
 #' }
