+++ R/posterior_sum.R
@@ -61,10 +61,12 @@
 #' Y_females <- subset(Y, gender == 2, select = -c(education, gender))[,1:5]
 #'
 #' # males
-#' fit_males <- estimate(Y_males, seed = 1)
+#' fit_males <- estimate(Y_males, seed = 1,
+#'                       progress = FALSE)
 #'
 #' # fit females
-#' fit_females <- estimate(Y_females, seed = 2)
+#' fit_females <- estimate(Y_females, seed = 2,
+#'                         progress = FALSE)
 #'
 #'
 #' sums <- pcor_sum(fit_males,
