+++ man/zero_order_cors.Rd
@@ -4,7 +4,13 @@
 \alias{zero_order_cors}
 \title{Zero-Order Correlations}
 \usage{
-zero_order_cors(Y, type = "continuous", iter = 5000, mixed_type = NULL)
+zero_order_cors(
+  Y,
+  type = "continuous",
+  iter = 5000,
+  mixed_type = NULL,
+  progress = TRUE
+)
 }
 \arguments{
 \item{Y}{Matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}
@@ -17,6 +23,8 @@ zero_order_cors(Y, type = "continuous", iter = 5000, mixed_type = NULL)
 \item{mixed_type}{Numeric vector. An indicator of length p for which varibles should be treated as ranks.
 (1 for rank and 0 to assume normality). The default is currently to treat all integer variables as ranks
 when \code{type = "mixed"} and \code{NULL} otherwise. See note for further details.}
+
+\item{progress}{Logical. Should a progress bar be included (defaults to \code{TRUE}) ?}
 }
 \value{
 \itemize{
