+++ man/weighted_adj_mat.Rd
@@ -23,10 +23,11 @@ and \code{\link{ggm_compare_explore}} objects.
 \examples{
 \donttest{
 # note: iter = 250 for demonstrative purposes
-Y <- bfi
+Y <- bfi[,1:5]
 
 # estimate
-fit <- estimate(Y, iter = 250)
+fit <- estimate(Y, iter = 250,
+                progress = FALSE)
 
 # select graph
 E <- select(fit)
