+++ man/plot.confirm.Rd
@@ -46,7 +46,8 @@ Yfemale <- subset(Y, gender == 2,
 test <- ggm_compare_confirm(Ymale,
                             Yfemale,
                             hypothesis = hypothesis,
-                            iter = 250)
+                            iter = 250,
+                            progress = FALSE)
 
 
 # plot

[…]
