+++ vignettes/control.Rmd
@@ -1,0 +1,225 @@
+---
+title: "Controlling for Variables"
+author: "Donny Williams"
+date: "5/25/2020"
+bibliography: ../inst/REFERENCES.bib
+output:
+rmarkdown::html_vignette:
+  toc: yes
+vignette: >
+  %\VignetteIndexEntry{Controlling for Variables}
+  %\VignetteEngine{knitr::rmarkdown}
+  \usepackage[utf8]{inputenc}
+---
+
+# Introduction
+This vignette describes how to control for variables. This is a new feature to **BGGM** (version `2.0.0`).
+
+# Example 1: Multivariate Regression
+When controlling for variables, a multivariate regression is fitted in **BGGM**. In fact, a GGM can be understood as a multivariate regression with intercepts only models for the predictors.
+
+## Notes about Implemention
+**BGGM** does not use the typical approach for multivariate regression in `R`. This avoids having to 
+write out each outcome variable, of which there are typically many in a GGM. In **BGGM**, it is assumed
+that the data matrix includes only the variables to be included in the GGM and the control variables.
+
+### Correct 
+Suppose that we want to control for education level, with five variables included in the 
+GGM.
+
+```r
+# data
+Y <- bfi[,c(1:5, 27)]
+
+# head
+head(Y)
+
+#>       A1 A2 A3 A4 A5 education
+#> 61617  2  4  3  4  4        NA
+#> 61618  2  4  5  2  5        NA
+#> 61620  5  4  5  4  4        NA
+#> 61621  4  4  6  5  5        NA
+#> 61622  2  3  3  4  5        NA
+#> 61623  6  6  5  6  5         3
+```
+
+Notice that `Y` includes **only** the five variables and `education`.
+
+
+
+
+
+## Fit Model
+This model can then be fitted with
+
+```
+fit <- explore(Y, formula = ~ as.factor(education))
+```
+
+To show this is indeed a multivariate regression, here are the summarized regression coefficients
+
+```
+summ_coef <- regression_summary(fit)
+
+# outcome one
+summ_coef$reg_summary[[1]]
+
+#>                       Post.mean Post.sd Cred.lb Cred.ub
+#> (Intercept)               0.256   0.095   0.072   0.442
+#> as.factor(education)2     0.073   0.128  -0.177   0.323
+#> as.factor(education)3    -0.202   0.104  -0.405  -0.001
+#> as.factor(education)4    -0.462   0.119  -0.691  -0.233
+#> as.factor(education)5    -0.578   0.117  -0.815  -0.346
+```
+
+And here are the coefficients from `lm` (a univariate regression for `A1`)
+
+```
+round(
+  cbind(
+    # summary: coef and se
+    summary( lm(scale(A1, scale = F) ~ as.factor(education), data = Y))$coefficients[,1:2],
+    # confidence interval
+    confint( lm(scale(A1, scale = F) ~ as.factor(education), data = Y))
+), 3)
+
+
+#>                       Estimate Std. Error  2.5 % 97.5 %
+#> (Intercept)              0.256      0.093  0.073  0.438
+#> as.factor(education)2    0.072      0.125 -0.172  0.316
+#> as.factor(education)3   -0.203      0.101 -0.401 -0.004
+#> as.factor(education)4   -0.461      0.116 -0.690 -0.233
+#> as.factor(education)5   -0.578      0.115 -0.804 -0.351
+```
+The estimate are very (very) similar.
+
+## Summary
+Note that all the other functions work just the same. For example, the relations controlling for education
+are summarized with
+
+```
+summary(fit)
+
[…]
