+++ _pkgdown.yml
@@ -49,6 +49,10 @@ reference:
     contents:
     - bggm_missing
   - title: "Estimation Based Methods"
+    desc: >
+      'Estimation' indicates that the methods to not employ Bayes factor testing. Rather,
+      the graph is determined with the posterior distribution. The prior distribtuion has a
+      minimal influence.
     contents:
     - estimate
     - coef.estimate
@@ -57,6 +61,9 @@ reference:
     - select.estimate
     - summary.estimate
   - title: "Exploratory Hypothesis Testing"
+    desc: >
+      Bayes factor testing to determine the graph. 'Exploratory' reflects that there
+      is not a specific hypothesis being test.
     contents:
     - explore
     - coef.explore
@@ -67,10 +74,14 @@ reference:
     - summary.explore
     - summary.select.explore
   - title: "Confirmatory Hypothesis Testing"
+    desc: >
+     Test (in)equality constrained hypotheses with the Bayes factor.
     contents:
     - confirm
     - plot.confirm
   - title: "Compare Gaussian Graphical Models"
+    desc: >
+      A variety of methods for comparing GGMs.
   - subtitle: "Posterior Predictive Check"
     desc: >
      Compare groups with a posterior predictive check, where the null model
@@ -99,22 +110,28 @@ reference:
     - summary.ggm_compare_explore
   - subtitle: "Confirmatory Hypothesis Testing"
     desc: >
-      Confirmatory hypothesis testing for groups.
+       Test (in)equality constrained hypotheses with the Bayes factor.
     contents:
     - ggm_compare_confirm
     - plot.confirm
   - title: "Predictability"
+    desc: >
+      Bayesian variance explained for each node in the model.
     contents:
     - predictability
     - plot.predictability
     - summary.predictability
   - title: "Network Statistics"
+    desc: >
+     Compute network statistics from a partial correlation matrix or a
+     weighted adjacency matrix.
     contents:
     - roll_your_own
     - plot.roll_your_own
   - title: "Partial Correlation Sums"
     desc: >
-      Compute the sum of partial correlations within (one group) or between (two groups) GGMs.
+      Compute the sum of partial correlations within (one group) or between (two groups) GGMs. This
+      can be used to compare sums.
     contents:
     - pcor_sum
     - plot.pcor_sum
@@ -141,7 +158,7 @@ reference:
     - zero_order_cors
   - title: "Data"
     desc: >
-     Example dataests and correlation matrices.
+     Example datasets and correlation matrices.
     contents:
     - asd_ocd
     - bfi
