+++ inst/REFERENCES.bib
@@ -33,7 +33,7 @@ year = {2017}
 
 
 @article{Williams2019,
-author = {Williams, Donald R.},
+author = {Williams, Donald R},
 doi = {10.31234/OSF.IO/X8DPR},
 journal = {arXiv},
 keywords = {Explanation,Gaussian Graphical Model,Partial Correlation,Prediction,Psychometrics,Quantitative Methods,Quantitative Psychology,Social and Behavioral Sciences,Statistical Methods},
@@ -43,7 +43,7 @@ year = {2018}
 }
 
 @article{Williams2019_bf,
-author = {Williams, Donald R. and Mulder, Joris},
+author = {Williams, Donald R and Mulder, Joris},
 doi = {10.31234/osf.io/ypxd8},
 file = {:C$\backslash$:/Users/wdona/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Williams, Mulder - 2019 - Bayesian Hypothesis Testing for Gaussian Graphical Models Conditional Independence and Order Constraints.pdf:pdf},
 journal = {PsyArXiv},
@@ -122,7 +122,7 @@ year = {2019}
 
 @article{williams2020comparing,
   title={Comparing Gaussian graphical models with the posterior predictive distribution and Bayesian model selection.},
-  author={Williams, Donald R. and Rast, Philippe and Pericchi, Luis R and Mulder, Joris},
+  author={Williams, Donald R and Rast, Philippe and Pericchi, Luis R and Mulder, Joris},
   journal={Psychological Methods},
   year={2020},
   publisher={American Psychological Association}
@@ -470,13 +470,6 @@ year = {2011}
   publisher={Taylor \& Francis}
 }
 
-@article{williams2020bggm,
-  title={BGGM: Bayesian Gaussian Graphical Models in R},
-  author={Williams, Donald R and Mulder, Joris},
-  year={2020},
-  publisher={PsyArXiv},
-   journal={PsyArXiv}
-}
 
 @article{meng1994posterior,
   title={Posterior predictive $ p $-values},
@@ -547,4 +540,12 @@ year = {2011}
   author={Mohammadi, Reza and Wit, Ernst C},
   journal={arXiv preprint arXiv:1501.05108},
   year={2015}
+}
+
+
+@article{williams2020bggm,
+  title={BGGM: Bayesian Gaussian Graphical Models in R},
+  author={Williams, Donald R and Mulder, Joris},
+  journal={PsyArXiv},
+  year={2020},
 }
