### 2020-05-27 0d8690d
website: remove part about analytic solution
