### 2020-05-27 68b7edf
documentation: finish by removing progress, lengthy examples, etc.
