+++ R/zero_order.R
@@ -75,7 +75,8 @@
 #' #################################
 #'
 #' fit <- zero_order_cors(Y, type = "continuous",
-#'                        iter = 250)
+#'                        iter = 250,
+#'                        progress = FALSE)
 #'
 #'
 #' #################################
@@ -83,7 +84,8 @@
 #' #################################
 #'
 #' fit <- zero_order_cors(Y+1, type = "ordinal",
-#'                        iter = 250)
+#'                        iter = 250,
+#'                        progress = FALSE)
 #'
 #'
 #' ###########################
@@ -91,7 +93,8 @@
 #' ###########################
 #'
 #' fit <- zero_order_cors(Y+1, type = "mixed",
-#'                        iter = 250)
+#'                        iter = 250,
+#'                        progress = FALSE)
 #'
 #' ############################
 #' ## example 4: tetrachoric ##
@@ -101,7 +104,8 @@
 #' Y <- women_math[,1:3]
 #'
 #' fit <- zero_order_cors(Y, type = "binary",
-#'                        iter = 250)
+#'                        iter = 250,
+#'                        progress = FALSE)
 #'
 #' }
 #' @export
