+++ R/select.estimate.R
@@ -47,11 +47,14 @@
 #'
 #' @examples
 #' \donttest{
+#' # note: iter = 250 for demonstrative purposes
+#'
 #' # data
-#' Y <- bfi[,1:25]
+#' Y <- bfi[,1:10]
 #'
 #' # estimate
-#' fit <- estimate(Y, iter = 250)
+#' fit <- estimate(Y, iter = 250,
+#'                 progress = FALSE)
 #'
 #'
 #' # select edge set
