### 2020-05-27 def452c
website: fix documentation (remove plots from roll your own, typos, etc.)
