+++ man/plot.roll_your_own.Rd
@@ -21,3 +21,38 @@ An object of class \code{ggplot}
 \description{
 Plot \code{roll_your_own} Objects
 }
+\examples{
+\donttest{
+####################################
+###### example 1: assortment #######
+####################################
+# assortment
+library(assortnet)
+
+Y <- BGGM::bfi[,1:10]
+membership <- c(rep("a", 5), rep("c", 5))
+
+# fit model
+fit <- estimate(Y = Y, iter = 250,
+                progress = FALSE)
+
+# membership
+membership <- c(rep("a", 5), rep("c", 5))
+
+# define function
+f <- function(x,...){
+ assortment.discrete(x, ...)$r
+}
+
+net_stat <- roll_your_own(object = fit,
+                          FUN = f,
+                          types = membership,
+                          weighted = TRUE,
+                          SE = FALSE, M = 1,
+                          progress = FALSE)
+
+# plot
+plot(net_stat)
+
+}
+}
