### 2020-05-27 beeca4a
website: update NEWS.md
