### 2020-05-28 donaldRwilliams/BGGM#6 @jayrobwilliams
I'm satisfied with this implementation, but I'll let @paulgovan decide whether to close the issue as he opened it.
