### 2020-05-28 donaldRwilliams/BGGM#3 @jayrobwilliams
I think a list of disciplines that commonly employ GGMs with a citation or two for canonical examples of articles that use GGMs would be sufficient for me. The state of the field discussion and comparison to existing packages goes a long way here, but I think a clearly stated example of a problem or two that GGMs are used for in the intro of the paper and README would make the need for BGGM crystal clear.
