+++ index.Rmd
@@ -1,6 +1,7 @@
 ---
 output: github_document
 bibliography: inst/REFERENCES.bib
+css: custom.css
 ---
 
 ```{r, echo = FALSE, message=F}
@@ -33,6 +34,10 @@ two general approaches for Bayesian inference: (1) estimation and (2) hypothesis
 testing. The key distinction is that the former focuses on either the posterior or posterior 
 predictive distribution [@Gelman1996a; see section 5 in @rubin1984bayesianly], whereas the 
 latter focuses on model comparison with the Bayes factor [@Jeffreys1961; @Kass1995].
+
+## <i class="fas fa-question-circle"></i> What is a Gaussian Graphical Model ?
+
+>A Gaussian graphical model comprises of a set of items or variables, depicted by circles, and a set of lines that visualize relationships between the items or variables...The thickness of these lines represents the strength of the relationships between items or variables...Notably, in the Gaussian graphical model, these lines capture partial correlations, that is, the correlation between two items or variables when controlling for all other items or variables included in the data set [@bhushan2019using].
 
 ## <i class="fas fa-cog"></i> Installation
 
