+++ inst/REFERENCES.bib
@@ -549,3 +549,15 @@ year = {2011}
   journal={PsyArXiv},
   year={2020},
 }
+
+@article{bhushan2019using,
+  title={Using a Gaussian graphical model to explore relationships between items
+  and variables in environmental psychology research},
+  author={Bhushan, Nitin and Mohnert, Florian and Jans, Lise and Sloot,
+  Daniel and Albers, Casper and Steg, Linda},
+  journal={Frontiers in psychology},
+  volume={10},
+  pages={1050},
+  year={2019},
+  publisher={Frontiers}
+}
