### 2020-05-29 f9d55de
bggm_missing: changed iteration default to save memory and updated the documentation accordingly.
