+++ R/explore.default.R
@@ -453,7 +453,7 @@ explore <- function(Y,
     prior_samp <- .Call('_BGGM_sample_prior',
                       PACKAGE = 'BGGM',
                       Y = Y_dummy,
-                      iter = iter,
+                      iter = iter + 50,
                       delta = delta,
                       epsilon = eps,
                       prior_only = 1,
