### 2020-05-31 e43248a
package: delete extra figures
