### 2020-05-31 cb6cbc1
package submission: remove old readme figures
