+++ README.Rmd
@@ -9,7 +9,7 @@ knitr::opts_chunk$set(
   comment = "#>",
   fig.path = "man/figures/README-",
   dev = "png",
-  dpi = 500,
+  dpi = 200,
   fig.align = "center",
   knitr::opts_chunk$set(comment = NA)
   )
