### 2020-06-01 donaldRwilliams/BGGM#3 @donaldRwilliams
I am having a bit of a hard time with a "statement of need". Will have to come back to it, as I have some other things to work on.
