### 2020-06-01 ea8c675
update readme: remove reference to dev version
