### 2020-06-02 e6e184e
new_feature: add VAR_estimate
