+++ man/VAR_estimate.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/var_estimate.R
+\name{VAR_estimate}
+\alias{VAR_estimate}
+\title{VAR: Estimation}
+\usage{
+VAR_estimate(Y, rho_sd = 0.5, beta_sd = 1, iter = 5000)
+}
+\arguments{
+\item{Y}{Matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}
+
+\item{rho_sd}{Numeric. Scale of the prior distribution for the partial correlations,
+approximately the standard deviation of a beta distribution
+(defaults to 0.50).}
+
+\item{beta_sd}{Numeric. Standard deviation of the prior distribution for the regression coefficients
+(defaults to 1).}
+
+\item{iter}{Number of iterations (posterior samples; defaults to 5000).}
+}
+\value{
+An object of class \code{VAR_estimate}
+}
+\description{
+VAR: Estimation
+}
