### 2020-06-02 f405b11
new_feature: add select.VAR_estimate
