+++ R/var_estimate.R
@@ -15,33 +15,45 @@ VAR_estimate <- function(Y, rho_sd = 0.50,
                          beta_sd = 1,
                          iter = 5000) {
 
+  # number of nodes
   p <- ncol(Y)
 
+  # Y lagged: add NA
   Y_lag <-   rbind(NA, Y)
 
+  # Y lagged: column names
   colnames(Y_lag) <- paste0(colnames(Y), ".l1")
 
+  # combine all data
   Y_all <- na.omit(cbind.data.frame(rbind(Y, NA), Y_lag))
 
+  # nodes in GGM
   Y <- scale(as.matrix(Y_all[,1:p]), scale = TRUE)
 
+  # predictors (lagged effects)
   X <- scale(as.matrix(Y_all[,(p+1):(p*2)]), scale = TRUE)
 
+  # delta: rho ~ beta(delta/2, delta/2)
   delta <- BGGM:::delta_solve(rho_sd)
 
+  # prior variance
   beta_var <- beta_sd^2
 
-  fit <- BGGM:::var(Y =  as.matrix( Y ),
-                    X = as.matrix( X ),
-                    delta = delta,
-                    epsilon = 0.001,
-                    beta_prior = diag(p) * (1 / beta_var),
-                    iter = iter + 50,
-                    start = diag(p),
-                    progress = TRUE)
+  fit <-.Call(
+      "_BGGM_var",
+      Y =  as.matrix(Y),
+      X = as.matrix(X),
+      delta = delta,
+      epsilon = 0.001,
+      beta_prior = diag(p) * (1 / beta_var),
+      iter = iter + 50,
+      start = solve(cor(Y)),
+      progress = TRUE
+    )
 
   returned_object <- fit
-  class(returned_object) <- c("BGGM", "VAR", "VAR_estimate")
-  returned_object
-
+  class(returned_object) <- c("BGGM",
+                              "VAR",
+                              "VAR_estimate")
+  return(returned_object)
 }
