+++ NAMESPACE
@@ -16,6 +16,7 @@ S3method(plot,summary.select.explore)
 S3method(predict,estimate)
 S3method(predict,explore)
 S3method(print,BGGM)
+S3method(select,VAR_estimate)
 S3method(select,estimate)
 S3method(select,explore)
 S3method(select,ggm_compare_estimate)
