### 2020-06-02 26b9fe4
new_feature: add select for VAR_estimate
