### 2020-06-02 5c0b81d
new_feature: change name from VAR to var
