+++ man/select.VAR_estimate.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/select.VAR_estimate.R
-\name{select.VAR_estimate}
-\alias{select.VAR_estimate}
+% Please edit documentation in R/select.var_estimate.R
+\name{select.var_estimate}
+\alias{select.var_estimate}
 \title{Graph Selection for \code{VAR.estimate} Object}
 \usage{
-\method{select}{VAR_estimate}(object, cred = 0.95, alternative = "two.sided")
+\method{select}{var_estimate}(object, cred = 0.95, alternative = "two.sided")
 }
 \arguments{
 \item{object}{An object of class \code{VAR.estimate}.}
