+++ man/VAR_estimate.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/var_estimate.R
-\name{VAR_estimate}
-\alias{VAR_estimate}
+\name{var_estimate}
+\alias{var_estimate}
 \title{VAR: Estimation}
 \usage{
-VAR_estimate(Y, rho_sd = 0.5, beta_sd = 1, iter = 5000)
+var_estimate(Y, rho_sd = 0.5, beta_sd = 1, iter = 5000)
 }
 \arguments{
 \item{Y}{Matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).}
