+++ NAMESPACE
@@ -16,11 +16,11 @@ S3method(plot,summary.select.explore)
 S3method(predict,estimate)
 S3method(predict,explore)
 S3method(print,BGGM)
-S3method(select,VAR_estimate)
 S3method(select,estimate)
 S3method(select,explore)
 S3method(select,ggm_compare_estimate)
 S3method(select,ggm_compare_explore)
+S3method(select,var_estimate)
 S3method(summary,coef)
 S3method(summary,estimate)
 S3method(summary,explore)
@@ -28,7 +28,6 @@ S3method(summary,ggm_compare_estimate)
 S3method(summary,ggm_compare_explore)
 S3method(summary,predictability)
 S3method(summary,select.explore)
-export(VAR_estimate)
 export(bggm_missing)
 export(confirm)
 export(convergence)
@@ -51,6 +50,7 @@ export(predictability)
 export(regression_summary)
 export(roll_your_own)
 export(select)
+export(var_estimate)
 export(weighted_adj_mat)
 export(zero_order_cors)
 import(Rcpp)
