### 2020-06-02 d333d05
new_feature: add print function to var_estimate
