+++ NAMESPACE
@@ -28,6 +28,7 @@ S3method(summary,ggm_compare_estimate)
 S3method(summary,ggm_compare_explore)
 S3method(summary,predictability)
 S3method(summary,select.explore)
+S3method(summary,var_estimate)
 export(bggm_missing)
 export(confirm)
 export(convergence)
