### 2020-06-03 dc55715
new_feature: add print summary for var_estimate
