### 2020-06-03 4f93611
new_feature: add pcor and beta mean to print.var_estimate
