### 2020-06-03 aa8ceb0
new_feature: finish functions for var_estimate. Not that plotting will take more time.
