+++ NAMESPACE
@@ -13,6 +13,7 @@ S3method(plot,summary.explore)
 S3method(plot,summary.ggm_compare_estimate)
 S3method(plot,summary.ggm_compare_explore)
 S3method(plot,summary.select.explore)
+S3method(plot,summary.var_estimate)
 S3method(predict,estimate)
 S3method(predict,explore)
 S3method(print,BGGM)
