### 2020-06-04 92615f4
new_feature: predictability for var_estimate
