+++ inst/REFERENCES.bib
@@ -561,3 +561,20 @@ year = {2011}
   year={2019},
   publisher={Frontiers}
 }
+
+@article{o2020use,
+  title={Use of Composites in Analysis of Individual Time Series: Implications for Person-Specific Dynamic Parameters},
+  author={O'Laughlin, Kristine D and Liu, Siwei and Ferrer, Emilio},
+  journal={Multivariate Behavioral Research},
+  pages={1--18},
+  year={2020},
+  publisher={Taylor \& Francis}
+}
+
+@article{williams2019bayesian,
+  title={Bayesian Multivariate Mixed-Effects Location Scale Modeling of Longitudinal Relations among Affective Traits, States, and Physical Activity},
+  author={Williams, Donald R and Liu, Siwei and Martin, Stephen R and Rast, Philippe},
+  year={2019},
+  publisher={PsyArXiv},
+  journal={PsyArXiv}
+}
