+++ NAMESPACE
@@ -16,6 +16,7 @@ S3method(plot,summary.select.explore)
 S3method(plot,summary.var_estimate)
 S3method(predict,estimate)
 S3method(predict,explore)
+S3method(predict,var_estimate)
 S3method(print,BGGM)
 S3method(select,estimate)
 S3method(select,explore)
