### 2020-06-04 84ec354
new_feature: change var_estimate
