### 2020-06-04 9f06e4b
website: add var model vignette
