+++ vignettes/var_model.Rmd
@@ -1,0 +1,269 @@
+---
+title: "Graphical VAR"
+author: "Donny Williams"
+date: "6/04/2020"
+bibliography: ../inst/REFERENCES.bib
+output:
+rmarkdown::html_vignette:
+  toc: yes
+vignette: >
+  %\VignetteIndexEntry{Graphical VAR}
+  %\VignetteEngine{knitr::rmarkdown}
+  \usepackage[utf8]{inputenc}
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE)
+```
+
+# Introduction
+This vignette describes the implement ion for a "graphical VAR" in `BGGM`. In `BGGM`, this is fitted as multivariate regression. The  key innovation is a novel prior distribution for the residual covariance matrix. There are a variety of much cooler names than a *mere* "multivariate regression", including "VAR" (vector autoregressive models) and "TSCGM" (yime series chain graphical model).
+
+## R package
+```
+# need the developmental version
+if (!requireNamespace("remotes")) { 
+  install.packages("remotes")   
+}   
+
+# install from github
+remotes::install_github("donaldRwilliams/BGGM")
+library(BGGM)
+
+# for comparsion
+library(vars)
+
+# for plotting
+library(qgraph)
+
+# combine plots
+library(cowplot)
+```
+
+## Data 
+I use data from the ifit (fit bit) study. The data were gathered over 100 consecutive days on a variety of variables, including the PANAS scale (positive and negative affect) and the number of steps each day. `BGGM` includes a subset of variables for two individuals.
+
+```r
+# data
+Y <- subset(ifit, id == 1)[,-1]
+
+# first 3 rows
+head(Y, n = 3)
+
+#>  interested disinterested excited upset strong stressed steps
+#>         72            10      50     2     50       16  7805
+#>         75             6      75     0     76        0 18248
+>#         36            58      38     5     45        1 12139
+```
+
+
+
+# Estimation 
+The methods in **BGGM** are organized around Bayesian "estimation" and "hypothesis testing". This is to reach
+a broader audience, as former is more similar to classical methods (those more familiar to researchers).
+
+## Fit Model
+With the data in hand, the model is fitted as follows
+
+```
+# fit model
+fit <- var_estimate(Y, beta_sd = 1)
+```
+Note that `beta_sd` is the prior distribution for the regression coefficients. A smaller value, say, `beta_sd = 0.25`, results in a Bayesian ridge regression. Note also this model, including 5000 draws from the posterior,
+
+The results can then be printed
+```r
+#> BGGM: Bayesian Gaussian Graphical Models 
+#> --- 
+#> Vector Autoregressive Model (VAR) 
+#> --- 
+#> Posterior Samples: 5000 
+#> Observations (n): 94 
+#> Nodes (p): 7 
+#> --- 
+#> Call: 
+#> var_estimate(Y = Y, beta_sd = 10)
+#> --- 
+#> Partial Correlations: 
+#> 
+#>               interested disinterested excited  upset strong stressed  steps
+#> interested         0.000        -0.170   0.388 -0.217  0.313    0.268  0.089
+#> disinterested     -0.170         0.000  -0.172 -0.029  0.094    0.160 -0.078
+#> excited            0.388        -0.172   0.000 -0.126  0.500   -0.161 -0.016
+#> upset             -0.217        -0.029  -0.126  0.000  0.118    0.350 -0.039
+#> strong             0.313         0.094   0.500  0.118  0.000   -0.010  0.176
+#> stressed           0.268         0.160  -0.161  0.350 -0.010    0.000 -0.038
[…]
