### 2020-06-04 b2dbf80
website: fix typos in var_model vignette
