### 2020-06-04 f89fd9c
website: update var_model to remove typos
