### 2020-06-04 1e92a77
website: fix typos in var_estimate
