### 2020-06-05 #2111 @arfon
@donaldRwilliams - did you latest updates address many of the comments opened by the reviewers (e.g. https://github.com/donaldRwilliams/BGGM/issues/3, https://github.com/donaldRwilliams/BGGM/issues/4, https://github.com/donaldRwilliams/BGGM/issues/6)? If so, perhaps you could leave comments on those issues (and close them if appropriate) to help us all understand what changes have been made?
