### 2020-06-05 6a2a7a4
readme: update with applications
