+++ inst/REFERENCES.bib
@@ -578,3 +578,37 @@ year = {2011}
   publisher={PsyArXiv},
   journal={PsyArXiv}
 }
+
+@article{zerenner2014gaussian,
+  title={A Gaussian graphical model approach to climate networks},
+  author={Zerenner, Tanja and Friederichs, Petra and Lehnertz, Klaus and Hense, Andreas},
+  journal={Chaos: An Interdisciplinary Journal of Nonlinear Science},
+  volume={24},
+  number={2},
+  pages={023103},
+  year={2014},
+  publisher={American Institute of Physics}
+}
+
+
+@article{chu2009graphical,
+  title={A graphical model approach for inferring large-scale networks integrating gene expression and genetic polymorphism},
+  author={Chu, Jen-hwa and Weiss, Scott T and Carey, Vincent J and Raby, Benjamin A},
+  journal={BMC systems biology},
+  volume={3},
+  number={1},
+  pages={55},
+  year={2009},
+  publisher={BioMed Central}
+}
+
+@article{millington2020partial,
+  title={Partial correlation financial networks},
+  author={Millington, Tristan and Niranjan, Mahesan},
+  journal={Applied Network Science},
+  volume={5},
+  number={1},
+  pages={11},
+  year={2020},
+  publisher={Springer}
+}
