### 2020-06-05 646355c
joss paper: add applications
