### 2020-06-05 92e8f27
joss paper: fix typo
