+++ paper.md
@@ -79,7 +79,7 @@ The central contribution of **BGGM** is to extend those approaches:
 ## Supported Data Types
 
 * **Continuous**: The continuous method was described in  @Williams2019. Note that 
-                  this is based on the customary[Wishartdistribution](https://en.wikipedia.org/wiki/Wishart_distribution).
+                  this is based on the customary [Wishartdistribution](https://en.wikipedia.org/wiki/Wishart_distribution).
 
 * **Binary**: The binary method builds directly upon @talhouk2012efficient
   that, in turn, built upon the approaches of @lawrence2008bayesian and
