### 2020-06-05 donaldRwilliams/BGGM#3 @donaldRwilliams
To address this comment, we added two sections. The first (What is a Gaussian Graphical Model) stated what a GGM is. And then the second ("Applications") listed several applications with references.
