### 2020-06-05 donaldRwilliams/BGGM#4 @donaldRwilliams
To address this comment, we replaced all the examples in the readme and documentation with actual data. In the readme in particular, now the examples show explicitly how to answer research questions with the package.
