### 2020-06-05 donaldRwilliams/BGGM#6 @donaldRwilliams
To address this comment, we added a section (Bug Reports, Feature Requests, and Contributing) to the README and the paper
