### 2020-06-05 #2111 @donaldRwilliams
@arfon I left comments on those issues and also closed them (I think this was fine, but can always re open if needed).
