### 2020-06-05 donaldRwilliams/BGGM#3 @jayrobwilliams
This is great! Clearly and concisely helps the reader understand what you might want to use a GGM for.
