### 2020-06-06 674b8ac
var_estimate: finish var_estimate documentation
