+++ R/var_estimate.R
@@ -1,5 +1,15 @@
 #' VAR: Estimation
 #'
+#' @description Estimate VAR(1) models by efficiently sampling from the posterior distribution. This
+#' provides two graphical structures: (1) a network of undirected relations (the GGM, controlling for the
+#' lagged predictors) and (2) a network of directed relations (the lagged coefficients). Note that
+#' in the graphical modeling literature, this model is also known as a time series chain graphical model
+#' \insertCite{abegaz2013sparse}{BGGM}.
+#'
+#' @name var_estimate
+#'
+#' @aliases tscgm_estimate
+#'
 #' @param Y Matrix (or data frame) of dimensions \emph{n} (observations) by  \emph{p} (variables).
 #'
 #' @param rho_sd Numeric. Scale of the prior distribution for the partial correlations,
@@ -7,16 +17,66 @@
 #' (defaults to 0.50).
 #'
 #' @param beta_sd Numeric. Standard deviation of the prior distribution for the regression coefficients
-#'        (defaults to 1).
+#'        (defaults to 1). The prior is by default centered at zero and follows a normal distribution
+#'        \insertCite{@Equation 9, @sinay2014bayesian}{BGGM}
 #'
 #' @param iter Number of iterations (posterior samples; defaults to 5000).
 #'
-#' @return An object of class \code{var_estimate}
+#' @param progress Logical. Should a progress bar be included (defaults to \code{TRUE}) ?
+#'
+#' @param seed An integer for the random seed (defaults to 1).
+#'
+#' @param ... Currently ignored.
+#'
+#' @details Each time series in \code{Y} is standardized (mean  = 0; standard deviation = 1).
+#'
+#' @note
+#' \strong{Regularization}:
+#'
+#' A Bayesian ridge regression can be fitted by decreasing \code{beta_sd}
+#' (e.g., \code{beta_sd = 0.25}). This could be advantageous for forecasting
+#' (out-of-sample prediction) in particular.
+#'
+#'
+#' @references
+#' \insertAllCited{}
+#'
+#' @return An object of class \code{var_estimate} containing a lot of information that is
+#' used for printing and plotting the results. For users of \strong{BGGM}, the following are the
+#' useful objects:
+#'
+#' \itemize{
+#'
+#' \item \code{beta_mu} A matrix including the regression coefficients (posterior mean).
+#'
+#' \item \code{pcor_mu} Partial correlation matrix (posterior mean).
+#'
+#' \item \code{fit} A list including the posterior samples.
+#'
+#' }
+#'
+#' @examples
+#' \donttest{
+#' # data
+#' Y <- subset(ifit, id == 1)[,-1]
+#'
+#' # use alias (var_estimate also works)
+#' fit <- tscgm_estimate(Y, progress = FALSE)
+#'
+#' fit
+#'
+#' }
 #' @export
 var_estimate <- function(Y, rho_sd = 0.50,
                          beta_sd = 1,
-                         iter = 5000) {
+                         iter = 5000,
+                         progress = TRUE,
+                         seed = 1,
+                         ...) {
 
+  old <- .Random.seed
+
+  set.seed(seed)
 
   Y <- scale(na.omit(Y))
   # number of nodes
@@ -41,12 +101,14 @@ var_estimate <- function(Y, rho_sd = 0.50,
   X <- as.matrix(Y_all[,(p+1):(p*2)])
 
   # delta: rho ~ beta(delta/2, delta/2)
-  delta <- BGGM:::delta_solve(rho_sd)
+  delta <- delta_solve(rho_sd)
 
   # prior variance
[…]
