+++ inst/REFERENCES.bib
@@ -612,3 +612,24 @@ year = {2011}
   year={2020},
   publisher={Springer}
 }
+
+
+@article{abegaz2013sparse,
+  title={Sparse time series chain graphical models for reconstructing genetic networks},
+  author={Abegaz, Fentaw and Wit, Ernst},
+  journal={Biostatistics},
+  volume={14},
+  number={3},
+  pages={586--599},
+  year={2013},
+  publisher={Oxford University Press}
+}
+
+@article{sinay2014bayesian,
+  title={Bayesian inference of a multivariate regression model},
+  author={Sinay, Marick S and Hsu, John SJ},
+  journal={Journal of Probability and Statistics},
+  volume={2014},
+  year={2014},
+  publisher={Hindawi}
+}

[…]
