### 2020-06-08 cdfb302
issue # 40: changed starting values
