+++ R/select.VAR_estimate.R
@@ -34,7 +34,7 @@
 #' Y <- subset(ifit, id == 1)[,-1]
 #'
 #' # fit model with alias (var_estimate also works)
-#' fit <- tscgm_estimate(Y, progress = FALSE)
+#' fit <- var_estimate(Y, progress = FALSE)
 #'
 #' # select graphs
 #' select(fit, cred = 0.95)
