+++ NAMESPACE
@@ -53,7 +53,6 @@ export(predictability)
 export(regression_summary)
 export(roll_your_own)
 export(select)
-export(tscgm_estimate)
 export(var_estimate)
 export(weighted_adj_mat)
 export(zero_order_cors)
