### 2020-06-08 7d10be8
remove tscgm from examples
