+++ R/predict.estimate.R
@@ -362,7 +362,7 @@ predict.explore <- function(object,
 #' Y <- subset(ifit, id == 1)[,-1]
 #'
 #' # fit model with alias (var_estimate also works)
-#' fit <- tscgm_estimate(Y, progress = FALSE)
+#' fit <- var_estimate(Y, progress = FALSE)
 #'
 #' # fitted values
 #' pred <- predict(fit, progress = FALSE)
