+++ man/predict.var_estimate.Rd
@@ -29,7 +29,7 @@ Model Predictions for \code{var_estimate} Objects
 Y <- subset(ifit, id == 1)[,-1]
 
 # fit model with alias (var_estimate also works)
-fit <- tscgm_estimate(Y, progress = FALSE)
+fit <- var_estimate(Y, progress = FALSE)
 
 # fitted values
 pred <- predict(fit, progress = FALSE)
