+++ R/confirm.R
@@ -473,7 +473,7 @@ confirm <- function(Y, hypothesis,
       z0_start = rank_vars$z0_start,
       levels = rank_vars$levels,
       K = rank_vars$K,
-      Sigma_start = rank_vars$Sigma_start,
+      Sigma_start = cov(Y),
       iter = iter + 50,
       delta = delta,
       epsilon = 0.01,
