### 2020-06-14 40c0147
new_feature: mvn imputation. added continuous and partial documentation
