### 2020-07-03 #2111 @paulgovan
Sorry for the delay. I see several issues related to the paper that are still open. My thought is that these should be closed before moving forward with publication. @arfon @akeshavan do you agree?
