### 2020-07-04 #2111 @arfon
> Sorry for the delay. I see several issues related to the paper that are still open. My thought is that these should be closed before moving forward with publication. @arfon @akeshavan do you agree?

Yes, I agree, although it looks like almost all of them have been addressed and that @donaldRwilliams might just need to close those issues. The one I see that is still unresolved is https://github.com/donaldRwilliams/BGGM/issues/10 - could you confirm what is happening there @donaldRwilliams?
