### 2020-07-15 #2111 @arfon
@donaldRwilliams - to help our reviewers (and me), could you state on the issues that are still open as part of this review what actions you have taken and (if appropriate) close them. The issues I still see open here are: https://github.com/donaldRwilliams/BGGM/issues/3 https://github.com/donaldRwilliams/BGGM/issues/7 https://github.com/donaldRwilliams/BGGM/issues/8

Thanks!
