### 2020-07-15 donaldRwilliams/BGGM#7 @jayrobwilliams
I am satisfied with the way these issues have been addressed.
