### 2020-07-15 donaldRwilliams/BGGM#7 @donaldRwilliams
To address these comments, we followed the suggestions of the reviewers to include a section comparing to other software.
