### 2020-07-15 donaldRwilliams/BGGM#3 @donaldRwilliams
To address this comment, we clearly stated the need for the package in the paper and the README. This included showcasing examples with actual data and comparisons to other packages
