### 2020-07-15 donaldRwilliams/BGGM#8 @donaldRwilliams
We address this by making the paper much more a "high level" overview which followed the suggestion of the reviewer.
