### 2020-07-15 #2111 @donaldRwilliams
@arfon I went through and noted the changes. I think we followed the advice in the reviewers in all cases, and we noted the changes accordingly in the issue. They are all now closed as well.
