### 2020-07-16 #2111 @arfon
> @arfon I went through and noted the changes. I think we followed the advice in the reviewers in all cases, and we noted the changes accordingly in the issue. They are all now closed as well.

Excellent, thanks @donaldRwilliams

@paulgovan - could ask you to take one more quick look at this (and update your checklist accordingly) now that the author has resolved all of the outstanding issues? 🙏
