### 2020-07-20 #2111 @arfon
@donaldRwilliams - could you take a look at the DOIs suggested by Whedon above and add them to your BibTeX file if they are correct.
