### 2020-07-20 #2111 @donaldRwilliams
@arfon I updated the .bib file. Quite nice that the dois were all provided. Thanks !
