### 2020-07-21 #2111 @donaldRwilliams
@arfon I uploaded the new release of the software.

Here is the doi:
10.5281/zenodo.3954503
